% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllCompounds}
\alias{AllCompounds}
\title{All possible compound IDs, suffixes, etc.}
\format{
a data.frame
}
\usage{
AllCompounds
}
\description{
All possible compound IDs, suffixes, etc.
}
\keyword{datasets}
