% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllRegCompounds}
\alias{AllRegCompounds}
\title{All regular compound IDs, suffixes, etc.}
\format{
a data.frame
}
\usage{
AllRegCompounds
}
\description{
All regular compound IDs, suffixes, etc.
}
\keyword{datasets}
