% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllStats}
\alias{AllStats}
\title{All names for statistics used in the package}
\format{
A data.frame with columns: \describe{ \item{InternalColNames}{how a
  stat is referred to internal to the package, which generally needs to be
  compatible with standard R column-naming rules.}
  
  \item{SimulatorNames}{how the Simulator names those stats}
  
  \item{ReportNames}{how the stat should show up in a report}}
}
\usage{
AllStats
}
\description{
INTERNAL USE: We needed something to make sure we were being consistent with
how columns were named in various situations and for how stats were described
in reports. This does that. \code{\link{renameStats}} calls on this.
}
\keyword{datasets}
