% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllTissues}
\alias{AllTissues}
\title{All possible tissues and tissue subtypes}
\format{
a data.frame
}
\usage{
AllTissues
}
\description{
All possible tissues and tissue subtypes
}
\keyword{datasets}
