% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllWorkspaceDetails}
\alias{AllWorkspaceDetails}
\title{All possible experimental details for \code{\link{extractExpDetails_XML}}}
\format{
A data.frame
}
\usage{
AllWorkspaceDetails
}
\description{
All possible experimental details that can be extracted from a simulator
workspace file using \code{\link{extractExpDetails_XML}}. This version includes
columns used only internally with coding.
}
\keyword{datasets}
