% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDetails.R
\name{annotateDetails}
\alias{annotateDetails}
\title{Annotate Simcyp Simulator experimental details and format into a long
data.frame instead of a wide one}
\usage{
annotateDetails(Deets)
}
\arguments{
\item{Deets}{output from \code{\link{extractExpDetails}} or
\code{\link{extractExpDetails_mult}}}
}
\value{

}
\description{
\code{annotateDetails} converts output from either
\code{\link{extractExpDetails}} or \code{\link{extractExpDetails_mult}} into
a long data.frame and adds to that data.frame columns for a) which compound
the information pertains to (substrate, inhibitor, etc.), b) which section of
the Simcyp Simulator this detail is found in (physchem, absorption,
distribution, etc.), c) notes describing what the detail is, and d) which
sheet in the Excel file the information was pulled from.
}
\examples{

annotateDetails(Deets)

}
