% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectors.R
\name{blues}
\alias{blues}
\title{Create a vector of blues}
\usage{
blues(ncolors, shade = "regular")
}
\arguments{
\item{ncolors}{number of colors desired}

\item{shade}{the shade of colors to use, which can be "regular" (default),
"darker", "lighter", or "muted" (although "muted" isn't working well atm
and it just making everything the same color)}
}
\value{
a character vector of colors
}
\description{
\code{blues} is useful for getting a set of pretty blues for graphs. These
blues come from the 4th through 9th blues from grDevices::blues9, just to
give credit where it's due.
}
\examples{
# Create a set of 10 blues.
blues(10)

# Try using scales::show_col() to visualize the colors, ex:
scales::show_col(blues(10))

}
