% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BufForestData_20mg}
\alias{BufForestData_20mg}
\title{Example output from \code{extractForestData}}
\format{
a data.frame
}
\usage{
BufForestData_20mg
}
\description{
Example output from running \code{extractForestData} on some simulations with
bufuralol dosed at 20 mg as the substrate and then 4 different perpetrators.
All PK parameters in any forest data will be for the victim compound listed.
The data.frame also includes columns for all the compounds that were included
for the simulation, but all PK are for the victim compound listed.
}
\keyword{datasets}
