% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BufObsForestData}
\alias{BufObsForestData}
\title{Example observed PK data for working with \code{forest_plot}}
\format{
a data.frame
}
\usage{
BufObsForestData
}
\description{
Fake observed bufuralol PK data for showing examples with the forest_plot
function.
}
\keyword{datasets}
