% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PK.R
\name{calc_PK}
\alias{calc_PK}
\title{Calculate basic PK parameters for the 1st and last doses of
concentration-time data}
\usage{
calc_PK(
  ct_dataframe,
  first_dose_time = NA,
  last_dose_time = NA,
  dose_interval = NA,
  fit_points_after_x_time = NA,
  fit_last_x_number_of_points = NA,
  returnAggregateOrIndiv = "both",
  return_graphs_of_fits = TRUE,
  save_graphs_of_fits = FALSE,
  ncol = NULL,
  nrow = NULL,
  fig_width = 8.5,
  fig_height = 11
)
}
\arguments{
\item{ct_dataframe}{a data.frame of concentration-time data in the same
format as those created by running \code{extractConcTime},
\code{extractConcTime_mult}, \code{extractObsConcTime}, or
\code{extractObsConcTime_mult}. NB: This was only set up to calculate PK
for the 1st and last doses; all other doses will be ignored. If you want PK
for a dose other than those two, we recommend subsetting your data to
include only the 1st dose and whatever dose number you do want and perhaps
running this multiple times in a loop to get all the PK you need. A member
of the R Working Group can help you set this up if you'd like.}

\item{first_dose_time}{the time at which the first dose was administered. If
this is left as NA, the default value, this will be set to 0.}

\item{last_dose_time}{the time at which the last dose was administered. If
this is left as NA, the default value, we'll assume that the last dose was
administered at the earliest time included in the data for the highest dose
number included in the data. It bears repeating here that this function
only works well if all the data included have the same dosing regimen.}

\item{dose_interval}{the dosing interval; default is NA which assumes that
all data assigned with a given dose number should be used in calculating PK
values. Cases where this wouldn't necessarily be true: When there's a
washout period included in the data that is longer than the dosing
interval. In that situation, this will assume that only data within the
dosing interval should be used. For example, say that the last QD dose
occurred at 168 h but the washout period lasted until 336 h. For
calculating the last-dose AUCtau, you'd only want the time from 168 to 192
h, so if you set \code{dose_interval = 24}, that will only calculate AUCtau
using data from 168 to 192 h rather than from 168 to 336 h. This is only
set up to accommodate a single dosing interval and not custom dosing
regimens.}

\item{fit_points_after_x_time}{optionally specify that you want to fit only
points after a certain time after the most-recent dose. Default of NA means
that we'll fit all the data after tmax. Keep in mind that this will apply
to ALL profiles.}

\item{fit_last_x_number_of_points}{optionally specify that you want to fit
the last X number of points for each dose (replace with whatever number
makes sense for your situation). Default of NA means that we'll fit all the
data after tmax. Keep in mind that this will apply to ALL profiles.
\emph{An important note for fitting simulated data:} If you supply
simulated data with a lot of points, we will only use 100 of those
concentration-time points to describe each dosing interval because
performing, say, 10 trials x 10 subjects = 100 nonlinear regressions with a
thousand points for each dose number requires excessive computing time
needlessly because the regression must minimize the distance between the
fitted curve and every one of those points. Simulated data are pretty
predictable; your fitted parameters will not be less accurate with this
approach. If you supply simulated data, it will probably be less confusing
and yield more predictable results if you specify a value for
\code{fit_points_after_x_time} rather than
\code{fit_last_x_number_of_points}.}

\item{returnAggregateOrIndiv}{return aggregate and/or individual PK
parameters? Options are "aggregate", "individual", or "both" (default).}

\item{return_graphs_of_fits}{TRUE (default) or FALSE for whether to return a
list of the graphs showing the fitted data any time the dose 1 AUC was
extrapolated to infinity.}

\item{save_graphs_of_fits}{TRUE or FALSE (default) for whether to save png
files of graphs showing the fitted data for any time the dose 1 AUC was
extrapolated to infinity. This will save one png per set of file,
compoundID, inhibitor, and tissue.}

\item{ncol}{number of columns to use for graphing the fitted data}

\item{nrow}{number of rows to use for graphing the fitted data}

\item{fig_width}{figure width in inches for saving graphs of fits (you may
want this to be huge if there are a lot of profiles). Defaults to 8.5.}

\item{fig_height}{figure height in inches for saving graphs of fits (you may
want this to be huge if there are a lot of profiles). Defaults to 11.}
}
\value{
returns a list of individual and/or aggregate PK data
}
\description{
\code{calc_PK} calculates AUCinf_dose1, AUCt_dose1, AUCtau_last, Cmax_dose1,
Cmax_last, tmax_dose1, tmax_last, CLinf_dose1, and CLtau_last for the
supplied concentration-time data and, when applicable, the same parameters in
the presence of an effector and the ratios of those values for effector /
baseline. This can accommodate multiple simulations and multiple compounds as
long as the dosing regimen was the same. For example, this will do fine with
calculating the last-dose PK for two simulations where the last dose of
substrate was at t = 168 h but not where one simulation last dose was at t =
168 h and one was at t = 48 h. If you've got a scenario like that, we
recommend running this function multiple times, one for each dosing regimen.
The input required for \code{ct_dataframe} is pretty specific; please see the
help file for that argument.
}
\examples{
# None yet

}
