% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PK_ratios_mult.R
\name{calc_PK_ratios_mult}
\alias{calc_PK_ratios_mult}
\title{Calculate the ratio of PK parameters between multiple pairs of simulations}
\usage{
calc_PK_ratios_mult(
  PKparameters = NA,
  compoundToExtract = NA,
  tissue = NA,
  sheet_PKparameters = NA,
  existing_exp_details = NA,
  paired = TRUE,
  match_subjects_by = "individual and trial",
  distribution_type = "t",
  mean_type = "geometric",
  include_num_denom_columns = TRUE,
  conf_int = 0.9,
  includeConfInt = TRUE,
  includeCV = TRUE,
  include_dose_num = NA,
  prettify_columns = TRUE,
  prettify_compound_names = TRUE,
  rounding = NA,
  checkDataSource = TRUE,
  extract_forest_data = FALSE,
  save_table = NA,
  highlight_gmr_colors = NA,
  single_table = TRUE,
  page_orientation = "landscape",
  fontsize = 11,
  sim_data_file_pairs = "deprecated"
)
}
\arguments{
\item{PKparameters}{the PK parameters to include. There are two main options
  for this: 1) supply a file to read or a data.frame (R speak for "a table")
  that lists which simulation files, compounds, tissues, and PK you want or
  2) supply a character vector of which PK parameters you want and then also
  specify what you need in terms of which tissue, which compound, which
  simulation files, and which tab to get the data from with the arguments
  \code{tissue}, \code{compoundToExtract}, \code{sim_data_file_numerator},
  \code{sim_data_file_denominator}, and \code{sheet_PKparameters}.
  \strong{Details on each option:} \describe{\item{\strong{Option 1: }a file to read or a data.frame}{This
  is the most versatile option and, we think, the clearest in terms of
  getting what you expected. Please try running \code{\link{make_example_PK_input}}
  to see examples for how to set up a csv or Excel file or data.frame to
  specify exactly which simulation file should get which PK parameter from
  which tissue and, when user-specified intervals are involved, from which
  tab in the Excel file those data should be pulled. Whatever you supply, the
  columns that will be read are: \describe{\item{"Numerator_File"}{this is the
  same thing as the argument \code{sim_data_file_numerator}}

  \item{"Denominator_File"}{This is the same thing as the argument \code{sim_data_file_denominator})}

  \item{"Numerator_Sheet" or "Denominator_Sheet"}{When it's a user-defined AUC interval you want,
  this specifies which sheet in the Simulator output Excel file to use for
  the corresponding PK parameter in the numerator or denominator,
  respectively. If it's a regular first-dose or last-dose PK parameter,
  leave this blank or as NA; we know which sheets to use for those
  values. You can specify as many different sheets as needed with one row
  for every new sheet.}

  \item{"Numerator_Tissue" or "Denominator_Tissue"}{This is the same as the
  argument \code{tissue} except you can specify it separately for each set of PK parameters.}

  \item{"Numerator_CompoundID" or "Denominator_CompoundID"}{This is the same
  as the argument \code{compoundToExtract} except that you can specify it
  separately for each set of PK parameters.}

  If you omit any of those columns, whatever you supply for their respective
  arguments will be used instead. Note that the respective arguments will
  use the \emph{same value} for both the numerator and the denominator
  simulations. If you supply something for one of them
  in the data.frame or file and \emph{also} something for its argument, the
  argument will be ignored. \cr

  Here is how to specify each of the possible
  inputs for Option 1:\describe{\item{a csv file}{list the file
  name, including the file extension ".csv" inside quotes, e.g., "PK needed.csv"}

  \item{an Excel file}{list the file name, including the file extension
  ".xlsx", inside quotes, e.g., "PK needed.xlsx". We will be looking for a
  tab titled "PKparameters" (all one word and with the same capitalization).}

  \item{a data.frame}{If you'd like to supply a data.frame with the same
  columns you would have had in the .csv or Excel file, that works just the
  same.}}}}

  \item{\strong{Option 2: }specify just the PK parameters you want}{This is
  a good option when you want the same information
  from all your simulations. List the PK parameters you want here and then,
  in the arguments
  \code{tissue}, \code{compoundToExtract}, and
  \code{sheet_PKparameters} specify what you want for each of those. If
  you're going this route, here are the two options
  you have for the argument \code{PKparameters}: \describe{

  \item{NA}{If you leave this as NA, by default, if you have a single-dose
  simulation, the parameters will
  include AUC and Cmax for dose 1, or, if you have a multiple-dose
  simulation, AUC and Cmax for the last dose. Also by default, if you have a
  perpetrator present, the parameters will include the AUC and Cmax values with
  and without the perpetrator as well as those ratios.}

  \item{a character vector of any combination of specific, individual
  parameters}{This character vector must contain SimcypConsultancy package
  coded names for each parameter you want, e.g., \code{c("Cmax_dose1",
  "AUCtau_last").} Be sure to encapsulate the parameters you want with
  \code{c(...)}. Please try running \code{\link{make_example_PK_input}} to
  see examples, or, to see the full set of all possible parameters to extract, enter
  \code{view(PKParameterDefinitions)} into the console.}}}}

  Parameters that don't make sense for your scenario -- such as asking for
  \code{AUCinf_dose1_withInhib} when your simulation did not include a
  perpetrator -- will be ignored.}

\item{compoundToExtract}{For which compound do you want to extract PK data?
Options are: \itemize{\item{"substrate" (default if left as NA),} \item{"primary
metabolite 1",} \item{"primary metabolite 2",} \item{"secondary
metabolite",} \item{"inhibitor 1" -- this can be an inducer, inhibitor,
activator, or suppressor, but it's labeled as "Inhibitor 1" in the
simulator,} \item{"inhibitor 2" for the 2nd perpetrator listed in the
simulation,} \item{"inhibitor 1 metabolite" for the primary metabolite of
inhibitor 1}} If you have want one compound for the numerator PK and a
different one for the denominator PK, that must be specified in a
data.frame or a csv file that you supply to the argument
\code{PKparameters}.}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are \itemize{\item{"plasma" (default if left as NA)}
\item{"unbound plasma"} \item{"blood"} \item{"unbound blood"}
\item{"peripheral plasma"} \item{"peripheral blood"}} If you want one
tissue for the numerator PK and a different one for the denominator PK,
that must be specified in a data.frame or a csv file that you supply to the
argument \code{PKparameters}.}

\item{sheet_PKparameters}{If you have a user-defined AUC interval and you
want the PK parameters for to be pulled from that specific tab in the
Simulator output Excel files, list that tab here. If you want standard
first-dose or last-dose PK parameters, leave this as the default NA; we
know where to find those. If you want one tab for the numerator simulation
and a different tab for the denominator simulation, that must be specified
in a data.frame or a csv file that you supply to the argument
\code{PKparameters}.}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} to get information about how the
simulations were set up, you can save some processing time by supplying
that object here, unquoted. If left as NA, this function will run
\code{extractExpDetails_mult} behind the scenes to figure out some
information about your experimental set up.}

\item{paired}{TRUE (default) or FALSE for whether the study design is paired,
as in, the subjects are \emph{identical} between the two simulations.
\strong{THIS IS AN IMPORTANT DISTINCTION AND WILL AFFECT HOW THE
CALCULATIONS ARE PERFORMED!} An example of a paired study would be a DDI
study where each subject has a measurement without the perpetrator of
interest and then has a second measurement \emph{with} the perpetrator. The
comparison is for repeated measurements of the \emph{same subject}. An
example of an unpaired study design would be comparing healthy volunteers
to subjects with hepatic impairment because those are measurements on
\emph{different} subjects. For paired study designs, the order of
operations is to calculate each subject's mean ratio and then to calculate
the mean of those ratios. For unpaired study designs, the order of
operations is to calculate the mean for the numerator simulation and then
divide it by the mean for the denominator simulation. Would this be clearer
if you could see the mathematical equations? We agree but can't easily
include equations in the help file. However, if you run this and save the
output to a Word file, the equations will be included in the output.}

\item{match_subjects_by}{For a paired study design, how would you like to
match your subjects? Options are "individual and trial" (default), which
matches by both the simulated individual ID number AND by the trial number,
or "individual only", which matches only by the individual ID number. This
will be ignored for unpaired study designs. Why are we even bothering with
this, you ask? Just to be totally safe, we normally match simulated
subjects by both the individual subject ID and by the trial number. We
thought that this would always work and would be the safest option, but we
discovered that, for some scenarios where you might expect the individuals
to be \emph{exactly} the same between two simulations, they actually were
randomly assigned to different trials. Mismatched subjects would lead to
inaccurate calculations, so we want to avoid that. If you use the default
setting of "individual and trial" but the trials are NOT the same between
simulations, you'll get a warning and no results, which we think is vastly
superior to getting \emph{incorrect} results.}

\item{distribution_type}{use a "t" distribution (default) or a "Z"
distribution. Note: The Simcyp Simulator calculates geometric confidence
intervals with a t distribution.}

\item{mean_type}{What kind of means and confidence intervals do you want
listed in the output table? Options are "arithmetic" or "geometric"
(default).}

\item{include_num_denom_columns}{TRUE (default) or FALSE for whether to
include columns in the output table for the numerator data alone and
columns for the denominator alone. For example, if you wanted to calculate
the dose 1 AUC ratio for cancer patients compared to healthy volunteers,
settting \code{include_num_denom_columns = TRUE} would give you that ratio
and also a column with summary statistics on the AUC for cancer patients
and a column with summary statistics on the AUC for healthy volunteers.
Setting it to FALSE would give you only the ratios.}

\item{conf_int}{confidence interval to use; default is 90\%}

\item{includeConfInt}{TRUE (default) or FALSE for whether to include whatever
confidence intervals were included in the simulator output file. Note that
the confidence intervals are geometric since that's what the simulator
outputs (see an AUC tab and the summary statistics; these values are the
ones for, e.g., "90\% confidence interval around the geometric mean(lower
limit)").}

\item{includeCV}{TRUE (default) or FALSE for whether to include rows for CV
in the table}

\item{include_dose_num}{NA (default), TRUE, or FALSE on whether to include
the dose number when listing the PK parameter. By default, the parameter
will be labeled, e.g., "Dose 1 Cmax ratio" or "Last dose AUCtau ratio", if
you have PK data for both the first dose and the last dose. Also by
default, if you have data only for the first dose or only for the last
dose, the dose number will be omitted and it will be labeled, e.g., "AUCtau
ratio" or "Cmax ratio". Set this to TRUE or FALSE as desired to override
the default behavior and get exactly what you want.}

\item{prettify_columns}{TRUE (default) or FALSE for whether to make easily
human-readable column names. TRUE makes pretty column names such as "AUCinf
(h*ng/mL)" whereas FALSE leaves the column with the R-friendly name from
\code{\link{extractPK}}, e.g., "AUCinf_dose1". We're still tweaking this to
make it look just right!}

\item{prettify_compound_names}{TRUE (default) or FALSE on whether to make
compound names prettier in the prettified column titles and in any Word
output files. This was designed for simulations where the substrate and any
metabolites, perpetrators, or perpetrator metabolites are among the
standard
options for the simulator, and leaving \code{prettify_compound_names =
TRUE} will make the name of those compounds something more human readable.
For example, "SV-Rifampicin-MD" will become "rifampicin", and
"Sim-Midazolam" will become "midazolam". Set each compound to the name
you'd prefer to see in your column titles if you would like something
different. For example, \code{prettify_compound_names = c("inhibitor" =
"teeswiftavir", "substrate" = "superstatin")}. Please note that "inhibitor"
includes \emph{all} the perpetrators and perpetrator metabolites present,
so, if you're setting the perpetrator name, you really should use something
like this
if you're including perpetrator metabolites: \code{prettify_compound_names =
c("inhibitor" = "teeswiftavir and 1-OH-teeswiftavir", "substrate" =
"superstatin")}.}

\item{rounding}{option for what rounding to perform, if any. Options are:
\describe{\item{NA or "Consultancy"}{All output will be rounded according
to Simcyp Consultancy Team standards: to three significant figures when the
value is < 100 or to the ones place if the value is >= 100. Please see the
function \code{\link{round_consultancy}}, which does the rounding here.}
\item{"none"}{No rounding will be performed.} \item{"significant X" where
"X" is a number}{Output will be rounded to X significant figures. "signif
X" also works fine.} \item{"round X" where "X" is a number}{Output will be
rounded to X digits} \item{"Word only"}{Output saved to Word or a csv file
will be rounded using the function \code{\link{round_consultancy}}, but
nothing will be rounded in the output R object. This can be useful when you
want to have nicely rounded and formatted output in a Word file but you
\emph{also} want to use the results from \code{calc_PK_ratios_mult} to make
forest plots, which requires numbers that are \emph{not} rounded.}}}

\item{checkDataSource}{TRUE (default) or FALSE for whether to include in the
output a data.frame that lists exactly where the data were pulled from the
simulator output file. Useful for QCing.}

\item{extract_forest_data}{TRUE or FALSE (default) to get forest-plot data at
the same time. If set to TRUE, this will return a list that includes data
formatted for use with the function \code{\link{forest_plot}}. This will
assume that the denominator is the baseline or control scenario and the
numerator is the comparison. In the output for this, the column "Dose_sub"
will contain the dose of the substrate in the denominator simualtions, and
the column "Dose_inhib" will contain the dose of the inhibitor (if there
was one) in the numerator simulations or the dose of the substrate in the
numerator simulations if there was not. If there was not an inhibitor, the
column "Inhibitor 1" will contain the file names for the numerator sims.}

\item{save_table}{optionally save the output table and, if requested, the QC
info, by supplying a file name in quotes here, e.g., "My nicely formatted
table.docx" or "My table.csv", depending on whether you'd prefer to have
the main PK table saved as a Word or csv file.  Do not include any slashes,
dollar signs, or periods in the file name. If you supply only the file
extension, e.g., \code{save_table = "docx"}, the name of the file will be
the file name plus "PK summary table" with that extension and output will
be located in the same folder as \code{sim_data_file}. If you supply
something other than just "docx" or just "csv" for the file name but you
leave off the file extension, we'll assume you want it to be ".csv". While
the main PK table data will be in whatever file format you requested, if
you set \code{checkDataSource = TRUE}, the QC data will be in a csv file on
its own and will have "- QC" added to the end of the file name.}

\item{highlight_gmr_colors}{optionally specify a set of colors to use for
highlighting geometric mean ratios for DDIs. Options are "yellow to red",
"green to red" or a vector of 4 colors of your choosing. If left as NA, no
highlighting for GMR level will be done.}

\item{single_table}{TRUE (default) or FALSE for whether to save all the PK
data in a single table or break the data up by tissue, compound ID, and
file into multiple tables. This only applies to the Word output.}

\item{page_orientation}{set the page orientation for the Word file output to
"portrait" or "landscape" (default)}

\item{fontsize}{the numeric font size for Word output. Default is 11 point.
This only applies when you save the table as a Word file.}

\item{sim_data_file_pairs}{DEPRECATED. This argument was used in the past to
specify which pairs of files to compare, but we are changing the way we're
asking you to tell us that. Now, we'd like you to give us this information
with the argument \code{PKarameters} and ONLY there. We apologize for the
inconvenience, but, to be frank, this is a pretty complicated function to
put together, and we needed to simplify things.}
}
\value{
A list or a data.frame of PK data that optionally includes where the
  data came from and data to use for making forest plots
}
\description{
\code{calc_PK_ratios_mult} matches PK data from pairs of simulator output
Excel files and calculates the mean and confidence intervals of the ratios of
the requested PK parameters. For detailed instructions and examples, please
see the SharePoint file "Simcyp PBPKConsult R Files - Simcyp PBPKConsult R
Files/SimcypConsultancy function examples and instructions/Calculating PK
ratios from separate simulations/Calculating-PK-ratios.docx". (Sorry, we are
unable to include a link to it here.)
}
\examples{
# No examples yet.

}
