% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PK_ratios_subfun.R
\name{calc_PK_ratios_subfun}
\alias{calc_PK_ratios_subfun}
\title{INTERNAL - subfunction for calc_PK_ratios and calc_PK_ratios_mult, which I
always should have made as just one function.}
\usage{
calc_PK_ratios_subfun(
  FilePairs_i,
  existing_exp_details,
  paired,
  match_subjects_by,
  mean_type,
  conf_int,
  distribution_type,
  conc_units,
  time_units
)
}
\arguments{
\item{FilePairs_i}{file pairs for this iteration}

\item{existing_exp_details}{existing_exp_details}

\item{paired}{TRUE (default) or FALSE}

\item{match_subjects_by}{"individual and trial" (default) or "individual
only", which matches only by the individual ID number.}

\item{mean_type}{"arithmetic" or "geometric" (default)}

\item{conf_int}{confidence interval to use; default 0.9}

\item{distribution_type}{use a "t" distribution (default) or a "Z"
distribution. Note: The Simcyp Simulator calculates geometric confidence
intervals with a t distribution.}

\item{PKparameters}{PK parameters as a tibble}
}
\value{
A list or a data.frame of PK data that optionally includes where the
  data came from
}
\description{
This will compare PK for ONE set of data, where that set is defined as ONE
compoundID, ONE tissue, ONE simulation file, and ONE interval (1st dose, last
dose, or a user-defined interval) for the numerator and ONE of each of those
for the denominator.
}
