% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKi.R
\name{calcKi}
\alias{calcKi}
\title{Calculate the Ki of an  inhibitor with the Cheng-Prusoff}
\usage{
calcKi(IC50, S, Km = NA)
}
\arguments{
\item{IC50}{The IC50 value for the inhibitor (uM)}

\item{S}{The concentration of the substrate used (uM)}

\item{Km}{The Michaelis constant for the affinity between the substrate and
the enzyme (uM)}
}
\value{
Returns the inhibitory constant Ki (uM) (numeric)
}
\description{
\code{calcKi} uses IC50, substrate concentration, and Km to calculate a Ki
value for an inhibitor
}
\examples{
calcKi(20, 5, 5)
# 10

}
