% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CertaraColors}
\alias{CertaraColors}
\title{Colors used in the Certara PowerPoint template}
\format{
A named character vector of colors
}
\usage{
CertaraColors
}
\description{
Colors used in the Certara PowerPoint template
}
\examples{
scales::show_col(CertaraColors)
}
\keyword{datasets}
