% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CertaraDir}
\alias{CertaraDir}
\title{Paths for Certara's large file directory and the SharePoint directory}
\format{
A two-item list:
\describe{
\item{LgFileDir}{the large-file directory}
\item{SharePtDir}{the SharePoint directory}}
}
\usage{
CertaraDir
}
\description{
Paths for Certara's large file directory and the SharePoint directory
}
\keyword{datasets}
