% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_wksz_interactions.R
\name{change_wksz_interactions}
\alias{change_wksz_interactions}
\title{Change interaction parameters in Simcyp Simulator workspace files}
\usage{
change_wksz_interactions(
  sim_workspace_files,
  new_sim_workspace_files = NA,
  interactions_to_set = NA,
  CompoundID = "inhibitor 1",
  CYP = NA,
  competitive_inhibition_switch = "off",
  Ki = NA,
  Ki_fumic = NA,
  TDI_switch = "off",
  Kapp = NA,
  kinact = NA,
  Kapp_fumic = NA,
  induction_IndC50_switch = "off",
  IndMax = NA,
  IndC50 = NA,
  IndC50_fuinc = NA,
  Ind_gamma = NA,
  induction_slope_switch = "off",
  Ind_slope = NA
)
}
\arguments{
\item{sim_workspace_files}{the set of workspace files to modify; must end in
".wksz"}

\item{new_sim_workspace_files}{optionally specify the new workspace file
names to use. If left as NA, the original workspace will be overwritten.
Otherwise, specify a character vector of file names to use, e.g.,
\code{new_sim_workspace_files = c("new file 1.wksz", "new file
2.wksz")}}

\item{interactions_to_set}{a data.frame of interaction parameters to use
instead of filling in each of the arguments below (CompoundID through
Ind_slope) with character vectors of values. Columns MUST match the names
of those arguments, e.g., "CompoundID", "CYP",
"competitive_inhibition_switch", "Ki", "Ki_fumic", "TDI_switch", "Kapp",
"kinact", "Kapp_fumic", "induction_IndC50_switch", "IndMax", "IndC50",
"IndC50_fuinc", "induction_slope_switch", "Ind_slope", and "Ind_gamma".}

\item{CompoundID}{Compound ID whose parameters you want to change. Options
are: \itemize{\item{"substrate",} \item{"primary metabolite 1",}
\item{"primary metabolite 2",} \item{"secondary metabolite",}
\item{"inhibitor 1" (default) -- this can be an inducer, inhibitor,
activator, or suppresesor, but it's labeled as "Inhibitor 1" in the
simulator,} \item{"inhibitor 2" for the 2nd inhibitor listed in the
simulation,} \item{"inhibitor 1 metabolite" for the primary metabolite of
inhibitor 1}}}

\item{CYP}{CYP whose interaction parameters you want to change. Note: This is
currently only set up to change CYP interaction parameters. If you need to
change interaction parameters for some other enzyme or transporter, please
talk to Laura Shireman.}

\item{competitive_inhibition_switch}{turn competitive inhibition "on" or
"off"}

\item{Ki}{Ki value to use for competitive inhibition (uM); if left as NA,
value will be set to 1E+06.}

\item{Ki_fumic}{fu,mic value to use for competitive inhibition; if left as
NA, value will be set to 1.}

\item{TDI_switch}{turn time-dependent inactivation "on" or "off" (default)}

\item{Kapp}{Kapp value to use for TDI (uM); if left as NA, value will be set
to 1E+06.}

\item{kinact}{kinact value to use for TDI (1/h); if left as NA, value will be
set to 0.}

\item{Kapp_fumic}{fu,mic value to use for TDI; if left as NA, value will be
set to 1.}

\item{induction_IndC50_switch}{turn "on" or "off" (default) the use of an
induction IndC50 value}

\item{IndMax}{maximum fold induction to use}

\item{IndC50}{IndC50 value to use (uM); if left as NA, value will be set to
1E+06.}

\item{IndC50_fuinc}{fu,mic value to use for the IndC50 incubation; if left as
NA, value will be set to 1.}

\item{Ind_gamma}{value to use for gamma aka the Hill coefficient for
induction; if left as NA, value will be set to 1.}

\item{induction_slope_switch}{turn "on" or "off" (default) the use of a slope
value for describing induction}

\item{Ind_slope}{value to use for slope in induction; if left as NA, value
will be set to 0.}
}
\value{
does not return anything in R but saves workspace files
}
\description{
\code{change_wksz_interactions} changes interaction parameters such as Kapp
or Ki or IndMax in workspace files (.wksz files) for then running with the
Simcyp Simulator. This will change \emph{all} the workspace files provided to
have the \emph{same} interaction parameters listed here. Please note that if
you want to change one interaction parameter for a given CYP, you must must
set \emph{all} the interaction parameters for that CYP to what you want. For
example, say your drug has a Ki of 1 uM and that it should have an IndMax of
2, but Indmax is not the correct value in your workspace file. You will need
to set \code{IndMax = 2} AND ALSO \code{Ki = 1}. Set the value to NA to use
the default Simcyp Simulator value. UNDER CONSTRUCTION.
}
\examples{
# None yet

}
