% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectors.R
\name{chartreuse}
\alias{chartreuse}
\title{Create a vector of yellow-greens}
\usage{
chartreuse(ncolors)
}
\arguments{
\item{ncolors}{number of colors desired}
}
\value{
a character vector of colors
}
\description{
\code{chartreuse} is useful for getting a set of yellows to greens for
graphs. These are from the RColorBrewer package, color set "YlGn", just to
give credit where it's due, but they can be any number in length rather than
only up to 9 colors.
}
\examples{
# Create a set of 10 yellows to chartreuses to greens.
blues(10)

# Try using scales::show_col() to visualize the colors, ex:
scales::show_col(blues(10))

}
