% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_accumulation.R
\name{check_accumulation}
\alias{check_accumulation}
\title{Graph for checking accumulation over multiple doses -- UNDER CONSTRUCTION}
\usage{
check_accumulation(
  CT,
  t0 = 0,
  timepoint = "Clast",
  mark_dosing = "red dotted",
  diff_cutoff = 0.05
)
}
\arguments{
\item{CT}{data.frame w/conc time data from extractConcTime}

\item{t0}{start time for compound being plotted}

\item{timepoint}{Options: Cmin, Cmax, C0, Clast -- Cmin would be whenever the
min was, C0 would be 1st time point available for that dose, and Clast
would be last time point available for that dose. Is that the best,
clearest way to refer to those? Clast generally means the last for *all*
times monitored, so I don't want it to be confusing.}

\item{mark_dosing}{"red dotted", "none", "blue dashed" etc. Any named R color
and any named linetype.}

\item{diff_cutoff}{}
}
\value{

}
\description{
UNDER CONSTRUCTION -- DoseNum must be complete. mark_dosing doesn't work for
custom dosing intervals at present. Haven't done much to make this "nice"
yet. Just using it for a project I'm working on and thought I'd add it. -LSh
}
\examples{

CT <- extractConcTime()


}
