% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_doseint.R
\name{check_doseint}
\alias{check_doseint}
\title{INTERNAL check whether AUC interval ends on a dosing interval}
\usage{
check_doseint(
  sim_data_file,
  existing_exp_details,
  compoundID,
  interval = NA,
  stop_or_warn_missing_file = "stop",
  warnings = "silent"
)
}
\arguments{
\item{sim_data_file}{sim_data_file}

\item{existing_exp_details}{existing_exp_details}

\item{compoundID}{compoundID}

\item{interval}{optionally include a specific interval that you want to check}

\item{stop_or_warn_missing_file}{stop or just warn if no info available. NB:
This is NOT stop or warn if the dose interval dosen't match.}

\item{warnings}{"silent" (default) or "show" (really, or anything else) to
see warnings in this function}
}
\value{
list of 1) message: "custom dosing" (as in, we can't tell whether the
  interval matched), "good", "mismatch last dose", "mismatch user-defined
  interval", or "can't check - missing file" and 2) data.frame of File,
  CompoundID, SimDuration, DoseInt_X, StartHr_X, NumDoses_X, LastDoseTime,
  IntervalRemaining, OneDoseIntRemaining, UserIntervalStart, UserIntervalEnd,
  UserIntervalStartGood, UserIntervalEndGood. If info is not available, e.g.,
  custom dosing or missing file, then not all of those columns will be
  included in the data.frame.
}
\description{
INTERNAL check whether AUC interval ends on a dosing interval
}
\examples{
# none yet
}
