% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_name.R
\name{check_file_name}
\alias{check_file_name}
\title{Check file names for naming standards}
\usage{
check_file_name(file_name)
}
\arguments{
\item{file_name}{name of data file}
}
\value{
Only prints messages
}
\description{
\code{check_file_name} checks that files are named according to USFDA
file-naming protocols using Cade's checklist.
}
\examples{
 
check_file_name(file_name = "Example simulator output MD.xlsx")
check_file_name(file_name = "abc1b-5mg-sd.xlsx")
check_file_name(file_name = "abc1b-5mg-sd-sdy101a-cohortb-increasedki-10x-decreased-indmax-2x.xlsx")



}
