% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_include_dose_num.R
\name{check_include_dose_num}
\alias{check_include_dose_num}
\title{INTERNAL PACKAGE USE ONLY. Check on whether to include the dose number in the
PK table headings}
\usage{
check_include_dose_num(PK, include_dose_num)
}
\arguments{
\item{PK}{table of PK data or col names of that table}

\item{include_dose_num}{meant to be a pass through when this is T or F and
only check when this is NA}
}
\value{
T or F for whether to include dose numbers
}
\description{
INTERNAL PACKAGE USE ONLY. Check on whether to include the dose number in the
PK table headings
}
\examples{
# none
}
