% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_simulator_initialized.R
\name{check_simulator_initialized}
\alias{check_simulator_initialized}
\title{check whether the Simcyp Simulator has been initialized}
\usage{
check_simulator_initialized()
}
\arguments{
\item{verbose}{}
}
\value{
TRUE or FALSE for whether the Simcyp Simulator has been initialized
  with Simcyp::Initialise.
}
\description{
INTERNAL. This is only for checking whether Simulator has been initialized
for functions that require the Simcyp package.
}
\examples{
check_simulator_initialized()

}
