% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_obs_data.R
\name{clean_obs_data}
\alias{clean_obs_data}
\title{Tidy up messy observed data. UNDER CONSTRUCTION!!!}
\usage{
clean_obs_data(
  untidy_data,
  untidy_data_sheet = NA,
  subject_column,
  time_column,
  DV_column,
  analyte_column,
  dose_column,
  day_column,
  age_column,
  weight_column,
  height_column,
  sex_column,
  SDSE_column,
  injection_site_column,
  dose_unit = "mg",
  weighting = 1,
  save_csv = NA
)
}
\arguments{
\item{untidy_data}{a data.frame to be tidied or a csv file with the untidy
data or an Excel file with the untidy data. This should be in quotes.}

\item{untidy_data_sheet}{if you supplied an Excel file with the untidy data,
specify what sheet to read here. This should be in quotes.}

\item{subject_column}{the column in the observed concentration-time
data.frame that contains the subject IDs, unquoted.}

\item{time_column}{the column in the untidy data.frame that contains times,
unquoted.}

\item{DV_column}{the column in the untidy data.frame that contains
concentration data, unquoted.}

\item{analyte_column}{the column in the untidy data.frame that contains the
analytes, unquoted.}

\item{dose_column}{the column in the untidy data.frame that contains the
doses, unquoted.}

\item{day_column}{the column in the untidy data.frame that contains the days,
unquoted.}

\item{age_column}{the column in the untidy data.frame that contains the ages,
unquoted.}

\item{weight_column}{the column in the untidy data.frame that contains the
weights, unquoted.}

\item{height_column}{the column in the untidy data.frame that contains
heights, unquoted.}

\item{sex_column}{the column in the untidy data.frame that contains sexes,
unquoted.}

\item{SDSE_column}{the column in the untidy data.frame that contains the
standard deviation or standard error, unquoted.}

\item{injection_site_column}{the column in the untidy data.frame that
contains the injection site, unquoted.}

\item{dose_unit}{the unit of dosing. Options are "mg" (default), "mg/m2", or
"mg/kg".}

\item{weighting}{weighting to use for parameter estimation. Defaults to 1.}

\item{save_csv}{optionally specify a file name for saving the tidied data as
a csv. If you have more than one dose level in the data, the files will be
split up by dose level, and we'll include a suffix with the dose amount at
the end of the file name.}
}
\value{
a list of tidied data, split by dose level
}
\description{
Tidy up messy observed data. UNDER CONSTRUCTION!!!
}
\examples{
# none yet 


}
