% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{confInt}
\alias{confInt}
\title{Calculate the X percent confidence interval}
\usage{
confInt(x, CI = 0.9, na.rm = TRUE, distribution_type = "t")
}
\arguments{
\item{x}{A vector of numbers}

\item{CI}{The confidence interval desired; default is 90%. Enter this
value as a decimal, e.g., 0.90.}

\item{na.rm}{Should NA values be removed? (logical)}

\item{distribution_type}{use a "t" distribution (default) or a "Z"
distribution. Note: The Simcyp Simulator calculates geometric confidence
intervals with a t distribution.}
}
\description{
\code{confInt} takes as input a numeric vector, the desired confidence
interval, and returns the lower and upper values for the confidence interval.
Note that, because calculating a confidence interval relies on an accurate
estimate of the mean and standard deviation, your data should be normally
distributed.
}
\examples{
x <- rnorm(100, 5, 1)
confInt(x)
confInt(x, CI = 0.95)

}
