% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_units.R
\name{convert_conc_units}
\alias{convert_conc_units}
\title{Convert the concentration units in a data.frame of concentration-time data}
\usage{
convert_conc_units(
  DF_to_convert,
  conc_units = "ng/mL",
  DF_with_good_units = NA,
  MW = NA
)
}
\arguments{
\item{DF_to_convert}{the data.frame of concentration-time data with units
that may need to be converted. If you're converting concentration units,
this must include the columns "Conc" and "Conc_units", and, if you're
converting time units, it must include the columns "Time", and
"Time_units". (Outputs from \code{\link{extractConcTime}},
\code{\link{extractConcTime_mult}}, and \code{\link{extractObsConcTime}}
work here.) If you want to convert between mass per volume and molar
concentrations and there are multiple compounds present, this must also
include the column "CompoundID".}

\item{conc_units}{concentration units to use, e.g. "ng/mL". Options for
concentration units are the same as the ones in the Excel form for PE data
entry.}

\item{DF_with_good_units}{if you would like to just match another data.frame,
supply here a data.frame that has the desired concentration and time units.
Options for concentration units are the same as the ones in the Excel form
for PE data entry, and options for time units are "hours", "days", and
"minutes".}

\item{MW}{optionally supply a molecular weight for the compound of interest
to enable conversions between mass per volume and molar concentrations. If
you have more than one compound, list the compoundID and the MW as a named
character vector, e.g., \code{MW = c("substrate" = 325.8, "inhibitor 1" =
705.6)} for a DDI simulation of midazolam with itraconazole. If you want to
make this type of conversion, the data.frame DF_to_convert must also
include the column "CompoundID". Acceptable compoundIDs are "substrate",
"primary metabolite 1", "primary metabolite 2", "secondary metabolite",
"inhibitor 1", "inhibitor 2", or "inhibitor 1 metabolite".}
}
\value{
a data.frame with converted concentration units
}
\description{
Convert the concentration units in a data.frame of concentration-time data
}
\examples{
# none yet

}
