% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_days_to_hours.R
\name{convert_days_to_hours}
\alias{convert_days_to_hours}
\title{Convert times in days to times in hours}
\usage{
convert_days_to_hours(days)
}
\arguments{
\item{days}{the days you want to convert to hours, which can be either a
character or numeric vector. This can include "day" or "days" and can be
from one day to a different day. For example, all of these are acceptable:
"day 1", "2", "days 5-6", "day 3 to day 4".}
}
\value{
a numeric vector
}
\description{
Convert times in days to times in hours
}
\examples{
convert_days_to_hours("day 1")
# Note that this will give you the full 24-hour period, so "day 1" will 
# give you the numeric vector of 0 and 24. 

# You don't need to include the "day" bit.
convert_days_to_hours(2)

# If you include days in a format anything like "day X to day Y" or
# "day X-Y", we'll give you a numeric vector that includes the full last day.
convert_days_to_hours("day 3 to 4")


}
