% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_unit_subfun}
\alias{convert_unit_subfun}
\title{INTERNAL. For use with pk_table, calc_PK_ratios, etc. This will convert time
and conc units as needed, including accounting for the time unit being in the
denominator for CL.}
\usage{
convert_unit_subfun(PKlist, existing_exp_details, time_units, conc_units)
}
\arguments{
\item{PKlist}{list of BOTH individual and aggregate PK}

\item{existing_exp_details}{existing_exp_details}

\item{time_units}{desired time units}

\item{conc_units}{desired conc units}
}
\value{
a list of two data.frames -- individual and aggregate -- PK
  parameters
}
\description{
INTERNAL. For use with pk_table, calc_PK_ratios, etc. This will convert time
and conc units as needed, including accounting for the time unit being in the
denominator for CL.
}
