% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_external_simulation_directory.R
\name{copy_report_files}
\alias{copy_report_files}
\title{Copy all the files in a draft simulation directory to a different folder}
\usage{
copy_report_files(
  draft_simulation_directory,
  sim_directory_folder,
  create_folder = "ask",
  overwrite = "ask"
)
}
\arguments{
\item{draft_simulation_directory}{This function was written with the idea
that what you supply here is the output from running either
\code{\link{make_simulation_directory}} or
\code{\link{find_report_file_locations}}, either of which will find all of
the simulation files (both workspaces and Excel outputs as well as any
observed XML overlay files) in a project folder. You can use this to copy
files in general, though, as long as what you supply for
draft_simulation_directory includes a column "File name" with the file
names and a column "Folder" that lists the path where that file is located
(relative paths from your current working directory are fine).}

\item{sim_directory_folder}{folder where you want to put the copies of your
simulation files.}

\item{create_folder}{If \code{sim_directory_folder} does not exist, would you
like it to be created? Options are "yes", "no", or "ask" (default).}

\item{overwrite}{If any of your simulation files already exist in
\code{sim_directory_folder}, would you like them to be overwritten? Options
are "yes", "no", and "ask" (default).}
}
\value{
Does not return any R objects; only copies files
}
\description{
We strongly recommend providing \strong{full file paths} here or
  you might not be copying what you think you're copying.
}
\examples{
copy_report_files(
   draft_simulation_directory = DraftSimDirectory,
   sim_directory_folder = "C:/Users/MyName/abc-1a/Modeling/Final sims")
}
