% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dose_rows.R
\name{create_dose_rows}
\alias{create_dose_rows}
\title{Create a csv file of dosing rows for XML files}
\usage{
create_dose_rows(
  dose_interval = 24,
  num_doses = NA,
  end_time = NA,
  custom_dosing_schedule = NA,
  dose_amount = 100,
  study_file = NA,
  subjID_column,
  subjIDs = NA,
  save_output = NA
)
}
\arguments{
\item{dose_interval}{the dosing interval in hours. Default is 24 for a QD
dosing regimen.}

\item{num_doses}{the number of doses to generate. If this is left as NA, then
the value for \code{end_time} will be used to determine the number of doses
administered.}

\item{end_time}{the end time of the dosing in hours. If \code{num_doses} is
filled out, that value will be used preferentially.}

\item{custom_dosing_schedule}{a custom dosing schedule to be used for each
subject in hours, e.g., \code{custom_dosing_schedule = c(0, 12, 24, 168,
180, 192)}; if this is filled out, values in \code{dose_interval},
\code{num_doses}, and \code{end_time} will all be ignored.}

\item{dose_amount}{the amount of the dose (units are not included). If this
amount varies, please use the \code{custom_dosing_schedule} option and
include one dose amount for each time. For example:
\code{custom_dosing_schedule = c(0, 24, 48), dose_amount = c(100, 50, 50)}
will generate doses of 100 mg at t = 0 and then 50 mg at t = 24 and 48
hours.}

\item{study_file}{optionally specify a file containing study information.
This will be used for figuring out the IDs for the subjects who will be
dosed.}

\item{subjID_column}{which column in \code{study_file} contains the subject
IDs; not quoted. This is the only column that matters in \code{study_file}.}

\item{subjIDs}{optionally specify subject IDs manually if you don't provide a
file for \code{study_file}}

\item{save_output}{the file name to use for saving the output as a csv; if
left as NA, this will generate a data.frame in R but no output will be
saved.}
}
\value{

}
\description{
\code{create_dose_rows} generates a data.frame of dosing times and amounts --
one for every subject if subject IDs are provided -- based on the dosing
regimen used. This is meant for generating XML files for use in Phoenix WNL.
}
\examples{

# QD dosing regimen of 100 mg
create_dose_rows(dose_interval = 24, num_doses = 4)

# QD dosing regimen of 100 mg for subjects A, B, and C
create_dose_rows(dose_interval = 24, num_doses = 4, 
                 subjIDs = c("A", "B", "C"))
                 
# Custom dosing regimen for subjects A, B, and C
create_dose_rows(custom_dosing_schedule = c(0, 12, 24, 48, 92, 168),
                 subjIDs = c("A", "B", "C"))
                 
                  
}
