% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot.R
\name{ct_plot}
\alias{ct_plot}
\title{ct_plot}
\usage{
ct_plot(
  sim_data_file,
  sim_data_cells,
  obs_data_file,
  obs_data_cells,
  xlab,
  ylab,
  save = FALSE,
  figname = NA
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulated
concentration-time data}

\item{sim_data_cells}{which cells in the Excel sheet "Conc Profiles
CSys(CPlasma)" contain your data, e.g., "D29:GU32". You don't need to
change this every time if the number of sampling stays the same.}

\item{obs_data_file}{names of the Excel file containing the observed
concentration-time data. This is the file that it is ready to be converted
to an XML file, not the file that contains only the digitized time and
concentration data.}

\item{obs_data_cells}{which cells contain the observed data, e.g., "A11:C259"}

\item{xlab}{text you want for the x-axis title}

\item{ylab}{text you want for the y-axis title}

\item{save}{TRUE or FALSE: Save the graph?}

\item{figname}{If \code{save} is TRUE, the file name for the output figure, e.g., "Study 001 SD.png"}
}
\value{

}
\description{
This function takes observed and simulated concentration-time data and makes
publication-quality graphs.
}
\examples{
# (These won't actually work yet. -LS)
sim_data_file <- "Delamanid output v8.xlsx"
sim_data_cells <- "D29:GU32"

obs_data_file <- "fig1-242-06-001-SD - for XML conversion.xlsx"
obs_data_cells <- "A11:C259"
ct_plot(sim_data_file, sim_data_cells, obs_data_file, obs_data_cells,
   xlab = "Time (hr)", ylab = "Concentration (uM)")

}
