% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot3.R
\name{ct_plot3}
\alias{ct_plot3}
\title{Concentration-time plots of the full time range, first dose, and last dose}
\usage{
ct_plot3(
  ct_dataframe,
  overlay = FALSE,
  save_graph = NA,
  fig_height = 6,
  fig_width = 5,
  ...
)
}
\arguments{
\item{ct_dataframe}{the input concentration-time data generated by
running either the function \code{\link{extractConcTime}} or the function
\code{\link{extractConcTime_mult}}}

\item{overlay}{TRUE or FALSE for whether to make overlaid graphs, i.e.,
whether to use \code{\link{ct_plot_overlay}} (TRUE) or use
\code{\link{ct_plot}} (FALSE) to generate the graphs}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My conc time graph.png". If you leave off ".png", it
will be saved as a png file, but if you specify a different file extension,
it will be saved as that file format. Acceptable extensions are "eps",
"ps", "jpeg", "jpg", "tiff", "png", "bmp", or "svg". Leaving this as NA
means the file will not be saved to disk.}

\item{fig_height}{figure height in inches; default is 6}

\item{fig_width}{figure width in inches; default is 5}

\item{...}{arguments that pass through to \code{\link{ct_plot}} or
\code{\link{ct_plot_overlay}}}
}
\value{

}
\description{
Conveniently group three concentration-time plots of multiple-dose data
together: \enumerate{\item{the full time range across the top,} \item{the
first dose in the lower left quadrant, and}  \item{the last dose in the lower
right quadrant.}} Calls on either \code{\link{ct_plot}} or
\code{\link{ct_plot_overlay}} to make the graphs. Requires the package
ggpubr. Does not work for single-dose scenarios.
}
\examples{

CT <- extractConcTime(sim_data_file = "Example simulator output.xlsx")
ct_plot3(ct_dataframe = CT, figure_type = "trial means",
         linear_or_log = "log")

CT <- extractConcTime_mult(sim_data_files = c("Simulator output 1.xlsx",
                                              "Simulator output 2.xlsx",
                                              "Simulator output 3.xlsx"),
                           tissues = c("plasma", "blood"))
ct_plot3(ct_dataframe = CT, overlay = TRUE, colorBy_column = File,
         facet1_column = Tissue)

}
