% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ct_plot.R
\name{shiny_ct_plot}
\alias{shiny_ct_plot}
\title{Run the ct_plot function through a graphical user interface}
\usage{
shiny_ct_plot()
}
\value{

}
\description{
Opens the shiny app version of the function \code{\link{ct_plot}}.
}
\examples{
# You don't need to specify any arguments to this function
# in the console. Just type the following:
shiny_ct_plot()

}
