% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot_mult.R
\name{ct_plot_mult}
\alias{ct_plot_mult}
\title{Make a single graphic with multiple, nicely arranged concentration-time
graphs}
\usage{
ct_plot_mult(
  sim_obs_dataframe,
  linear_or_log = "semi-log",
  file_order = NA,
  include_title = TRUE,
  graph_labels = TRUE,
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{sim_obs_dataframe}{the data.frame with multiple sets of
concentration-time data}

\item{linear_or_log}{the type of graph to be returned. Options: "semi-log",
"linear", "both vertical" (default, graphs are stacked vertically), or
"both horizontal" (graphs are side by side).}

\item{file_order}{optionally specify the order in which the files are graphed
with a character vector of the files in the order you would like. If you
would like to include graph titles to make it clear which graph is which,
you can also specify what the title should be for that file. Example of
acceptable input: \code{c("simfile3.xlsx" = "10 mg SD", "simfile1.xlsx" =
"50 mg SD", "simfile1.xlsx" = "100 mg SD")}. If you get an order that you
didn't think you specified, please double check that you have specified the
file names \emph{exactly} as they appear in \code{sim_obs_dataframe}.}

\item{include_title}{TRUE or FALSE on whether to include a title for each
small graph. NOTE: This is not currently working well when you choose to
get both linear and semi-log plots; the title overlaps the graph. UNDER
CONSTRUCTION.}

\item{graph_labels}{TRUE or FALSE for whether to include labels (A, B, C,
etc.) for each of the small graphs.}

\item{nrow}{number of rows of small graphs in the arrangement. If left as NA,
a reasonable guess will be made.}

\item{ncol}{number of columns of small graphs in the arrangement. If left as
NA, a reasonable guess will be made.}

\item{...}{arguments that pass through to \code{\link{ct_plot}}}
}
\value{

}
\description{
This function was designed for making nicely arranged concentration-time
graphs from several Simcyp Simulator output files all together.
}
\details{
\strong{A note on the order of the graphs:} This function arranges graphs
first by file, then by compound ID, and then by tissue, and all sorting is
alphabetical. However, since sorting alphabetically might not be the optimal
graph arrangement for your scenario, you \emph{can} specify the order of the
graphs using either the \code{file_order} argument or, if you're comfortable
with setting factors in R, by making any of File, CompoundID, Tissue, and
subsection_ADAM factor rather than character data and setting the levels how
you wish. If you're unfamiliar with setting factor levels in R and setting
\code{file_order} isn't achieving what you want, please ask a member of the R
Working Group for assistance.
}
\examples{
# No examples yet

}
