% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot_overlay.R
\name{ct_plot_overlay}
\alias{ct_plot_overlay}
\title{Overlay multiple data sets onto a single concentration-time graph}
\usage{
ct_plot_overlay(
  conctime_DF,
  colorBy = File,
  facet_column1 = NA,
  facet_column2 = NA
)
}
\arguments{
\item{conctime_DF}{the data.frame with multiple sets of concentration-time
data}

\item{colorBy}{What column in \code{conctime_DF} should be used for coloring
the lines and/or points on the graph? This should be unquoted, e.g.,
\code{colorBy = Tissue}.}

\item{facet_column1}{If you would like to break up your graph into small
multiples, you can break the graphs up by up to two columns in
\code{conctime_DF}. What should be the 1st column to break up the data by?
This should be unquoted.}

\item{facet_column2}{What should be the 2nd column to break up the data into
small multiples by? This should be unquoted.}
}
\value{

}
\description{
\code{ct_plot_overlay} is meant to be used in conjunction with
\code{\link{extractConcTime_mult}} to create single graphs with overlaid
concentration-time data from multiple Simcyp Simulator output files for easy
comparisons. UNDER CONSTRUCTION.
}
\examples{
# No examples yet


}
