% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_x_axis.R
\name{ct_x_axis}
\alias{ct_x_axis}
\title{Set up x axis in a conc-time plot}
\usage{
ct_x_axis(Data, time_range, t0)
}
\arguments{
\item{Data}{the data.frame containing conc-time data; this is the output from
extractConcTime}

\item{time_range}{the user-supplied input for time_range}

\item{t0}{the user-supplied input for t0}
}
\value{

}
\description{
This function is specifically for setting options for the x axis in a
concentration-time graph and is NOT meant to be called on its own.
}
