% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_y_axis.R
\name{ct_y_axis}
\alias{ct_y_axis}
\title{Set up y axis in a conc-time plot}
\usage{
ct_y_axis(
  y_compound_info,
  IsADAM = FALSE,
  subsection_ADAM = "free compound in lumen",
  is_enz_plot = FALSE,
  prettify_compound_names = TRUE
)
}
\arguments{
\item{y_compound_info}{a data.frame containing columns for subsection_ADAM
and Conc_units}

\item{IsADAM}{T or F for whether ADAM model used}

\item{subsection_ADAM}{which ADAM-model subsection the data include}

\item{is_enz_plot}{T or F for whether this was a plot of enzyme abundance}
}
\value{

}
\description{
This function is specifically for setting options for the y axis in a
SimcypConsultancy package concentration-time graph and is NOT meant to be
called on its own. The function \code{\link{scale_y_conc}} is more generic
and \emph{is} meant to be called on its own, so check that out if you're
looking to scale y axes quickly and easily and similar to how they look in
SimcypConsultancy concentration-time plots.
}
