% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP3A4_liver}
\alias{CYP3A4_liver}
\title{CYP3A4 enzyme-abundance data in the liver}
\format{
A data.frame
}
\usage{
CYP3A4_liver
}
\description{
Enzyem-abundance levels for liver CYP3A4 in the presence and absence of
multiple-dose ritonavir. Format of the data are the same as the output from
\code{\link{extractEnzAbund}}.
}
\keyword{datasets}
