% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deannotateDetails.R
\name{deannotateDetails}
\alias{deannotateDetails}
\title{De-annotate a data.frame of experimental details that was previously
annotated}
\usage{
deannotateDetails(existing_exp_details, apply_class = TRUE)
}
\arguments{
\item{existing_exp_details}{the annotated output from \code{extractExpDetails} or
\code{extractExpDetails_mult} where the argument "annotate_details" was set
to TRUE or the output from \code{annotateDetails}}

\item{apply_class}{TRUE (default) or FALSE for whether to set the class of
each column to numeric or character data based on the data classes listed
in \code{AllExpDetails}. If set to FALSE, all data classes will be
character.}
}
\value{
a list of simulator experimental details
}
\description{
\code{deannotateDetails} takes a data.frame that contains annotated
experimental details and converts it from having columns for each file and
rows for each experimental detail to having rows for each file and columns
for each detail.
}
\examples{

Annotated <- annotateDetails(MDZdetails)
Deannotated <- deannotateDetails(Annotated)

}
