% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_table.R
\name{demog_table}
\alias{demog_table}
\title{Make a table of demographics for a set of simulations}
\usage{
demog_table(
  demog_dataframe,
  demog_parameters = NA,
  mean_type = "geometric",
  variability_type = "90\% CI",
  variability_format = "to",
  break_down_by_sex = TRUE,
  rounding = NA,
  save_table = NA,
  sort_column,
  page_orientation = "landscape",
  fontsize = 11
)
}
\arguments{
\item{demog_dataframe}{the output from running \code{\link{extractDemog}}.
Optionally (and we recommend) with added observed demographic data, perhaps
from observed overlay XML files.}

\item{demog_parameters}{Which demographic parameters do you want to include?
Options are NA to include all of the parameters in demog_dataframe or a
chraacter vector of the columns in demog_dataframe that you want to
include. No need to include any columns that are not a demographic
parameter because we'll ignore them, e.g., you don't need to tell us to
include the columns "File" or "Individual", etc.)}

\item{mean_type}{What kind of means and CVs or confidence intervals do you
want listed in the output table? Options are "arithmetic" or "geometric"
(default).}

\item{variability_type}{What statistic would you like to use for reporting
the variability? Options are: \describe{\item{"90\% CI" (default)}{90\% confidence
interval; this will be geometric or arithmetic based on your choice for
\code{mean_type}} \item{"CV"}{coefficient of variation; this will be
geometric or arithmetic based on your choice for \code{mean_type}}
\item{"SD"}{arithmetic standard deviation} \item{"none"}{to get no 
variability stats included in the table}}}

\item{variability_format}{formatting used to indicate the variability When
the variability is concatenated. Options are "to" (default) to get output
like "X to Y", "hyphen" to get output like "X - Y", "brackets" to get
output like "[X, Y]", or "parentheses" for the eponymous symbol if you're
an American and a bracket if you're British, e.g., "(X, Y)". (Sorry for the
ambiguity; this was written by an American who didn't originally realize
that there was another name for parentheses.)}

\item{break_down_by_sex}{TRUE (default) or FALSE for whether to break down
the data by sex.}

\item{rounding}{option for what rounding to perform, if any. Options are:
\describe{\item{NA or "Consultancy"}{All output will be rounded according
to Simcyp Consultancy Team standards: to three significant figures when the
value is < 100 or to the ones place if the value is >= 100. Please see the
function \code{\link{round_consultancy}}, which does the rounding here.}
\item{"none"}{No rounding will be performed.} \item{"significant X" where
"X" is a number}{Output will be rounded to X significant figures. "signif
X" also works fine.} \item{"round X" where "X" is a number}{Output will be
rounded to X digits} \item{"Word only"}{Output saved to Word or a csv file
will be rounded using the function \code{\link{round_consultancy}}, but
nothing will be rounded in the output R object. This can be useful when you
want to have nicely rounded and formatted output in a Word file but you
\emph{also} want to use the results from \code{pksummary_mult} to make
forest plots, which requires numbers that are \emph{not} rounded.}}}

\item{save_table}{optionally save the output table and, if requested, the QC
info, by supplying a file name in quotes here, e.g., "My nicely formatted
table.docx" or "My table.csv", depending on whether you'd prefer to have
the table saved as a Word or csv file.  Do not include any slashes, dollar
signs, or periods in the file name. (You can also save the table to a Word
file later with the function \code{\link{formatTable_Simcyp}}.) If you
supply only the file extension, e.g., \code{save_table = "docx"}, the name
of the file will be "PK summary table" with that extension. If you supply
something other than just "docx" or just "csv" for the file name but you
leave off the file extension, we'll assume you want it to be ".csv". All PK
info will be included in a single Word or csv file, and, if
\code{checkDataSource = TRUE}, that will be saved in a single csv file.}

\item{sort_column}{optionally specify a column to sort by. If none are
supplied, the table will not be sorted. If you would like to sort by more
than one column, we recommend sorting \emph{before} using this function,
e.g., \code{MyPKTable <- MyPKTable \%>\% arrange(Study, Dose)} to sort by
the column "Study" and then by the column "Dose" and \emph{then} supply
"MyPKTable" to \code{formatTable_Simcyp}. (This is just an example; your
table must include those two columns for that to work.)}

\item{page_orientation}{set the page orientation for the Word file output to
"portrait" (default) or "landscape"}

\item{fontsize}{the numeric font size for Word output. Default is 11 point.
This only applies when you save the table as a Word file.}

\item{shading_column}{If you would like to alternate the shading of the rows
in the output table, supply here the unquoted name of the column to check
for when to change the shading; every time that column's value changes, the
shading will alternate between white and light gray. For example, if you
have a table with PK values for multiple files and you have more than one
row per file (an example of this would be the output from the function
\code{\link{pksummary_mult}}), setting \code{shading_column = File} will
cause the shading of the rows to alternate between white and light gray
whenever the file changes. Please see the examples at the bottom of this
help file.}

\item{merge_shaded_cells}{TRUE (default) or FALSE for whether to merge the
cells that have the same shade. This only applies when one of the columns
in the input data.frame is used for deciding when to alternate shading,
that is, \code{shading_column} has a value.}

\item{merge_columns}{a vector of quoted column names or of numeric column
positions that should be merged vertically whenever the values are the
same. For example, \code{merge_columns = c("File", "Tissue")} will cause
the cells in the columns "File" and "Tissue" to merge vertically whenever
the same value shows up in consecutive rows. Similarly, \code{merge_columns
= c(1, 3, 5)} will merge vertically the 1st, 3rd, and 5th columns whenever
the values are the same. Note: This is different from most other functions
in the SimcypConsultancy package, which require unquoted column names.
Honestly, we just don't know how code things for you to supply a variable
number of unquoted column names for a single argument; we've just hit a
coding knowledge limitation here!}
}
\value{
a formatted table
}
\description{
\code{demog_table} makes a table of the demographics of a set of simulations
and, if appliable, observed data. UNDER CONSTRUCTION.
}
\examples{
# none yet

}
