% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_myeffector.R
\name{determine_myeffector}
\alias{determine_myeffector}
\title{Determine the (optionally pretty) name of all effectors included in a
simulation, all concatenated nicely}
\usage{
determine_myeffector(Deets, prettify_compound_names)
}
\arguments{
\item{Deets}{experimental details for only the single sim_data_file in
question rather than for all possible files that may have been contained in
existing_exp_details}

\item{prettify_compound_names}{pass through from parent function}
}
\description{
Determine the (optionally pretty) name of all effectors included in a
simulation, all concatenated nicely
}
\examples{
# None yet

}
