% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difftime_sim.R
\name{difftime_sim}
\alias{difftime_sim}
\title{Calculate the time difference between two times formatted as in simulator
output}
\usage{
difftime_sim(time1, time2, units = "hours")
}
\arguments{
\item{time1}{the earlier time in the simulator output, e.g., "Day 1, 09:00"}

\item{time2}{the later time in the simulator output, e.g., "Day 2, 09:00"}

\item{units}{units for the time. Options are "hours" (default), "minutes", or
"days".}
}
\value{
Returns a number
}
\description{
Calculate the time difference between two times formatted as in simulator
output
}
\examples{
difftime_sim(time1 = "Day 1, 09:00", time2 = "Day 14, 21:00")
difftime_sim(time1 = "Day 1, 09:00", time2 = "Day 1, 21:00",
             units = "minutes")
difftime_sim(time1 = "Day 1, 09:00", time2 = "Day 42, 21:00",
             units = "days")

}
