% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_DDI_section.R
\name{draft_DDI_section}
\alias{draft_DDI_section}
\title{Create a draft of DDI experimental and results sections in a report -- UNDER
CONSTRUCTION}
\usage{
draft_DDI_section(
  sim_data_file,
  exp_details,
  ct_dataframe,
  word_file_name,
  victim_sim = FALSE,
  PKparameters = NA,
  observed_PK = NA,
  tissue = "plasma",
  mean_type_PK = "geometric",
  mean_type_graph = "arithmetic",
  time_range_sub = NA,
  x_axis_interval_sub = NA,
  y_axis_label_sub = NA,
  y_axis_limits_log_sub = NA,
  legend_label_sub = NA,
  time_range_inhib = NA,
  x_axis_interval_inhib = NA,
  y_axis_label_inhib = NA,
  y_axis_limits_log_inhib = NA,
  clin_study_name = "XXX",
  prettify_compound_names = TRUE,
  default_cmpd_file = TRUE
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes}

\item{exp_details}{the output from running either
\code{\link{extractExpDetails}} or \code{\link{extractExpDetails_mult}} --
either is fine as long as it contains \code{sim_data_file}. Not quoted.}

\item{ct_dataframe}{concentration-time data generated by running the function
\code{\link{extractConcTime}} or \code{\link{extractConcTime_mult}} --
either is fine as long as it contains \code{sim_data_file} and the tissue
you want. Not quoted.}

\item{word_file_name}{the name of the Word file to use for saving the output}

\item{victim_sim}{TRUE (default) or FALSE for whether this was a victim DDI
simulation}

\item{PKparameters}{(optional) the PK parameters to include as a character
 vector. \itemize{

 \item{By default, if you have a single-dose simulation, the parameters will
 include AUC and Cmax for dose 1, and, if you have a multiple-dose
 simulation, AUC and Cmax for the last dose. Also by default, if you have an
 effector present, the parameters will include the AUC and Cmax values with
 and without the effector as well as those ratios.}

 \item{Alternatively, you can specify a vector of any combination of
 specific, individual parameters, e.g., \code{c("Cmax_dose1",
 "AUCtau_last").} Be sure to encapsulate the parameters you want with
 \code{c(...)}! To see the full set of possible parameters to extract, enter
 \code{view(PKParameterDefinitions)} into the console.}

 \item{If you supply observed data using either the argument
 \code{report_input_file} or the argument \code{observed_PK}, the only PK
 parameters that will be included are those available for the observed data.}

 \item{Parameters that don't make sense for your scenario -- such as asking
 for \code{AUCinf_dose1_withInhib} when your simulation did not include an
 inhibitor or effector -- will not be included.}

 \item{tmax will be listed as median, min, and max rather than mean, lower
 and higher X\% confidence interval or X percentiles. Similarly, if you
 request trial means, the values for tmax will be the range of medians for
 the trials rather than the range of means.}}

 An example of acceptable input here: \code{PKparameters = c("AUCtau_last",
 "AUCtau_last_withInhib", "Cmax_last", "Cmax_last_withInhib",
 "AUCtau_ratio_last", "Cmax_ratio_last")}.}

\item{observed_PK}{(optional) If you have a data.frame, a named numeric
vector, or an Excel or csv file with observed PK parameters, supply the full
file name in quotes or the data.frame or vector here, and the
simulated-to-observed mean ratios will be calculated. If you supply an Excel
file, it should have only one tab. We prefer supplying csv files here since
they're faster to read in anyway. The supplied data.frame or file must
include columns for each of the PK parameters you would like to compare, and
those column names \emph{must} be among the PK parameter options listed in
\code{PKParameterDefinitions}. If you would like the output table to include
the observed data CV for any of the parameters, add "_CV" to the end of the
parameter name, e.g., "AUCinf_dose1_CV". Please see the "Example" section of
this help file for examples of how to set this up.}

\item{tissue}{the tissue to use for extracting PK data and for graphing.
Options are "plasma" (default) or "blood".}

\item{mean_type_PK}{What kind of means and CVs do you want listed in the
output PK summary table? Options are "arithmetic" or "geometric" (default).}

\item{mean_type_graph}{graph "arithmetic" (default) or "geometric" means or
"median" for median concentrations in the concentration-time graph. If that
option was not included in the output, you'll get a warning and the graph
will include one that was.}

\item{time_range_sub}{time range to show relative to the start of the
 simulation for the graph of the substrate. Options: \describe{

 \item{NA}{(default) entire time range of data}

 \item{a start time and end time in hours}{only data in that time range, e.g.
 \code{c(24, 48)}. Note that there are no quotes around numeric data.}

 \item{"first dose"}{only the time range of the first dose}

 \item{"last dose"}{only the time range of the last dose}

 \item{"penultimate dose"}{only the time range of the 2nd-to-last dose, which
 can be useful for BID data where the end of the simulation extended past the
 dosing interval or data when the substrate was dosed BID and the effector
 was dosed QD}

 \item{a specific dose number with "dose" or "doses" as the prefix}{the time
 range encompassing the requested doses, e.g., \code{time_range = "dose 3"}
 for the 3rd dose or \code{time_range = "doses 1 to 4"} for doses 1 to 4}

 \item{"all obs" or "all observed" if you feel like spelling it out}{Time
 range will be limited to only times when observed data are present.}

 \item{"last dose to last observed" or "last obs" for short}{Time range will
 be limited to the start of the last dose until the last observed data
 point.}


 }}

\item{x_axis_interval_sub}{optionally set the x-axis major tick-mark interval
for the graph of the substrate. Acceptable input: any number or leave as NA
to accept default values, which are generally reasonable guesses as to
aesthetically pleasing and PK-relevant intervals.}

\item{y_axis_label_sub}{optionally supply a character vector or an expression
to use for the y axis label for the graph of the substrate}

\item{y_axis_limits_log_sub}{optionally set the Y axis limits for the semi-log
plot for the graph of the substrate, e.g., \code{c(10, 1000)}. Values will
be rounded down and up, respectively, to a round number. If left as the
default NA, the Y axis limits for the semi-log plot will be automatically
selected.}

\item{legend_label_sub}{optionally indicate on the legend of the graph of the
substrate whether the effector is an inhibitor, inducer, activator, or
suppressor. Input will be used as the label in the legend for the line style
and the shape. If left as the default NA when a legend is included and an
effector is present, the label in the legend will be "Inhibitor".}

\item{time_range_inhib}{time range to show relative to the start of the
simulation for the graph of the inhibitor. Options are the same as for the
substrate.}

\item{x_axis_interval_inhib}{optionally set the x-axis major tick-mark interval
for the graph of the inhibitor. Acceptable input: any number or leave as NA
to accept default values, which are generally reasonable guesses as to
aesthetically pleasing and PK-relevant intervals.}

\item{y_axis_label_inhib}{optionally supply a character vector or an expression
to use for the y axis label for the graph of the inhibitor}

\item{y_axis_limits_log_inhib}{optionally set the Y axis limits for the semi-log
plot for the graph of the inhibitor, e.g., \code{c(10, 1000)}. Values will
be rounded down and up, respectively, to a round number. If left as the
default NA, the Y axis limits for the semi-log plot will be automatically
selected.}

\item{clin_study_name}{the name of the clinical study for any observed data;
this only shows up in the text describing the table and graph. Any quoted
value is fine here.}

\item{prettify_compound_names}{TRUE (default) or FALSE on whether to make
compound names prettier in legend entries and in any Word output files. This
was designed for simulations where the substrate and any metabolites,
effectors, or effector metabolites are among the standard options for the
simulator, and leaving \code{prettify_compound_names = TRUE} will make the
name of those compounds something more human readable. For example,
"SV-Rifampicin-MD" will become "rifampicin", and "Sim-Midazolam" will become
"midazolam". Set each compound to the name you'd prefer to see in your
legend and Word output if you would like something different. For example,
\code{prettify_compound_names = c("inhibitor" = "teeswiftavir", "substrate"
= "superstatin")}. Please note that "inhibitor" includes \emph{all} the
effectors and effector metabolites present, so, if you're setting the
effector name, you really should use something like this if you're including
effector metabolites: \code{prettify_compound_names = c("inhibitor" =
"teeswiftavir and 1-OH-teeswiftavir", "substrate" = "superstatin")}.}

\item{default_cmpd_file}{Was one of the default compound files used for the
substrate (if this was a perpetrator simulation) or the effector (if this
was a victim simulation)? TRUE (default) or FALSE.}
}
\value{

}
\description{
\code{draft_DDI_section} outputs a Word document that can be copied
}
\examples{
# No examples yet. Under construction.


}
