% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_methods_text.R
\name{draft_methods_text}
\alias{draft_methods_text}
\title{Create some draft text to insert into the methods section of a report
describing the trial design of a simulation}
\usage{
draft_methods_text(
  sim_data_file,
  existing_exp_details,
  prettify_compound_names = TRUE,
  default_cmpd_file = TRUE,
  victim_sim = FALSE
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes}

\item{existing_exp_details}{the output from running either
\code{\link{extractExpDetails}} or \code{\link{extractExpDetails_mult}} --
either is fine as long as it contains details for \code{sim_data_file}.}

\item{prettify_compound_names}{TRUE (default) or FALSE for whether to make
compound names prettier in legend entries and in any Word output files.
This was designed for simulations where the substrate and any metabolites,
perpetrators, or perpetrator metabolites are among the standard options for
the simulator, and leaving \code{prettify_compound_names = TRUE} will make
the name of those compounds something more human readable. For example,
"SV-Rifampicin-MD" will become "rifampicin", and "Sim-Midazolam" will
become "midazolam".}

\item{default_cmpd_file}{Was one of the default compound files used for the
substrate (if this was a perpetrator simulation) or the perpetrator (if
this was a victim simulation)? TRUE (default) or FALSE. The only thing this
affects is the sentence in the template report text, "The default compound
library file for XXX was used."}

\item{victim_sim}{TRUE (default) or FALSE for whether this was a victim DDI
simulation, so "TRUE" means that the client's drug was the victim. The only
thing this affects is the sentence in the template report text "Performance
verification of the XXX model is provided in Appendix B - Performance
Verification of CYP3A Substrates, Inhibitors and Inducers." If this was a
victim DDI simulation, then this sentence will replace "XXX" with the name
of the perpetrator. If not, it will replace "XXX" with the name of the
substrate.}
}
\value{
list of generic methods text
}
\description{
Create some draft text to insert into the methods section of a report
describing the trial design of a simulation
}
\examples{
# none yet
}
