% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_results_text.R
\name{draft_results_text}
\alias{draft_results_text}
\title{Create some draft text to insert into the results section of a report
describing the trial design of a simulation}
\usage{
draft_results_text(
  sim_data_file,
  existing_exp_details,
  prettify_compound_names = TRUE,
  mean_type = "geometric"
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes}

\item{existing_exp_details}{the output from running either
\code{\link{extractExpDetails}} or \code{\link{extractExpDetails_mult}} --
either is fine as long as it contains details for \code{sim_data_file}.}

\item{prettify_compound_names}{TRUE (default) or FALSE for whether to make
compound names prettier in legend entries and in any Word output files.
This was designed for simulations where the substrate and any metabolites,
perpetrators, or perpetrator metabolites are among the standard options for
the simulator, and leaving \code{prettify_compound_names = TRUE} will make
the name of those compounds something more human readable. For example,
"SV-Rifampicin-MD" will become "rifampicin", and "Sim-Midazolam" will
become "midazolam".}

\item{mean_type}{"arithmetic" or "geometric" (default) means in PK tables}
}
\value{
list of generic results text
}
\description{
Create some draft text to insert into the results section of a report
describing the trial design of a simulation
}
\examples{
# none yet

}
