% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_text_mult.R
\name{draft_text_mult}
\alias{draft_text_mult}
\title{Title}
\usage{
draft_text_mult(
  sim_data_files,
  existing_exp_details,
  prettify_compound_names = TRUE,
  default_cmpd_file = TRUE,
  client_drug_regex = NA,
  mean_type = "geometric",
  save_text = NA
)
}
\arguments{
\item{sim_data_files}{files}

\item{existing_exp_details}{details}

\item{prettify_compound_names}{T or F}

\item{default_cmpd_file}{T or F}

\item{mean_type}{geometric or arithmetic}

\item{save_text}{word doc file name}

\item{victim_sim}{T or F}
}
\value{
list of study design info for a report and, optionall, a Word
  document with that info
}
\description{
Title
}
\examples{
none yet
}
