% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eCT_harmonize.R
\name{eCT_harmonize}
\alias{eCT_harmonize}
\title{Harmonize the compound names on sim_data_xl to made finding the correct rows
easier later.}
\usage{
eCT_harmonize(
  sim_data_xl,
  compoundToExtract,
  AllCompoundsPresent,
  tissue,
  PerpPresent,
  Deets,
  ADAM,
  AdvBrainModel
)
}
\arguments{
\item{sim_data_xl}{sim_data_xl}

\item{compoundToExtract}{compoundToExtract}

\item{PerpPresent}{T or F}

\item{ADAM}{T or F}

\item{AdvBrainModel}{T or F}
}
\value{
a data.frame of sim_data_xl with compound names and compound codes
  switched to, e.g., SUBSTRATE
}
\description{
INTERNAL USE ONLY.
}
\examples{
# none
}
