% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eCT_pulldata.R
\name{eCT_pulldata}
\alias{eCT_pulldata}
\title{Pull the appropriate data from the harmonized Excel tab}
\usage{
eCT_pulldata(
  sim_data_xl,
  cmpd,
  AllPerpsPresent,
  pull_interaction_data,
  fromMultFunction,
  Deets,
  ADAM,
  ss,
  AdvBrainModel,
  tissue,
  SimTimeUnits,
  SimConcUnits,
  returnAggregateOrIndiv
)
}
\arguments{
\item{sim_data_xl}{sim_data_xl}

\item{cmpd}{cmpd}

\item{AllPerpsPresent}{AllPerpsPresent}

\item{pull_interaction_data}{T or F}

\item{fromMultFunction}{T or F}

\item{Deets}{1 row}

\item{ADAM}{T or F}

\item{ss}{tissue subtype}

\item{AdvBrainModel}{T or F}

\item{tissue}{tissue}

\item{SimConcUnits}{SimConcUnits}

\item{returnAggregateOrIndiv}{returnAggregateOrIndiv}
}
\value{
data.frame of conc time data
}
\description{
INTERNAL USE ONLY.
}
\examples{
# none

}
