% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eCT_readxl.R
\name{eCT_readxl}
\alias{eCT_readxl}
\title{Read into memory an Excel sheet for extracting conc-time data.}
\usage{
eCT_readxl(
  sim_data_file,
  Deets,
  compoundToExtract,
  CompoundType,
  tissue,
  TissueType,
  SheetNames
)
}
\arguments{
\item{sim_data_file}{sim_data_file}

\item{Deets}{1 row data.frame of MainDetails for that sim_data_file}

\item{compoundToExtract}{compoundID}

\item{tissue}{tissue}

\item{TissueType}{systemic or tissue}

\item{SheetNames}{all sheets in sim_data_file}
}
\value{
sim_data_xl - the entire, unedited contents of the sheet.
}
\description{
INTERNAL USE ONLY. Determines correct sheet to read based on tissue and
compound ID requested.
}
\examples{
# none 

}
