% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enz_plot.R
\name{enz_plot}
\alias{enz_plot}
\title{Plots of enzyme abundance changes over time to match Consultancy template}
\usage{
enz_plot(
  sim_enz_dataframe = NA,
  figure_type = "percentiles",
  ...,
  sim_data_file = NA,
  enzyme = "CYP3A4",
  tissue = "liver"
)
}
\arguments{
\item{sim_enz_dataframe}{the data.frame of enzyme abundance data obtained
from running the function \code{\link{extractEnzAbund}}}

\item{figure_type}{The type of figure to graph. Default is "percentiles". See
\code{figure_type} options and their explanations for the function
\code{\link{ct_plot}}.}

\item{...}{other arguments passed to the function \code{\link{ct_plot}}}

\item{sim_data_file}{HISTORICAL, BACK-COMPATIBILITY PURPOSES ONLY: name of
the Excel file containing the simulated enzyme-abundance data}

\item{enzyme}{HISTORICAL, BACK-COMPATIBILITY PURPOSES ONLY: the enzyme to
plot, e.g., "CYP3A4" (default), "UGT1A1", etc. Spaces or hyphens in enzyme
names will be ignored. Not case sensitive.}

\item{tissue}{HISTORICAL, BACK-COMPATIBILITY PURPOSES ONLY: the tissue to
plot. Options are "liver" (default), "small intestine", "colon", or
"kidney".}
}
\description{
Using simulated enzyme-abundance data, make publication-quality graphs that
match the consultancy template formatting instructions. The options listed in
the function \code{\link{ct_plot}}, e.g., specifying the figure type or the
axis limits, etc., work here as long as they make sense for plotting
enzyme-abundance data.
}
\examples{
enz_plot(sim_data_file = "../Example simulator output MD.xlsx",
         enzyme = "CYP3A4", tissue = "liver", line_type = "dotted")
}
