% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sample_size_DDI.R
\name{estimate_sample_size_DDI}
\alias{estimate_sample_size_DDI}
\title{Estimate the sample size necessary for determining whether there's a
significant DDI - UNDER CONSTRUCTION}
\usage{
estimate_sample_size_DDI(
  alpha = 0.05,
  power = 0.8,
  sim_data_file = NA,
  existing_exp_details = NA,
  PKparameters = c("AUCinf_dose1", "AUCt_dose1", "Cmax_dose1"),
  compoundToExtract = "substrate",
  tissue = "plasma",
  sheet_PKparameters = NA
)
}
\arguments{
\item{alpha}{alpha required, default is 0.05 for the 95\% confidence level}

\item{power}{power (1-beta) required, default is 0.8.}

\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes. \strong{If you want more than one,
please supply a data.frame or .csv file to the argument \code{PKparameters}.}}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} or \code{\link{extractExpDetails}} to
get all the details from the "Input Sheet" (e.g., when you ran
extractExpDetails you said \code{exp_details = "Input Sheet"} or
\code{exp_details = "all"}), you can save some processing time by supplying
that object here, unquoted. If left as NA, this function will run
\code{extractExpDetails} behind the scenes to figure out some information
about your experimental set up.}

\item{PKparameters}{the PK parameters to include. For this, you only need to
  list baseline parameters; we'll find the matching DDI one for each. There
  are two main options for supplying this information: 1) supply a file to
  read or a data.frame (R speak for "a table") that lists which simulation
  files, compounds, tissues, and PK you want or 2) supply a character vector
  of which PK parameters you want and then also specify what you need in
  terms of which tissues, which compounds, which simulation files, and which
  tab to get the data from with the arguments \code{tissues},
  \code{compoundsToExtract}, \code{sim_data_files}, and
  \code{sheet_PKparameters}.
  \strong{Details on each option:} \describe{

  \item{\strong{Option 1: }a file to read or a data.frame}{This
  is the most versatile option and, we think, the clearest in terms of
  getting what you expected. Please try running \code{\link{make_example_PK_input}}
  to see examples for how to set up a csv or Excel file or data.frame to
  specify exactly which simulation file should get which PK parameter from
  which tissue and, when user-specified intervals are involved, from which
  tab in the Excel file those data should be pulled. Whatever you supply, the
  columns that will be read are: \itemize{\item{"File" (same thing as the argument
  \code{sim_data_files})} \item{"Sheet" (same thing as the argument
  \code{sheet_PKparameters})} \item{"Tissue" (same as the argument \code{tissues})}
  \item{"CompoundID" (same as the argument \code{compoundsToExtract})
  \item{"ObsValue" for any observed data (no equivalent argument)}
  \item{"Variability" for any observed variability (no equivalent argument
  here, either)}} If you
  omit any of those columns, whatever you supply for their respective
  arguments will be used instead. If you supply something for one of them
  in the data.frame or file and \emph{also} something for its argument, the
  argument will be ignored. \cr

  Here is how to specify each of the possible
  inputs for Option 1:\describe{\item{a csv file}{list the file
  name, including the file extension ".csv" inside quotes, e.g., "PK needed.csv"}

  \item{an Excel file}{list the file name, including the file extension
  ".xlsx", inside quotes, e.g., "PK needed.xlsx". We will be looking for a
  tab titled "PKparameters" (all one word and with the same capitalization).}

  \item{a data.frame}{If you'd like to supply a data.frame with the same
  columns you would have had in the .csv or Excel file, that works just the
  same.}}}}

  \item{\strong{Option 2: }specify the PK parameters you want for all your
  simulations}{This is a good option when you want the same information
  from all your simulations. List the PK parameters you want here and then,
  in the arguments
  \code{tissues}, \code{compoundsToExtract}, \code{sim_data_files}, and
  \code{sheet_PKparameters} specify which of each of those items you want.
  You'll get all possible combinations of these, so, for example, if you say
  \code{PKparameters = c("AUCinf_dose1", "Cmax_dose1")} and
  \code{tissues = c("blood", "plasma")}, you'll get the dose 1 AUCinf and
  Cmax for both blood and plasma for all the simulation files you list with
  \code{sim_data_files}. If you're going this route, here are the two options
  you have for the argument \code{PKparameters}: \describe{

  \item{NA}{If you leave this as NA, by default, if you have a single-dose
  simulation, the parameters will
  include AUC and Cmax for dose 1, or, if you have a multiple-dose
  simulation, AUC and Cmax for the last dose. Also by default, if you have a
  perpetrator present, the parameters will include the AUC and Cmax values with
  and without the perpetrator as well as those ratios.}

  \item{a character vector of any combination of specific, individual
  parameters}{This character vector must contain SimcypConsultancy package
  coded names for each parameter you want, e.g., \code{c("Cmax_dose1",
  "AUCtau_last").} Be sure to encapsulate the parameters you want with
  \code{c(...)}. Please try running \code{\link{make_example_PK_input}} to
  see examples, or, to see the full set of all possible parameters to extract, enter
  \code{view(PKParameterDefinitions)} into the console.}}}}

  Parameters that don't make sense for your scenario -- such as asking for
  \code{AUCinf_dose1_withInhib} when your simulation did not include a
  perpetrator -- will be ignored.}

\item{compoundToExtract}{For which compound do you want to extract
PK data? Options are: \itemize{\item{"substrate"
(default),} \item{"primary metabolite 1",} \item{"primary metabolite 2",}
\item{"secondary metabolite",} \item{"inhibitor 1" -- this can be an
inducer, inhibitor, activator, or suppresesor, but it's labeled as
"Inhibitor 1" in the simulator,} \item{"inhibitor 2" for the 2nd inhibitor
listed in the simulation,} \item{"inhibitor 1 metabolite" for the primary
metabolite of inhibitor 1}} \strong{If you want more than one,
please supply a data.frame or .csv file to the argument \code{PKparameters}.}}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are "plasma" (default), "unbound plasma", "blood", "unbound blood",
"peripheral plasma", or "peripheral blood". \strong{If you want more than one,
please supply a data.frame or .csv file to the argument \code{PKparameters}.}}

\item{sheet_PKparameters}{(optional) If you want the PK parameters to be
pulled from a \strong{user-defined interval tab} in the simulator output
file, list that tab here. Otherwise, this should be left as NA.
\code{sheet_PKparameters} can only have a \emph{single value}, though. If
you want some parameters from a custom-interval tab and others from the
regular tabs, you must supply that as part of a data.frame or csv file for
the argument \code{PKparameters}. Please try running
\code{\link{make_example_PK_input}} to see examples for how to do this.
\strong{If you want more than one,
please supply a data.frame or .csv file to the argument \code{PKparameters}.}}
}
\value{
a data.frame of the sample sizes needed for each combination of
  simulation file, tissue, compound ID, and PK parameter.
}
\description{
\code{estimate_sample_size_DDI} uses a DDI simulation with
  paired subjects to determine how large your sample size would need to be to
  reject the null hypothesis of no DDI. If alpha is 0.05, i.e., a 95\%
  confidence level, and 1 - beta is 0.8, i.e., 80\% power, this function
  answers the question: What is the estimated sample size needed such that,
  at the 95\% confidence level, the null hypothesis should be rejected 80\%
  percent of the time?
}
