% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExpDetailListItems}
\alias{ExpDetailListItems}
\title{The names of the items that should be in the list output from running
\code{\link{extractExpDetails}} or \code{\link{extractExpDetails_mult}}}
\format{
a character vector
}
\usage{
ExpDetailListItems
}
\description{
INTERNAL USE.
}
\keyword{datasets}
