% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fu.R
\name{extract_fu}
\alias{extract_fu}
\title{Extract fu,plasma values that change with concentration or with time from a
Simulator output Excel file. UNDER CONSTRUCTION.}
\usage{
extract_fu(sim_data_files, existing_exp_details = NA)
}
\arguments{
\item{sim_data_files}{the Simcyp Simulator Excel results files containing the
simulated time-dependent fu data, in quotes}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} or \code{\link{extractExpDetails}} to
get all the details from the "Input Sheet" (e.g., when you ran
extractExpDetails you said \code{exp_details = "Input Sheet"} or
\code{exp_details = "all"}), you can save some processing time by supplying
that object here, unquoted. If left as NA, this function will run
\code{extractExpDetails} behind the scenes to figure out some information
about your experimental set up.}
}
\value{
a data.frame
}
\description{
\code{extract_fu} extracts the fraction unbound in plasma from a
simulator output Excel file. A tab named something like "fu Profile (Sub)"
must be present. 

\itemize{\item{This currently only extracts data for the substrate, but if
you have an example with a different compound, please talk to Laura Shireman.}

\item{This has only been set up for baseline simulations without 
perpetrators. Please tell Laura Shireman if you need a DDI version.}}
}
