% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractConcTime.R
\name{extractConcTime}
\alias{extractConcTime}
\title{extractConcTime}
\usage{
extractConcTime(
  sim_data_file,
  obs_data_file = NA,
  returnAggregateOrIndiv = c("aggregate", "individual")
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulated
concentration-time data}

\item{obs_data_file}{name of the Excel file containing the observed
concentration-time data. If the observed data you want to plot were already
included in the Excel output from the simulator, leave this as NA.
Otherwise, this is the file that it is ready to be converted to an XML
file, not the file that contains only the digitized time and concentration
data.}

\item{returnAggregateOrIndiv}{Return aggregate and/or individual simulated
concentration-time data? Options are one or both of "aggregate" and
"individual".}
}
\value{
A data.frame of concentration-time data
}
\description{
Extracts concentration-time data from simulater output Excel files and,
optionally, a separately specified clinical data file, and puts all data into
a single, tidy data.frame.
}
\examples{
# sim_data_file <- "../Example simulator output MD.xlsx"
extractConcTime("../Example simulator output MD.xlsx")

}
