% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractConcTime_mult.R
\name{extractConcTime_mult}
\alias{extractConcTime_mult}
\title{Pull concentration-time data from multiple Simcyp Simulator output files}
\usage{
extractConcTime_mult(
  sim_data_files = NA,
  obs_data_files = NA,
  ct_dataframe = ConcTime,
  overwrite = FALSE,
  tissues = "plasma",
  compoundsToExtract = "all",
  conc_units_to_use = "ng/mL",
  time_units_to_use = "hours",
  returnAggregateOrIndiv = "aggregate",
  ...
)
}
\arguments{
\item{sim_data_files}{a character vector of simulator output files, each in
quotes and encapsulated with \code{c(...)}, or, if left as NA, \emph{all}
the Excel files in the current directory. Example of acceptable input:
\code{c("sim1.xlsx", "sim2.xlsx")}. The path should be included with the
file names if they are located somewhere other than your working directory.
If some of your Excel files are not regular simulator output, e.g. they are
sensitivity analyses or a file where you were doing some calculations,
those files will be skipped.}

\item{obs_data_files}{a character vector of the observed data files, each in
quotes and encapsulated with \code{c(...)}, that you'd like to compare,
e.g., \code{c("obsdata1.xlsx", "obsdata2.xlsx")}. The path should be
included with the file names if they are located somewhere other than your
working directory. This is the file that it is ready to be converted to an
XML file, not the file that contains only the digitized time and
concentration data. This function assumes that the dosing intervals for the
observed data match those in the simulated data. See "Details" for more
info.}

\item{ct_dataframe}{(optional) a data.frame that contains previously
extracted concentration-time data. This should NOT be in quotes. Because we
can see scenarios where you might want to extract some concentration-time
data, play around with those data, and then later decide you want to pull
more concentration-time data for comparisons, this data.frame can already
exist. When that is the case, this function will \emph{add} data to that
data.frame. It will \emph{not} overwrite existing data unless
\code{overwrite} is set to TRUE.}

\item{overwrite}{TRUE or FALSE (default) on whether to re-extract the
concentration-time data from output files that are already included in
\code{ct_dataframe}. Since pulling data from Excel files is slow, by
default, this will \emph{not} overwrite existing data and instead will only
add data from any Excel files that aren't already included. A situation
where you might want to set this to TRUE would be when you have changed
input parameters for simulations and re-run them.}

\item{tissues}{From which tissue(s) should the desired concentrations be
extracted? The default is plasma for typical plasma concentration-time
data. Other options are "blood" or any tissues included in "Sheet Options",
"Tissues" in the simulator. All possible options: "plasma", "blood",
"unbound blood", "unbound plasma", "additional organ", "adipose", "bone",
"brain", "feto-placenta", "GI tissue", "heart", "kidney", "liver", "lung",
"muscle", "pancreas", "peripheral blood", "peripheral plasma", "peripheral
unbound blood", "peripheral unbound plasma", "portal vein blood", "portal
vein plasma", "portal vein unbound blood", "portal vein unbound plasma",
"skin", or "spleen". Not case sensitive. Acceptable input is all tissues
desired as a character vector, e.g., \code{c("plasma", "blood", "liver")}.}

\item{compoundsToExtract}{For which compound(s) do you want to extract
concentration-time data? Options are: \itemize{\item{"substrate"
(default),} \item{"primary metabolite 1",} \item{"primary metabolite 2",}
\item{"secondary metabolite",} \item{"inhibitor 1" (this can be an inducer,
inhibitor, activator, or suppresesor, but it's labeled as "Inhibitor 1" in
the simulator),} \item{"inhibitor 2" for the 2nd inhibitor listed in the
simulation,} \item{"inhibitor 1 metabolite" for the primary metabolite of
inhibitor 1, or} \item{"all" for all possible compounds in the
simulation.}} Input to this argument should be all desired compounds as a
character vector, e.g., \code{c("substrate", "primary metabolite 1")}.}

\item{conc_units_to_use}{concentration units to use so that all data will be
comparable. Options are the same as the ones in the Excel form for PE data
entry. Default is "ng/mL". Note: ADAM model data concentration units are
not converted because there are simply too many units to manage easily, so
please check that the units are what you expected in the end.}

\item{time_units_to_use}{time units to use so that all data will be
comparable. Options are "hours" (default) or "minutes".}

\item{returnAggregateOrIndiv}{Return aggregate and/or individual simulated
concentration-time data? Options are "aggregate" (default), "individual",
or "both". Aggregated data are not calculated here but are pulled from the
simulator output rows labeled as "Population Statistics".}

\item{...}{other arguments passed to the function
\code{\link{extractConcTime}}}
}
\value{
Returns a large data.frame with multiple sets of concentration-time
  data, formatted the same way as output from the function
  \code{\link{extractConcTime}}
}
\description{
\code{extractConcTime_mult} is meant to be used in conjunction with
\code{\link{ct_plot_overlay}} or \code{\link{ct_plot_mult}} to create graphs
from multiple Simcyp Simulator output files. \strong{A couple of
notes:}\enumerate{\item{If ANY of the Excel files you wish to extract data
from are saved on SharePoint and are open, this WILL CRASH and WILL NOT save
whatever progress it has made so far. Be sure to close all of the source
Excel files.} \item{If you list multiple files, multiple tissues, and/or
multiple compounds to extract (see options below), this will extract
\emph{all} possible variations of them. For example, if you ask for
"sim1.xlsx" and "sim2.xlsx" and then also ask for "substrate" and "primary
metabolite 1", you will get the substrate and primary metabolite 1 data from
\emph{both} files.}}
}
\details{
\strong{Regarding dose intervals for observed data:} The observed data files
don't include information on dosing intervals or dose numbers, which makes it
a little tricky to figure out which dose number a given time in an observed
data file should have. If the compound IDs in the simulated data match those
in the observed data, we will assume that the dosing intervals are the same.
This was coded with the assumption that the dosing interval would be a round
number (subjects aren't getting dosed on the half hour, for example), so if
that's not the case, these dose number assignments will be off.
}
\examples{
ConcTimeData <-
      extractConcTime_mult(
            sim_data_files = c("MyFile1.xlsx", "MyFile2.xlsx"),
            ct_dataframe = "ConcTimeData",
            overwrite = FALSE,
            tissue = "unbound plasma") # Note that "tissue" is passed to "extractConcTime".

}
