% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDemog.R
\name{extractDemog}
\alias{extractDemog}
\title{Extract simulated demographic data from the "Demographic Data" tab of a
Simcyp Simulator ouptut Excel file}
\usage{
extractDemog(sim_data_files = NA, demog_dataframe = NA, overwrite = FALSE)
}
\arguments{
\item{sim_data_files}{a character vector of simulator output files, each in
quotes and encapsulated with \code{c(...)}, NA to extract
concentration-time data for \emph{all} the Excel files in the current
folder, or "recursive" to extract concentration-time data for \emph{all}
the Excel files in the current folder and \emph{all} subfolders. Example of
acceptable input: \code{c("sim1.xlsx", "sim2.xlsx")}. The path should be
included with the file names if they are located somewhere other than your
working directory. If some of your Excel files are not regular simulator
output, e.g. they are sensitivity analyses or a file where you were doing
some calculations, those files will be skipped.}

\item{demog_dataframe}{an existing data.frame of demographic data}

\item{overwrite}{TRUE or FALSE for whether to overwrite any existing data}
}
\value{
a data.frame of demographic data for simulated subjects. Columns
  included: File, Trial, Individual, Population, Sex, Age, Weight_kg,
  Height_cm, BSA_m2, BrainWt_g, KidneyWt_g, LiverWt_g, BMI_kgm2, CardiacOut,
  Haematocrit, HSA_gL, AGP_gL, Other_uM, Creatinine_umolL, GFR_mLminm2,
  RenalFunction, AllometricScalar, SimDuration, GFR_mLmin, ObsFile,
  SerumCreatinine_umolL, PhenotypeCYP2D6, and Simulated (TRUE for simulated
  data)
}
\description{
Extract simulated demographic data from the "Demographic Data" tab of a
Simcyp Simulator ouptut Excel file
}
\examples{
# none yet

}
