% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEnzAbund.R
\name{extractEnzAbund}
\alias{extractEnzAbund}
\title{Extract enzyme abundance data from a simulator output Excel file}
\usage{
extractEnzAbund(
  sim_data_file,
  enzyme = "CYP3A4",
  tissue = "liver",
  returnAggregateOrIndiv = c("aggregate", "individual")
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulated
enzyme-abundance-time data}

\item{enzyme}{the enzyme, e.g., "CYP3A4", "UGT1A1", etc. Spaces or hyphens in
enzyme names will be ignored. Not case sensitive.}

\item{tissue}{From which tissue should the desired enzyme abundance be
extracted? Options are "liver" (default), "gut", or "kidney".}

\item{returnAggregateOrIndiv}{Return aggregate and/or individual simulated
enzyme abundance data? Options are one or both of "aggregate" and
"individual". Aggregated data are not calculated here but are pulled from
the simulator output rows labeled as "mean".}
}
\value{
A data.frame of enzyme abundance with time with the following
  columns: \describe{

  \item{Enzyme}{the enzyme whose abundance is listed}

  \item{Tissue}{the tissue}

  \item{EffectorPresent}{TRUE or FALSE for whether the inhibitor was present
  at that time point}

  \item{Time}{the time since the first dose}

  \item{Abundance}{abundance of the enzyme listed}

  \item{Time_units}{units used for time}
}
}
\description{
Extracts enzyme abundance data from a simulator output Excel file. The
appropriate tab must be present in the output file.
}
\examples{
extractEnzAbund(sim_data_file = "../Example simulator output MD.xlsx",
                enzyme = "CYP3A4", tissue = "liver")
}
