% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpDetails.R
\name{extractExpDetails}
\alias{extractExpDetails}
\title{Extract details about the experimental design}
\usage{
extractExpDetails(
  sim_data_file,
  exp_details = "Summary tab",
  annotate_output = FALSE,
  save_output = NA
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes}

\item{annotate_output}{TRUE or FALSE (default) on whether to transpose the
rows and columns in the output, making the output table longer instead of
wider, and adding columns to the output for a) which compound the
information pertains to (substrate, inhibitor, etc.), b) which section of
the Simcyp Simulator this detail is found in (physchem, absorption,
distribution, etc.), c) notes describing what the detail is, and d) which
sheet in the Excel file the information was pulled from.}

\item{save_output}{optionally save the output by supplying a file name in
quotes here, e.g., "My experimental details.csv". If you leave off ".csv",
it will still be saved as a csv file.}
}
\value{
Returns a named list of the experimental details
}
\description{
\code{extractExpDetails} looks up experimental design details from a Simcyp
Simulator output file.
}
\examples{

extractExpDetails(sim_data_file = "../Example simulator output.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output MD + inhibitor.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output.xlsx",
                  exp_details = "all")



}
