% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpDetails_VBE.R
\name{extractExpDetails_VBE}
\alias{extractExpDetails_VBE}
\title{Extract details about the experimental design}
\usage{
extractExpDetails_VBE(sim_data_file)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes}
}
\value{
a named list
}
\description{
\code{extractExpDetails} looks up experimental design details
  from a Simcyp Simulator output file.

  For detailed instructions and examples, please see the SharePoint file
  "Simcyp PBPKConsult R Files - Simcyp PBPKConsult R Files/SimcypConsultancy
  function examples and instructions/Checking simulation experimental
  details/Checking-simulation-experimental-details.docx". (Sorry, we are
  unable to include a link to it here.)
}
