% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpDetails_XML.R
\name{extractExpDetails_XML}
\alias{extractExpDetails_XML}
\title{Extract experimental details for multiple Simcyp Simulator workspace files at
once -- UNDER CONSTRUCTION}
\usage{
extractExpDetails_XML(
  sim_workspace_files,
  compoundsToExtract = "all",
  exp_details = "all",
  save_output = NA
)
}
\arguments{
\item{sim_workspace_files}{a character vector of simulator output files, each
in quotes and encapsulated with \code{c(...)}, or NA to extract
experimental details for \emph{all} the Simulator workspace files in the
current folder. Example of acceptable input: \code{c("sim1.wksz",
"sim2.wksz")}.}

\item{compoundsToExtract}{For which compound do you want to extract
concentration-time data? Options are: \itemize{\item{"all" (default) for
all the possible compounds in the simulation (substrate, metabolites,
inhibitors, and ADC-related compounds)} \item{"substrate" (default),}
\item{"primary metabolite 1",} \item{"primary metabolite 2",}
\item{"secondary metabolite",} \item{"inhibitor 1" -- this can be an
inducer, inhibitor, activator, or suppresesor, but it's labeled as
"Inhibitor 1" in the simulator,} \item{"inhibitor 2" for the 2nd inhibitor
listed in the simulation,} \item{"inhibitor 1 metabolite" for the primary
metabolite of inhibitor 1,} \item{"conjugated protein" for DAR1-DARmax for
an antibody-drug conjugate; observed data with DV listed as "Conjugated
Protein Plasma Total" will match these simulated data,} \item{"total
protein" for DAR0-DARmax for an ADC; observed data with DV listed as "Total
Protein Conjugate Plasma Total" will match these simulated data, or}
\item{"released payload" for the released drug from an ADC, which shows up
as primary metabolite 1 in Simulator output files}} Input to this argument
should be all desired compounds as a character vector, e.g.,
\code{c("substrate", "primary metabolite 1")}. \strong{Note: If your
compound is a therapeutic protein or ADC, we haven't tested this very
thoroughly, so please be extra careful to check that you're getting the
correct data.}}

\item{exp_details}{experimental details you want to extract from the
simulator workspace files; currently "all" is the only acceptable input and
anything else will be ignored. These are much more limited than the options
for the function \code{\link{extractExpDetails}} and currently only include
calculated Peff,human, predicted Vss, the XML overlay file used for
observed data, and the XML file used for a fixed-trial design.}

\item{save_output}{optionally save the output by supplying a csv or Excel
file name in quotes here, e.g., "Simulation details.csv" or "Simulation
details.xlsx". If you leave off the file extension, it will be saved as a
csv file.}
}
\value{
Returns a data.frame of the experimental details
}
\description{
\code{extractExpDetails_XML} takes a character vector of Simcyp Simulator
workspaces -- or all the workspace files in the current directory if no files
are specified -- and collects experimental details for the simulations into a
single table. It optionally saves that table to a csv or Excel file.
}
\examples{

# None yet

}
