% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFmFe.R
\name{extractFmFe}
\alias{extractFmFe}
\title{Extract dynamic fm and fe values from a Simulator output Excel file}
\usage{
extractFmFe(
  sim_data_file,
  returnAggregateOrIndiv = "both",
  existing_exp_details = NA
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulated dynamic
fm and fe data, in quotes}

\item{returnAggregateOrIndiv}{Return aggregate and/or individual simulated
enzyme abundance data? Options are "individual", "aggregate", or "both"
(default). Aggregated data are not calculated here but are pulled from the
simulator output rows labeled as "mean".}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} or \code{\link{extractExpDetails}} to
get all the details from the "Input Sheet" (e.g., when you ran
extractExpDetails you said \code{exp_details = "Input Sheet"} or
\code{exp_details = "all"}), you can save some processing time by supplying
that object here, unquoted. If left as NA, this function will run
\code{extractExpDetails} behind the scenes to figure out some information
about your experimental set up.}
}
\value{
a data.frame
}
\description{
\code{extractFmFe} extracts enzyme abundance data from a simulator output
Excel file. A tab named something like "Time variant \%fm and fe" must be
present. A note: This has not yet been set up to deal with custom dosing
regimens, so if you have a simulation that involves that, please contact
Laura Shireman.
}
\examples{
# None yet

}
