% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInputTab.R
\name{extractInputTab}
\alias{extractInputTab}
\title{INTERNAL: Extract info from a tab formatted like the main "Input Sheet" tab
in Simulator output Excel files}
\usage{
extractInputTab(
  deets = "all",
  sim_data_file,
  sheet,
  VBE = FALSE,
  CustomDosing = FALSE
)
}
\arguments{
\item{deets}{details to extract (the coded name)}

\item{sim_data_file}{sim_data_file}

\item{sheet}{what sheet to read}
}
\value{
list of details from a tab formatted like the main "Input Sheet" tab
  in Simulator output Excel files
}
\description{
INTERNAL: Extract info from a tab formatted like the main "Input Sheet" tab
in Simulator output Excel files
}
