% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractObsConcTime_mult.R
\name{extractObsConcTime_mult}
\alias{extractObsConcTime_mult}
\title{Extract observed concentration-time data from multiple Excel files}
\usage{
extractObsConcTime_mult(obs_data_files = NA, returnDosingInfo = FALSE)
}
\arguments{
\item{obs_data_files}{a character vector of the names of the Excel files
containing the observed concentration-time data, each in quotes, NA
(default) to extract data from all possible observed data Excel files in
that folder, or "recursive" to extract data from all possible observed data
Excel files in the current folder and any subfolders as well. These are the
files that are \emph{ready} to be converted to XML files, not files that
contain only the digitized time and concentration data and not the XML
files themselves that you would include in a Simulator workspace for
observed data.}

\item{returnDosingInfo}{TRUE or FALSE (default) for whether to return a
second data.frame with dosing and demographic information from the Excel
file.}
}
\value{
a data.frame with the following columns:
  \describe{\item{Individual}{the individual ID}

  \item{CompoundID}{the compound ID listed in the observed file, e.g., "Sub
  Plasma", "Sub PM1 Plasma", "Sub (Inb) Plasma"}

  \item{Tissue}{the tissue}

  \item{Dose_sub, Dose_inhib, and/or Dose_inhib2}{the dose of the substrate
  and any perpetrators present}

  \item{Time}{time since dosing}

  \item{Conc}{concentration}

  \item{Time_units}{the units of measurement for the time column}

  \item{Conc_units}{the units of measurement for the concentration column}

  \item{Period, Age, Weight_kg, Height_cm, Sex, SerumCreatinine_umolL,
  HSA_gL, Haematocrit, PhenotypeCYP2D6, SmokingStatus}{the columns in the
  template for "Period" and "Covariates" but with R-friendly names.
  (Currently, no dosing information is pulled because the data format is
  different from what we need for other functions related to
  concentration-time data.)}}
}
\description{
Extract observed data from Excel files that follow the Simcyp Simulator
template for converting concentration-time data into an XML file.
}
\examples{
extractObsConcTime(obs_data_file = "My observed data.xlsx")

}
