% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractObsConcTime_xlsx-vdipbpkcons-83.R,
%   R/extractObsConcTime_xlsx.R
\name{extractObsConcTime_xlsx}
\alias{extractObsConcTime_xlsx}
\title{INTERNAL - Extract observed conc-time data from an xlsx file}
\usage{
extractObsConcTime_xlsx(obs_data_file)

extractObsConcTime_xlsx(obs_data_file)
}
\arguments{
\item{obs_data_file}{xlsx file used to make observed data overlay XML file}
}
\value{
a data.frame

a data.frame
}
\description{
INTERNAL - Extract observed conc-time data from an xlsx file

INTERNAL - Extract observed conc-time data from an xlsx file
}
