% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractObsConcTime_XML.R
\name{extractObsConcTime_XML}
\alias{extractObsConcTime_XML}
\title{INTERNAL - Extract observed conc-time data from an xml file}
\usage{
extractObsConcTime_XML(obs_data_file)
}
\arguments{
\item{obs_data_file}{xml file used for observed data in a workspace}
}
\value{
a data.frame of observed conc time data
}
\description{
INTERNAL - Extract observed conc-time data from an xml file
}
