% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPK_DB.R
\name{extractPK_DB}
\alias{extractPK_DB}
\title{Extract PK from a Simcyp Simulator database file -- UNDER CONSTRUCTION!!!!}
\usage{
extractPK_DB(
  sim_data_file,
  compoundToExtract = "substrate",
  PKparameters = NA,
  tissue = "plasma",
  existing_exp_details
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes}

\item{compoundToExtract}{For which compound do you want to extract
PK data? Options are: \itemize{\item{"substrate"
(default),} \item{"primary metabolite 1",} \item{"primary metabolite 2",}
\item{"secondary metabolite",} \item{"inhibitor 1" -- this can be an
inducer, inhibitor, activator, or suppresesor, but it's labeled as
"Inhibitor 1" in the simulator,} \item{"inhibitor 2" for the 2nd inhibitor
listed in the simulation,} \item{"inhibitor 1 metabolite" for the primary
metabolite of inhibitor 1}}}

\item{PKparameters}{PK parameters you want to extract from the simulator
  output file. Options are: \describe{

  \item{"all"}{all possible parameters}

  \item{"AUC tab"}{only those parameters on the "AUC" tab (default). For V21
  or later, this will look for tabs named something like
  "Int AUC 1st_CI(Sub)(CPlasma)" for the 1st dose of the substrate, for
  example, and pull all possible parameters there. For V21,
  and earlier, this will pull data from the tab literally named "AUC"
  or the "AUC_CI" tab or "AUC_SD" tab if a tab named "AUC"is not present.}

  \item{"Absorption tab"}{only those parameters on the "Absorption" or
  "Overall Fa Fg" tab}

  \item{"Regional ADAM"}{regional fraction absorbed and fraction metabolized
  from intestinal segments; only applies to ADAM models where the tab
  "Regional ADAM Fractions (Sub)" is included in the Excel file and currently
  only applies to substrate}

  \item{a vector of any combination of specific, individual parameters, each
  surrounded by quotes and encapsulated with \code{c(...)}}{An example:
  \code{c("Cmax_dose1", "AUCtau_last")}. To see the full set of possible
  parameters to extract, enter \code{view(PKParameterDefinitions)} into the
  console. Not case sensitive. If you use "_first" instead of "_dose1", that
  will also work.}}}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are "plasma" (default), "unbound plasma", "blood", "unbound blood",
"peripheral plasma", or "peripheral blood". \strong{NOTE: PK for peripheral
sampling is not as well tested as for other tissues and is only set up for 
V21+. Please check your results carefully.}}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} or \code{\link{extractExpDetails}} to
get all the details from the "Input Sheet" (e.g., when you ran
extractExpDetails you said \code{exp_details = "Input Sheet"} or
\code{exp_details = "all"}), you can save some processing time by supplying
that object here, unquoted. If left as NA, this function will run
\code{extractExpDetails} behind the scenes to figure out some information
about your experimental set up.}
}
\value{
a list: "individual" = individual PK data, "aggregate" = aggregate PK
  data, "TimeInterval" = the time intervals used
}
\description{
\code{extractPK_DB} pulls PK data from a database file and calculates summary
statistics for the data
}
\examples{
# none yet
}
