% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sims.R
\name{filter_sims}
\alias{filter_sims}
\title{Selectively include or omit specific simulations from an R object}
\usage{
filter_sims(which_object, which_sims, include_or_omit)
}
\arguments{
\item{which_object}{the R object for which you want to filter to remove or
retain only certain simulation files}

\item{which_sims}{which simulations are affected? Must be a character vector
of the file names \emph{exactly} as they appear in the column "File" if the
object is a data.frame or in the data.frames inside the list if the object
is a list}

\item{include_or_omit}{Do you want to "include" or "omit" these simulations?}
}
\value{
the R object with only those simulations if \code{include_or_omit} is
  set to "include" or that R object with those simulations omitted if
  \code{include_or_omit} is set to "omit".
}
\description{
Selectively include or omit specific simulations from an R object
}
\examples{
Details <- filter_sims(which_object = Details, 
                       which_sims = c("file 1.xlsx", "file 2.xlsx"), 
                       include_or_omit = "omit")


}
