% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_external_simulation_directory.R
\name{find_report_file_locations}
\alias{find_report_file_locations}
\title{Find the location of all the simulation files included in a report}
\usage{
find_report_file_locations(
  sims_in_report,
  project_folder = NA,
  save_draft_simulation_directory = NA,
  overwrite = "ask"
)
}
\arguments{
\item{sims_in_report}{output from running \code{\link{find_sims_in_report}},
a data.frame with columns titled "Index" (paragraph number), "Text" (the
text in that paragraph), "File" the simulation Excel output file name, and
"Table/Figure" for the matched figure or table where the simulation
appears.}

\item{project_folder}{location of the project folder to search. Please note
that R requires forward slashes ("/") in file paths rather than the back
slashes Microsoft uses. If left as NA (default), we'll assume your current
working directory is the top level of your project folder.}

\item{save_draft_simulation_directory}{optionally specify an Excel file name
for saving your simulation directory.}

\item{overwrite}{Should we overwrite if your Excel file already exists and
already has a tab named "Simulation directory"? Options are "yes" to always
overwrite, "no" to never overwrite, or "ask" (default), which means that we
will ask you whether to overwrite and give you a chance to supply a
different file name if the one you supplied already exists.}
}
\value{
a draft simulation directory as a data.frame
}
\description{
\code{find_report_file_locations} looks for simulation files
  that were included in a report. To do this, it relies on input from running
  \code{\link{find_sims_in_report}}.
}
\examples{
DraftSimDirectory <- find_report_file_locations(
   sims_in_report = ReportRefs,
   project_folder = "C:/Users/MyName/abc-1a/Modeling",
   save_draft_simulation_directory = "abc-1a draft sim directory.xlsx")


}
