% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_external_simulation_directory.R
\name{find_sims_in_report}
\alias{find_sims_in_report}
\title{Find references to simulations in a Word report}
\usage{
find_sims_in_report(report_name, save_sims_in_report = NA)
}
\arguments{
\item{report_name}{the file name of the report, including the path if it's
not in your current working directory}

\item{save_sims_in_report}{optionally save the list of the references it
found to a csv file so that you can check that the simulations and which
figure or table they refer to appear to be correct. For example, if you
spot a typo here, you can find it and fix it in your report and then re-run
this.}
}
\value{
an object containing the report text
}
\description{
\code{find_sims_in_report} reads a Word report into memory,
  finds all references to simulations in that Word report, and then looks at
  the text preceding each reference to determine which figure or table shows
  the results of each simulation listed in that report.

  For finding simulations, it looks for the text "Source simulated data:" and
  then all Excel files listed in that same paragraph. References to
  simulations MUST include the ".xlsx" file extension for R to find it.

  For determining which figure or table each reference is for, it looks for
  the immediate paragraph before each simulation reference that starts with
  "Figure X" or "Table X" where X is a number.
}
\examples{
ReportRefs <- find_sims_in_report(report_name = "abc-1a PBPK report.docx")

}
