% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDetails.R
\name{findDetails}
\alias{findDetails}
\title{Find specific details or sets of details from the output from
extractExpDetails}
\usage{
findDetails(
  Deets,
  compoundID = NA,
  compound = NA,
  show_compound_col = TRUE,
  detail_set = NA,
  simulator_section = NA
)
}
\arguments{
\item{Deets}{output from \code{\link{extractExpDetails}} or
\code{\link{extractExpDetails_mult}}}

\item{compoundID}{optionally supply one or more of "substrate", "primary
metabolite 1", "primary metabolite 2", "secondary metabolite", "inhibitor
1", "inhibitor 2", or "inhibitor 1 metabolite" to get information on
that/those compound(s). Remember to contain more than one value with
\code{c(...)}!}

\item{compound}{optionally supply a specific compound name or part of a
specific compound name to get all possible compounds that match that.
Regular expressions are acceptable here, e.g., \code{compound =
"midaz|keto"} to find any compound with either "midaz" or "keto" in the
name. Not case sensitive.}

\item{show_compound_col}{TRUE (default), FALSE, or "concatenate" for whether
to include in the results the column "Compound", which is the compound's
specific name in each simulation. Why would you ever omit this? If you have
a compound with a slightly different name across multiple simulations,
e.g., "DrugX" and "Drug X", and "Drug X - reduced Ki", you'll get a new row
for every possible combination of "Compound" and "Detail", which might not
make for easy comparisons. For example, a Ki value for "DrugX" will be in
one row and the same Ki value for "Drug X" will be on a separate row. Try
setting this to TRUE when you have similarly named compounds that really
should be compared and see how that compares to the default. If you set
this to "concatenate", you'll get all the possible compound names together;
for example, you might see "DrugX, Drug X, or Drug X - reduced Ki" listed
as the compound.}

\item{detail_set}{optionally supply a set of details to extract where options
  are the same as the ones listed as possiblilities for \code{exp_details}
  for \code{\link{extractExpDetails}} or
  \code{\link{extractExpDetails_mult}}: \describe{

  \item{"Summary tab"}{Extract details available from the "Summary tab"
  (default)}

  \item{"Input Sheet"}{Extract details available from the "Input Sheet" tab}

  \item{"population tab"}{Extract details about the population used (data
  come from the tab with the same name as the population simulated)}

  \item{"Simcyp inputs"}{Extract all the details that you normally fill out
  on the "Simcyp inputs (and QC)" tab of a compound data sheet}

  \item{"all"}{Extract all possible parameters}

  \item{a string of the specific parameters you want, each in quotes and
  encapsulated with \code{c(...)},}{For a complete list, type
  \code{data(ExpDetailDefinitions); view(ExpDetailDefinitions)} into the
  console. Parameters are reported with a suffix depending on which compound
  they pertain to: "_sub" for the substrate, "_met1" for the primary
  metabolite, "_met2" for the second primary metabolite, "_secmet" for the
  secondary metabolite, "_inhib" for the 1st inhibitor or inducer listed,
  "_inhib2" for the 2nd inhibitor or inducer listed, or "_inh1met" for the
  inhibitor 1 metabolite. An example of acceptable input: \code{c("pKa1_sub",
  "fa_inhib2", "Regimen_sub")}}  Not case sensitive.}#'}

\item{simulator_section}{optionally supply a specific simulator section from
which to find simulation setup details. Options are "Absorption",
"Distribution", "Elimination", "Interaction", "Phys Chem and Blood
Binding", "Population", "Transport", or "Trial Design". Not case sensitive.}
}
\value{

}
\description{
Nota bene: If you ran \code{\link{annotateDetails}} on your data first and,
when you ran that, you asked to concatenate the values in the compound
column, please understand that you cannot go backwards here an un-concatenate
the compound names. You'll still get concatenated compound names in the
output.
}
\details{
\code{findDetails} takes output from \code{\link{extractExpDetails}} or
\code{\link{extractExpDetails_mult}}, annotates it, and finds specific
details or sets of details from that output. For example, it can find all the
details on anything that was used as "inhibitor 1" or all the details for a
compound with "midazo" in the name. You can also search by Simulator section
to get everything on absorption, for example, or everything on interactions.
}
\examples{

# Find details regarding absorption.
findDetails(Deets = MyDetails, simulator_section = "absorption")

# Find details on whatever was used as inhibitor 1
findDetails(Deets = MyDetails, compoundID = "inhibitor 1")

# Find details for any compounds with "midaz" or "keto" in the name.
findDetails(Deets = MyDetails, compound = "midaz|keto")

# Find (most of) the details to put into a table of simulation inputs:
findDetails(Deets = MyDetails, detail_set = "Simcyp inputs")

# Combine multiple options to get just a few specific details:
findDetails(Deets = MyDetails, simulator_section = "absorption",
            compound = "midaz|keto", compoundID = "substrate")

}
