% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_scripts.R
\name{format_scripts}
\alias{format_scripts}
\title{Format common PBPK model parameters to have appropriate subscripts and
superscripts. UNDER CONSTRUCTION.}
\usage{
format_scripts(DF, FT = NA, parameter_column)
}
\arguments{
\item{DF}{a data.frame with a column containing, e.g., "ka (h^-1)", "fu,p",
or other PBPK model parameters with syntax that matches what's in the
column "Notes" in the object AllExpDetails, which is included with this
package and is used extensively for figuring out how simulations were set
up.}

\item{FT}{optionally supply a flextable if you've already got one started.
This MUST be the flextable version of DF because DF is what we'll use to
figure out which cells should be changed to what values and NOT the
flextable itself. If you don't supply a flextable for FT, we'll make DF
into a generic flextable and use that, and that -- NOT supplying anything
for FT -- is actually the preferable option. Honestly, we're tempted to
just not let you supply a separate flextable and ONLY use DF to make the
output, but we're trying to be flexible here. So please play nice and ONLY
supply a value for FT that has EVERYTHING in the same position as it is in
DF.}

\item{parameter_column}{the name of the column in DF that contains your PK
parameters, unquoted. For example, supply \code{parameter_column = Notes}}
}
\value{
a flextable
}
\description{
Format common PBPK model parameters to have appropriate subscripts and
superscripts. UNDER CONSTRUCTION.
}
\examples{
DF <- data.frame(Parameter = c("MW", "ka (h^-1)",
                               "fu,p", "fu,gut", "fu,mic CYP3A4"),
                  Value = c(350, 0.5, 0.01, 0.8, 1))

format_scripts(DF = DF)

}
