% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table_simple.R
\name{format_table_simple}
\alias{format_table_simple}
\title{Format a table rather simply to look nice in a Word file}
\usage{
format_table_simple(
  DF,
  shading_column,
  merge_shaded_cells = TRUE,
  merge_columns = NA,
  font = "Palatino Linotype",
  fontsize = 11,
  column_widths = NA,
  alignment = "left",
  save_table = NA,
  page_orientation = "portrait",
  title_document = NA,
  table_caption = NA
)
}
\arguments{
\item{DF}{a data.frame}

\item{shading_column}{If you would like to alternate the shading of the rows
in the output table, supply here the unquoted name of the column to check
for when to change the shading; every time that column's value changes, the
shading will alternate between white and light gray. For example, if you
have a table with PK values for multiple files and you have more than one
row per file (an example of this would be the output from the function
\code{\link{pksummary_mult}}), setting \code{shading_column = File} will
cause the shading of the rows to alternate between white and light gray
whenever the file changes. Please see the examples at the bottom of this
help file.}

\item{merge_shaded_cells}{TRUE (default) or FALSE for whether to merge the
cells that have the same shade. This only applies when one of the columns
in the input data.frame is used for deciding when to alternate shading,
that is, \code{shading_column} has a value.}

\item{merge_columns}{a vector of quoted column names or of numeric column
positions that should be merged vertically whenever the values are the
same. For example, \code{merge_columns = c("File", "Tissue")} will cause
the cells in the columns "File" and "Tissue" to merge vertically whenever
the same value shows up in consecutive rows. Similarly, \code{merge_columns
= c(1, 3, 5)} will merge vertically the 1st, 3rd, and 5th columns whenever
the values are the same. Note: This is different from most other functions
in the SimcypConsultancy package, which require unquoted column names.
Honestly, we just don't know how code things for you to supply a variable
number of unquoted column names for a single argument; we've just hit a
coding knowledge limitation here!}

\item{font}{font to use. Default is "Palatino Linotype" and any fonts
available on your machine in either Word or PowerPoint should be
acceptable. If you get Times New Roman in your table when you asked for
something else, it means that that font isn't available or maybe wasn't
spelled the way R is expecting it. For example, "Calibri" works but
"Calibri (Body)" doesn't even though the latter is listed in PowerPoint and
Word.}

\item{fontsize}{the numeric font size for the output table. Default is 11
point.}

\item{column_widths}{optionally specify what the widths of the columns should
be with a numeric vector of the widths in inches, e.g., \code{column_widths
= c(1.5, 2, 0.5, 3)}}

\item{alignment}{alignment of text throughout table. Options are "left"
(default), "right", "center", or "justify".}

\item{save_table}{optionally save the output table by supplying a file name
in quotes here, e.g., "My nicely formatted table.docx".  Do not include any
slashes, dollar signs, or periods in the file name. If you leave off the
file extension, we'll assume you want it to be ".docx". If there is a
column titled "File" in your table, we'll add a caption listing which files
were included.}

\item{page_orientation}{set the page orientation for the Word file output to
"portrait" (default) or "landscape"}

\item{title_document}{optionally specify a title for the Word document
output. If you don't save the table, this will be ignored.}

\item{table_caption}{optionally add some text for a table caption. If the
table you supply contains a column titled "File", there will already be a
caption listing the source files; this would add some additional text
before that.}
}
\value{
a formatted table
}
\description{
\code{format_table_simple} is meant to for use with writing compound summary
files. It is a wrapper function for formatTable_Simcyp and includes fewer
options for formatting, providing a somewhat simpler output style of table.
}
\examples{
# None yet

}
