% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTable_Simcyp.R
\name{formatTable_Simcyp}
\alias{formatTable_Simcyp}
\title{Format tables according to Simcyp Consultancy Team specfications}
\usage{
formatTable_Simcyp(DF, fontsize = 11, shading_column)
}
\arguments{
\item{DF}{a data.frame, usually output from \code{\link{pksummary_table}} or
\code{\link{pksummary_mult}}}

\item{fontsize}{the numeric font size for the output. Default is 11 point.}

\item{shading_column}{If you would like to alternate the shading of the rows
in the output data.frame, supply here the unquoted name of the column to
check for when to change the shading; everytime that column's value
changes, the shading will alternate between white and light gray. For
example, if you have a table with PK values for multiple files and you have
more than one row per file (an example of this would be the output from the
function \code{\link{pksummary_mult}}), if you set \code{shading_column =
File}, the shading of the rows will alternate between white and light gray
whenever the file changes.}
}
\value{

}
\description{
\code{formatTable_Simcyp} was designed to work with output from
\code{\link{pksummary_table}} and \code{\link{pksummary_mult}}, and it
formats tables so that the column headings and the first column are bold, and
the second through the last columns are centered. Column headings with, e.g.,
"AUCinf" or "Cmax" will have the "inf" or the "max" subscripted, and the
table will automatically expand to fit the contents.
}
\examples{
MyData <- data.frame(ColA = rep(LETTERS[1:3], each = 2),
                     ColB = 1:6)
formatTable_Simcyp(MyData)

}
