% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_assignments.R
\name{get_obs_assignments}
\alias{get_obs_assignments}
\title{INTERNAL - Figure out observed CT files to user for extractConcTime}
\usage{
get_obs_assignments(
  sim_data_files,
  obs_to_sim_assignment = NA,
  existing_exp_details = NA
)
}
\arguments{
\item{sim_data_files}{sim_data_files}

\item{obs_to_sim_assignment}{obs_to_sim_assignment or obs_data_file}

\item{existing_exp_details}{existing_exp_details}
}
\value{
data.frame
}
\description{
INTERNAL - Figure out observed CT files to user for extractConcTime
}
\examples{
# none
}
