% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_file.R
\name{get_obs_file}
\alias{get_obs_file}
\title{Get the observed XML overlay file from a Simcyp Simulator workspace}
\usage{
get_obs_file(workspace)
}
\arguments{
\item{workspace}{the workspace file}
}
\value{
a character vector of the observed XML overlay file and its path
  relative to the working directory
}
\description{
Get the observed XML overlay file from a Simcyp Simulator workspace
}
