% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReferences.R
\name{getReferences}
\alias{getReferences}
\title{Pull references from a "DataRec" folder}
\usage{
getReferences(main_directory)
}
\arguments{
\item{main_directory}{the main directory from which you want a list of all
the files}
}
\value{
A data.frame containing columns "Dir" for the directory and, where
  applicable, the subdirectory of a file, and "File" for the file name.
}
\description{
Get a list of all the pdf, Word, and Excel files in, e.g., the "DataRec"
folder and subfolder. Can be useful for the bibliography section of a report.
}
\examples{

main_directory <- "\\\\\\\\s08sharepoint.certara.com@SSL/DavWWWRoot/sites/consult/dndi-1a/DataRec"
getReferences(main_directory)


}
