% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSectionInfo.R
\name{getSectionInfo}
\alias{getSectionInfo}
\title{Get information about a section for a report}
\usage{
getSectionInfo(report_input_file = NA, sheet = NA, section_input_DF = NA)
}
\arguments{
\item{report_input_file}{the name of the Excel file formatted exactly like
"Report input template.xlsx", including the path if it's in any other
directory than the current one}

\item{sheet}{the sheet in the Excel file that contains information about this
section of the report}

\item{section_input_DF}{a data.frame object that contains information about
this section of the report and is the filled-out version of
ReportInputForm[["Section input form"]]. If this is NOT set to NA, then
this will be used for section details instead of reading the Excel sheet
listed. You still must fill out \code{report_input_file}, though, because
that will be used to read in data about the clinical study.}
}
\value{
a list object
}
\description{
\code{getSectionInfo} pulls information from an Excel file that is formatted
\emph{exactly} like the file "Report input template.xlsx" and organizes those
data plus some further data it pulls from the simulator output Excel file.
The output of this function serves as part of the input for the function
\code{\link{so_table}}.
}
\examples{
# No examples yet.

}
