% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_CV}
\alias{gm_CV}
\title{Calculate the geometric coefficient of variation (CV)}
\usage{
gm_CV(x, na.rm = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{A vector of numbers}

\item{na.rm}{Should NA values be removed? (logical)}

\item{zero.propagate}{Should zeroes be propagated? (logical)}
}
\description{
\code{gm_CV} takes as input a numeric vector and returns the geometric
CV, calculated as: sqrt(exp(sd(log(x))^2)-1)
}
\examples{

gm_CV(rnorm(10, 5, 1))
gm_CV(c(5, 3, 6, 10, 2, 0), zero.propagate = TRUE)

}
