% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_boxplot.R
\name{graph_boxplot}
\alias{graph_boxplot}
\title{Make boxplots or boxplots overlaid with individual points}
\usage{
graph_boxplot(
  DF,
  category_column,
  value_column,
  facet_column1 = NA,
  facet_column2 = NA,
  graph_type = "boxplot",
  include_errorbars = FALSE,
  xlabel = NA,
  ylabel = NA,
  color = "default",
  outfile = NA,
  outwidth = 6,
  outheight = 4
)
}
\arguments{
\item{DF}{the data.frame you want to graph}

\item{category_column}{the name of the column with categorical data in quotes}

\item{value_column}{the name of the column with value data in quotes}

\item{facet_column1}{(optional) the name of a column by which you might break
up your graph into small multiples. Please see the example if you're
uncertain what this does.}

\item{facet_column2}{(optional) the name of a second column by which you
might break up the graph into small multiples.}

\item{graph_type}{the type of graph to plot. Options: \describe{
  \item{"boxplot"}{standard boxplots or box-and-whisker plots}

  \item{"jittered points"}{boxplots overlaid with points depicting each
  individual observation.}

  \item{"jittered points, filled boxes"}{boxplots with fill color according
  to \code{category_column} and open circles for the jittered points}}}

\item{include_errorbars}{TRUE or FALSE on whether to include horizontal error
bars on the whiskers}

\item{xlabel}{the label to  use for the x axis}

\item{ylabel}{the label to use for the y axis}

\item{color}{the set of colors to use. Options: \describe{

  \item{"default"}{colors selected from the color brewer palette "set 1"}

  \item{"ggplot2 default"}{the default set of colors used in ggplot2 graphs
  (ggplot2 is an R package for graphing.)}

  \item{"black"}{black and white}

  \item{"rainbow"}{colors selected from a rainbow palette. The default
  palette is limited to something like 6 groups, so if you have more than
  that, that's when this palette is most useful.}

  \item{"blue-green"}{a set of blues and greens}

  \item{"Brewer set 2"}{a set of colors from Cynthia Brewer et al. from Penn
  State that are friendly to those with red-green colorblindness}

  \item{"Tableau"}{uses the standard Tableau palette; requires the "ggthemes"
  package}}}

\item{outfile}{the file name for the output graph, e.g., "My pretty
boxplot.png". If you don't want to automatically save the graph, leave this
as NA.}

\item{outwidth}{the width in inches of the output graph}

\item{outheight}{the height in inches of the output graph}
}
\value{

}
\description{
\code{graph_boxplot} takes as input a data.frame of numeric values and the
categories those values are in and creates boxplots of those data. Options
are included for type of graph to make and some aesthetics.
}
\examples{
AUCs <- data.frame(AgeGroup = LETTERS[1:5],
                   AUC = rnorm(n = 500, mean = 20000, sd = 2000))

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC")

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
               color = "rainbow")

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
               color = "blue-green", graph_type =  "jittered points")

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
              color = "Tableau", graph_type =  "jittered points, filled boxes")

# Adding a couple of example columns to use the "facet" options.
AUCs$Sex <- c("M", "F")
AUCs$Metabolizer <- sample(c("poor", "extensive"), 100, replace = TRUE)

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
               facet_column1 = "Sex",
               color = "blue-green", graph_type =  "jittered points")

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
               facet_column1 = "Sex", facet_column2 = "Metabolizer",
               color = "blue-green", graph_type =  "jittered points")

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
              graph_type = "jittered points, filled boxes")

# Saving the output
graph_boxplot(AUCs \%>\% filter(AgeGroup \%in\% c("A", "B")),
              category_column = "AgeGroup", value_column = "AUC",
              graph_type = "jittered points, filled boxes",
              color = "Brewer set 2", include_errorbars = TRUE,
              outfile = "test boxplot.png")


}
