% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectors.R
\name{green_to_red}
\alias{green_to_red}
\title{Make a vector of colors that match the "green to red" option for highlighting
GMR values}
\usage{
green_to_red()
}
\value{
a named character vector
}
\description{
Make a vector of colors that match the "green to red" option for highlighting
GMR values
}
