% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectors.R
\name{greens}
\alias{greens}
\title{Create a vector of greens}
\usage{
greens(ncolors, shade = "regular")
}
\arguments{
\item{ncolors}{number of colors desired}

\item{shade}{the shade of colors to use, which can be "regular" (default),
"darker", "lighter", or "muted"}
}
\value{
a character vector of colors
}
\description{
\code{greens} is useful for getting a set of colors from chartreuse to
forest green for graphs. These are from the RColorBrewer package, color set
"YlGn", just to give credit where it's due, but they can be any number in
length rather than only up to 9 colors.
}
\examples{
# Create a set of 10 colors that are deepening shades of green.
greens(10)

# Try using scales::show_col() to visualize the colors, ex:
scales::show_col(greens(ncolors = 10))
scales::show_col(greens(ncolors = 10, shade = "darker"))

}
