% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_col_widths.R
\name{guess_col_widths}
\alias{guess_col_widths}
\title{INTERNAL guess appropriate column widths}
\usage{
guess_col_widths(DF)
}
\arguments{
\item{DF}{data.frame to write to Excel}
}
\value{
numeric vector of column widths
}
\description{
INTERNAL guess appropriate column widths
}
\examples{
# none
}
