% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_demog.R
\name{harmonize_demog}
\alias{harmonize_demog}
\title{INTERNAL - Harmonize inputs for demog_x functions}
\usage{
harmonize_demog(demog_dataframe, demog_parameters)
}
\arguments{
\item{demog_dataframe}{data.frame of demographic data}

\item{demog_parameters}{which demographic parameters user wants}
}
\value{
list of harmonized demog_dataframe and demog_parameters
}
\description{
INTERNAL - Harmonize inputs for demog_x functions
}
