% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_dosing.R
\name{harmonize_dosing}
\alias{harmonize_dosing}
\title{INTERNAL - Harmonize the dosing information for an object with details about
simulations. Use information from MainDetails and CustomDosing to add an item
to the list called "Dosing" that will include ALL dosing info.}
\usage{
harmonize_dosing(existing_exp_details)
}
\arguments{
\item{existing_exp_details}{output from running extractExpDetails_mult}
}
\value{
harmonized existing_exp_details
}
\description{
INTERNAL - Harmonize the dosing information for an object with details about
simulations. Use information from MainDetails and CustomDosing to add an item
to the list called "Dosing" that will include ALL dosing info.
}
