% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_PK_names.R
\name{harmonize_PK_names}
\alias{harmonize_PK_names}
\title{Harmonize user input for PK parameters}
\usage{
harmonize_PK_names(PKparameters)
}
\arguments{
\item{PKparameters}{user input for PK parameters that might not perfectly
follow required syntax for PK parameters from \code{AllPKParameters}}
}
\value{
harmonized PK parameters
}
\description{
INTERNAL PACKAGE USE
}
\examples{
# None

}
