% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IndData}
\alias{IndData}
\title{Example induction data for use with the inductFit function}
\format{
A data.frame with 24 rows and 3 columns:
\describe{
\item{DonorID}{Unique ID for the donor}
\item{Concentration_uM}{Concentration of the inducer (uM)}
\item{FoldInduction}{Fold induction observed compared to control}}
}
\usage{
IndData
}
\description{
Example induction data for use with the inductFit function
}
\keyword{datasets}
