% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_copyright.R
\name{insert_copyright}
\alias{insert_copyright}
\title{Insert the standard Certara UK Limited copyright box that appears in compound
summary files}
\usage{
insert_copyright()
}
\value{
a flextable with the current copyright information in it
}
\description{
Insert the standard Certara UK Limited copyright box that appears in compound
summary files
}
\examples{
insert_copyright()

}
