% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IVConcTime}
\alias{IVConcTime}
\title{Dummy plasma concentration-time data after an IV dose}
\format{
A data.frame with 90 rows and 3 columns:
\describe{
\item{SubjectID}{a unique ID for each subject}
\item{Time}{time since dose (units not specified)}
\item{Concentration}{plasma drug concentration (units not specified)}}
}
\usage{
IVConcTime
}
\description{
Dummy plasma concentration-time data after an IV dose
}
\keyword{datasets}
