% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_package_index.R
\name{launch_package_index}
\alias{launch_package_index}
\title{Launch the shiny app for showing an annotated index of what the
SimcypConsultancy package can do}
\usage{
launch_package_index()
}
\description{
\code{launch_package_index} will launch a graphical user interface for
showing various things you can do with the SimcypConsultancy package.
}
\examples{
launch_package_index()

}
