% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LMVct_mean}
\alias{LMVct_mean}
\title{Plasma concentration-time data for a simulation of letermovir}
\format{
A data.frame
}
\usage{
LMVct_mean
}
\description{
Plasma concentration-time data after 8 days of 60 mg PO QD letermovir,
including arithmetic-mean observed data for the last dose. Format of the data
are the same as the output from \code{\link{extractConcTime}}.
}
\keyword{datasets}
