% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log10_breaks.R
\name{log10_breaks}
\alias{log10_breaks}
\title{Create nice breaks for graphs of log-transformed data}
\usage{
log10_breaks(min, max, prettify = FALSE)
}
\arguments{
\item{min}{minimum value to use for breaks}

\item{max}{maximum value to use for breaks}
}
\value{
a numeric vector
}
\description{
\code{log10_breaks} creates a vector of numbers fairly evenly spaced on a log 10 scale with 10
points. (They're more evenly spaced if "min" and "max" span 1 order of
magnitude.) The decimals aren't necessarily pretty, though, so you may want
to set "prettify" to TRUE if they're ugly.
}
\examples{
data("ConcTime")
ggplot(ConcTime, aes(x = Time, y = Conc, color = ID)) +
   geom_line() +
   scale_y_log10(breaks = log10_breaks(min = min(ConcTime$Conc[ConcTime$Conc > 0]),
                                       max = max(ConcTime$Conc)))

ggplot(ConcTime, aes(x = Time, y = Conc, color = ID)) +
   geom_line() +
   scale_y_log10(breaks = log10_breaks(min = min(ConcTime$Conc[ConcTime$Conc > 0]),
                                       max = max(ConcTime$Conc),
                                       prettify = TRUE))


}
