% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ct_caption.R
\name{make_ct_caption}
\alias{make_ct_caption}
\title{INTERNAL - make caption for conc-time plots}
\usage{
make_ct_caption(
  ct_dataframe,
  single_or_multiple = "single",
  existing_exp_details = NA,
  mean_type = "arithmetic",
  linear_or_log = "linear",
  tissue = "plasma",
  compoundID = "substrate",
  prettify_compound_names = TRUE
)
}
\arguments{
\item{single_or_multiple}{whether this is for a ct_plot graph (only 1
compound, tissue, file, etc.) or a graph with multiple profiles (multipel
compounds, tissues, files, etc.). Options are "single" (default) or
"multiple".}

\item{existing_exp_details}{OR supply: a data.frame with the following
columns: File, Population, Substrate, Inhibitor1, Inhibitor1Metabolite,
Inhibitor2, ... If left as NA, we'll note some generic parameters and make
them bold to note that they should be changed.}

\item{mean_type}{"arithmetic" (default) or "geometric"}

\item{linear_or_log}{"linear" (default), "both", "both horizontal", "both
vertical", "log", "semi-log", "horizontal and vertical"}

\item{tissue}{default is plasma}

\item{compoundID}{default is substrate}

\item{prettify_compound_names}{default is TRUE}
}
\value{
a string of text for making a caption
}
\description{
INTERNAL - make caption for conc-time plots
}
\examples{
# none 
}
