% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example_PK_csv.R
\name{make_example_PK_csv}
\alias{make_example_PK_csv}
\title{Create an example csv file for supplying PK parameters to the functions
\code{\link{pksummary_table}} or \code{\link{pksummary_mult}}}
\usage{
make_example_PK_csv(study_type = "DDI")
}
\arguments{
\item{study_type}{For what kind of study do you want to specify PK
  parameters? Options are: \describe{

  \item{"DDI"}{DDI studies}

  \item{"single dose"}{studies with a single dose or when you only want the
  1st dose PK}

  \item{"multiple dose"}{studies with multiple doses}

  \item{"specific sheets"}{when you've got a user-defined AUC interval
  and want to pull data only from there}

  \item{"specific tissues"}{when you want PK for a specific tissue or
  tissues}

  \item{"specific compounds"}{when you want PK for a specific compound or
  compounds}

  \item{"ADAM"}{when you want ADAM-related parameters}

  \item{"I want it all"}{This actually saves a csv file of all the possible PK 
  parameters you can extract automatically. It doesn't actually give you 
  examples of each of them but, instead, is meant to give you an overview of 
  what's possible.}}}
}
\value{
saves a csv file of example PK parameters
}
\description{
Create an example csv file for supplying PK parameters to the functions
\code{\link{pksummary_table}} or \code{\link{pksummary_mult}}
}
\examples{
make_example_PK_csv("specific sheets")


}
