% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example_PK_input.R
\name{make_example_PK_input}
\alias{make_example_PK_input}
\title{Examples for supplying PK parameters to the pk_table or calc_PK_ratios
functions for the argument \code{PKparameters}}
\usage{
make_example_PK_input()
}
\value{
saves a csv file of example PK parameters for supplying to the
  argument \code{PKparameters}
}
\description{
\code{make_example_PK_input} gives examples of how to request specific PK
parameters or supply specific PK data to the functions
\code{\link{pk_table}}, \code{\link{calc_PK_ratios}}, or
\code{\link{calc_PK_ratios_mult}}. You can either just get a printout in the
console of the examples or you can get an example csv file to use a as a
template.
}
\examples{
# None. This is an interactive function.


}
