% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_external_simulation_directory.R
\name{make_external_simulation_directory}
\alias{make_external_simulation_directory}
\title{Make a simulation directory of just the files in a project folder and nothing
else}
\usage{
make_external_simulation_directory(
  project_folder,
  draft_simulation_directory,
  save_table = NA,
  overwrite = "ask"
)
}
\arguments{
\item{project_folder}{location of the project folder to search. Please note
that R requires forward slashes ("/") in file paths rather than the back
slashes Microsoft uses. If left as NA (default), we'll assume your current
working directory is the top level of your project folder.}

\item{draft_simulation_directory}{the output from running either
  \code{\link{make_simulation_directory}} or
  \code{\link{find_report_file_locations}}, either of which will find all of
  the simulation files (both workspaces and Excel outputs as well as any
  observed XML overlay files) in a project folder. This MUST include the
  following columns:

  \itemize{\item{"File name" with the simulation file names}

  \item{"File type" with the file extensions}

  \item{"XML file used" with the XML observed-data overlay file}

  \item{"Table/Figure" with the table and figure that simulation appears in}}

  If you have only run \code{\link{make_simulation_directory}}, this will not
  include the figures and tables unless you have added those manually.}

\item{save_table}{optionally specify an Excel file name for saving your
simulation directory.}

\item{overwrite}{Should we overwrite if your Excel file already exists and
already has a tab named "Simulation directory"? Options are "yes" to always
overwrite, "no" to never overwrite, or "ask" (default), which means that we
will ask you whether to overwrite and give you a chance to supply a
different file name.}
}
\value{
a simulation directory as a data.frame; optionally saves that
  data.frame to Excel
}
\description{
\code{make_external_simulation_directory} is a wrapper for
  \code{\link{make_simulation_directory}} that sets all the arguments up
  correctly when you have all your simulations and XML observed-data-overlay
  files in a single folder and you don't need to search for any observed data
  files or anything fancy. Instead, this is the scenario where you just want
  a simulation directory of \emph{exactly} what's in the project folder.
}
\examples{
make_external_simulation_directory(
   project_folder = "C:/Users/Buffy/Project ABC/Modelling",
   draft_sim_directory = DraftSimDirectory,
   save_table = "Simulation directory for project ABC.xlsx")
}
