% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_log_breaks.R
\name{make_log_breaks}
\alias{make_log_breaks}
\title{Make pretty breaks and labels for a log axis}
\usage{
make_log_breaks(axis_range, axis_limits_log = NA)
}
\arguments{
\item{axis_limits_log}{desired y axis limits for a semi-log
concentration-time plot}

\item{time_range_to_use}{optionally specify a time range for the conc-time
data you're plotting and we'll find min and max concs within only that time
range. If left as NA, we'll use the min and max concs for the overall data.}
}
\value{
a list of log axis breaks and log axis labels
}
\description{
\code{make_log_breaks} is mainly meant for internal SimcypConsultancy package
use and is thus not terribly annotated in terms of warnings if you do
something wrong.
}
\examples{
# none


}
