% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_text_legos.R
\name{make_text_legos}
\alias{make_text_legos}
\title{INTERNAL - Assemble pieces of text for headings, captions and trial design
descriptions}
\usage{
make_text_legos(sim_data_file, existing_exp_details, prettify_compound_names)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes}

\item{existing_exp_details}{the output from running either
\code{\link{extractExpDetails}} or \code{\link{extractExpDetails_mult}} --
either is fine as long as it contains details for \code{sim_data_file}.}

\item{prettify_compound_names}{TRUE (default) or FALSE for whether to make
compound names prettier in legend entries and in any Word output files.
This was designed for simulations where the substrate and any metabolites,
perpetrators, or perpetrator metabolites are among the standard options for
the simulator, and leaving \code{prettify_compound_names = TRUE} will make
the name of those compounds something more human readable. For example,
"SV-Rifampicin-MD" will become "rifampicin", and "Sim-Midazolam" will
become "midazolam".}
}
\value{
list of text strings
}
\description{
INTERNAL - Assemble pieces of text for headings, captions and trial design
descriptions
}
\examples{
# none

}
