% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trial_design_table.R
\name{make_trial_design_table}
\alias{make_trial_design_table}
\title{Create a table describing your trial design.}
\usage{
make_trial_design_table(
  existing_exp_details,
  sim_data_files = NA,
  cols_to_include = "default",
  fontsize = 11,
  shading_column,
  merge_shaded_cells = TRUE,
  merge_columns = NA,
  sort_column,
  bold_cells = list(c(0, NA), c(NA, 1)),
  center_1st_column = FALSE,
  highlight_cells = NA,
  highlight_color = "yellow",
  save_table = NA,
  title_document = NA,
  table_caption = NA
)
}
\arguments{
\item{existing_exp_details}{output from running extractExpDetails_mult}

\item{sim_data_files}{optionally specify which simulation files you want to
inlude.}

\item{cols_to_include}{optionally specify which columns you want. Current
options are "default" for a handful of standard columns or "all trial
design" to get all possible trial design columns. You can also specify a
character vector of columns you want from the "MainDetails" item from
existing_exp_details.}

\item{...}{Arguments that pass through to formatTable_Simcyp.}
}
\value{
a formatted table and optionally, a saved Word file with that table
}
\description{
UNDER CONSTRUCTION. Proof of concept only for now.
}
\examples{
make_trial_design_table(existing_exp_details = MDZdetails, 
                        shading_column = File, 
                        save_table = "MDZ trial design info.docx")

}
