% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_units.R
\name{match_units}
\alias{match_units}
\title{Match concentration and time units as needed}
\usage{
match_units(DF_to_adjust, goodunits, MW = NA)
}
\arguments{
\item{DF_to_adjust}{the data.frame of concentration-time data with units that
may need to be adjusted. This must include the columns "Conc",
"Conc_units", "Time", and "Time_units". (Outputs from
\code{\link{extractConcTime}}, \code{\link{extractConcTime_mult}}, and
\code{\link{extractObsConcTime}} work here.) If you want to convert between
mass per volume and molar concentrations and there are multiple compounds
present, this must also include the column "CompoundID".}

\item{goodunits}{either a data.frame that has the desired concentration and
time units OR a named list with the desired units. Example:
\code{list("Conc_units" = "ng/mL", "Time_units" = "hours")}. Names of the
list must be "Conc_units" and "Time_units", although you can provide only
one or the other if you only want to convert one of them. Options for
concentration units are the same as the ones in the Excel form for PE data
entry, and options for time units are "hours", "days", and "minutes".}

\item{MW}{optionally supply a molecular weight for the compound of interest
to enable conversions between mass per volume and molar concentrations. If
you have more than one compound, list the compoundID and the MW as a named
character vector, e.g., \code{MW = c("substrate" = 325.8, "inhibitor 1" =
705.6)} for a DDI simulation of midazolam with itraconazole. If you want to
make this type of conversion, the data.frame DF_to_adjust must also include
the column "CompoundID". Acceptable compoundIDs are "substrate", "primary
metabolite 1", "primary metabolite 2", "secondary metabolite", "inhibitor
1", "inhibitor 2", or "inhibitor 1 metabolite".}
}
\value{
a data.frame with the corrected units
}
\description{
\code{match_units} adjusts the units for a data.frame containing
concentration-time data as necessary to match some desired set of
concentration and time units.
}
\examples{
DF_to_adjust <- match_units(DF_to_adjust = SimulatedData,
                            goodunits = ObsData)

}
