% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MDZConcTime}
\alias{MDZConcTime}
\title{Dummy plasma concentration-time data after a 5 mg PO MDZ dose. Format of the
data are the same as the output from \code{extractConcTime}.}
\format{
A data.frame with 2211 rows and 7 columns:
\describe{
\item{Compound}{the name of the compound}
\item{Individual}{a unique ID for each simulated individual}
\item{Simulated}{TRUE or FALSE for whether the data were simulated (all TRUE
in this dummy example)}
\item{Time}{time since dose}
\item{Conc}{plasma drug concentration}
\item{Time_units}{units for the time column}
\item{Conc_units}{units for the concentration column}}
}
\usage{
MDZConcTime
}
\description{
Dummy plasma concentration-time data after a 5 mg PO MDZ dose. Format of the
data are the same as the output from \code{extractConcTime}.
}
\keyword{datasets}
