% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MDZct}
\alias{MDZct}
\title{Plasma concentration-time data after a single dose of 5 mg MDZ but with
varying fa values, one for each of 4 simulator output files. Format of the
data are the same as the output from \code{\link{extractConcTime_mult}}.}
\format{
A data.frame with 15 columns "MDZct"
}
\usage{
MDZct
}
\description{
Plasma concentration-time data after a single dose of 5 mg MDZ but with
varying fa values, one for each of 4 simulator output files. Format of the
data are the same as the output from \code{\link{extractConcTime_mult}}.
}
\keyword{datasets}
