% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MDZdetails}
\alias{MDZdetails}
\title{Example output from \code{extractExpDetails_mult}}
\format{
a data.frame with rows for each of 4 simulator files and columns for
  each simulation experimental detail
}
\usage{
MDZdetails
}
\description{
Example output from running \code{extractExpDetails_mult} on some midazolam
simulations
}
\keyword{datasets}
