% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_asterisks.R
\name{move_asterisks}
\alias{move_asterisks}
\title{INTERNAL - Move asterisks around so that you can put asterisks around that
string and things will be italicized appropriately}
\usage{
move_asterisks(string, num_asterisks_to_match = NA)
}
\arguments{
\item{string}{a string of text, e.g., a figure or table caption for use in an
Rmarkdown document that will be italicized}

\item{num_asterisks_to_match}{number of asterisks to match and move. If left as NA,
this will move sets of 3, 2, and then 1 asterisks. If set to 3, 2, or 1,
though, it will \emph{only} look for that number to match and move.}
}
\value{
a string
}
\description{
\code{move_asterisks} finds all locations of an asterisk plus a
  space or a space plus an asterisk and swaps the positions of the spaces and
  the asterisks. It also makes sure that your string doesn't end in two
  asterisks by removing the terminal asterisk.

  \strong{How to use this:} Say you've got a string that is a figure caption
  with two italicized references included inside the string. If you try to
  put the whole caption in italics in a .Rmd file with \code{*`r
  MyCaption`*}, the italics will be messed up and you will likely have an
  extra asterisk at the end of the final output in the knitted document. This
  fixes that.
}
