% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OntogenyEquations}
\alias{OntogenyEquations}
\title{Equations used for calculating drug-metabolizing enzyme and drug transport
ontogenies in pediatric populations}
\format{
a data.frame
}
\usage{
OntogenyEquations
}
\description{
Equations used for calculating drug-metabolizing enzyme and drug transport
ontogenies in pediatric populations
}
\keyword{datasets}
