% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OurFavoriteColors}
\alias{OurFavoriteColors}
\title{A list of the SimcypConsultancy Team's favorite colors}
\format{
a named list of character vector of colors
}
\usage{
OurFavoriteColors
}
\description{
A list of the SimcypConsultancy Team's favorite colors
}
\examples{
scales::show_col(OurFavoriteColors$Shinji)
}
\keyword{datasets}
