% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_csv.R
\name{pdf_to_csv}
\alias{pdf_to_csv}
\title{Convert a page from a pdf file to a csv file}
\usage{
pdf_to_csv(pdf_file, page, save_csv = NA)
}
\arguments{
\item{pdf_file}{the pdf file you want to read, in quotes, e.g.,
\code{pdf_file = "Table data that I want to graph but is in pdf form.pdf"}}

\item{page}{page number of the pdf you want to read (this function only works
with one page at a time)}

\item{save_csv}{optionally specify a file name for saving the output. If left
as NA, no file will be saved. If set to just "csv", the file name will be
the pdf file name plus the page number requested.}
}
\value{

}
\description{
\code{pdf_to_csv} takes as input a pdf file and a page number and converts
that to a data.frame and, optionally, saves that to a csv file. This can be
useful for getting client-supplied data into a useful format. NOTE: This
requires the package pdftools, which you may have to download and install
with \code{install.packages("pdftools")}. You only have to do that once.
}
\examples{
# No examples yet.

}
