% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_table.R
\name{pk_table}
\alias{pk_table}
\title{Title}
\usage{
pk_table(
  sim_data_files = NA,
  compoundsToExtract = "substrate",
  tissues = "plasma",
  PKparameters = NA,
  PKorder = "default",
  sheet_PKparameters = NA,
  existing_exp_details = NA,
  mean_type = NA,
  use_median_for_tmax = TRUE,
  includeCV = TRUE,
  includeSD = FALSE,
  includeConfInt = TRUE,
  includeMedian = FALSE,
  includeRange = FALSE,
  includePerc = FALSE,
  includeTrialMeans = FALSE,
  concatVariability = FALSE,
  variability_format = "to",
  convert_conc_units = NA,
  include_dose_num = NA,
  add_header_for_DDI = TRUE,
  rounding = NA,
  prettify_columns = TRUE,
  prettify_compound_names = TRUE,
  extract_forest_data = FALSE,
  checkDataSource = TRUE,
  highlight_gmr_colors = NA,
  highlight_so_cutoffs = NA,
  highlight_so_colors = "yellow to red",
  save_table = NA,
  single_table = FALSE,
  page_orientation = "portrait",
  fontsize = 11,
  return_PK_pulled = FALSE,
  ...
)
}
\arguments{
\item{sim_data_files}{}

\item{compoundsToExtract}{}

\item{tissues}{}

\item{PKparameters}{}

\item{PKorder}{}

\item{sheet_PKparameters}{}

\item{existing_exp_details}{}

\item{mean_type}{}

\item{use_median_for_tmax}{}

\item{includeCV}{}

\item{includeSD}{}

\item{includeConfInt}{}

\item{includeMedian}{}

\item{includeRange}{}

\item{includePerc}{}

\item{includeTrialMeans}{}

\item{concatVariability}{}

\item{variability_format}{}

\item{convert_conc_units}{}

\item{include_dose_num}{}

\item{add_header_for_DDI}{}

\item{rounding}{}

\item{prettify_columns}{}

\item{extract_forest_data}{}

\item{checkDataSource}{}

\item{highlight_gmr_colors}{}

\item{highlight_so_cutoffs}{}

\item{highlight_so_colors}{}

\item{save_table}{}

\item{single_table}{}

\item{page_orientation}{}

\item{fontsize}{}

\item{...}{}
}
\value{
a data.frame
}
\description{
Title
}
\examples{
# none yet

}
