% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_table_subfun.R
\name{pk_table_subfun}
\alias{pk_table_subfun}
\title{Subfunction for pk_table. FOR INTERNAL PACKAGE USE.}
\usage{
pk_table_subfun(
  sim_data_file,
  PKparameters,
  existing_exp_details,
  conc_units,
  MeanType,
  GMR_mean_type,
  includeTrialMeans,
  use_median_for_tmax
)
}
\arguments{
\item{sim_data_file}{sim_data_file}

\item{PKparameters}{data.frame that includes Tissue, CompoundID, and Sheet}

\item{existing_exp_details}{harmonized existing_exp_details}

\item{conc_units}{how to convert conc units, if desired}

\item{MeanType}{main mean type}

\item{GMR_mean_type}{GMR mean type}

\item{includeTrialMeans}{determines whether to get individual data
w/extractPK}

\item{use_median_for_tmax}{T or F}
}
\value{
a list of 1) "PK" - a data.frame of unformatted simulated and, when
  supplied, observed PK data with the columns Stat, PKParam, and Sim. NOTHING
  has been rounded. 2) "PKrequested" - which PK parameters were requested, 3)
  "PKpulled" - which PK parameters were successfully pulled, 4) "QC" - QC
  info, 5) "ForestData" - forest data formatted appropriately for the
  forest_plot function, and 6) "CheckDoseInt" - list of the message and
  dosing interval data
}
\description{
Subfunction for pk_table. FOR INTERNAL PACKAGE USE.
}
\examples{
# none
}
