% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PKexamples}
\alias{PKexamples}
\title{Examples for specifying PK parameters with pk_table}
\format{
a data.frame
}
\usage{
PKexamples
}
\description{
Examples for specifying PK parameters with pk_table
}
\keyword{datasets}
