% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PKexamples_ratios}
\alias{PKexamples_ratios}
\title{Examples for specifying PK parameters with calc_PK_ratios and
calc_PK_ratios_mult}
\format{
a data.frame
}
\usage{
PKexamples_ratios
}
\description{
Examples for specifying PK parameters with calc_PK_ratios and
calc_PK_ratios_mult
}
\keyword{datasets}
