% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PKParameterDefinitions}
\alias{PKParameterDefinitions}
\title{All possible PK parameters for \code{\link{extractPK}}}
\format{
A data.frame with 6 columns: \describe{ \item{PKparameter}{the PK
  parameter name to use with \code{\link{extractPK}} for the argument
  \code{PKparameters}} \item{Sheet}{the sheet in a simulator output file
  where the PK parameter will be extracted, if possible}
  \item{AppliesToSingleDose}{TRUE or FALSE for whether this item applies only
  to single-dose data} \item{AppliesOnlyWhenPerpPresent}{TRUE or FALSE
  for whether this item only applies when a perpetrator is present in the
  simulation} \item{SortOrder}{the order to arrange columns for pk_table and
  pk_table} \item{Notes}{an explanation of what the parameter is}}
}
\usage{
PKParameterDefinitions
}
\description{
All possible PK parameters that can be extracted from a simulator
output file using \code{\link{extractPK}}. For the version that
includes additional columns only used for coding purposes, see
\code{\link{AllPKParameters}}.
}
\keyword{datasets}
