% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pksummary_mult.R
\name{pksummary_mult}
\alias{pksummary_mult}
\title{Make PK summary tables from multiple simulator output files at once}
\usage{
pksummary_mult(
  sim_data_files,
  observed_PK = NA,
  PKparameters = NA,
  ...,
  checkDataSource = TRUE,
  save_table = NA
)
}
\arguments{
\item{observed_PK}{(optional) If you have a data.frame or an Excel or csv
file with observed PK parameters, supply the data.frame or the full file
name in quotes here, and the simulated-to-observed mean ratios will be
calculated. The supplied data.frame or file \emph{must} include columns for
the simulator output Excel file (title this "File") and each of the PK
parameters you would like to compare, and those column names \emph{must}
match the PK parameter options listed in
\code{data(PKParameterDefinitions)}. If you would like the output table to
include the observed data CV for any of the parameters, add "_CV" to the
end of the parameter name, e.g., "AUCinf_dose1_CV". Note: Whatever you list
for "File" will override anything specified for the argument
\code{sim_data_file}.}

\item{PKparameters}{(optional) the PK parameters to include as a character
  vector. Notes: \itemize{

  \item{By default, if you have a single-dose simulation, the parameters will
  include AUC and Cmax for dose 1, and, if you have a multiple-dose
  simulation, AUC and Cmax for the last dose. Also by default, if you have an
  effector present, the parameters will include the AUC and Cmax values with
  and without the effector as well as those ratios.}

  \item{Alternatively, you can specify a vector of any combination of
  specific, individual parameters, e.g., \code{c("Cmax_dose1",
  "AUCtau_last").} Be sure to encapsulate the parameters you want with
  \code{c(...)}! To see the full set of possible parameters to extract, enter
  \code{data(PKParameterDefinitions); view(PKParameterDefinitions)} into the
  console. Not case sensitive. If you use "_first" instead of "_dose1", that
  will also work.}

  \item{If you supply observed data using either the argument
  \code{report_input_file} or the argument \code{observed_PK}, the PK
  parameters included are only those available for the observed data.}

  \item{Parameters that don't make sense for your scenario -- such as asking
  for \code{AUCinf_dose1_withInhib} when your simulation did not include an
  inhibitor or effector -- will not be included.}

  \item{tmax will be listed as median, min, and max rather than mean, lower
  and higher X\% confidence interval or X percentiles. Similarly, if you
  request trial means, the values for tmax will be the range of medians for
  the trials rather than the range of means.}} An example of acceptable input
  here: \code{c("AUCtau_last", "AUCtau_last_withInhib", "Cmax_last",
  "Cmax_last_withInhib", "AUCtau_ratio_last", "Cmax_ratio_last")}.}

\item{...}{other arguments passed to the function
\code{\link{pksummary_table}}}

\item{checkDataSource}{TRUE (default) or FALSE for whether to include in the
output a data.frame that lists exactly where the data were pulled from the
simulator output file. Useful for QCing.}

\item{save_table}{optionally save the output table and, if requested, the QC
info, by supplying a file name in quotes here, e.g., "My nicely formatted
table.csv". If you leave off ".csv", it will still be saved as a csv file.
If you requested both the table and the QC info, the QC file will have "-
QC" added to the end of the file name.#'}
}
\value{
Returns a data.frame with summary PK parameters from multiple
  simulator output files
}
\description{
Make PK summary tables from multiple simulator output files at once
}
\examples{

# No examples yet

}
