% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify_column_names.R
\name{prettify_column_names}
\alias{prettify_column_names}
\title{Convert the column names in a table of PK parameters into more human-readable
names}
\usage{
prettify_column_names(PKtable, prettify_compound_names = TRUE)
}
\arguments{
\item{PKtable}{a table of PK data with column names such as "AUCinf_dose1" or
"Cmax_last"}

\item{prettify_compound_names}{TRUE (default) or FALSE on whether to make
compound names prettier in the prettified column titles and in any Word
output files. This was designed for simulations where the substrate and any
metabolites, perpetrators, or perpetrator metabolites are among the standard
options for the simulator, and leaving \code{prettify_compound_names =
TRUE} will make the name of those compounds something more human readable.
For example, "SV-Rifampicin-MD" will become "rifampicin", and
"Sim-Midazolam" will become "midazolam". Set each compound to the name
you'd prefer to see in your column titles if you would like something
different. For example, \code{prettify_compound_names = c("perpetrator" =
"teeswiftavir", "substrate" = "superstatin")}. Please note that "perpetrator"
includes \emph{all} the perpetrators and perpetrator metabolites present, so, if
you're setting the perpetrator name, you really should use something like this
if you're including perpetrator metabolites: \code{prettify_compound_names =
c("perpetrator" = "teeswiftavir and 1-OH-teeswiftavir", "substrate" =
"superstatin")}.}
}
\value{
a PK table with prettier column names
}
\description{
\code{prettify_column_names} converts columns in a table of PK values from
names such as "AUCinf_dose1" or "Cmax_last" to prettier names such as "Dose 1
AUCinf (ng/mL.h)" or "Last dose Cmax (ng/mL)"
}
\examples{
# None yet

}
