% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectors.R
\name{purples}
\alias{purples}
\title{Create a vector of purples}
\usage{
purples(ncolors, shade = "regular")
}
\arguments{
\item{ncolors}{number of colors desired}

\item{shade}{the shade of colors to use, which can be "regular" (default),
"darker", "lighter", or "muted"}
}
\value{
a character vector of colors
}
\description{
\code{purples} is useful for getting a set of purples from lavender to
aubergine for graphs.
}
\examples{
# Create a set of 10 purples.
purples(10)

# Try using scales::show_col() to visualize the colors, ex:
scales::show_col(purples(10))

}
