% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_sims.R
\name{qc_sims}
\alias{qc_sims}
\title{Make an Excel file for QCing simulations}
\usage{
qc_sims(
  existing_exp_details,
  template_sim = NA,
  compoundID = NA,
  compound = NA,
  save_output = NA
)
}
\arguments{
\item{existing_exp_details}{output from \code{\link{extractExpDetails}} or
\code{\link{extractExpDetails_mult}}}

\item{template_sim}{optionally include a specific file name -- it must be one
of the files included in the object you supply for
\code{existing_exp_details} -- that can be used as a template simulation to
compare all the other simulations to. Any details in any simulations that
do NOT match the template simulation details will be highlighted in red if
you save the output to an Excel file using \code{save_output}.
\strong{NOTE:} If you use a template simulation, we \emph{strongly
recommend} you also set \code{compoundID} to be only the compound ID you
want, e.g., \code{compoundID = "substrate"} and/or set the compound name to
be only the compound you want, e.g., \code{compound = "midaz"} (this will
match any compound names that include the letters "midaz" and is NOT case
sensitive), and set \code{show_compound_col = FALSE}. This will result in a
much clearer Excel file.}

\item{compoundID}{optionally supply one or more of "substrate", "primary
metabolite 1", "primary metabolite 2", "secondary metabolite", "inhibitor
1", "inhibitor 2", or "inhibitor 1 metabolite" to return information
\emph{only} on that/those compound(s). Remember to contain more than one
compound ID with \code{c(...)}.}

\item{compound}{optionally supply a specific compound name or part of a
specific compound name to get all possible compounds that match that and
\emph{only} compounds that match that. Regular expressions are acceptable
here, e.g., \code{compound = "midaz|keto"} to find any compound with either
"midaz" or "keto" in the name. Not case sensitive.}

\item{save_output}{optionally save the output by supplying an Excel file name
in quotes here, e.g., "Simulation details.xlsx". If you leave off the file
extension, it will be saved as an xlsx file. This will save the output with
one tab per compound ID (one for substrate, one for inhibitor 1, etc.)}
}
\value{
Returns a list of data.frames, broken up by the compound ID, and
  saves each data.frame in a formatted Excel file with one tab per compound
  ID (one for substrate, one for inhibitor 1, etc.)
}
\description{
\code{qc_sims} will create an Excel file specifically designed for general
QCing
}
\examples{
# No examples yet.

}
