% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectors.R
\name{reds}
\alias{reds}
\title{Create a vector of reds}
\usage{
reds(ncolors, shade = "regular")
}
\arguments{
\item{ncolors}{number of colors desired}

\item{shade}{the shade of colors to use, which can be "regular" (default),
"darker", "lighter", or "muted" (although "muted" isn't working well atm
and it just making everything the same color)}
}
\value{
a character vector of colors
}
\description{
\code{reds} is useful for getting a set of reds from pink to
brick for graphs.
}
\examples{
# Create a set of 10 reds. 
reds(10)

# Try using scales::show_col() to visualize the colors, ex:
scales::show_col(reds(10))
scales::show_col(reds(10, shade = "lighter"))

}
