% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ReportInputForm}
\alias{ReportInputForm}
\title{Report input forms to be used in conjunction with
\code{\link{getSectionInfo}} and \code{\link{so_table}}}
\format{
A named list of two forms: "Observed data form" for entering
  information about a clinical study and "Section input form" for entering
  data about a specific simulation file to be described in this section of a
  report
}
\usage{
ReportInputForm
}
\description{
Report input forms to be used in conjunction with
\code{\link{getSectionInfo}} and \code{\link{so_table}}
}
\keyword{datasets}
