% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{round_consultancy}
\alias{round_consultancy}
\title{Rounding to match the SimcypConsultancy report template requirements}
\usage{
round_consultancy(x)
}
\arguments{
\item{x}{}
}
\value{
Returns character data with the appropriate number of digits
}
\description{
If the number is > 100, round to the ones place. If the number is less, round
to 3 significant figures. Output is character data.
}
\examples{
round_consultancy(1239.236)
round_consultancy(0.23513)
round_consultancy(c(1239.236, 0.23513))

}
