% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{round_down_unit}
\alias{round_down_unit}
\title{Round a number or vector of numbers up or down to nice intervals}
\usage{
round_down_unit(x, b)
}
\arguments{
\item{x}{the number or vector of numbers to be rounded}

\item{b}{the unit to round the number to (applies to
\code{\link{round_up_unit}} and \code{\link{round_down_unit}} only)}
}
\value{
a rounded number
}
\description{
Round numbers to nice intervals with the following options: \describe{
\item{round_up}{round up to the nearest power of 10} \item{round_down}{round
down to the nearest power of 10} \item{round_up_nice}{round up to the nearest
"nice" value. From
\url{http://stackoverflow.com/questions/6461209/how-to-round-up-to-the-nearest-10-or-100-or-x)}
} \item{round_up_unit}{round up x to the nearest b}
\item{round_down_unit}{round x down to the nearest b} \item{round_unit}{round
x up or down to the nearest b; direction depends on how close x is to b}
\item{log10_breaks}{creates a vector evenly spaced on a log 10 scale: 10
points across each order of magnitude} }
}
\examples{
round_up(c(4, 1, 101, 99, 1002))
round_down(c(4, 1, 101, 99, 1002))
round_up_nice(0.24562)
round_up_nice(2936.24562)
round_up_nice(c(0.2356, 1.52, 102549.2))
round_up_unit(33.5, 2)
round_up_unit(33.5, 5)
round_up_unit(x = c(4, 1, 101, 99, 1002), b = 5)
round_down_unit(1.23, 0.1)
round_down_unit(x = c(4, 1, 101, 99, 1002), b = 25)


}
