% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Launch the shiny app for making PK summary tables and concentration-time
plots}
\usage{
runShiny()
}
\description{
\code{runShiny} will launch a graphical user interface for making PK summary
tables and concentration-time plots. All the arguments will be specified
within the app itself.
}
\examples{
runShiny()

}
