% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_table_to_Excel.R
\name{save_table_to_Excel}
\alias{save_table_to_Excel}
\title{Save a data.frame or tibble to Excel with some standard formatting}
\usage{
save_table_to_Excel(
  table,
  save_table,
  output_tab_name = "Sheet1",
  freeze_top_row = TRUE,
  center_top_row = TRUE,
  wrap = TRUE
)
}
\arguments{
\item{table}{the table (the data.frame or tibble, really) that you want to
save}

\item{save_table}{file name for saving your table. \emph{This does not work
if you have included a "." anywhere in your file name except in front of
the file extension.} Good: \code{save_table = "my file 1.xlsx"} Bad:
\code{save_table = "my.file.1.xlsx"}}

\item{output_tab_name}{name of the tab to save your table to; defaults to
"Sheet1".}

\item{freeze_top_row}{TRUE (default) or FALSE for whether to freeze the view
in Excel so that the top row will always be visible}

\item{center_top_row}{TRUE (default) or FALSE for whether to center the text
in the top row}

\item{wrap}{TRUE (default) or FALSE for whether to wrap text}
}
\value{
does not return a new object; only saves an existing object to Excel
}
\description{
Please be aware that this \emph{does} overwrite existing data.
}
\examples{
save_table_to_Excel(table = starwars,
                    save_table = "starwars characters.xlsx")
                    
}
