% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_conc.R
\name{scale_y_conc}
\alias{scale_y_conc}
\title{Automatically scale a ggplot graph y axis for concentrations}
\usage{
scale_y_conc(
  conc_range = NA,
  is_enz_plot = FALSE,
  linear_or_log = "linear",
  pad_y_axis = TRUE
)
}
\arguments{
\item{conc_range}{user input for y axis limits for a linear plot}

\item{linear_or_log}{default "linear"}

\item{pad_y_axis}{user-specified value for pad_y_axis}
}
\value{

}
\description{
Automatically scale a ggplot graph y axis for concentrations
}
\examples{

# Working on it


}
