% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_plot.R
\name{sensitivity_plot}
\alias{sensitivity_plot}
\title{Make graphs of sensitivity analysis results - UNDER CONSTRUCTION}
\usage{
sensitivity_plot(SA_file, dependent_variable, title = NA)
}
\arguments{
\item{SA_file}{sensitivity analysis Excel file}

\item{dependent_variable}{dependent variable to plot. Options are: "CL",
"Dose over AUC", "AUC over Dose", "Vss", "Fg", "Fh", "fa", "CLpo", "Cmax",
"AUC", "tmax", or "Plasma Concentration", and, other than "Plasma
Concentration", that parameter \emph{must} be one of the ones you requested
when you ran the sensitivity analysis. Currently case sensitive.}

\item{title}{(optional) a title to include on your graph in quotes}
}
\value{

}
\description{
Doesn't detect units yet. Would need to check that sheet names are
consistent. I bet they're not. Would like to add options for axis breaks and
limits as well as color schemes. Will need to add more options for labeling
the sensitivity parameter in the graphs. Does not do 3D graphs and I have no
plans to add option for 3D graphs b/c A) they're just not easy to interpret
and B) there aren't great R packages out there for 3D graphs.
}
\examples{

SA_file = "../SA example.xlsx"
dependent_variable = "CL"

}
