% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aesthet.R
\name{set_aesthet}
\alias{set_aesthet}
\title{INTERNAL PACKAGE USE: Set up certain aesthetics for a conc-time plot}
\usage{
set_aesthet(
  line_type,
  figure_type,
  MyPerpetrator,
  MyCompoundID,
  obs_shape,
  line_color,
  obs_color,
  obs_line_trans,
  obs_fill_trans
)
}
\arguments{
\item{line_type}{user-specified line type}

\item{figure_type}{user-specified figure type}

\item{MyPerpetrator}{concatenated string of all perpetrators involved}

\item{MyCompoundID}{compound user wants}

\item{obs_shape}{user-specified observed data shape}

\item{line_color}{user-specified line color}

\item{obs_color}{user-specified observed-data color}
}
\value{
several objects to use with setting graph aesthetics
}
\description{
First (not last) step in determinining line colors, line types, point shapes,
and point colors. Subsequent steps will adjust the exact number of all
shapes, linetypes, colors, etc. in the graph.
}
