% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_code_example.R
\name{show_code_example}
\alias{show_code_example}
\title{IN PROGRESS - Show some examples for the code to use for making PK tables or
concentration-time plots}
\usage{
show_code_example()
}
\value{
Doesn't return anything; only shows messages with code examples
}
\description{
IN PROGRESS - Show some examples for the code to use for making PK tables or
concentration-time plots
}
\examples{
# none 
}
