% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colors.R
\name{show_colors}
\alias{show_colors}
\title{Show color sets included in the SimcypConsultancy package}
\usage{
show_colors(color_set, num_colors = "original", shade = NA)
}
\arguments{
\item{color_set}{the color set or color sets you would like to see. Options
are: "all" to see all possibilities or "reds", "blues", "purples",
"blue-greens", "greens", "rainbow", and/or the name of a specific
consultant who has supplied their favorite colors to the R Working Group to
see specific sets of colors. To see which consultants have included their
favorite colors here, please run \code{names(OurFavoriteColors)} in the
console.}

\item{num_colors}{optionally specify the number of colors you wish to see for
each color set. Setting this to "original", the default, will show you the
number of colors that were originally specified for that color set. Set
this to some other number and, if there were not originally that number of
colors, we'll make a gradient between them until you have the number you
want.}

\item{shade}{optionally specify whether you want the shade to be "regular",
"darker", or "lighter". This only applies to "reds", "blues", "purples",
"blue-greens", "greens", "reds", and "rainbow" color sets.}
}
\value{
a graph
}
\description{
Show color sets included in the SimcypConsultancy package
}
\examples{
show_colors(color_set = c("Elle", "Laura Sh", "reds", "blues"),
            num_colors = "original")

show_colors(color_set = "traffic", num_colors = "original")
}
