% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimcypColors}
\alias{SimcypColors}
\title{Simcyp colors used in the PowerPoint template}
\format{
A named character vector of colors
}
\usage{
SimcypColors
}
\description{
Simcyp colors used in the PowerPoint template
}
\examples{
scales::show_col(SimcypColors)
}
\keyword{datasets}
