% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimcypDir}
\alias{SimcypDir}
\title{Paths for Simcyp's large file directory and the SharePoint directory}
\format{
A two-item list:
\describe{
\item{LgFileDir}{the large-file directory}
\item{SharePtDir}{the SharePoint directory}}
}
\usage{
SimcypDir
}
\description{
Paths for Simcyp's large file directory and the SharePoint directory
}
\keyword{datasets}
