% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_graph.R
\name{so_graph}
\alias{so_graph}
\title{Graph of simulated vs. observed PK}
\usage{
so_graph(
  PKtable,
  PKparameters = NA,
  color_set = "red green",
  boundary_indicator = "line",
  line_width = 0.8,
  save_graph = NA,
  fig_width = 8,
  fig_height = 6
)
}
\arguments{
\item{PKtable}{a table in the same format as output from the function
\code{\link{pksummary_mult}}}

\item{PKparameters}{any of the PK parameters included in the output from
\code{\link{pksummary_mult}}; if left as NA, this will make graphs for each
parameter included in \code{PKtable}. To see the full set of possible
parameters, enter \code{view(PKParameterDefinitions)} into the console.}

\item{color_set}{set of colors to use for indicating the 1.5-fold and 2-fold
boundaries of the simulated / observed ratio. The default is "red green",
"muted red green" will result in lighter, somewhat more muted red and green
that works better for indicating the boundary with shading instead of
lines, and "black" will result in only black lines or shading. You also can
set this to any two colors you'd like, e.g., \code{color_set = c("yellow",
"blue")}}

\item{boundary_indicator}{how to indicate the boundaries of 1.5-fold and
2-fold comparisons of simulated / observed. Options are "lines" (default)
"fill" to get a shaded area, or "none" to remove any indicators of those
boundaries.}

\item{line_width}{line width; default is 0.8}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My conc time graph.png"}

\item{fig_width}{figure width in inches; default is 6}

\item{fig_height}{figure height in inches; default is 8}
}
\value{

}
\description{
\code{so_graph} makes a graph of simulated vs. observed PK, including
indicating where the predicted parameters fell within 1.5- or 2-fold of the
observed.
}
\examples{

}
