% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SOdata}
\alias{SOdata}
\title{Example simulated vs. observed PK data for use with the so_graph function}
\format{
a data.frame
}
\usage{
SOdata
}
\description{
Example simulated vs. observed PK data for use with the so_graph function
}
\keyword{datasets}
