% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_sims.R
\name{sort_sims}
\alias{sort_sims}
\title{Sort the simulation file order in an object}
\usage{
sort_sims(which_object, sim_order)
}
\arguments{
\item{which_object}{the R object for which you want to sort simulation files}

\item{sim_order}{what should the simulation file order be? Must be a
character vector of the file names \emph{exactly} as they appear in the
column "File" if the object is a data.frame or in the data.frames inside
the list if the object is a list. Note that this means that "File" must be
a column in the data.frames. If it is not, the object will be returned
without any sorting. If any simulation files were omitted from
\code{sort_order}, they will be included at the end in the same order they
originally appeared.}
}
\value{
the R object with all simulations sorted.
}
\description{
Sort the simulation file order in an object
}
\examples{
Details <- sort_sims(which_object = Details,
                     sim_order = c("file 1.xlsx", "file 2.xlsx"))


}
