% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_comma.R
\name{str_comma}
\alias{str_comma}
\title{Collapse a vector of character strings into one readable string, separated by
commas and by "and"}
\usage{
str_comma(x, oxford = TRUE, conjunction = "and")
}
\arguments{
\item{x}{vector to collapse}

\item{oxford}{TRUE or FALSE for whether to use the Oxford comma}

\item{conjunction}{the conjunction to use to join the character strings.
Options are "and" (default) and "or".}
}
\value{
Returns a single, collapsed character string
}
\description{
\code{str_comma} combines multiple character strings into one and makes them
more readable by separating those strings with commas and "and" or "or" as
appropriate. This removes missing values from x before collapsing.
}
\examples{
str_comma(LETTERS[1:2])
str_comma(LETTERS[1:4])
str_comma(LETTERS[1:4], oxford = FALSE)


}
