% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_consultancy.R
\name{theme_consultancy}
\alias{theme_consultancy}
\title{Apply the standard Consultancy Team graphing styles to your graph}
\usage{
theme_consultancy()
}
\value{

}
\description{
\code{theme_consultancy} sets the following: \itemize{ \item{Plot background
is white; this includes the background for the legend and any facets and
their labels.} \item{Axis ticks, lines, text, and titles are black.} \item{If
it's a concentration-time plot or an enzyme-abundance plot that does
\emph{not} have any facets, then there will be a line on the x axis and a
line on the y axis but no overall border around the graph.} \item{If the
graph is facetted, the labels for the facets are not enclosed with a border.}
\item{Axis titles and, when applicable, the plot title are bold face.}
\item{Font sizes and margins are the same as in theme_bw in
\code{\link{ggplot2::ggtheme}}} }
}
\examples{
# Some example data to work with
MyData <- data.frame(X = rnorm(10), Y = runif(10))

# Without applying this theme to a standard ggplot2 graph
ggplot(MyData, aes(x = X, y = Y)) + geom_point()

# With application of this theme
ggplot(MyData, aes(x = X, y = Y)) + geom_point() +
       theme_consultancy()

}
