% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_input_PK.R
\name{tidy_input_PK}
\alias{tidy_input_PK}
\title{Tidy input PK data as needed}
\usage{
tidy_input_PK(
  PKparameters,
  sim_data_files = NA,
  compoundsToExtract = "substrate",
  tissues = "plasma",
  sheet_PKparameters = NA,
  report_input_file = NA,
  sheet_report = NA,
  existing_exp_details = NA
)
}
\arguments{
\item{PKparameters}{whatever they have supplied for PKparameters for
pksummary_table or pksummary_mult. This can be a csv file, a character
vector, or NA.}

\item{sim_data_files}{a character vector of file names. This should no longer
be NA or "recursive".}

\item{compoundsToExtract}{the values they supplied for compoundsToExtract in
the parent function.}

\item{tissues}{the values they supplied for tissues in the parent function}

\item{sheet_PKparameters}{the values they supplied for sheet_PKparameters in
the parent function}

\item{report_input_file}{probably not in use}

\item{sheet_report}{probably not in use}
}
\value{
a data.frame
}
\description{
INTERNAL PACKAGE USE ONLY. Reads (as needed), harmonizes and tidies observed
PK input. Output is a data.frame of PK parameters with columns File, Sheet,
CompoundID, Tissue, Value, and Variability. The columns Sheet, Value, and
Variability may contain NAs.
}
\examples{
# nope 
}
