% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_PKparameters_names.R
\name{tidy_PKparameters_names}
\alias{tidy_PKparameters_names}
\title{INTERNAL USE. tidy the column names of PKparameters supplied to, e.g.,
pk_table.}
\usage{
tidy_PKparameters_names(PKparameters)
}
\arguments{
\item{PKparameters}{a data.frame of PK parameters}
}
\value{
a data.frame of PK parameters with harmonized column names
}
\description{
INTERNAL USE. tidy the column names of PKparameters supplied to, e.g.,
pk_table.
}
