% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_means_plot.R
\name{trial_means_plot}
\alias{trial_means_plot}
\title{Make graphs comparing center statistic and variability in PK across trials
and, optionally, against observed PK data as well}
\usage{
trial_means_plot(
  sim_data_file,
  PKparameter = "Cmax_dose1",
  compoundToExtract = "substrate",
  tissue = "plasma",
  sheet = NA,
  mean_type = "geometric",
  variability_type = "90\% CI",
  y_axis_limits_lin = NA,
  include_dose_num = FALSE,
  existing_exp_details = NA,
  observed_PK = NA,
  return_caption = FALSE,
  name_clinical_study = NA
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes}

\item{PKparameter}{PK parameter you want to extract from the simulator output
file. To see the full set of possible parameters to extract, enter
\code{view(PKParameterDefinitions)} into the console. Not case sensitive.
If you use "_first" instead of "_dose1", that will also work.}

\item{compoundToExtract}{For which compound do you want to extract
PK data? Options are: \itemize{\item{"substrate"
(default),} \item{"primary metabolite 1",} \item{"primary metabolite 2",}
\item{"secondary metabolite",} \item{"inhibitor 1" -- this can be an
inducer, inhibitor, activator, or suppresesor, but it's labeled as
"Inhibitor 1" in the simulator,} \item{"inhibitor 2" for the 2nd inhibitor
listed in the simulation,} \item{"inhibitor 1 metabolite" for the primary
metabolite of inhibitor 1}}}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are "plasma" (default), "unbound plasma", "blood", "unbound blood",
"peripheral plasma", or "peripheral blood". \strong{NOTE: PK for peripheral
sampling is not as well tested as for other tissues and is only set up for
V21+. Please check your results carefully.}}

\item{sheet}{optionally specify the name of the sheet where you'd like to
pull the PK data, in quotes; for example, specify the tab where you have a
user-defined AUC integration. \emph{Note:} Unless you want a very specific
Excel sheet that's not what the usual sheet name would be for a first or
last dose, this function will work best if this is left as NA. Also, since
we don't know which dose these data were for, you'll see that the output
parameter names do not include the suffixes "_last" or "_dose1".}

\item{mean_type}{What kind of means do you want to use for the center point
on the graph? Options are "arithmetic" or "geometric" (default).}

\item{variability_type}{What variability statistic to you want to use for the
error bars? Options are "90% CI", "SD", "CV", "GCV" (for geometric CV), or
"range".}

\item{y_axis_limits_lin}{optionally set the Y axis limits for the linear
plot, e.g., \code{c(10, 1000)}. If left as the default NA, the Y axis
limits for the linear plot will be automatically selected. (Setting up
semi-log plot y axis intervals manually is a bit tricky and is not
currently supported.)}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} to get all the details from the "Input
Sheet" (e.g., when you ran extractExpDetails you said \code{exp_details =
"Summary and Input"} or \code{exp_details = "all"}), you can save some processing
time by supplying that object here, unquoted. If left as NA, this function
will run \code{extractExpDetails} behind the scenes anyway to figure out
some information about your experimental set up.}

\item{observed_PK}{a data.frame of observed PK that you would like to
compare. This must include the columns "ObsValue" for the center value for
the observed data and "ObsVariability" for the error bars for the observed
data. We'll assume that observed PK mean type and variability type match
what's supplied here. This will probably be buggy if they don't. UNDER
CONSTRUCTION. LIKELY SUPER BUGGY.}

\item{return_caption}{TRUE or FALSE (default) for whether to return any
caption text to use with the graph. This works best if you supply something
for the argument \code{existing_exp_details}. If set to TRUE, you'll get as
output a list of the graph, the figure heading, and the figure caption.}

\item{name_clinical_study}{optionally specify the name(s) of the clinical
study or studies for any observed data. This only affects the caption of
the graph. For example, specifying \code{name_clinical_study = "101, fed
cohort"} will result in a figure caption that reads in part "clinical study
101, fed cohort". If you have more than one study, that's fine; we'll take
care of stringing them together appropriately. Just list them as a
character vector, e.g., \code{name_clinical_study = c("101",
"102", "103")} will become "clinical studies 101, 102, and 103."}
}
\value{
a ggplot2 graph
}
\description{
UNDER CONSTRUCTION. For now, must supply observed_PK as a
  data.frame.
}
\examples{
# None yet

}
