% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VBE_disso_example}
\alias{VBE_disso_example}
\title{Examples for setting up data for use with the VBE_safe_space_plot function}
\format{
a data.frame
}
\usage{
VBE_disso_example
}
\description{
Examples for setting up data for use with the VBE_safe_space_plot function
}
\keyword{datasets}
