% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBE_safe_space_plot.R
\name{VBE_safe_space_plot}
\alias{VBE_safe_space_plot}
\title{Make dissolution safe space plots}
\usage{
VBE_safe_space_plot(
  VBE_dataframe,
  color_set = NA,
  safe_space_color = NA,
  safe_space_trans = NA,
  linetypes = NA,
  linewidths = NA,
  point_shapes = NA,
  point_sizes = NA,
  save_graph = NA,
  fig_height = NA,
  fig_width = NA
)
}
\arguments{
\item{VBE_dataframe}{data.frame}

\item{color_set}{color set}

\item{safe_space_color}{color}

\item{safe_space_trans}{transparency}

\item{linetypes}{the line type(s) to use for the graph. Possible options
can be seen by typing \code{ggpubr::show_line_types()} into the console.}

\item{linewidths}{must be numeric}

\item{point_shapes}{the point shape(s) to use for the graph. Possible options
can be seen by typing \code{ggpubr::show_point_shapes()} into the console.}

\item{point_sizes}{point sizes. must be numeric.}

\item{save_graph}{file name}

\item{fig_height}{default is 6}

\item{fig_width}{default is 8}
}
\value{
a ggplot2 graph
}
\description{
The function \code{VBE_safe_space_plot} makes dissolution ...
UNDER CONSTRUCTION
}
\examples{
# None yet

}
