% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapn.R
\name{wrapn}
\alias{wrapn}
\title{INTERNAL PACKAGE USE. Wrap warning text and add a carriage return at the end.}
\usage{
wrapn(x)
}
\arguments{
\item{x}{warning text to wrap}
}
\value{
a character string
}
\description{
INTERNAL PACKAGE USE. Wrap warning text and add a carriage return at the end.
}
\examples{
warning(wrapn("This is a long warning that should be wrapped nicely with pretty line breaks and end with a carriage return."))
}
