% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectors.R
\name{yellow_to_red}
\alias{yellow_to_red}
\title{Make a vector of colors that match the "yellow to red" option for
highlighting GMR values}
\usage{
yellow_to_red()
}
\value{
a named character vector
}
\description{
Make a vector of colors that match the "yellow to red" option for
highlighting GMR values
}
