% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sims.R
\name{add_sims}
\alias{add_sims}
\title{Add simulations to an R object}
\usage{
add_sims(object_A, object_B)
}
\arguments{
\item{object_A}{one R object}

\item{object_B}{a second R object whose results you want to combine with the
first R object. Bothe object_A and object_B must have the same data class,
i.e., both must be lists or both data.frames.}
}
\value{
the combined R object
}
\description{
\code{add_sims} was designed with the idea of adding information from two
objects that were created by running extractExpDetails_mult, but you can also
use it with other R objects. For example, say you ran extractExpDetails_mult
with some initial development simulations and then, later, you ran it with
some other development simulations another time and now, you want both of
those sets of results be combined, maybe because you want to run
\code{\link{annotateDetails}} to compare them.
}
\examples{
Details <- add_sims(object_A = Details_dev,
                    object_B = Detrails_final)


}
