% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addObsPoints.R
\name{addObsPoints}
\alias{addObsPoints}
\title{INTERNAL PACKAGE USE: Add observed data points to a concentration-time plot}
\usage{
addObsPoints(
  obs_dataframe,
  A,
  figure_type,
  AESCols,
  obs_shape,
  line_type,
  obs_color,
  line_color,
  obs_size,
  obs_line_trans,
  obs_fill_trans,
  connect_obs_points,
  line_width,
  LegCheck
)
}
\arguments{
\item{obs_dataframe}{observed data as a data.frame}

\item{A}{the existing ggplot2 graph to which the observed data will be added}

\item{figure_type}{figure type}

\item{AESCols}{which aesthetics are mapped to which columns}

\item{obs_shape}{shape}

\item{line_type}{line types used for simulated data b/c obs shape needs to be
named the same for ct_plot legends}

\item{obs_color}{color}

\item{line_color}{line colors used for simulated data b/c obs color needs to
be named the same for ct_plot legends}

\item{obs_size}{size}

\item{obs_line_trans}{line transparency}

\item{obs_fill_trans}{fill transparency}

\item{connect_obs_points}{TRUE or FALSE (default) for whether to add
connecting lines between observed data points from the same individual}

\item{line_width}{line width}

\item{LegCheck}{TRUE or FALSE for whether to include the legend. From
ct_plot_overlay originally and should always be TRUE for ct_plot.}
}
\value{
a ggplot2 layer with observed data
}
\description{
\code{addObsPoints} adds observed data to \code{\link{ct_plot}} or
\code{\link{ct_plot_overlay}} and sets the color and shape of those points
appropriately. This is NOT meant to be used as a stand-alone function.
}
