% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllExpDetails}
\alias{AllExpDetails}
\title{All possible experimental details that can be extracted from a simulator
output file using \code{\link{extractExpDetails}}.}
\format{
A data.frame with 9 columns: \describe{

  \item{Detail}{the experimental detail name to use with
  \code{\link{extractExpDetails}} for the argument \code{exp_details}}

  \item{Compound}{the specific compound this experimental detail applies to}

  \item{Notes}{an explanation of what the experimental detail is}

  \item{NameColDetect}{FOR INTERNAL USE}

  \item{Class}{Data class}

  \item{Sheet}{Which simulator output sheet this detail is extracted from}

  \item{NameCol}{Which column in the simulator output tab will be searched
  for this detail}

  \item{ValueCol}{Which column in the simulator output tab will contains the
  value used for this detail}

  \item{CDSInputMatch}{FOR INTERNAL USE: Which compound data sheet item on
  the "Simcyp Inputs and QC" tab match this experimental detail} }
}
\usage{
AllExpDetails
}
\description{
All possible experimental details that can be extracted from a simulator
output file using \code{\link{extractExpDetails}}.
}
\keyword{datasets}
