% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllExpDetails}
\alias{AllExpDetails}
\title{All possible experimental details for \code{\link{extractExpDetails}}
(version for coding purposes)}
\format{
A data.frame with 10 columns: \describe{

  \item{Detail}{the experimental detail name to use with
  \code{\link{extractExpDetails}} for the argument \code{exp_details}}

  \item{CompoundID}{the specific compound ID this experimental detail applies
  to, e.g., "substrate", "inhibitor 1", etc.}

  \item{Notes}{an explanation of what the experimental detail is}

  \item{NameColDetect}{FOR INTERNAL USE. When searching the Input Sheet for
  the set of columns to use, what regular expression should be used.}

  \item{Class}{Data class}

  \item{Sheet}{Which simulator output sheet this detail is extracted from}

  \item{NameCol}{Which column in the simulator output tab will be searched
  for this detail}

  \item{ValueCol}{Which column in the simulator output tab will contains the
  value used for this detail}

  \item{CDSInputMatch}{FOR INTERNAL USE: Which compound data sheet item on
  the "Simcyp Inputs and QC" tab match this experimental detail}

  \item{SimulatorSection}{FOR INTERNAL USE: For matching with the CDS, which
  section does this detail belong? Options: Absorption, Distribution,
  Elimination, Interaction, Phys Chem and Blood Binding.}

  }
}
\usage{
AllExpDetails
}
\description{
All possible experimental details that can be extracted from a simulator
output file using \code{\link{extractExpDetails}}. This version includes
columns used only internally with coding. For the more user-friendly version,
see \code{\link{ExpDetailDefinitions}}.
}
\keyword{datasets}
