% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllExpDetails}
\alias{AllExpDetails}
\title{All possible experimental details for \code{\link{extractExpDetails}}}
\format{
A data.frame with columns: \describe{

  \item{Detail}{the experimental detail name to use with
  \code{\link{extractExpDetails}} for the argument \code{exp_details}}

  \item{Notes}{an explanation of what the experimental detail is}

  \item{Regex_row}{When searching the specified column, what regular
  expression should be used for finding the correct row.}

  \item{Class}{Data class}

  \item{Sheet}{simulator output sheet this detail is extracted from}

  \item{NameCol}{the index of the column in the simulator output tab that will
  be searched for matches to the text in the column "Regex_row".}

  \item{ValueCol}{index of the column in the simulator output tab that
  contains the value used for this detail. This applies when the column
  doesn't move, e.g., \emph{not} Input Sheet compound-specific details.}

  \item{ColsChangeWithCmpd}{TRUE or FALSE for whether NameCol and/or ValueCol
  change depending on what compound ID it is}

  \item{CmpdsForWhichDeetAvail}{Compounds for which this detail is available.
  Many compound-specific details are available for all compounds (entry here
  is "all"), but some are only available for the substrate (entry here would
  be "substrate") or only for the substrate and inhibitor 1
  ("substrate, inhibitor 1").}

  \item{CDSInputMatch}{compound data sheet item on
  the "Simcyp Inputs and QC" tab that matches this experimental detail}

  \item{SimulatorSection}{For matching with the CDS, in which
  section does this detail belong? Options: Absorption, Distribution,
  Elimination, Interaction, Phys Chem and Blood Binding.}

  \item{ADAMParameter}{TRUE or FALSE for whether the
  parameter only comes into play when it's an ADAM model}

  \item{DiscoveryParameter}{Is this detail available for Simcyp Discovery
  simulations? Options are "Simulator only", "Discovery only", or "Simulator
  and Discovery". This is used to determine which details to check and which
  regex to use for those details.}

  \item{OffsetRows}{When there just isn't good regex in the
  specific row for this detail, we need to look for a value and then go down
  this number of rows to get to the actual value we want.}

  \item{SortOrder}{Order in which this detail should show
  up in the output of `annotateDetails`}

  \item{Level1, Level2, etc.}{the level in the XML file
  where the information lives}

  \item{XMLswitch}{If there is a switch involved in the
  XML file, what is the tag to use for checking on that switch}

  \item{SwitchTo}{If the XML switch is turned to "1" or "true"
  in the XML file, this is the tag to switch to for looking up the value for
  this detail.}

  \item{CodingNotes}{Notes specifically to coders}

  \item{CompoundID}{CompoundID (a.k.a. compound position in the Simulator)}

  \item{Suffix}{suffix to append for this detail when it's for a specific
  compound ID}

  }
}
\usage{
AllExpDetails
}
\description{
INTERNAL USE: All possible experimental details that can be extracted from a
simulator output file using \code{\link{extractExpDetails}}. This version
includes columns used only internally with coding. For the more user-friendly
version, see \code{\link{ExpDetailDefinitions}}.
}
\keyword{datasets}
