% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AllPKParameters}
\alias{AllPKParameters}
\title{All possible PK parameters that can be extracted from a simulator output file
using \code{\link{extractPK}}.}
\format{
A data.frame with 3 columns: \describe{ \item{PKparameter}{the PK
  parameter name to use with \code{\link{extractPK}} for the argument
  \code{PKparameters}} \item{Sheet}{the sheet in a simulator output file
  where the PK parameter will be extracted, if possible} \item{Notes}{an
  explanation of what the parameter is}}
}
\usage{
AllPKParameters
}
\description{
All possible PK parameters that can be extracted from a simulator output file
using \code{\link{extractPK}}.
}
\keyword{datasets}
