% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDetails.R
\name{annotateDetails}
\alias{annotateDetails}
\title{Annotate Simcyp Simulator experimental details and format into a long
data.frame instead of a wide one}
\usage{
annotateDetails(
  Deets,
  show_compound_col = TRUE,
  compoundID = NA,
  compound = NA,
  detail_set = NA,
  simulator_section = NA,
  omit_all_missing = FALSE,
  save_output = NA
)
}
\arguments{
\item{Deets}{output from \code{\link{extractExpDetails}} or
\code{\link{extractExpDetails_mult}}}

\item{show_compound_col}{TRUE (default), FALSE, or "concatenate" for whether
to include in the results the column "Compound", which is the compound's
specific name in each simulation. Why would you ever omit this? If you have
a compound with a slightly different name across multiple simulations,
e.g., "DrugX" and "Drug X", and "Drug X - reduced Ki", you'll get a new row
for every possible combination of "Compound" and "Detail", which might not
make for easy comparisons. For example, a Ki value for "DrugX" will be in
one row and the same Ki value for "Drug X" will be on a separate row. Try
setting this to TRUE when you have similarly named compounds that really
should be compared and see how that compares to the default. If you set
this to "concatenate", you'll get all the possible compound names together;
for example, you might see "DrugX, Drug X, or Drug X - reduced Ki" listed
as the compound.}

\item{compoundID}{optionally supply one or more of "substrate", "primary
metabolite 1", "primary metabolite 2", "secondary metabolite", "inhibitor
1", "inhibitor 2", or "inhibitor 1 metabolite" to return information
\emph{only} on that/those compound(s). Remember to contain more than one
compound ID with \code{c(...)}.}

\item{compound}{optionally supply a specific compound name or part of a
specific compound name to get all possible compounds that match that and
\emph{only} compounds that match that. Regular expressions are acceptable
here, e.g., \code{compound = "midaz|keto"} to find any compound with either
"midaz" or "keto" in the name. Not case sensitive.}

\item{detail_set}{optionally supply a set of details to return \emph{only}
  those details. Options are the same as the ones listed as possibilities for
  \code{exp_details} for \code{\link{extractExpDetails}} or
  \code{\link{extractExpDetails_mult}}: \describe{

  \item{"Summary tab"}{details available from the "Summary tab" (default)}

  \item{"Input Sheet"}{details available from the "Input Sheet" tab}

  \item{"population tab"}{details about the population used (data come from
  the tab with the same name as the population simulated)}

  \item{"Simcyp inputs"}{the details that you normally fill out on the
  "Simcyp inputs (and QC)" tab of a compound data sheet}

  \item{"all"}{all possible details}

  \item{a string of the specific details you want, each in quotes and
  encapsulated with \code{c(...)},}{For a complete list of possibilities,
  type \code{data(ExpDetailDefinitions); view(ExpDetailDefinitions)} into the
  console. Parameters are reported with a suffix depending on which compound
  they pertain to: "_sub" for the substrate, "_met1" for the primary
  metabolite, "_met2" for the second primary metabolite, "_secmet" for the
  secondary metabolite, "_inhib" for the 1st inhibitor or inducer listed,
  "_inhib2" for the 2nd inhibitor or inducer listed, or "_inh1met" for the
  inhibitor 1 metabolite. An example of acceptable input: \code{c("pKa1_sub",
  "fa_inhib2", "Regimen_sub")}}  Not case sensitive.}}

\item{simulator_section}{optionally supply a specific simulator section or
sections from which to find simulation experimental details and then return
\emph{only} those details. Options are "Absorption", "Distribution",
"Elimination", "Interaction", "Phys Chem and Blood Binding", "Population",
"Transport", or "Trial Design". Not case sensitive. If you want more than
one, enclose them with \code{c(...)}.}

\item{omit_all_missing}{TRUE or FALSE (default) for whether to omit a detail
if the values are NA for all files}

\item{save_output}{optionally save the output by supplying a file name in
quotes here, e.g., "My experimental details.csv". If you leave off ".csv",
it will still be saved as a csv file.}
}
\value{

}
\description{
\code{annotateDetails} converts output from either
\code{\link{extractExpDetails}} or \code{\link{extractExpDetails_mult}} into
a long data.frame and adds to that data.frame columns for
\enumerate{\item{which compound the information pertains to (substrate,
inhibitor, etc.),} \item{which section of the Simcyp Simulator this detail is
found in (physchem, absorption, distribution, etc.),} \item{notes describing
what the detail is, and} \item{which sheet in the Excel file the information
was pulled from.}} It will also optionally filter the data to return only
specifically requested information.
}
\details{
\emph{Nota bene:} When you initially extracted your simulation experimental
details, if you set \code{annotate_output = TRUE} and \code{show_compound_col
= "concatenate"}, please understand that you cannot go backwards here an
un-concatenate the compound names. You'll still get concatenated compound
names in the output.
}
\examples{

annotateDetails(Deets)

# Get annotated details regarding absorption.
annotateDetails(Deets = MyDetails, simulator_section = "absorption")

# Get annotated details on whatever was used as inhibitor 1
annotateDetails(Deets = MyDetails, compoundID = "inhibitor 1")

# Get annotated details for any compounds with "midaz" or "keto" in the name.
annotateDetails(Deets = MyDetails, compound = "midaz|keto")

# Get (most of) the details to put into a table of simulation inputs:
annotateDetails(Deets = MyDetails, detail_set = "Simcyp inputs")

# Combine multiple options to get just a few specific details:
annotateDetails(Deets = MyDetails, 
            simulator_section = "absorption",
            compound = "midaz|keto", 
            compoundID = "substrate")


}
