% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDetails.R
\name{annotateDetails}
\alias{annotateDetails}
\title{Annotate Simcyp Simulator experimental details}
\usage{
annotateDetails(
  Deets,
  compoundID = NA,
  compound = NA,
  detail_set = NA,
  find_matching_details = NA,
  simulator_section = NA,
  show_compound_col = TRUE,
  omit_all_missing = TRUE,
  save_output = NA
)
}
\arguments{
\item{Deets}{output from \code{\link{extractExpDetails}} or
\code{\link{extractExpDetails_mult}}}

\item{compoundID}{optionally supply one or more of "substrate", "primary
metabolite 1", "primary metabolite 2", "secondary metabolite", "inhibitor
1", "inhibitor 2", or "inhibitor 1 metabolite" to return information
\emph{only} on that/those compound(s). Remember to contain more than one
compound ID with \code{c(...)}.}

\item{compound}{optionally supply a specific compound name or part of a
specific compound name to get all possible compounds that match that and
\emph{only} compounds that match that. Regular expressions are acceptable
here, e.g., \code{compound = "midaz|keto"} to find any compound with either
"midaz" or "keto" in the name. Not case sensitive.}

\item{detail_set}{optionally supply a set of details to return \emph{only}
  those details. Options are the same as the ones listed as possibilities for
  \code{exp_details} for \code{\link{extractExpDetails}} or
  \code{\link{extractExpDetails_mult}}: \describe{

  \item{"Summary tab"}{details available from the "Summary tab" (default)}

  \item{"Input Sheet"}{details available from the "Input Sheet" tab}

  \item{"population tab"}{details about the population used (data come from
  the tab with the same name as the population simulated)}

  \item{"Simcyp inputs"}{the details that you normally fill out on the
  "Simcyp inputs (and QC)" tab of a compound data sheet}

  \item{"all"}{all possible details}

  \item{a string of the specific details you want, each in quotes and
  encapsulated with \code{c(...)},}{For a complete list of possibilities,
  type \code{view(ExpDetailDefinitions)} into the console. Parameters are
  reported with a suffix depending on which compound they pertain to: "_sub"
  for the substrate, "_met1" for the primary metabolite, "_met2" for the
  second primary metabolite, "_secmet" for the secondary metabolite, "_inhib"
  for the 1st inhibitor or inducer listed, "_inhib2" for the 2nd inhibitor or
  inducer listed, or "_inh1met" for the inhibitor 1 metabolite. An example of
  acceptable input: \code{c("pKa1_sub", "fa_inhib2", "Regimen_sub")}}  Not
  case sensitive.}}

\item{find_matching_details}{optionally supply a string of text to search for
  in the column "Detail". Regular expressions are supported here, so use,
  e.g., \itemize{

  \item{"|" to mean "or",}

  \item{".*" to mean "match any character any number of times",}

  \item{"^" to mean "starts with...", and}

  \item{"$" to mean "ends with...".}} An example: \code{find_matching_details
  = "^CLint.*CYP3A.*sub$"} to find all details that start with "CLint", then,
  somewhere in the name, include "CYP3A", and then end with "sub" (for
  "substrate"). Case sensitive.}

\item{simulator_section}{optionally supply a specific simulator section or
sections from which to find simulation experimental details and then return
\emph{only} those details. Options are "Absorption", "Distribution",
"Elimination", "Interaction", "Phys Chem and Blood Binding", "Population",
"Transport", or "Trial Design". Not case sensitive. If you want more than
one, enclose them with \code{c(...)}.}

\item{show_compound_col}{TRUE (default), FALSE, or "concatenate" for whether
to include in the results the column "Compound", which is the compound's
specific name in each simulation. Why would you ever omit this? If you have
a compound with a slightly different name across multiple simulations,
e.g., "DrugX" and "Drug X", and "Drug X - reduced Ki", you'll get a new row
for every possible combination of "Compound" and "Detail", which might not
make for easy comparisons. For example, a Ki value for "DrugX" will be in
one row and the same Ki value for "Drug X" will be on a separate row. Try
setting this to TRUE when you have similarly named compounds that really
should be compared and see how that compares to the default. If you set
this to "concatenate", you'll get all the possible compound names together;
for example, you might see "DrugX, Drug X, or Drug X - reduced Ki" listed
as the compound.}

\item{omit_all_missing}{TRUE (default) or FALSE for whether to omit a detail
if the values are NA for all files}

\item{save_output}{optionally save the output by supplying a file name in
quotes here, e.g., "My experimental details.csv". If you leave off ".csv",
it will still be saved as a csv file.}
}
\value{
Returns a data.frame of simulation experimental details including the
  following columns: \describe{

  \item{SimulatorSection}{the simulator section this detail is from, e.g.,
  "absorption" or "elimination"}

  \item{Sheet}{the sheet in the Excel output where the data were found}

  \item{Notes}{an explanation of what this detail is}

  \item{CompoundID}{the simulator compound ID that this information pertains
  to, e.g., "substrate" or "inhibitor 1"}

  \item{Compound}{the name of the compound in the simulator, e.g.,
  "Sim-Midazolam"}

  \item{Detail}{the specific experimental detail}

  \item{All files have this value for this compound and compound ID}{If all
  of the files have the \emph{exact same value} for this particular
  combination of detail, compound, and compound ID, that value will show up
  in this column. The idea is that this makes it easy to check that
  everything that \emph{should} be the same in a plethora of simulations
  actually \emph{is} the same.}

  \item{one column for every file included}{the extracted simulator
  experimental details for each file will show up in their own columns}}
}
\description{

}
\details{
\emph{Nota bene:} When you initially extracted your simulation experimental
details, if you set \code{annotate_output = TRUE} and \code{show_compound_col
= "concatenate"}, please understand that you cannot go backwards here and
un-concatenate the compound names because, once we combine everything, we
don't know which piece of data belonged to which compound. You'll still get
concatenated compound names in the output.
}
\examples{

annotateDetails(Deets = MDZdetails)

# Get annotated details regarding absorption.
annotateDetails(Deets = MDZdetails, simulator_section = "absorption")

# Get annotated details on whatever was used as inhibitor 1
annotateDetails(Deets = MDZdetails, compoundID = "inhibitor 1")

# Get annotated details for any compounds with "midaz" or "keto" in the name.
annotateDetails(Deets = MDZdetails, compound = "midaz|keto")

# Get (most of) the details to put into a table of simulation inputs:
annotateDetails(Deets = MDZdetails, detail_set = "Simcyp inputs")

# Combine multiple options to get just a few specific details:
annotateDetails(Deets = MDZdetails,
            simulator_section = "absorption",
            compound = "midaz|keto",
            compoundID = "substrate")


}
