% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BufForestData}
\alias{BufForestData}
\title{Example output from \code{extractForestData}}
\format{
a data.frame
}
\usage{
BufForestData
}
\description{
Example output from running \code{extractForestData} on some simulations with
bufuralol as the substrate and then 4 different effectors.
}
\keyword{datasets}
