% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dosenumber.R
\name{calc_dosenumber}
\alias{calc_dosenumber}
\title{Calculate the dose number for each time in a set of concentration-time data
given the experimental details for that simulation}
\usage{
calc_dosenumber(ct_dataframe, existing_exp_details)
}
\arguments{
\item{ct_dataframe}{a data.frame of concentration-time data from running
either \code{\link{extractConcTime}} or \code{\link{extractConcTime_mult}}}

\item{existing_exp_details}{the output from running
\code{\link{extractExpDetails}} or \code{\link{extractExpDetails_mult}} on
the same simulations that were used for making \code{ct_dataframe}. Note:
We have not set up this function to check that your files match in order to
allow for some flexibility here. If you supply \code{existing_exp_details}
with more than one simulator output file, we'll just use the 1st one. If
you would like to have this calculate dose numbers on other data where you
wouldn't have simulation experimental details (example: observed data),
then supply a single-row data.frame with the following columns: \code{File}
(character, just set this to "all" or some other placeholder text),
\code{DoseInt_sub} (numeric), \code{StartHr_sub} (numeric; probably 0),
\code{NumDoses_sub} (numeric), and \code{Regimen_sub} (character;
presumably "Multiple Dose" here). If you want the dose number for other
compound IDs, then replace "_sub" with, e.g., "_inhib". Please run
\code{view(ExpDetailDefinitions)} to see examples of acceptable suffixes.}
}
\value{
Output is a data.frame of concentration-time data with
}
\description{
\code{calc_dosenumber} is mainly for internal use in the SimcypConsultancy
package and will calculate which dose number a given time should be if you
provide the ct_dataframe and the extracted simulation experimental details
(the output from \code{\link{extractExpDetails}}). This is meant to be used
for a single simulation at a time or a set of concentration-time data where
the dosing regimen is IDENTICAL for all simulations.
}
\examples{
# None yet

}
