% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PK_ratios.R
\name{calc_PK_ratios}
\alias{calc_PK_ratios}
\title{Calculate the ratio of PK parameters between the two simulations -- UNDER
CONSTRUCTION}
\usage{
calc_PK_ratios(
  sim_data_file_numerator,
  sim_data_file_denominator,
  paired = TRUE,
  PKparameters = "AUC tab",
  sheet_PKparameters_num = NA,
  sheet_PKparameters_denom = NA,
  tissue = "plasma",
  mean_type = "geometric",
  conf_int = 0.9,
  includeCV = TRUE,
  includeConfInt = TRUE,
  prettify_columns = TRUE,
  prettify_compound_names = TRUE,
  checkDataSource = TRUE,
  save_table = NA,
  fontsize = 11
)
}
\arguments{
\item{sim_data_file_numerator}{a simulator output Excel file that will
provide the numerator for the calculated ratios.}

\item{sim_data_file_denominator}{a simulator output Excel file that will
provide the denominator for the calculated ratios}

\item{paired}{TRUE (default) or FALSE for whether the study design is paired,
as in, the subjects are \emph{identical} between the two simulations. For
paired study designs, the order of operations is to calculate each
subject's ratio and then to calculate the mean of those ratios. For
unpaired study designs, the order of operations is to calculate the mean of
the parameter of interest for the numerator file and then divide it by the
mean of the parameter of interest for the denominator file. \strong{A
caveat:} We're checking on how best to calculate the CV and confidence
intervals for the unpaired data, so please check our work before using
those! REMINDER: THIS FUNCTION IS UNDER CONSTRUCTION.}

\item{PKparameters}{PK parameters you want to extract from the simulator
  output file. Options are: \describe{

  \item{"all"}{all possible parameters}

  \item{"AUC tab"}{only those parameters on the "AUC" tab (default). The
  "AUC_CI" tab or "AUC_SD" tab will be used if "AUC" tab is not present.}

  \item{"Absorption tab"}{only those parameters on the "Absorption" tab.
  Please note that we haven't developed this function for output in the
  "Overall Fa Fg" tab for ADAM-model simulations yet.}

  \item{a vector of any combination of specific, individual parameters, each
  surrounded by quotes and encapsulated with \code{c(...)}}{An example:
  \code{c("Cmax_dose1", "AUCtau_last")}. To see the full set of possible
  parameters to extract, enter \code{view(PKParameterDefinitions)} into the
  console. Not case sensitive. If you use "_first" instead of "_dose1", that
  will also work.}

  \item{a vector of individual parameters with one parameter for the
  numerator and whatever parameter you want from the other file for the
  denominator, separated by "/"}{The previous options are all for when you
  want to take the ratio of the \emph{same} parameter for file 1 / file 2.
  However, if you want to compare one PK parameter from file 1 with a
  \emph{different} parameter for file 2, you can do that with this option.
  Here's an example of how to input the parameters so that you can calculate
  the dose 1 AUCinf with an inhibitor present for file 1 divided by the
  AUCinf for dose 1 with no inhibitor (baseline) for file 2:
  \code{PKparameters = c("AUCinf_dose1_withInhib / AUCinf_dose1")} Please
  note that the quotes are around \emph{both} of the two parameters!}}

  Currently, the PK data are only for the substrate unless noted, although
  you can sometimes hack around this by supplying a specific sheet to extract
  for a compound other than the substrate, e.g. sheet = "AUC(Sub Pri Met1)".
  This has NOT been as well tested, though, so be sure to check that you're
  getting what you expected!}

\item{sheet_PKparameters_num}{(optional) If you want the PK parameters for
the numerator to be pulled from a specific tab in
\code{sim_data_file_numerator}, list that tab here. Most of the time, this
should be left as NA.}

\item{sheet_PKparameters_denom}{(optional) If you want the PK parameters for
the numerator to be pulled from a specific tab in
\code{sim_data_file_denominator}, list that tab here. Most of the time,
this should be left as NA.}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are "plasma" (default) or "blood" (possible but not as thoroughly
tested).}

\item{mean_type}{What kind of means and confidence intervals do you want
listed in the output table? Options are "arithmetic" or "geometric"
(default).}

\item{includeCV}{TRUE (default) or FALSE for whether to include rows for CV
in the table}

\item{includeConfInt}{TRUE (default) or FALSE for whether to include whatever
confidence intervals were included in the simulator output file. Note that
the confidence intervals are geometric since that's what the simulator
outputs (see an AUC tab and the summary statistics; these values are the
ones for, e.g., "90\% confidence interval around the geometric mean(lower
limit)").}

\item{prettify_columns}{TRUE (default) or FALSE for whether to make easily
human-readable column names. TRUE makes pretty column names such as "AUCinf
(h*ng/mL)" whereas FALSE leaves the column with the R-friendly name from
\code{\link{extractPK}}, e.g., "AUCinf_dose1". \emph{Note:} This does not
currently do anything if the supplied PK parameters are ratios of one
parameter for the numerator file and a different parameter for the
denominator file; we haven't figured out how best to set that up yet.}

\item{prettify_compound_names}{TRUE (default) or FALSE on whether to make
compound names prettier in the prettified column titles and in any Word
output files. This was designed for simulations where the substrate and any
metabolites, effectors, or effector metabolites are among the standard
options for the simulator, and leaving \code{prettify_compound_names =
TRUE} will make the name of those compounds something more human readable.
For example, "SV-Rifampicin-MD" will become "rifampicin", and
"Sim-Midazolam" will become "midazolam". Set each compound to the name
you'd prefer to see in your column titles if you would like something
different. For example, \code{prettify_compound_names = c("inhibitor" =
"teeswiftavir", "substrate" = "superstatin")}. Please note that "inhibitor"
includes \emph{all} the effectors and effector metabolites present, so, if
you're setting the effector name, you really should use something like this
if you're including effector metabolites: \code{prettify_compound_names =
c("inhibitor" = "teeswiftavir and 1-OH-teeswiftavir", "substrate" =
"superstatin")}.}

\item{checkDataSource}{TRUE (default) or FALSE for whether to include in the
output a data.frame that lists exactly where the data were pulled from the
simulator output file. Useful for QCing.}

\item{save_table}{optionally save the output table and, if requested, the QC
info, by supplying a file name in quotes here, e.g., "My nicely formatted
table.docx" or "My table.csv", depending on whether you'd prefer to have
the main PK table saved as a Word or csv file. If you supply only the file
extension, e.g., \code{save_table = "docx"}, the name of the file will be
the file name plus "PK summary table" with that extension and output will
be located in the same folder as \code{sim_data_file}. If you supply
something other than just "docx" or just "csv" for the file name but you
leave off the file extension, we'll assume you want it to be ".csv". While
the main PK table data will be in whatever file format you requested, if
you set \code{checkDataSource = TRUE}, the QC data will be in a csv file on
its own and will have "- QC" added to the end of the file name.
\strong{WARNING:} SAVING TO WORD DOES NOT WORK ON SHAREPOINT. This is a
Microsoft permissions issue, not an R issue. If you try to save on
SharePoint, you will get a warning that R will save your file to your
Documents folder instead.}

\item{fontsize}{the numeric font size for Word output. Default is 11 point.
This only applies when you save the table as a Word file.}
}
\value{

}
\description{
\code{calc_PK_ratios} matches PK data from two simulator output Excel files
by subject ID and trial and calculates the ratio of those parameters for each
individual. It then calculates either the geometric or arithmetic mean and
desired confidence interval for those data.
}
\examples{
# No examples yet.

}
