% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PK_ratios.R
\name{calc_PK_ratios}
\alias{calc_PK_ratios}
\title{Calculate the ratio of PK parameters between two simulations}
\usage{
calc_PK_ratios(
  sim_data_file_numerator,
  sim_data_file_denominator,
  paired = TRUE,
  distribution_type = "t",
  compoundToExtract = "substrate",
  tissue = "plasma",
  PKparameters = "AUC tab",
  sheet_PKparameters_num = NA,
  sheet_PKparameters_denom = NA,
  existing_exp_details = NA,
  mean_type = "geometric",
  include_num_denom_columns = TRUE,
  conf_int = 0.9,
  includeCV = TRUE,
  includeConfInt = TRUE,
  include_dose_num = NA,
  prettify_columns = TRUE,
  prettify_compound_names = TRUE,
  rounding = NA,
  checkDataSource = TRUE,
  returnExpDetails = FALSE,
  save_table = NA,
  fontsize = 11
)
}
\arguments{
\item{sim_data_file_numerator}{a simulator output Excel file that will
provide the numerator for the calculated ratios.}

\item{sim_data_file_denominator}{a simulator output Excel file that will
provide the denominator for the calculated ratios}

\item{paired}{TRUE (default) or FALSE for whether the study design is paired,
as in, the subjects are \emph{identical} between the two simulations.
\strong{THIS IS AN IMPORTANT DISTINCTION AND WILL AFFECT HOW THE
CALCULATIONS ARE PERFORMED!} An example of a paired study would be a DDI
study where each subject has a measurement without the effector of interest
and then has a second measurement \emph{with} the effector. The comparison
is for repeated measurements of the \emph{same subject}. An example of an
unpaired study design would be comparing healthy volunteers to subjects
with hepatic impairment because those are measurements on \emph{different}
subjects. For paired study designs, the order of operations is to calculate
each subject's mean ratio and then to calculate the mean of those ratios.
For unpaired study designs, the order of operations is to calculate the
mean for the numerator simulation and then divide it by the mean for the
denominator simulation. Would this be clearer if you could see the
mathematical equations? We agree but can't easily include equations in the
help file. However, if you run this and save the output to a Word file, the
equations will be included in the output.}

\item{distribution_type}{use a "t" distribution (default) or a "Z"
distribution. Note: The Simcyp Simulator calculates geometric confidence
intervals with a t distribution.}

\item{compoundToExtract}{For which compound do you want to extract PK data?
Options are: \itemize{\item{"substrate" (default),} \item{"primary
metabolite 1",} \item{"primary metabolite 2",} \item{"secondary
metabolite",} \item{"inhibitor 1" -- this can be an inducer, inhibitor,
activator, or suppressor, but it's labeled as "Inhibitor 1" in the
simulator,} \item{"inhibitor 2" for the 2nd effector listed in the
simulation,} \item{"inhibitor 1 metabolite" for the primary metabolite of
inhibitor 1}}}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are "plasma" (default) or "blood" (possible but not as thoroughly
tested).}

\item{PKparameters}{PK parameters you want to extract from the simulator
  output file. Options are: \describe{

  \item{"all"}{all possible parameters}

  \item{"AUC tab"}{only those parameters on the "AUC" tab (default). The
  "AUC_CI" tab or "AUC_SD" tab will be used if "AUC" tab is not present.}

  \item{"Absorption tab"}{only those parameters on the "Absorption" tab.
  Please note that we haven't developed this function for output in the
  "Overall Fa Fg" tab for ADAM-model simulations yet.}

  \item{a vector of any combination of specific, individual parameters, each
  surrounded by quotes and encapsulated with \code{c(...)}}{An example:
  \code{c("Cmax_dose1", "AUCtau_last")}. To see the full set of possible
  parameters to extract, enter \code{view(PKParameterDefinitions)} into the
  console. Not case sensitive. If you use "_first" instead of "_dose1", that
  will also work.}

  \item{a vector of individual parameters with one parameter for the
  numerator and whatever parameter you want from the other file for the
  denominator, separated by "/"}{The previous options are all for when you
  want to take the ratio of the \emph{same} parameter for file 1 / file 2.
  However, if you want to compare one PK parameter from file 1 with a
  \emph{different} parameter for file 2, you can do that with this option.
  Here's an example of how to input the parameters so that you can calculate
  the dose 1 AUCinf with an inhibitor present for file 1 divided by the
  AUCinf for dose 1 with no inhibitor (baseline) for file 2:
  \code{PKparameters = c("AUCinf_dose1_withInhib / AUCinf_dose1")} Please
  note that the quotes are around \emph{both} of the two parameters!}}}

\item{sheet_PKparameters_num}{(optional) If you want the PK parameters for
the numerator to be pulled from a specific tab in
\code{sim_data_file_numerator}, list that tab here. Most of the time, this
should be left as NA.}

\item{sheet_PKparameters_denom}{(optional) If you want the PK parameters for
the numerator to be pulled from a specific tab in
\code{sim_data_file_denominator}, list that tab here. Most of the time,
this should be left as NA.}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} to get all the details from the "Input
Sheet" (e.g., when you ran extractExpDetails_mult you said
\code{exp_details = "Input Sheet"} or \code{exp_details = "all"}), you can
save some processing time by supplying that object here, unquoted. If left
as NA, this function will run \code{extractExpDetails} behind the scenes to
figure out some information about your experimental set up.}

\item{mean_type}{What kind of means and confidence intervals do you want
listed in the output table? Options are "arithmetic" or "geometric"
(default).}

\item{include_num_denom_columns}{TRUE (default) or FALSE for whether to
include columns in the output table for the numerator data alone and
columns for the denominator alone. For example, if you wanted to calculate
the dose 1 AUC ratio for cancer patients compared to healthy volunteers,
settting \code{include_num_denom_columns = TRUE} would give you that ratio
and also a column with summary statistics on the AUC for cancer patients
and a column with summary statistics on the AUC for healthy volunteers.
Setting it to FALSE would give you only the ratios.}

\item{conf_int}{confidence interval to use; default is 90\%}

\item{includeCV}{TRUE (default) or FALSE for whether to include rows for CV
in the table}

\item{includeConfInt}{TRUE (default) or FALSE for whether to include whatever
confidence intervals were included in the simulator output file. Note that
the confidence intervals are geometric since that's what the simulator
outputs (see an AUC tab and the summary statistics; these values are the
ones for, e.g., "90\% confidence interval around the geometric mean(lower
limit)").}

\item{include_dose_num}{NA (default), TRUE, or FALSE on whether to include
the dose number when listing the PK parameter. By default, the parameter
will be labeled, e.g., "Dose 1 Cmax ratio" or "Last dose AUCtau ratio", if
you have PK data for both the first dose and the last dose. Also by
default, if you have data only for the first dose or only for the last
dose, the dose number will be omitted and it will be labeled, e.g., "AUCtau
ratio" or "Cmax ratio". Set this to TRUE or FALSE as desired to override
the default behavior and get exactly what you want.}

\item{prettify_columns}{TRUE (default) or FALSE for whether to make easily
human-readable column names. TRUE makes pretty column names such as "AUCinf
(h*ng/mL)" whereas FALSE leaves the column with the R-friendly name from
\code{\link{extractPK}}, e.g., "AUCinf_dose1". We're still tweaking this to
make it look just right!}

\item{prettify_compound_names}{TRUE (default) or FALSE on whether to make
compound names prettier in the prettified column titles and in any Word
output files. This was designed for simulations where the substrate and any
metabolites, effectors, or effector metabolites are among the standard
options for the simulator, and leaving \code{prettify_compound_names =
TRUE} will make the name of those compounds something more human readable.
For example, "SV-Rifampicin-MD" will become "rifampicin", and
"Sim-Midazolam" will become "midazolam". Set each compound to the name
you'd prefer to see in your column titles if you would like something
different. For example, \code{prettify_compound_names = c("inhibitor" =
"teeswiftavir", "substrate" = "superstatin")}. Please note that "inhibitor"
includes \emph{all} the effectors and effector metabolites present, so, if
you're setting the effector name, you really should use something like this
if you're including effector metabolites: \code{prettify_compound_names =
c("inhibitor" = "teeswiftavir and 1-OH-teeswiftavir", "substrate" =
"superstatin")}.}

\item{rounding}{option for what rounding to perform, if any. Options are:
\describe{\item{NA or "Consultancy"}{All output will be rounded according
to Simcyp Consultancy Team standards: to three significant figures when the
value is < 100 or to the ones place if the value is >= 100. Please see the
function \code{\link{round_consultancy}}, which does the rounding here.}
\item{"none"}{No rounding will be performed.} \item{"significant X" where
"X" is a number}{Output will be rounded to X significant figures. "signif
X" also works fine.} \item{"round X" where "X" is a number}{Output will be
rounded to X digits} \item{"Word only"}{Output saved to Word or a csv file
will be rounded using the function \code{\link{round_consultancy}}, but
nothing will be rounded in the output R object. This can be useful when you
want to have nicely rounded and formatted output in a Word file but you
\emph{also} want to use the results from \code{calc_PK_ratios} to make
forest plots, which requires numbers that are \emph{not} rounded.}}}

\item{checkDataSource}{TRUE (default) or FALSE for whether to include in the
output a data.frame that lists exactly where the data were pulled from the
simulator output file. Useful for QCing.}

\item{returnExpDetails}{TRUE or FALSE (default) for whether to return the
simulator experimental details, which this function will look up anyway
behind the scenes. If TRUE, this will return a list, and each set of
simulation details will be an item in that list.}

\item{save_table}{optionally save the output table and, if requested, the QC
info, by supplying a file name in quotes here, e.g., "My nicely formatted
table.docx" or "My table.csv", depending on whether you'd prefer to have
the main PK table saved as a Word or csv file.  Do not include any slashes,
dollar signs, or periods in the file name. If you supply only the file
extension, e.g., \code{save_table = "docx"}, the name of the file will be
the file name plus "PK summary table" with that extension and output will
be located in the same folder as \code{sim_data_file}. If you supply
something other than just "docx" or just "csv" for the file name but you
leave off the file extension, we'll assume you want it to be ".csv". While
the main PK table data will be in whatever file format you requested, if
you set \code{checkDataSource = TRUE}, the QC data will be in a csv file on
its own and will have "- QC" added to the end of the file name.}

\item{fontsize}{the numeric font size for Word output. Default is 11 point.
This only applies when you save the table as a Word file.}
}
\value{
A list or a data.frame of PK data that optionally includes where the
  data came from
}
\description{
\code{calc_PK_ratios} matches PK data from a pair of simulator output Excel
files and calculates the mean and confidence intervals of the ratios of the
requested PK parameters. To do this for multiple pairs of simulator output
files, please see the function \code{\link{calc_PK_ratios_mult}}. For
detailed instructions and examples, please see the SharePoint file "Simcyp
PBPKConsult R Files - Simcyp PBPKConsult R Files/SimcypConsultancy function
examples and instructions/Calculating PK ratios from separate
simulations/Calculating-PK-ratios.docx". (Sorry, we are unable to include a
link to it here.)
}
\examples{
# No examples yet.

}
