% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcKi.R
\name{calcKi}
\alias{calcKi}
\title{calcKi}
\usage{
calcKi(IC50, S, Km = NA)
}
\arguments{
\item{IC50}{The IC50 value for the inhibitor (uM)}

\item{S}{The concentration of the substrate used (uM)}

\item{Km}{The Michaelis constant for the affinity between the substrate and
the enzyme (uM)}
}
\value{
Returns the inhibitory constant Ki (uM) (numeric)
}
\description{
Function uses the Cheng-Prusoff equation to calculate a Ki value.
}
\examples{
calcKi(20, 5, 5)
# 10

}
