% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_wksz_interactions.R
\name{change_wksz_interactions}
\alias{change_wksz_interactions}
\title{Change interaction parameters in Simcyp Simulator workspace files}
\usage{
change_wksz_interactions(
  sim_workspace_files = NA,
  new_sim_workspace_files = NA,
  fix_xml_paths = TRUE,
  interactions_to_set = NA,
  compoundID = "inhibitor 1",
  enzymes,
  tissues = NA,
  competitive_inhibition_switch = "no change",
  Ki = "no change",
  Ki_fumic = "no change",
  TDI_switch = "no change",
  Kapp = "no change",
  kinact = "no change",
  Kapp_fumic = "no change",
  induction_IndMax_switch = "no change",
  IndMax = "no change",
  induction_IndC50_switch = "no change",
  IndC50 = "no change",
  IndC50_fuinc = "no change",
  Ind_gamma = "no change",
  Ind_slope = "no change"
)
}
\arguments{
\item{sim_workspace_files}{the set of workspace files to modify; must end in
".wksz" if you're specifying individual files. Leave as NA to change all
the workspaces in the current folder or set to "recursive" to change all
the workspaces in the current folder and \emph{all} the subfolders below
it. BE CAREFUL. This function changes workspaces, so please be certain
you're making the changes you want. We recommend keeping a backup copy of
the original workspaces until you're sure the new ones are set up how you
want.}

\item{new_sim_workspace_files}{optionally specify the new workspace file
names to use. If left as NA, the original workspace will be overwritten.
Otherwise, specify a character vector of file names to use, e.g.,
\code{new_sim_workspace_files = c("new file 1.wksz", "new file
2.wksz")}}

\item{fix_xml_paths}{TRUE (default) or FALSE to automatically fix any
discrepancies in user name on the SharePoint folder path for observed
overlay data files or fixed trial design files. This can be useful if, say,
someone else ran the development and verification simulations and you're
now running the application simulations.}

\item{interactions_to_set}{a data.frame of interaction parameters to use
instead of filling in each of the arguments below (compoundID through
Ind_slope) with character vectors of values. The only columns that will
this function will pay any attention to are "sim_workspace_files",
"new_sim_workspace_files", or any of "compoundID", "enzymes",
"competitive_inhibition_switch", "Ki", "Ki_fumic", "TDI_switch", "Kapp",
"kinact", "Kapp_fumic", "induction_IndC50_switch", "IndMax", "IndC50",
"IndC50_fuinc", "induction_IndMax_switch", "Ind_slope", and "Ind_gamma".
Please see the notes on each of the arguments below to see what values to
use for this data.frame. This can be an easier way to set things up when
you have a lot of information to keep track of or when you're setting
parameters for multiple enzymes or when you want some parameters for some
workspaces and other parameters for other workspaces. If you leave off the
column "sim_workspace_files", we'll assume you want the same parameters for
all the workspaces. If you leave off the column "new_sim_workspace_files",
we'll assume you want to overwrite the existing workspaces.}

\item{compoundID}{compoundID whose parameters you want to change. Options
are: \itemize{\item{"substrate",} \item{"primary metabolite 1",}
\item{"primary metabolite 2",} \item{"secondary metabolite",}
\item{"inhibitor 1" (default) -- this can be an inducer, inhibitor,
activator, or suppresesor, but it's labeled as "Inhibitor 1" in the
simulator,} \item{"inhibitor 2" for the 2nd perpetrator listed in the
simulation,} \item{"inhibitor 1 metabolite" for the primary metabolite of
inhibitor 1}}}

\item{enzymes}{enzymes whose interaction parameters you want to change, e.g.,
\code{enzymes = c("CYP3A4", "CYP3A5")}. Note: This is currently only set up
to change CYP, some UGT, and P-gp interaction parameters. If you need to
change interaction parameters for other enzymes or transporters, please
talk to Laura Shireman.}

\item{tissues}{tissues where the enzyme of interest may be found. Leave as NA
if not applicable (CYPs, for exaple) or set to the tissue as listed in the
Simulator, e.g., "Intestine" or "Liver" for P-gp.}

\item{competitive_inhibition_switch}{turn competitive inhibition "on" or
"off" or set to "no change" to leave the original value}

\item{Ki}{Ki value to use for competitive inhibition (uM) or set to "no
change" to leave the original value. If set to NA, value will be set to
1E+06.}

\item{Ki_fumic}{fu,mic value to use for competitive inhibition or set to "no
change" to leave the original value. If set to NA, value will be set to 1.}

\item{TDI_switch}{turn time-dependent inactivation "on" or "off" or set to
"no change" (default) to leave the original value}

\item{Kapp}{Kapp value to use for TDI (uM) or set to "no change" (default) to
leave the original value. If set to NA, value will be set to 1E+06.}

\item{kinact}{kinact value to use for TDI (1/h) or set to "no change"
(default) to leave the original value. If set to NA, value will be set to
0.}

\item{Kapp_fumic}{fu,mic value to use for TDI or set to "no change" (default)
to leave as the original value. If set to NA, value will be set to 1.}

\item{induction_IndMax_switch}{set to "on" when there is either a) no
induction (IndMax is 1) or b) when there \emph{is} induction and you're
using any induction model except the slope model. Set to "off" when there
is induction but you're using the "slope" model to describe it. Set to "no
change" (default) to leave as the original value.}

\item{IndMax}{a number for the maximum fold induction to use or set to "no
change" (default) to leave the original value.}

\item{induction_IndC50_switch}{turn "on" or "off" the use of an induction
IndC50 value or set to "no change" (default) to leave the original value}

\item{IndC50}{IndC50 value to use (uM) or set to "no change" (default) to
leave the original value. If set to NA, value will be set to 1E+06.}

\item{IndC50_fuinc}{fu,mic value to use for the IndC50 incubation or set to
"no change" (default) to leave the original value. If set to NA, value will
be set to 1.}

\item{Ind_gamma}{value to use for gamma aka the Hill coefficient for
induction or set to "no change" (default) to leave the original value. If
set to NA, value will be set to 1.}

\item{Ind_slope}{value to use for slope in induction or set to "no change"
(default) to leave the original value. If left as NA, value will be set to
0.}
}
\value{
does not return anything in R but saves workspace files
}
\description{
\code{change_wksz_interactions} changes interaction parameters such as Kapp
or Ki or IndMax in workspace files (.wksz files) for then running with the
Simcyp Simulator. This will change \emph{all} the workspace files provided to
have the \emph{same} interaction parameters you list here. Currently only set
up to change CYP, P-gp, and a few UGT parameters, but we'll add more options
soon. \strong{Pay attention to the "switch" arguments, which are set up just
like in the Simulator but are easy to overlook.} Also, this will
automatically fix any XML file paths such as observed data overlay files or
fixed trial design XML files to match the current user's path on SharePoint.
If you don't want that, set \code{fix_xml_paths} to FALSE. UNDER
CONSTRUCTION.
}
\examples{
# None yet

}
