% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_wksz_trial_design.R
\name{change_wksz_trial_design}
\alias{change_wksz_trial_design}
\title{Change a limited set of trial-design parameters in Simcyp Simulator workspace
files. UNDER CONSTRUCTION. USE WITH CAUTION.}
\usage{
change_wksz_trial_design(
  sim_workspace_files = NA,
  new_sim_workspace_files = NA,
  trial_design_parameters_to_set = NA,
  NumTrials = "no change",
  NumSubjTrial = "no change",
  PercFemale = "no change",
  AgeMin = "no change",
  AgeMax = "no change",
  SimDuration = "no change",
  activate_inhibitor1 = "no change",
  Dose_sub = "no change",
  Dose_inhib = "no change",
  DoseInt_sub = "no change",
  DoseInt_inhib = "no change",
  StartHr_sub = "no change",
  StartHr_inhib = "no change",
  NumTimePts = "no change",
  fix_xml_paths = TRUE
)
}
\arguments{
\item{sim_workspace_files}{the set of workspace files to modify; must end in
".wksz" if you're specifying individual files. Leave as NA to change all
the workspaces in the current folder or set to "recursive" to change all
the workspaces in the current folder and \emph{all} the subfolders below
it. BE CAREFUL. This function changes workspaces, so please be certain
you're making the changes you want. We recommend keeping a backup copy of
the original workspaces until you're sure the new ones are set up how you
want.}

\item{new_sim_workspace_files}{optionally specify the new workspace file
names to use. If left as NA, the original workspace will be overwritten.
Otherwise, specify a character vector of file names to use, e.g.,
\code{new_sim_workspace_files = c("new file 1.wksz", "new file
2.wksz")}}

\item{trial_design_parameters_to_set}{optionally specify a data.frame of
trial design parameters to set. Acceptable column names: NumTrials,
NumSubjTrial, PercFemale, AgeMin, AgeMax, SimDuration, activate_inhibitor1,
Dose_sub, Dose_inhib, DoseInt_sub, DoseInt_inhib, StartHr_sub, NumTimePts.}

\item{NumTrials}{number of trials to use}

\item{NumSubjTrial}{number of subjects per trial to use}

\item{PercFemale}{percent of subjects who are female}

\item{AgeMin}{minimum age}

\item{AgeMax}{maximum age}

\item{SimDuration}{study duration in hours}

\item{activate_inhibitor1}{TRUE or FALSE for whether to turn on inhibitor 1.
(At this point, this function will only se parametrs for substrate or
inhibitor 1.)}

\item{Dose_sub}{dose of substrate to use (mg)}

\item{Dose_inhib}{dose of inhibitor 1 to use (mg)}

\item{DoseInt_sub}{dose interval for the substrate in hours. If you would
like to have a single-dose regimen, set this to "single dose". Does not
allow for custom-dosing regimens.}

\item{DoseInt_inhib}{dose interval for inhibitor 1 in hours. If you would
like to have a single-dose regimen, set this to "single dose". Does not
allow for custom-dosing regimens.}

\item{StartHr_sub}{start time for administering substrate in hours -- This
may be tricky to do! May need to also adjust StartDayTime_sub!}

\item{StartHr_inhib}{start time for administering inhibitor 1 in hours}

\item{NumTimePts}{number of time points to use ("Number of Time Samples" in
the Simulator)}

\item{fix_xml_paths}{TRUE (default) or FALSE to automatically fix any
discrepancies in user name on the SharePoint folder path for observed
overlay data files or fixed trial design files. This can be useful if, say,
someone else ran the development and verification simulations and you're
now running the application simulations.}
}
\value{
does not return anything in R but saves workspace files
}
\description{
\code{change_wksz_trial_design} changes a few trial-design parameters such as
the number of subjects, the percent female, the dose or dose interval, etc.
It does \emph{not} change populations because there are simply too many
parameters to change in an automatic faction like this. (At least, not at
this point.) It also doesn't turn on or off a fixed trial design. This will
change \emph{all} the workspace files provided to have the \emph{same}
trial-design parameters you list here. UNDER CONSTRUCTION.
}
\examples{
# None yet

}
