% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_accumulation.R
\name{check_accumulation}
\alias{check_accumulation}
\title{Create a graph of simulated concentrations to check for accumulation}
\usage{
check_accumulation(
  ct_dataframe,
  accum_compoundID = "inhibitor 1",
  overlay_compoundID = "substrate",
  t0 = 0,
  conc_point = "Clast",
  mark_dosing = "none",
  diff_cutoff = 0.05,
  mean_type = "arithmetic",
  x_axis_interval = NA
)
}
\arguments{
\item{ct_dataframe}{the input concentration-time data generated by running
the function \code{\link{extractConcTime}} or, if you'd also like to see an
overlay of the substrate with or without any inhibitors, by running
\code{\link{extractConcTime_mult}}. This function assumes there's data for
only one tissue and that it's appropriate to compare all the data included,
so you'll get graph artifacts if those are not the case.}

\item{accum_compoundID}{the compound ID to monitor for accumulation. Defaults
to "inhibitor 1". The time point requested will be shown as points and will
be colored by percent difference from the previous point.}

\item{overlay_compoundID}{the compound ID to overlay the complete
concentration-time data for. Defaults to "substrate"; use "none" for no
overlaid plot. This was designed for the following scenario: Monitor
Ctrough for an inhibitor to make sure that it's at steady state, plot that
as points, and then plot the concentration-time profile of a substrate over
that to make sure that the inhibitor is present the whole time the
substrate is being eliminated.}

\item{t0}{start time for compound being plotted}

\item{conc_point}{Concentration point to plot. Options are: \describe{

  \item{"C0"}{concentration for the first time point available for that dose}

  \item{"Clast"}{(default) concentration for the last time point available
  for that dose}

  \item{"Cmin"}{the minimum concentration for that dosing interval, which may
  not be the last time point; this is somewhat prone to sampling artifacts
  and we generally recommend using "Clast" instead}

  \item{"Cmax"}{the maximum concentration for that dose number}}}

\item{mark_dosing}{optionally mark dosing intervals on the graph as "none"
(default) to have no marks for the dosing intervals or a combination of a
color in R and a named linetype, e.g., "red dotted" or "blue dashed" or
even "#FFBE33 longdash".}

\item{diff_cutoff}{what percent difference cutoff would you like to use to
color the points? The default is for points with less than a 5\% difference
from the previous point to be blue and points with a larger percent
difference to be red.}

\item{mean_type}{the mean type to use since this function only displays
summary data; defaults to the arithmetic mean}
}
\value{

}
\description{
Create a graph with time on the x axis and then some useful concentration
point -- C0, Cmax, Cmin (see notes on useage), or Clast for each dose -- on
the y axis. Optionally overlay this graph with the concentration-time data
for that same compound or a different one. For example, you could create a
graph with points for Clast for an inhibitor and then lines showing the
substrate concentration-time data overlaid on top of that. This way, you can
check whether the substrate is dosed after the inhibitor reaches steady
state, and you could also check whether the inhibitor was present the entire
time the substrate was getting eliminated.
}
\examples{

check_accumulation(ct_dataframe = MDZ_Keto)


}
