% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_accumulation.R
\name{check_accumulation}
\alias{check_accumulation}
\title{Graph for checking accumulation over multiple doses -- UNDER CONSTRUCTION}
\usage{
check_accumulation(
  ct_dataframe,
  accum_compoundID = "inhibitor 1",
  overlay_compoundID = "substrate",
  t0 = 0,
  timepoint = "tlast",
  mark_dosing = "none",
  diff_cutoff = 0.05,
  mean_type = "arithmetic",
  x_axis_interval = NA
)
}
\arguments{
\item{ct_dataframe}{the input concentration-time data generated by running
the function \code{\link{extractConcTime}} or, if you'd also like to see an
overlay of the substrate with or without any inhibitors, by running
\code{\link{extractConcTime_mult}}. This function assumes only one tissue
is included, so you'll get graph artifacts if that's not the case.}

\item{accum_compoundID}{the compound ID to monitor for accumulation. Defaults
to "inhibitor 1". The time point requested will be shown as points and will
be colored by percent difference from the previous point.}

\item{overlay_compoundID}{the compound ID to overlay the complete
concentration-time data for. Defaults to "substrate"; use "none" for no
overlaid plot. This was designed for the following scenario: Monitor
Ctrough for an inhibitor to make sure that it's at steady state, plot that
as points, and then plot the concentration-time profile of a substrate over
that to make sure that the inhibitor is present the whole time you're
monitoring the substrate.}

\item{t0}{start time for compound being plotted}

\item{timepoint}{Time point to plot. Options are:
\describe{\item{"tmin"}{whenever the minimum concentration occurred for
that dosing interval, which may not be the last time point}
\item{"tmax"}{whenever the maximum observed concentration for that dose
number occurred} \item{"t0"}{the first time point available for that dose}
\item{"tlast"}{the last time point available for that dose}}}

\item{mark_dosing}{set how to mark dosing intervals on the graph as "none"
(default) to have no marks for the dosing intervals or a combination of a
named color in R and a named linetype, e.g., "red dotted" or "blue dashed".}

\item{diff_cutoff}{what percent difference cutoff would you like to color the
points by? The default is for points with less than a 5\% difference from
the previous point to be blue and points with a larger percent difference
to be red.}

\item{mean_type}{the mean type to use since this function only displays
summary data; defaults to the arithmetic mean}
}
\value{

}
\description{
UNDER CONSTRUCTION -- DoseNum must be complete. Doesn't currently detect conc
units.
}
\examples{

check_accumulation(ct_dataframe = CT)


}
