% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_doseint.R
\name{check_doseint}
\alias{check_doseint}
\title{INTERNAL check whether AUC interval ends on a dosing interval}
\usage{
check_doseint(
  sim_data_file,
  existing_exp_details,
  compoundID,
  stop_or_warn = "stop"
)
}
\arguments{
\item{sim_data_file}{sim_data_file}

\item{existing_exp_details}{existing_exp_details}

\item{compoundID}{compoundID}
}
\value{
character vector of length 1: "custom dosing" (as in, we can't tell
  whether the interval matched), "good", "mismatch", or "can't check -
  missing file"
}
\description{
INTERNAL check whether AUC interval ends on a dosing interval
}
\examples{
# none yet
}
