% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_doses_from_file.R
\name{create_doses_from_file}
\alias{create_doses_from_file}
\title{Create a csv file of dosing rows for XML files using a study file to get
subject information}
\usage{
create_doses_from_file(
  dose_interval = 24,
  num_doses = NA,
  end_time = NA,
  custom_dosing_schedule = NA,
  study_file = NA,
  subj_ID = NA,
  compound = "Substrate",
  admin_route = "Oral",
  dose_unit = NA,
  dose_amount = 100,
  subj_age = NA,
  subj_weight = NA,
  subj_height = NA,
  subj_sex = NA,
  subjIDs = NA,
  save_output = NA
)
}
\arguments{
\item{dose_interval}{the dosing interval in hours. Default is 24 for a QD
dosing regimen.}

\item{num_doses}{the number of doses to generate. If this is left as NA, then
the value for \code{end_time} will be used to determine the number of doses
administered.}

\item{end_time}{the end time of the dosing in hours. If \code{num_doses} is
filled out, that value will be used preferentially.}

\item{custom_dosing_schedule}{a custom dosing schedule to be used for each
subject in hours, e.g., \code{custom_dosing_schedule = c(0, 12, 24, 168,
180, 192)}; if this is filled out, values in \code{dose_interval},
\code{num_doses}, and \code{end_time} will all be ignored.}

\item{study_file}{optionally specify a file containing study information.
  This will be used for determining: \itemize{

  \item{subject IDs,}

  \item{compound IDs,}

  \item{routes of administration,}

  \item{dose units,}

  \item{dose amounts,}

  \item{ages,}

  \item{weights,}

  \item{heights,}

  \item{and sexes}} for the subjects who will be dosed. If a study file is
  specified, the values for this information will \emph{only} be taken from
  the file, so you \emph{must} use unquoted column titles for the arguments
  for \code{subj_ID}, \code{compound}, \code{admin_route}, \code{dose_unit},
  \code{dose_amount}, \code{subj_age}, \code{subj_weight},
  \code{subj_height}, and \code{subj_sex}.}

\item{subj_ID}{optionally specify subject IDs as a) an unquoted column name
in \code{study_file} or b) a vector, e.g., \code{subj_ID = c("101-001",
"101-002", "101-003")}}

\item{compound}{optionally specify the compound that's being dosed as a) an
unquoted column name in \code{study_file} or b) a vector. Options are
"Substrate" (default if unspecified), "Inhibitor 1", "Inhibitor 2", or
"Inhibitor 3". Not case sensitive.}

\item{admin_route}{optionally specify the route of administration as a) an
unquoted column name in \code{study_file} or b) a vector. Options are
"Oral" (default if unspecified), "Intravenous", "Dermal", "Inhaled",
"SC-First Order", "SC-Mechanistic", or "Auto-detect". Not case sensitive.}

\item{dose_unit}{the unit of dosing. Options are "mg" (default), "mg/m2", or
"mg/kg".}

\item{dose_amount}{the amount of the dose (units are not included). If this
amount varies, please use the \code{custom_dosing_schedule} option and
include one dose amount for each time. For example:
\code{custom_dosing_schedule = c(0, 24, 48), dose_amount = c(100, 50, 50)}
will generate doses of 100 mg at t = 0 and then 50 mg at t = 24 and 48
hours.}

\item{subj_age}{age (years); either list them manually, enclosed in
\code{c(...)} if there's more than one, or specify the unquoted name of the
column in \code{study_file} to use. Example of input if you specify this
manually: \code{subj_age = c(25, 67, 45)}. Example of input if you are
instead specifying which column in study_file to use: \code{subj_age = Age}}

\item{subj_weight}{weight (kg); either specify -- unquoted -- which column in
\code{study_file} to use or list specific numeric weights, enclosing
multiple weights with \code{c(...)}. Example of input if you specify this
manually: \code{subj_weight = c(72, 87, 91)}. Example of input if you are
instead specifying which column in study_file to use: \code{subj_weight =
SubjWt}}

\item{subj_height}{height (cm)}

\item{subj_sex}{sex; options are "F" or "M"}

\item{save_output}{the file name to use for saving the output as a csv; if
left as NA, this will generate a data.frame in R but no output will be
saved.}

\item{subjID_column}{which column in \code{study_file} contains the subject
IDs; not quoted. This is the only column that matters in \code{study_file}.}
}
\value{

}
\description{
\code{create_doses_from_file} uses an Excel or csv file with subject
information to generate a data.frame of dosing times and amounts -- one for
every subject if subject IDs are provided -- based on the dosing regimen
used. This is meant for generating XML files for use in Phoenix WNL.
}
\examples{

# QD dosing regimen of 100 mg
create_dose_rows(dose_interval = 24, num_doses = 4)

# QD dosing regimen of 100 mg for subjects A, B, and C
create_dose_rows(dose_interval = 24, num_doses = 4,
                 subjIDs = c("A", "B", "C"))

# QD dosing regimen of 100 mg for subjects A, B, and C and save output
create_dose_rows(dose_interval = 24, num_doses = 4,
                 subjIDs = c("A", "B", "C"),
                 save_output = "My doses.csv")

# Custom dosing regimen for subjects A, B, and C
create_dose_rows(custom_dosing_schedule = c(0, 12, 24, 48, 92, 168),
                 subjIDs = c("A", "B", "C"))

                  
}
