% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot.R
\name{ct_plot}
\alias{ct_plot}
\title{ct_plot}
\usage{
ct_plot(
  sim_data_file,
  obs_data_file = NA,
  figure_type = "method development",
  save = FALSE,
  figname = NA,
  return_data = FALSE
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulated
concentration-time data}

\item{obs_data_file}{name of the Excel file containing the observed
concentration-time data. If the observed data you want to plot were already
included in the Excel output from the simulator, leave this as NA.
Otherwise, this is the file that it is ready to be converted to an XML
file, not the file that contains only the digitized time and concentration
data.}

\item{figure_type}{type of figure to plot. Options are: \describe{
\item{method development}{plots a black line for the mean data, gray lines
for the individual simulated data, and open circles for the observed data}
\item{method verification}{plots a black line for the mean data, gray lines
for the 5th and 95th percentiles of the simulated data, and open circles
for the observed data} }}

\item{save}{TRUE or FALSE: Save the graph?}

\item{figname}{If \code{save} is TRUE, the file name for the output figure,
e.g., "Study 001 SD.png"}

\item{return_data}{TRUE or FALSE: Return the data used in the graphs? If
TRUE, this will return a named list of: \describe{ \item{Graphs}{the set of
graphs} \item{Data}{a data.frame of the concentration-time data used in the
set of graphs} }}
}
\value{

}
\description{
This function takes observed and simulated concentration-time data and makes
publication-quality graphs.
}
\examples{
sim_data_file <- "../Example simulator output.xlsx"

obs_data_file <- "../fig1-242-06-001-MD - for XML conversion.xlsx"

ct_plot(sim_data_file)
ct_plot(sim_data_file, obs_data_file, figure_type = "method verification")
ct_plot(sim_data_file, return_data = TRUE)

}
