% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot.R
\name{ct_plot}
\alias{ct_plot}
\title{Using observed and simulated concentration-time data, make
publication-quality graphs.}
\usage{
ct_plot(
  sim_data_file,
  obs_data_file = NA,
  obs_effector_data_file = NA,
  sim_obs_dataframe = NA,
  tissue = "plasma",
  figure_type = "trial means",
  substrate_or_effector = "substrate",
  adjust_obs_time = TRUE,
  time_range = NA,
  mean_type = "arithmetic",
  return_data = FALSE,
  return_indiv_graphs = FALSE
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulated
concentration-time data}

\item{obs_data_file}{name of the Excel file containing the observed
concentration-time data. If the observed data you want to plot were already
included in the Excel output from the simulator, leave this as NA.
Otherwise, this is the file that it is ready to be converted to an XML
file, not the file that contains only the digitized time and concentration
data.}

\item{obs_effector_data_file}{name of the Excel file containing the observed
concentration-time data in the presence of an effector. This is the file
that is ready to be converted to an XML file.}

\item{sim_obs_dataframe}{If you have already extracted the concentration-time
data using the function \code{extractConcTime}, you can enter the name of
the output data.frame from that function instead of re-reading the Excel
file.}

\item{tissue}{the tissue to plot. Default is plasma for typical plasma
concentration-time data. Other tissues are acceptable, e.g., "lung",
"brain", etc., as long as the tissue is one of the options included in
"Sheet Options", "Tissues" in the simulator.}

\item{figure_type}{type of figure to plot. Options are:

  \describe{

  \item{trial means}{plots a black line for the mean data, gray lines for the
  mean of each trial of simulated data, and open circles for the observed
  data. If an effector was present, gray dashed lines indicate the mean of
  each trial of simulated data in the presence of the effector. (At present,
  if an effector were present, this does NOT graph observed data because I
  haven't yet figured out the best way to indicate whether those data are
  with or without the effector. -LS)}

  \item{trial percentiles}{plots a black line for the mean data, gray lines
  for the 5th and 95th percentiles of the simulated data, and open circles
  for the observed data}

  \item{means only}{plots a black line for the mean data and, if an effector
  was modeled, a dashed line for the concentration-time data with the the
  effector. (At present, if an effector were present, this does NOT graph
  observed data because I haven't yet figured out the best way to indicate
  whether those data are with or without the effector. -LS)}

  }}

\item{substrate_or_effector}{"substrate" or "effector" for whether to plot
the substrate (default) or the effector concentration-time data}

\item{adjust_obs_time}{TRUE or FALSE: Adjust the time listed in the observed
data file to match the last dose administered? This only applies to
multiple-dosing regimens. If TRUE, the graph will show the observed data
overlaid with the simulated data such that the dose in the observed data
was administered at the same time as the last dose in the simulated data.
If FALSE, the observed data will start at whatever times are listed in the
Excel file.}

\item{time_range}{time range to graph. Options: \describe{

  \item{NA}{entire time range of data}

  \item{a start time and end time in hours}{only data in that time range,
  e.g. \code{c(24, 48)}}

  \item{"first dose"}{only the time range of the first dose}

  \item{"last dose"}{only the time range of the last dose}

  \item{"penultimate dose"}{only the time range of the 2nd-to-last dose, which
  can be useful for BID data where the end of the simulation extended past
  the dosing interval or data when the substrate was dosed BID and the
  effector was dosed QD} }}

\item{mean_type}{"geometric" or "arithmetic" for which type of mean to
calculate}

\item{return_data}{TRUE or FALSE: Return the data used in the graphs? If
TRUE, this will return a named list of: \describe{ \item{Graphs}{the set of
graphs} \item{Data}{a data.frame of the concentration-time data used in the
set of graphs} }}

\item{return_indiv_graphs}{TRUE or FALSE: Return each of the two individual
graphs? This can be useful if you want to modify the graphs further or only
use one, etc.}
}
\value{
Depending on the options selected, returns either a set of graphs or
  a list of the set of graphs (named "Graphs" in the  output), the individual
  graphs ("Linear graph" and "Semi-log graph"), and/or the data used for
  creating the graphs ("Data").
}
\description{
Using observed and simulated concentration-time data, make
publication-quality graphs.
}
\examples{
sim_data_file <- "../Example simulator output.xlsx"
obs_data_file <- "../fig1-242-06-001-MD - for XML conversion.xlsx"

ct_plot(sim_data_file)
ct_plot(sim_data_file, figure_type = "trial percentiles")
ct_plot(sim_data_file, return_data = TRUE)
ct_plot(sim_data_file, return_indiv_graphs = TRUE)

# Here's where adjusting or not adjusting the observed data time comes
# into play:
ct_plot(sim_data_file = "../Example simulator output MD.xlsx",
        obs_data_file = "../fig1-242-06-001-MD - for XML conversion.xlsx")
ct_plot(sim_data_file = "../Example simulator output MD.xlsx",
        obs_data_file = "../fig1-242-06-001-SD - for XML conversion.xlsx",
        adjust_obs_time = FALSE)

# Perhaps you don't want to show *all* the data but instead want to
# limit the time interval that is graphed. Use \code{time_range} here:
ct_plot(sim_data_file = "../Example simulator output MD.xlsx",
        obs_data_file = "../fig1-242-06-001-SD - for XML conversion.xlsx",
        adjust_obs_time = FALSE, time_range = c(0, 24))

# Or you can let it automatically calculate the time frame for the first or
# last dose simulated
ct_plot(sim_data_file = "../Example simulator output MD.xlsx",
        obs_data_file = "../fig1-242-06-001-SD - for XML conversion.xlsx",
        time_range = "first dose")
ct_plot(sim_data_file = "../Example simulator output MD.xlsx",
        obs_data_file = "../fig1-242-06-001-MD - for XML conversion.xlsx",
        time_range = "last dose")

# These may be too busy when you've got an effector present:
ct_plot(sim_data_file = "../Example simulator output MD + inhibitor.xlsx")
ct_plot(sim_data_file = "../Example simulator output MD + inhibitor.xlsx",
        figure_type = "trial percentiles")
# so you may want to consider only plotting means as an alternative:
ct_plot(sim_data_file = "../Example simulator output MD + inhibitor.xlsx",
        figure_type = "means only")

# You can also add a separate observed concentration-time file for the
# compound with the effector present. The graph will show those points
# as open triangles.
ct_plot(sim_data_file = "../Example simulator output MD.xlsx",
        obs_data_file = "../fig1-242-06-001-SD - for XML conversion.xlsx",
        obs_effector_data_file = "../Mallikaarjun_2016_RTV-fig1-100mg-BID-DLM+Kaletra - for XML conversion.xlsx",
        time_range = "last dose")

# If you've already got a data.frame formatted like the output
# from extractConcTime...
data(ConcTime)
ct_plot(sim_obs_dataframe = ConcTime)

}
