% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot.R
\name{ct_plot}
\alias{ct_plot}
\title{Concentration-time plots to match Consultancy template}
\usage{
ct_plot(
  ct_dataframe = NA,
  figure_type = "percentiles",
  subsection_ADAM = "free compound in lumen",
  mean_type = "arithmetic",
  time_range = NA,
  x_axis_interval = NA,
  pad_x_axis = TRUE,
  pad_y_axis = TRUE,
  adjust_obs_time = FALSE,
  t0 = "simulation start",
  y_axis_limits_lin = NA,
  y_axis_limits_log = NA,
  obs_color = NA,
  obs_shape = NA,
  showBLQ = FALSE,
  line_type = NA,
  line_transparency = NA,
  line_color = NA,
  line_width = NA,
  legend_position = "none",
  legend_label = NA,
  prettify_compound_names = TRUE,
  linear_or_log = "both vertical",
  graph_labels = TRUE,
  save_graph = NA,
  fig_height = 6,
  fig_width = 5
)
}
\arguments{
\item{ct_dataframe}{the input concentration-time data generated by running
the function \code{\link{extractConcTime}}. Not quoted.}

\item{figure_type}{type of figure to plot. Options are:

  \describe{

  \item{"trial means"}{plots an opaque line for the mean data, lighter lines
  for the mean of each trial of simulated data, and open circles for the
  observed data. If an effector were present, lighter dashed lines indicate
  the mean of each trial of simulated data in the presence of the effector.}

  \item{"percentiles"}{(default) plots an opaque line for the mean data,
  lighter lines for the 5th and 95th percentiles of the simulated data, and
  open circles for the observed data. If an effecter were present, the
  default is dashed lines for the data in the presence of an effector.}

  \item{"percentile ribbon"}{plots an opaque line for the mean data,
  transparent shading for the 5th to 95th percentiles of the simulated data,
  and open circles for the observed data. If an effector were present, the
  default is to show the data without the effector in blue and the data in
  the presence of the effector in red. Note: You may sometimes see some
  artifacts -- especially for semi-log plots -- where the ribbon gets partly
  cut off. For arcane reasons we don't want to bore you with here, we can't
  easily prevent this. However, a possible fix is to set your y axis limits
  for the semi-log plot to be wider using \code{y_axis_limits_log}.}

  \item{"means only"}{plots a black line for the mean data and, if an
  effector was modeled, a dashed line for the concentration-time data with
  Inhibitor 1.}

  \item{"Freddy"}{Freddy's favorite style of plot with trial means in light
  gray, the overall mean in thicker black, the 5th and 95th percentiles in
  dashed lines, and the observed data in semi-transparent purple-blue. Graphs
  with an effector present lose the trial means, and the percentiles switch
  to solid, gray lines. \strong{An editorial comment:} While this does not
  align with the officially sanctioned template at this time, this looks
  \emph{sharp}, makes it easy to see the defining characteristics of the
  data, and I recommend checking it out, even just for your own purposes of
  examining your data. If the color is too much for you but you like the
  rest, try setting \code{obs_color = "none"}. -LS}}}

\item{subsection_ADAM}{ADAM model only: For ADAM model output, multiple types
of data are available. Which of those concentration types do you want to
show? Options are whatever were available for that tissue, which could be
"solid compound", "free compound in lumen", "Heff", "absorption rate",
"unreleased substrate in faeces", "unreleased inhibitor in faeces",
"dissolved compound", or "luminal CLint of compound". Default is "free
compound in lumen" but is ignored when ct_dataframe isn't for ADAM data.}

\item{mean_type}{graph "arithmetic" (default) or "geometric" means or
"median" for median concentrations}

\item{time_range}{time range to show relative to the start of the simulation.
  Options: \describe{

  \item{NA}{(default) entire time range of data}

  \item{a start time and end time in hours}{only data in that time range,
  e.g. \code{c(24, 48)}. Note that there are no quotes around numeric data.}

  \item{"first dose"}{only the time range of the first dose}

  \item{"last dose"}{only the time range of the last dose}

  \item{"penultimate dose"}{only the time range of the 2nd-to-last dose,
  which can be useful for BID data where the end of the simulation extended
  past the dosing interval or data when the substrate was dosed BID and the
  effector was dosed QD}

  \item{a specific dose number with "dose" or "doses" as the prefix}{the time
  range encompassing the requested doses, e.g., \code{time_range = "dose 3"}
  for the 3rd dose or \code{time_range = "doses 1 to 4"} for doses 1 to 4}

  \item{"all obs" or "all observed" if you feel like spelling it out}{Time
  range will be limited to only times when observed data are present.}

  \item{"last dose to last observed" or "last obs" for short}{Time range will
  be limited to the start of the last dose until the last observed data
  point.}


  }}

\item{x_axis_interval}{optionally set the x-axis major tick-mark interval.
Acceptable input: any number or leave as NA to accept default values, which
are generally reasonable guesses as to aesthetically pleasing and
PK-relevant intervals.}

\item{pad_x_axis}{optionally add a smidge of padding to the the x axis
(default is TRUE, which includes some generally reasonable padding). If
changed to FALSE, the y axis will be placed right at the beginning of your
time range and all data will end \emph{exactly} at the end of the time
range specified. If you want a \emph{specific} amount of x-axis padding,
set this to a number; the default is \code{c(0.02, 0.04)}, which adds 2\%
more space to the left side and 4\% more to the right side of the x axis.
If you only specify one number, we'll assume that's the percent you want
added to the left side.}

\item{pad_y_axis}{optionally add a smidge of padding to the y axis (default
is TRUE, which includes some generally reasonable padding). As with
\code{pad_x_axis}, if changed to FALSE, the x axis will be placed right at
the bottom of your data, possible cutting a point in half. If you want a
\emph{specific} amount of y-axis padding, set this to a number; the default
is \code{c(0.02, 0)}, which adds 2\% more space to the bottom and nothing
to the top of the y axis. If you only specify one number, we'll assume
that's the percent you want added to the bottom.}

\item{adjust_obs_time}{TRUE or FALSE (default) for whether to adjust the time
listed in the observed data file to match the last dose administered. This
only applies to multiple-dosing regimens. If TRUE, the graph will show the
observed data overlaid with the simulated data such that the dose in the
observed data was administered at the same time as the last dose in the
simulated data. If FALSE, the observed data will start at whatever times
are listed in the Excel file.}

\item{t0}{What event should be used for time zero? Options are: "simulation
start" (default), "dose 1", "penultimate dose", or "last dose". \emph{This
does not change which data are included in the graph;} instead, this
determines whether the x axis numbers are offset so that, e.g., the last
dose is administered at time 0.}

\item{y_axis_limits_lin}{optionally set the Y axis limits for the linear
plot, e.g., \code{c(10, 1000)}. If left as the default NA, the Y axis
limits for the linear plot will be automatically selected.}

\item{y_axis_limits_log}{optionally set the Y axis limits for the semi-log
plot, e.g., \code{c(10, 1000)}. Values will be rounded down and up,
respectively, to a round number. If left as the default NA, the Y axis
limits for the semi-log plot will be automatically selected.}

\item{obs_color}{If you would like the observed data points to be in color,
either specify a color here or set this to "default". Points will be
displayed in semi-transparent blue-purple for "default" and the
semi-transparent version of whatever other color you list otherwise.
Setting this to "none" will make sure that the symbols are black outlines
only with no fill. Hex color codes are also ok to use. If left as NA, all
observed data will be in black.}

\item{obs_shape}{optionally specify what shapes are used to depict observed
data for 1. the substrate drug alone and 2. the substrate drug in the
presence of an effector. Input should look like this, for example:
\code{c(1, 2)} to get an open circle and an open triangle. To see all the
possible shapes and what number corresponds to which shape, see
\url{https://r-graphics.org/recipe-scatter-shapes} (there's a graph around
the middle of that page). If left as NA, substrate alone will be an open
circle and substrate + inhibitor 1 will be an open triangle.}

\item{showBLQ}{TRUE or FALSE (default) to display observed concentrations
that were clearly below the lower limit of quantitation, that is,
concentrations equal to 0 after time 0. The default (FALSE) removes these
values so that they will not show up on graphs.}

\item{line_type}{Optionally specify what types of lines are used to depict
\enumerate{\item{the substrate drug alone and} \item{the substrate drug in
the presence of an effector (when applicable).}} Input should look like
this, for example: \code{c("solid", "dashed")} to get a solid line for the
substrate drug and a dashed line for inhibitor 1. \itemize{ \item{To see
all possible \code{line_type} options: \code{ggpubr::show_line_types()}}
\item{If left as NA, substrate alone will be a solid line and substrate +
inhibitor 1 will be a dashed line.} \item{If \code{figure_type} is "Freddy"
and there's no effector present, which is a slightly different scenario
than the other graph types, the 1st line type specified will be for the
mean simulated concentration and the trial means, and the 2nd line type
specified will be for the 5th and 95th percentiles.}}}

\item{line_transparency}{optionally specify the transparency for the trial
mean or percentile lines. Acceptable values are from 0 (fully transparent,
so no line at all) to 1 (completely opaque or black). If left as the
default NA, this value will be automatically determined.}

\item{line_color}{optionally specify what colors to use for the lines.
Acceptable input for, e.g., the substrate alone to be blue and the
substrate + Inhibitor 1 to be red: \code{c("blue", "red")}. If left as the
default NA, lines will be black or gray. Hex color codes are also
acceptable to use.}

\item{line_width}{optionally specify how thick to make the lines. Acceptable
input is a number; the default is 1 for most lines and 0.8 for some, to
give you an idea of where to start.}

\item{legend_position}{specify where you want the legend to be. Options are
"left", "right", "bottom", "top", or "none" (default) if you don't want one
at all.}

\item{legend_label}{optionally indicate on the legend whether the effector is
an inhibitor, inducer, activator, or suppressor. Input will be used as the
label in the legend for the line style and the shape. If left as the
default NA when a legend is included and an effector is present, the label
in the legend will be "Inhibitor".}

\item{prettify_compound_names}{TRUE (default) or FALSE on whether to make
compound names prettier in legend entries and in any Word output files.
This was designed for simulations where the substrate and any metabolites,
effectors, or effector metabolites are among the standard options for the
simulator, and leaving \code{prettify_compound_names = TRUE} will make the
name of those compounds something more human readable. For example,
"SV-Rifampicin-MD" will become "rifampicin", and "Sim-Midazolam" will
become "midazolam". Set each compound to the name you'd prefer to see in
your legend and Word output if you would like something different. For
example, \code{prettify_compound_names = c("inhibitor" = "defartinib",
"substrate" = "superstatin")}. Please note that "inhibitor" includes
\emph{all} the effectors and effector metabolites present, so, if you're
setting the effector name, you really should use something like this if
you're including effector metabolites: \code{prettify_compound_names =
c("inhibitor" = "defartinib and 1-OH-defartinib", "substrate" =
"superstatin")}.}

\item{linear_or_log}{the type of graph to be returned. Options: \describe{
  \item{"semi-log"}{y axis is log transformed}

  \item{"linear"}{no axis transformation}

  \item{"both vertical"}{(default) both the linear and the semi-log graphs
  will be returned, and graphs are stacked vertically}

  \item{"both horizontal"}{both the linear and the semi-log graphs will be
  returned, and graphs are side by side horizontally}

  \item{"horizontal and vertical"}{both the linear and the semi-log graphs
  will be returned, and graphs are side by side horizontally (one graph; file
  name will end in "- horizontal") and stacked vertically (second graph; file
  name will end in "- vertical"). This option, which was designed to create
  the vertically stacked version of a graph for a report and the horizontal,
  side-by-side version for a presentation, is a bit different from the others
  since it will return two separate files. In the RStudio "Plots"
  window, you'll only see the vertically stacked version. Setting
  \code{fig_height} and \code{fig_width} will adjust only the dimensions of
  the horizontal figure; the default values will be used for the vertical
  one. If you request Word output, only the vertical plot will be saved in
  Word format; the horizontal plot will be saved as a png file.}}}

\item{graph_labels}{TRUE or FALSE for whether to include labels (A, B, C,
etc.) for each of the small graphs. (Not applicable if only outputting
linear or only semi-log graphs.)}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My conc time graph.png" or "My conc time
graph.docx". If you leave off ".png" or ".docx" from the file name, it will
be saved as a png file, but if you specify a different graphical file
extension, it will be saved as that file format. Acceptable graphical file
extensions are "eps", "ps", "jpeg", "jpg", "tiff", "png", "bmp", or "svg".
Leaving this as NA means the file will not be saved to disk.
\strong{WARNING:} SAVING TO WORD DOES NOT WORK ON SHAREPOINT. This is a
Microsoft permissions issue, not an R issue. If you try to save on
SharePoint, you will get a warning that R will save your file instead to
your Documents folder.}

\item{fig_height}{figure height in inches; default is 6}

\item{fig_width}{figure width in inches; default is 5}
}
\value{
Output is a graph.
}
\description{
Using observed and simulated concentration-time data, make
publication-quality graphs that match the consultancy template formatting
instructions. We've tried to include a fair number of options here for
flexibility, but many of the function arguments are optional; most of the
time, you'll get decent-looking graphs while only setting a minimal number of
arguments. If you want to plot enzyme abundance data, please see
\code{\link{enz_plot}}. \strong{A few notes:} \enumerate{\item{Not all
substrate metabolites, inhibitors, or inhibitor metabolites are available in
all tissues.}\item{When observed data are included in a simulator output
file, because the simulator output does not explicitly say whether those
observed data were in the presence of an effector, this function cannot tell
the difference and will thus assume all observed data included in the
simulator output were for the substrate in the \emph{absence} of any
effector. It will further assume that the compound the observed data is for
is the same as \code{compoundToExtract}. If you want to specify which
compound the observed data were for and/or whether any effectors were
present, please supply an observed data file to the argument
\code{obs_data_file}.} \item{If you attempt to use data generated by
\code{\link{extractConcTime_mult}} here, you can get weird results if you're
not careful to only include one tissue and one compound ID. If you see
something odd, try including the legend to help decipher what might have gone
awry.}}
}
\examples{
# # Normal syntax with a fake file:
# sim_data_file <- "../Example simulator output.xlsx"
# obs_data_file <- "../fig1-242-06-001-MD - for XML conversion.xlsx"

# However, since we can't load an Excel file as an example, here are
# some data to play with:
data(LMVct)

ct_plot(ct_dataframe = LMVct)
ct_plot(ct_dataframe = LMVct, figure_type = "percentiles")
ct_plot(ct_dataframe = LMVct)
ct_plot(ct_dataframe = LMVct)

# Perhaps you don't want to show *all* the data but instead want to
# limit the time interval that is graphed. Use \code{time_range} here:
ct_plot(ct_dataframe = LMVct, time_range = c(0, 24))

# Or you can let it automatically calculate the time frame
# for a given set of doses
ct_plot(ct_dataframe = LMVct
        time_range = "first dose")
ct_plot(ct_dataframe = LMVct,
        time_range = "last dose")

# The default graph may be too busy when Inhibitor 1 is present,
# so you may want to consider only plotting means as an alternative:
ct_plot(ct_dataframe = LMVct,
        figure_type = "means only")

# Add some further options for the look of your graph -- especially useful
# if the default settings are clipping your data.
ct_plot(ct_dataframe = LMVct,
        obs_color = "red",
        line_color = "blue",
        y_axis_limits_log = c(50, 2000),
        pad_x_axis = TRUE,
        legend_label = "Inhibitor")
}
