% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot3.R
\name{ct_plot3}
\alias{ct_plot3}
\title{Concentration-time plots of the full time range, first dose, and last dose}
\usage{
ct_plot3(sim_obs_dataframe, overlay = FALSE, ...)
}
\arguments{
\item{sim_obs_dataframe}{the input concentration-time data generated by
running either the function \code{\link{extractConcTime}} or the function
\code{\link{extactConcTime_mult}}}

\item{overlay}{TRUE or FALSE for whether to make overlaid graphs, i.e.,
whether to use \code{\link{ct_plot_overlay}} (TRUE) or use
\code{\link{ct_plot}} (FALSE) to generate the graphs}

\item{...}{arguments that pass through to \code{\link{ct_plot}} or
\code{\link{ct_plot_overlay}}}
}
\value{

}
\description{
Conveniently group 3 concentration-time plots together: 1) the full time
range across the top, 2) the first dose in the lower left quadrant, and 3)
the last dose in the lower right quadrant. Calls on either
\code{\link{ct_plot}} or \code{\link{ct_plot_overlay}} to make the graphs.
Requires the package ggpubr.
}
\examples{

CT <- extractConcTime(sim_data_file = "Example simulator output.xlsx")
ct_plot3(sim_obs_dataframe = CT, figure_type = "trial means", 
         linear_or_log = "log")

CT <- extractConcTime_mult(sim_data_files = c("Simulator output 1.xlsx",
                                              "Simulator output 2.xlsx",
                                              "Simulator output 3.xlsx"),
                           tissues = c("plasma", "blood"))
ct_plot3(sim_obs_dataframe = CT, colorBy_column = File,
         facet1_column = Tissue)

}
