% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot3.R
\name{ct_plot3}
\alias{ct_plot3}
\title{Concentration-time plots of the full time range, first dose, and last dose}
\usage{
ct_plot3(
  ct_dataframe,
  overlay = FALSE,
  figure_type = "means only",
  mean_type = "arithmetic",
  linear_or_log = "semi-log",
  legend_position = "none",
  save_graph = NA,
  fig_height = 6,
  fig_width = 5,
  ...
)
}
\arguments{
\item{ct_dataframe}{the input concentration-time data generated by running
either the function \code{\link{extractConcTime}} or the function
\code{\link{extractConcTime_mult}}}

\item{overlay}{TRUE or FALSE for whether to make overlaid graphs, i.e.,
whether to use \code{\link{ct_plot_overlay}} (TRUE) or use
\code{\link{ct_plot}} (FALSE) to generate the graphs}

\item{figure_type}{type of figure to plot. Options are:

  \describe{

  \item{"percentiles"}{(default) plots an opaque line for the mean data,
  lighter lines for the 5th and 95th percentiles of the simulated data, and
  open circles for the observed data. If an effecter were present, the
  default is dashed lines for the data in the presence of an effector.}

  \item{"trial means"}{plots an opaque line for the mean data, lighter lines
  for the mean of each trial of simulated data, and open circles for the
  observed data. If an effector were present, lighter dashed lines indicate
  the mean of each trial of simulated data in the presence of the effector.}

  \item{"percentile ribbon"}{plots an opaque line for the mean data,
  transparent shading for the 5th to 95th percentiles of the simulated data,
  and open circles for the observed data. If an effector were present, the
  default is to show the data without the effector in blue and the data in
  the presence of the effector in red. Note: You may sometimes see some
  artifacts -- especially for semi-log plots -- where the ribbon gets partly
  cut off. For arcane reasons we don't want to bore you with here, we can't
  easily prevent this. However, a possible fix is to set your y axis limits
  for the semi-log plot to be wider using \code{y_axis_limits_log}.}

  \item{"means only"}{plots a black line for the mean data and, if an
  effector was modeled, a dashed line for the concentration-time data with
  Inhibitor 1.}

  \item{"Freddy"}{Freddy's favorite style of plot with trial means in light
  gray, the overall mean in thicker black, the 5th and 95th percentiles in
  dashed lines, and the observed data in semi-transparent purple-blue. Graphs
  with an effector present lose the trial means, and the percentiles switch
  to solid, gray lines. \strong{An editorial comment:} While this does not
  align with the officially sanctioned template at this time, this looks
  \emph{sharp}, makes it easy to see the defining characteristics of the
  data, and I recommend checking it out, even just for your own purposes of
  examining your data. If the color is too much for you but you like the
  rest, try setting \code{obs_color = "black", obs_shape = c(1, 2)}. -LSh}}}

\item{mean_type}{graph "arithmetic" (default) or "geometric" means or
"median" for median concentrations. If that option was not included in the
output, you'll get a warning and the graph will include one that was.}

\item{linear_or_log}{the type of graph to be returned. Options: \describe{
  \item{"semi-log"}{y axis is log transformed}

  \item{"linear"}{no axis transformation}

  \item{"both vertical"}{(default) both the linear and the semi-log graphs
  will be returned, and graphs are stacked vertically}

  \item{"both horizontal"}{both the linear and the semi-log graphs will be
  returned, and graphs are side by side horizontally}

  \item{"horizontal and vertical"}{both the linear and the semi-log graphs
  will be returned, and graphs are side by side horizontally (one graph; file
  name will end in "- horizontal") and stacked vertically (second graph; file
  name will end in "- vertical"). This option, which was designed to create
  the vertically stacked version of a graph for a report and the horizontal,
  side-by-side version for a presentation, is a bit different from the others
  since it will return two separate files. In the RStudio "Plots" window,
  you'll only see the vertically stacked version. Setting \code{fig_height}
  and \code{fig_width} will adjust only the dimensions of the horizontal
  figure; the default values will be used for the vertical one. If you
  request Word output, only the vertical plot will be saved in Word format;
  the horizontal plot will be saved as a png file.}}}

\item{legend_position}{Specify where you want the legend to be. Options are
"left", "right", "bottom", "top", or "none" (default) if you don't want one
at all. If you include the legend but then some graphs do have a legend and
some graphs do not (e.g., some have effectors and some do not so there's
nothing to put in a legend), the alignment between sets of graphs will be a
bit off.}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My conc time graph.png" or "My conc time
graph.docx". If you leave off ".png" or ".docx" from the file name, it will
be saved as a png file, but if you specify a different graphical file
extension, it will be saved as that file format. Acceptable graphical file
extensions are "eps", "ps", "jpeg", "jpg", "tiff", "png", "bmp", or "svg".
Leaving this as NA means the file will not be saved to disk.
\strong{WARNING:} SAVING TO WORD DOES NOT WORK ON SHAREPOINT. This is a
Microsoft permissions issue, not an R issue. If you try to save on
SharePoint, you will get a warning that R will save your file instead to
your Documents folder.}

\item{fig_height}{figure height in inches; default is 6}

\item{fig_width}{figure width in inches; default is 5}

\item{...}{arguments that pass through to \code{\link{ct_plot}} or
\code{\link{ct_plot_overlay}}}
}
a set of 3 arranged ggplot2 graphs
}
\description{
Conveniently group three concentration-time plots of multiple-dose regimen
data together: \enumerate{\item{the full time range across the top,}
\item{the first dose in the lower left quadrant, and}  \item{the last dose in
the lower right quadrant.}} This uses either \code{\link{ct_plot}} or
\code{\link{ct_plot_overlay}} behind the scenes to make the graphs, so most
of the options available for those functions will work here.
}
\examples{

data(LMVct)
ct_plot3(ct_dataframe = LMVct, linear_or_log = "log")

data(MDZ_Keto)
ct_plot3(ct_dataframe = MDZ_Keto, overlay = TRUE, colorBy_column = CompoundID)

}
