% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot3.R
\name{ct_plot3}
\alias{ct_plot3}
\title{Concentration-time plots of the full time range, first dose, and last dose}
\usage{
ct_plot3(
  ct_dataframe,
  overlay = FALSE,
  save_graph = NA,
  fig_height = 6,
  fig_width = 5,
  ...
)
}
\arguments{
\item{ct_dataframe}{the input concentration-time data generated by running
either the function \code{\link{extractConcTime}} or the function
\code{\link{extractConcTime_mult}}}

\item{overlay}{TRUE or FALSE for whether to make overlaid graphs, i.e.,
whether to use \code{\link{ct_plot_overlay}} (TRUE) or use
\code{\link{ct_plot}} (FALSE) to generate the graphs}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My conc time graph.png" or "My conc time
graph.docx". If you leave off ".png" or ".docx" from the file name, it will
be saved as a png file, but if you specify a different graphical file
extension, it will be saved as that file format. Acceptable graphical file
extensions are "eps", "ps", "jpeg", "jpg", "tiff", "png", "bmp", or "svg".
Leaving this as NA means the file will not be saved to disk.
\strong{WARNING:} SAVING TO WORD DOES NOT WORK ON SHAREPOINT OR THE LARGE
FILE STORE. This is a Microsoft permissions issue, not an R issue. If you
temporarily change your working directory to a local folder, it will work
fine and you can copy those files later back to SharePoint or the Large
File Store. We wish we had a better solution for this!}

\item{fig_height}{figure height in inches; default is 6}

\item{fig_width}{figure width in inches; default is 5}

\item{...}{arguments that pass through to \code{\link{ct_plot}} or
\code{\link{ct_plot_overlay}}}
}
\value{

}
\description{
Conveniently group three concentration-time plots of multiple-dose regimen
data together: \enumerate{\item{the full time range across the top,}
\item{the first dose in the lower left quadrant, and}  \item{the last dose in
the lower right quadrant.}} This uses either \code{\link{ct_plot}} or
\code{\link{ct_plot_overlay}} behind the scenes to make the graphs, so most
of the options available for those functions will work here.
}
\examples{

data(LMVct)
ct_plot3(ct_dataframe = LMVct, linear_or_log = "log")

data(MDZ_Keto)
ct_plot3(ct_dataframe = MDZ_Keto, overlay = TRUE, colorBy_column = CompoundID)

}
