% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot_gui.R
\name{ct_plot_gui}
\alias{ct_plot_gui}
\title{Run the ct_plot function through a graphical user interface}
\usage{
ct_plot_gui()
}
\value{

}
\description{
Opens the shiny app version of the function \code{\link{ct_plot}}.
}
\examples{
# You don't need to specify any arguments to this function
# in the console. Just type the following:
shiny_ct_plot()

}
