% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot_mult.R
\name{ct_plot_mult}
\alias{ct_plot_mult}
\title{Make graphs for multiple Simulator output files at once}
\usage{
ct_plot_mult(
  ct_dataframe,
  single_out_file = TRUE,
  linear_or_log = "semi-log",
  graph_labels = TRUE,
  include_title = FALSE,
  file_labels = NA,
  legend_position = "none",
  nrow = NULL,
  ncol = NULL,
  save_graph = NA,
  fig_height = 8,
  fig_width = 8,
  ...
)
}
\arguments{
\item{ct_dataframe}{the data.frame with multiple sets of concentration-time
data}

\item{single_out_file}{TRUE (default) or FALSE for whether to make multiple
output files. TRUE means that there will be a single file with all of the
graphs nicely arranged together. FALSE means that each Simulator output
file will have its own output file, and that graph's output file will be
named to match the Simulator output file name.}

\item{linear_or_log}{the type of graph to be returned. Options: "semi-log"
(default), "linear", "both vertical" (graphs are stacked vertically), or
"both horizontal" (graphs are side by side).}

\item{graph_labels}{TRUE or FALSE for whether to include labels (A, B, C,
etc.) for each of the small graphs. (Not applicable if
\code{single_out_file = FALSE}.)}

\item{include_title}{TRUE or FALSE (default) on whether to include a title
for each small graph.}

\item{file_labels}{optionally specify a label to be used for the file name in
the graphs (this applies when \code{include_title = TRUE}) and/or specify
the order in which the files are graphed with a named character vector of
the files in the order you would like. (Not applicable if
\code{single_out_file = FALSE}.) The file name must \emph{perfectly} match
the file name listed in ct_dataframe or it won't be used. An example of how
this might be specified: \code{alt_title = c("My file 1.xlsx" = "Healthy
volunteers", "My file 2.xlsx" = "Mild hepatic impairment")}  If you get an
order that you didn't think you specified, please double check that you
have specified the file names \emph{exactly} as they appear in
\code{ct_dataframe}.}

\item{legend_position}{Specify where you want the legend to be. Options are
"left", "right", "bottom", "top", or "none" (default) if you don't want one
at all. If you include the legend but then some graphs do have a legend and
some graphs do not (e.g., some have effectors and some do not so there's
nothing to put in a legend), the alignment between sets of graphs will be a
bit off.}

\item{nrow}{number of rows of small graphs in the arrangement. If left as NA,
a reasonable guess will be made. (Not applicable if \code{single_out_file =
FALSE}.)}

\item{ncol}{number of columns of small graphs in the arrangement. If left as
NA, a reasonable guess will be made. (Not applicable if
\code{single_out_file = FALSE}.)}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My conc time graph.png". If you leave off ".png", it
will be saved as a png file, but if you specify a different file extension,
it will be saved as that file format. Acceptable extensions are "eps",
"ps", "jpeg", "jpg", "tiff", "png", "bmp", or "svg". Leaving this as NA
means the file will not be automatically saved to disk. (Not applicable if
\code{single_out_file = FALSE} because each graph will be named according
to its Simulator output file.)}

\item{fig_height}{figure height in inches; default is 8}

\item{fig_width}{figure width in inches; default is 8}

\item{...}{arguments that pass through to \code{\link{ct_plot}}}
}
\value{

}
\description{
This function was designed for making nicely arranged concentration-time
graphs from several Simcyp Simulator output files all together \emph{or} for
making multiple files -- one for each Simulator file -- all at once.
}
\details{
\strong{A note on the order of the graphs:} This function arranges graphs
first by file, then by compound ID, and then by tissue, and all sorting is
alphabetical. However, since sorting alphabetically might not be the optimal
graph arrangement for your scenario, you \emph{can} specify the order of the
graphs using either the \code{file_labels} argument or, if you're comfortable
with setting factors in R, by making any of File, CompoundID, Tissue, and
subsection_ADAM factor rather than character data and setting the levels how
you wish. If you're unfamiliar with setting factor levels in R and setting
\code{file_labels} isn't achieving what you want, please ask a member of the
R Working Group for assistance.
}
\examples{

data(MDZct)
ct_plot_mult(ct_dataframe = MDZct) 

ct_plot_mult(ct_dataframe = MDZct, include_title = TRUE)

ct_plot_mult(ct_dataframe = MDZct, include_title = TRUE, 
   file_labels = c("mdz-5mg-sd-fa1.xlsx" = "fa = 1", 
                   "mdz-5mg-sd-fa0_8.xlsx" = "fa = 0.8", 
                   "mdz-5mg-sd-fa0_6.xlsx" = "fa = 0.6", 
                   "mdz-5mg-sd-fa0_4.xlsx" = "fa = 0.4"))

}
