% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot_overlay.R
\name{ct_plot_overlay}
\alias{ct_plot_overlay}
\title{Overlay multiple data sets onto a single concentration-time graph}
\usage{
ct_plot_overlay(
  conctime_DF,
  aggregate_option = "geomean",
  colorBy = File,
  linetypeBy = Inhibitor,
  facet_column1,
  facet_column2
)
}
\arguments{
\item{conctime_DF}{the data.frame with multiple sets of concentration-time
data, including observed data where appropriate}

\item{aggregate_option}{What type of aggregate measure should be shown?
Options are any combination of "mean" (arithmetic mean), "geomean"
(geometric mean), "median", "per5" (5th percentile), "per95" (95th
percentile), "per10" (10th percentile), or "per90" (90th percentile).
Default is "geomean". Example of useage: \code{aggregate_option = c("mean",
"per5", "per95")}}

\item{colorBy}{What column in \code{conctime_DF} should be used for coloring
the lines and/or points on the graph? This should be unquoted, e.g.,
\code{colorBy = Tissue}.}

\item{facet_column1}{If you would like to break up your graph into small
multiples, you can break the graphs up by up to two columns in
\code{conctime_DF}. What should be the 1st column to break up the data by?
This should be unquoted.}

\item{facet_column2}{What should be the 2nd column to break up the data into
small multiples by? This should be unquoted.}
}
\value{

}
\description{
\code{ct_plot_overlay} is meant to be used in conjunction with
\code{\link{extractConcTime_mult}} to create single graphs with overlaid
concentration-time data from multiple Simcyp Simulator output files for easy
comparisons. UNDER CONSTRUCTION.
}
\examples{
# No examples yet
# ct_plot_overlay(conctime_DF = CT, facet_column1 = Compound,
#                 facet_column2 = Tissue)


}
