% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_plot_overlay.R
\name{ct_plot_overlay}
\alias{ct_plot_overlay}
\title{Overlay multiple data sets onto a single concentration-time graph}
\usage{
ct_plot_overlay(
  sim_obs_dataframe,
  aggregate_option = "geomean",
  colorBy = File,
  facet_column1,
  facet_column2
)
}
\arguments{
\item{sim_obs_dataframe}{the data.frame with multiple sets of
concentration-time data. At the moment, this function does not plot
observed data, though; that's still under construction.}

\item{colorBy}{What column in \code{sim_obs_dataframe} should be used for
coloring the lines and/or points on the graph? This should be unquoted,
e.g., \code{colorBy = Tissue}.}

\item{facet_column1}{If you would like to break up your graph into small
multiples, you can break the graphs up by up to two columns in
\code{sim_obs_dataframe}. What should be the 1st column to break up the
data by? This should be unquoted.}

\item{facet_column2}{What should be the 2nd column to break up the data into
small multiples by? This should be unquoted.}

\item{mean_type}{show "geometric" (default) or "arithmetic" means for the
main (thickest) line for each data set. If this summary stat is not
available in the simulator output, it will not be plotted.}
}
\value{

}
\description{
\code{ct_plot_overlay} is meant to be used in conjunction with
\code{\link{extractConcTime_mult}} to create single graphs with overlaid
concentration-time data from multiple Simcyp Simulator output files for easy
comparisons. Note: This is generally workable but is admittedly not one of
our most maturely developed functions. :-)
}
\examples{
# No examples yet
# ct_plot_overlay(sim_obs_dataframe = CT, facet_column1 = Compound,
#                 facet_column2 = Tissue)


}
