% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_plot.R
\name{demog_plot}
\alias{demog_plot}
\title{Make a set of plots showing the demographics used in a set of simulations and
optionally compare them to observed demographics. \strong{UNDER
CONSTRUCTION}.}
\usage{
demog_plot(
  demog_dataframe,
  sim_data_file = NA,
  plot_title = "Demographics",
  variables = NA,
  obs_alpha = 0.8,
  sim_alpha = 0.4,
  ncol = NULL,
  nrow = NULL,
  facet_by_sex = FALSE,
  border_facets = TRUE,
  graph_labels = TRUE
)
}
\arguments{
\item{demog_dataframe}{the output from running \code{\link{extractDemog}}.
Optionally (and we recommend) with added observed demographic data, perhaps
from observed overlay XML files.}

\item{sim_data_file}{the simulator output Excel file to use. If left as NA,
all the files in demog_dataframe will be included.}

\item{plot_title}{title to use on the plots}

\item{variables}{variables to include. We're starting with a limited set:
"Age", "Weight_kg", "Height_cm", "Weight vs Height", "Height vs Age",
"Weight vs Age", "HSA_gL", "AGP_gL", "Sex", "Sex vs Age", "BMI_kgm2", and
"RenalFunction". If you want only a subset of those,
list them in a character vector, e.g., \code{variables = c("Age",
"Height_cm", "Weight_kg")}. Plots will be in the order listed.}

\item{obs_alpha}{how transparent to make the observed data, with 0 being
completely transparent and invisible so I don't know why you'd want that
but, hey, you do you, dude, to 1, which is fully opaque.}

\item{sim_alpha}{how transparent to make the simulated data. See notes on the
obs_alpha argument.}

\item{ncol}{optionally specify the number of columns. If left as NULL, a
reasonable guess will be used.}

\item{nrow}{optionally specify the number of rows. If left as NULL, a
reasonable guess will be used.}

\item{facet_by_sex}{TRUE or FALSE (default) for whether to break up the
graphs into facets based on the sex of the subjects}

\item{border_facets}{TRUE (default) or FALSE for whether to include a border
around the facets if the graphs are broken up by the sex of the subjects}

\item{graph_labels}{TRUE or FALSE for whether to include labels (A, B, C,
etc.) for each of the small graphs.}
}
\value{
a set of graphs. This does not yet save the graphs for you, so you'll
  need to run ggsave(...) to do that.
}
\description{
Make a set of plots showing the demographics used in a set of simulations and
optionally compare them to observed demographics. \strong{UNDER
CONSTRUCTION}.
}
\examples{
# none yet
}
