% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_plot.R
\name{demog_plot_sim}
\alias{demog_plot_sim}
\title{TO BE DEPRECATED: Make plots comparing demographic parameters}
\usage{
demog_plot_sim(...)
}
\arguments{
\item{...}{}
}
\value{
graphs
}
\description{
Please use the function demog_plot, which works exactly the same as
demog_plot_sim.
}
