% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_methods_text.R
\name{draft_methods_text}
\alias{draft_methods_text}
\title{Title}
\usage{
draft_methods_text(
  sim_data_file,
  existing_exp_details,
  prettify_compound_names = TRUE,
  default_cmpd_file = TRUE,
  victim_sim = FALSE
)
}
\arguments{
\item{sim_data_file}{file}

\item{existing_exp_details}{details}

\item{prettify_compound_names}{T or F}

\item{default_cmpd_file}{T or F - only affects text "The default compound
library file for XXX was used" when it's a DDI sim.}

\item{victim_sim}{T or F for whether client drug was a victim of DDI. Ignored
when the simulation did not include any perpetrator drugs.}
}
\value{
list of generic methods text
}
\description{
Title
}
\examples{
# none yet
}
