% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_results_text.R
\name{draft_results_text}
\alias{draft_results_text}
\title{Title}
\usage{
draft_results_text(
  sim_data_file,
  existing_exp_details,
  prettify_compound_names = TRUE,
  mean_type = "geometric"
)
}
\arguments{
\item{sim_data_file}{file}

\item{existing_exp_details}{details}

\item{prettify_compound_names}{T or F}

\item{mean_type}{arithmetic or geometric (default)}
}
\value{
list of generic results text
}
\description{
Title
}
\examples{
# none yet

}
