% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sample_size.R
\name{estimate_sample_size}
\alias{estimate_sample_size}
\title{Estimate the sample size necessary for determining whether there's a
significant difference between the PK under control and test conditions -
UNDER CONSTRUCTION}
\usage{
estimate_sample_size(
  sim_data_files = NA,
  PKparameters = c("AUCinf_dose1", "AUCt_dose1"),
  alpha = 0.05,
  power = 0.8,
  paired = TRUE,
  alternative_hypothesis = "two sided",
  existing_exp_details = NA,
  compoundToExtract = "substrate",
  tissue = "plasma",
  sheet_PKparameters = NA,
  PK_data = NA,
  round_up_N = TRUE,
  sim_trials_to_include = "all",
  save_result = NA
)
}
\arguments{
\item{sim_data_files}{name of the Excel file(s) containing the simulator
output, in quotes.}

\item{PKparameters}{the PK parameters to include. There are two main options
  for supplying this information: 1) supply a file to read or a data.frame (R
  speak for "a table") that lists which simulation files, compounds, tissues,
  and PK you want or 2) supply a character vector of which PK parameters you
  want and then also specify what you need in terms of which tissues, which
  compounds, which simulation files, and which tab to get the data from with
  the arguments \code{tissues}, \code{compoundsToExtract},
  \code{sim_data_files}, and \code{sheet_PKparameters}.
  \strong{Details on each option:} \describe{

  \item{\strong{Option 1:} a file to read or a data.frame}{This
  is the most versatile option and, we think, the clearest in terms of
  getting what you expected. Please try running \code{\link{make_example_PK_input}}
  to see examples for how to set up a csv or Excel file or data.frame to
  specify exactly which simulation file should get which PK parameter from
  which tissue and, when user-specified intervals are involved, from which
  tab in the Excel file those data should be pulled. If you want to make
  comparisons across anything other than a standard DDI simulation, you
  should use this option to specify exactly what should be compared for each
  simulation. You'll need to set this up as if you're comparing ratios of
  PK parameters (answer "2" to the 1st question if you run
  \code{make_example_PK_input}), and you'll set the control scenario to be 
  the denominator and the test scenario to be the numerator. 
  
  Whatever you supply, the
  columns that will be read are: \itemize{\item{"File" (same thing as the argument
  \code{sim_data_files})} \item{"Sheet" (same thing as the argument
  \code{sheet_PKparameters})} \item{"Tissue" (same as the argument \code{tissues})}
  \item{"CompoundID" (same as the argument \code{compoundsToExtract})
  \item{"ObsValue" for any observed data (no equivalent argument)}
  \item{"Variability" for any observed variability (no equivalent argument
  here, either)}} If you
  omit any of those columns, whatever you supply for their respective
  arguments will be used instead. If you supply something for one of them
  in the data.frame or file and \emph{also} something for its argument, the
  argument will be ignored. \cr

  Here is how to specify each of the possible
  inputs for Option 1:\describe{\item{a csv file}{list the file
  name, including the file extension ".csv" inside quotes, e.g., "PK needed.csv"}

  \item{an Excel file}{list the file name, including the file extension
  ".xlsx", inside quotes, e.g., "PK needed.xlsx". We will be looking for a
  tab titled "PKparameters" (all one word and with the same capitalization).}

  \item{a data.frame}{If you'd like to supply a data.frame with the same
  columns you would have had in the .csv or Excel file, that works just the
  same.}}}}

  \item{\strong{Option 2: }specify the PK parameters you want for all your
  simulations}{This is a good option when you want the same information
  from all your simulations. List the PK parameters you want here and then,
  in the arguments
  \code{tissues}, \code{compoundsToExtract}, \code{sim_data_files}, and
  \code{sheet_PKparameters} specify which of each of those items you want.
  You'll get all possible combinations of these, so, for example, if you say
  \code{PKparameters = c("AUCinf_dose1", "Cmax_dose1")} and
  \code{tissues = c("blood", "plasma")}, you'll get the dose 1 AUCinf and
  Cmax for both blood and plasma for all the simulation files you list with
  \code{sim_data_files}. If you're going this route, here are the two options
  you have for the argument \code{PKparameters}: \describe{

  \item{NA}{If you leave this as NA, by default, if you have a single-dose
  simulation, the parameters will
  include AUC and Cmax for dose 1, or, if you have a multiple-dose
  simulation, AUC and Cmax for the last dose. Also by default, if you have a
  perpetrator present, the parameters will include the AUC and Cmax values with
  and without the perpetrator as well as those ratios.}

  \item{a character vector of any combination of specific, individual
  parameters}{This character vector must contain SimcypConsultancy package
  coded names for each parameter you want, e.g., \code{c("Cmax_dose1",
  "AUCtau_last").} Be sure to encapsulate the parameters you want with
  \code{c(...)}. Please try running \code{\link{make_example_PK_input}} to
  see examples, or, to see the full set of all possible parameters to extract, enter
  \code{view(PKParameterDefinitions)} into the console.}}}}

  Parameters that don't make sense for your scenario -- such as asking for
  \code{AUCinf_dose1_withInhib} when your simulation did not include a
  perpetrator -- will be ignored.}

\item{alpha}{alpha required, default is 0.05 for the 95\% confidence level}

\item{power}{power (1-beta) required, default is 0.8.}

\item{paired}{TRUE (default) or FALSE for whether the study design was
paired, as in, the subjects were \emph{identical} between the two
simulations.
\strong{THIS IS AN IMPORTANT DISTINCTION AND WILL AFFECT HOW THE
CALCULATIONS ARE PERFORMED!} An example of a paired study would be a DDI
study where each subject has a measurement without the perpetrator of
interest and then has a second measurement \emph{with} the perpetrator. The
comparison is for repeated measurements of the \emph{same subject}. An
example of an unpaired study design would be comparing healthy subjects to
subjects with hepatic impairment because those are measurements on
\emph{different} subjects.}

\item{alternative_hypothesis}{Is the test condition expected to be greater
  than or less than the control or is that unknown? Acceptable values:
  \describe{\item{"two sided" (default)}{unknown \emph{a priori} whether the
  test condition will be greater or less than control, i.e., a two-sided
  t test would be used to compare the groups}

  \item{"less"}{test condition is expected to be less than control and a
  one-sided t test would be used to compare the groups. Example: Comparing
  the AUC in a DDI study with an inducer.}

  \item{"greater"}{test condition is expected to be greater than control and a
  one-sided t test would be used to compare the groups. Example: Comparing
  the AUC in a DDI study with an inhibitor.}}}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} or \code{\link{extractExpDetails}} to
get all the details from the "Input Sheet" (e.g., when you ran
extractExpDetails you said \code{exp_details = "Input Sheet"} or
\code{exp_details = "all"}), you can save some processing time by supplying
that object here, unquoted. If left as NA, this function will run
\code{extractExpDetails} behind the scenes to figure out some information
about your experimental set up.}

\item{compoundToExtract}{For which compound do you want to extract
PK data? Options are: \itemize{\item{"substrate"
(default),} \item{"primary metabolite 1",} \item{"primary metabolite 2",}
\item{"secondary metabolite",}} \strong{If you want more than one,
please supply a data.frame or .csv file to the argument \code{PKparameters}.}}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are "plasma" (default), "unbound plasma", "blood", "unbound blood",
"peripheral plasma", or "peripheral blood". \strong{If you want more than one,
please supply a data.frame or .csv file to the argument \code{PKparameters}.}}

\item{sheet_PKparameters}{(optional) If you want the PK parameters to be
pulled from a \strong{user-defined interval tab} in the simulator output
file, list that tab here. Otherwise, this should be left as NA.
\code{sheet_PKparameters} can only have a \emph{single value}, though. If
you want some parameters from a custom-interval tab and others from the
regular tabs, you must supply that as part of a data.frame or csv file for
the argument \code{PKparameters}. Please try running
\code{\link{make_example_PK_input}} to see examples for how to do this.
\strong{If you want more than one,
please supply a data.frame or .csv file to the argument \code{PKparameters}.}}

\item{PK_data}{You can either supply the PK data we'll use for estimating the
sample size or you can have us extract it for you. The PK data you supply
here MUST be in the format you'd get from running \code{\link{extractPK}}
and it MUST have both the baseline and the DDI version of each PK
parameter, e.g., both AUCinf_dose1 and AUCinf_dose1_withInhib. If you
supply something here, we'll only pay attention to the arguments
sim_data_files, existing_exp_details, PKparameters, compoundToExtract,
tissue, and sheet_PKparameters if what you supply doesn't have that same
structure. This ONLY works with the simplest of possible comparisons: a
standard DDI simulation.}

\item{round_up_N}{TRUE (default) or FALSE for whether to round up to the
nearest whole human}

\item{sim_trials_to_include}{Which simulated trials should be included in the
calculation? Options are all (default) to include all the data or any
numbers in the trials included in your simulation, e.g., 1 or 5:10.}

\item{save_result}{file name for optionally saving the results as either an
Excel file (file name must end in .xlsx) or a Word document (file name must
end in .docx). May not contain special characters or symbols. If you save
the output as a Word file, it will include background information on how
the calculations were performed and possible wording for including in a
report.}
}
\value{
a data.frame of the sample sizes needed for each combination of
  simulation file, tissue, compound ID, and PK parameter.
}
\description{
\code{estimate_sample_size} compares simulated PK data under a
  control condition with PK data under a test condition to determine how
  large your sample size would need to be to reject the null hypothesis of no
  difference between the two. The comparison may be paired -- e.g., comparing
  baseline to DDI -- or unpaired -- e.g., comparing healthy subjects to those
  with hepatic impairment. If alpha is 0.05, i.e., a 95\% confidence level,
  and 1 - beta is 0.8, i.e., 80\% power, this function answers the question:
  What is the estimated sample size needed such that, at the 95\% confidence
  level, the null hypothesis should be rejected 80\% percent of the time?

  \strong{Notes on the statistics being used:} Under the hood, this function
  will estimate the Cohen's d effect size based on the simulated data,
  correct with the Hedge's correction g, and then use the function pwr.t.test
  from the pwr package to calculate N. For references, please see:
  \itemize{\item{Cohen, J. (1988). Statistical power analysis for the
  behavioral sciences (2nd ed.). Hillsdale,NJ: Lawrence Erlbaum.}

  \item{Lakens, D. (2013). “Calculating and
  reporting effect sizes to facilitate cumulative science: a practical primer
  for t-tests and ANOVAs.” Frontiers in Psychology, 4:863. doi: 10.3389/fpsyg.2013.00863}}

  If you save the results from running this as a Word file, that document
  will contain more information on the methods and possible text to include
  in a report.
}
