% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAUCXtab.R
\name{extractAUCXtab}
\alias{extractAUCXtab}
\title{FOR INTERNAL USE ONLY: Extract PK data from a generic AUC tab, e.g., NOT one
formatted like the "AUC" tab from Simcyp V21 and earlier}
\usage{
extractAUCXtab(
  PKparameters,
  PKparameters_orig,
  sim_data_file,
  Sheet,
  PKset,
  UserSpecified,
  Deets,
  DataCheck,
  includeTrialInfo
)
}
\arguments{
\item{PKparameters}{parameters to extract; character vector only (a set of
parameters like in extractPK is not acceptable here)}

\item{sim_data_file}{simulator output Excel file to read}

\item{Sheet}{sheet to read from the Excel file}

\item{PKset}{which set of PK parameters to use. Options: AUC0, AUClast, AUCX}

\item{UserSpecified}{TRUE or FALSE for whether this was a user-specified
sheet, in which case we don't know which dose it is}

\item{Deets}{output from running extractExpDetails on the provided
sim_data_file}

\item{DataCheck}{DataCheck data.frame}

\item{includeTrialInfo}{TRUE or FALSE}
}
\value{
a list for use with the extractPK function
}
\description{
FOR INTERNAL USE ONLY: Extract PK data from a generic AUC tab, e.g., NOT one
formatted like the "AUC" tab from Simcyp V21 and earlier
}
\examples{
# none

}
