% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractConcTime.R
\name{extractConcTime}
\alias{extractConcTime}
\title{Extract concentration-time data from a simulator output Excel file}
\usage{
extractConcTime(
  sim_data_file,
  obs_data_file = NA,
  obs_effector_data_file = NA,
  adjust_obs_time = TRUE,
  tissue = "plasma",
  returnAggregateOrIndiv = c("aggregate", "individual")
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulated
concentration-time data}

\item{obs_data_file}{name of the Excel file containing the observed
concentration-time data. If the observed data you want to plot were already
included in the Excel output from the simulator, leave this as NA.
Otherwise, this is the file that it is ready to be converted to an XML
file, not the file that contains only the digitized time and concentration
data.}

\item{adjust_obs_time}{TRUE or FALSE: Adjust the time listed in the observed
data file to match the last dose administered? This only applies to
multiple-dosing regimens. If TRUE, the graph will show the observed data
overlaid with the simulated data such that the dose in the observed data
was administered at the same time as the last dose in the simulated data.
If FALSE, the observed data will start at whatever times are listed in the
Excel file.}

\item{tissue}{From which tissue should the desired concentrations be
extracted? Default is plasma for typical plasma concentration-time data.
Other tissues are acceptable, e.g., "lung", "brain", etc., as long as the
tissue is one of the options included in "Sheet Options", "Tissues" in the
simulator.}

\item{returnAggregateOrIndiv}{Return aggregate and/or individual simulated
concentration-time data? Options are one or both of "aggregate" and
"individual". Aggregated data are not calculated here but are pulled from
the simulator output rows labeled as "mean".}
}
\value{
A data.frame of concentration-time data with the following columns:
  \describe{

  \item{Compound}{the compound whose concentration is listed; this matches
  whatever you named your substrate or inhibitor in the simulator}

  \item{Effector (as applicable)}{the effector of interest; this matches
  whatever you named "Inhibitor 1" in the simulator}

  \item{Individual}{the individual for the given profile, which will be a
  number for a simulated individual or will be "obs" or "obs+effector" for
  observed data, "mean" for the mean data, or "per5" or "per95" for the 5th
  and 95th percentile data.}

  \item{Trial}{the trial number for that set of simulations or "obs", "mean",
  etc. for the observed or aggregate data}

  \item{Simulated}{TRUE or FALSE for whether the data were simulated}

  \item{Time}{the time since the first dose}

  \item{Conc}{concentration of the compound listed}

  \item{Time_units}{units used for time}

  \item{Conc_units}{units used for concentrations}}
}
\description{
Extracts concentration-time data from simulater output Excel files and,
optionally, a separately specified clinical data file, and puts all data into
a single, tidy data.frame. \emph{Note: Currently, only set up to extract
concentration-time data for one substrate and up to one inhibitor as
applicable.}
}
\examples{
extractConcTime(sim_data_file = "../Example simulator output MD.xlsx")

extractConcTime(sim_data_file = "../Example simulator output MD.xlsx",
                returnAggregateOrIndiv = "individual")

extractConcTime(sim_data_file = "../Example simulator output MD.xlsx",
                obs_data_file = "../fig1-242-06-001-MD - for XML conversion.xlsx")

extractConcTime(sim_data_file = "../Example simulator output MD + inhibitor.xlsx",
                returnAggregateOrIndiv = c("aggregate", "individual"))

extractConcTime(sim_data_file = "../Example simulator output MD + inhibitor.xlsx",
                obs_data_file = "../fig1-242-06-001-MD - for XML conversion.xlsx",
                obs_effector_data_file = "../Mallikaarjun_2016_RTV-fig1-100mg-BID-DLM+Kaletra - for XML conversion.xlsx",
                returnAggregateOrIndiv = c("aggregate", "individual"))

extractConcTime(sim_data_file = "../Example simulator output MD + inhibitor.xlsx",
                tissue = "lung")


}
