% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractConcTime_DB.R
\name{extractConcTime_DB}
\alias{extractConcTime_DB}
\title{Extract concentration-time data from a Simcyp Simulator database file --
UNDER CONSTRUCTION!!!}
\usage{
extractConcTime_DB(
  sim_data_file,
  compoundToExtract = "substrate",
  tissue = "plasma",
  existing_exp_details = NA,
  returnAggregateOrIndiv = "aggregate"
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulated
concentration-time data, in quotes; must be an output file from the Simcyp
simulator}

\item{compoundToExtract}{For which compound do you want to extract
concentration-time data? Options are: \itemize{\item{"substrate"
(default),} \item{"primary metabolite 1",} \item{"primary metabolite 2",}
\item{"secondary metabolite",} \item{"inhibitor 1" -- this can be an
inducer, inhibitor, activator, or suppresesor, but it's labeled as
"Inhibitor 1" in the simulator,} \item{"inhibitor 2" for the 2nd inhibitor
listed in the simulation,} \item{"inhibitor 1 metabolite" for the primary
metabolite of inhibitor 1} \item{"conjugated protein" for DAR1-DARmax for
an antibody-drug conjugate; observed data with DV listed as "Conjugated
Protein Plasma Total" will match these simulated data,} \item{"total
protein" for DAR0-DARmax for an ADC; observed data with DV listed as "Total
Protein Conjugate Plasma Total" will match these simulated data,}
\item{"released payload" for the released drug from an ADC, which shows up
as primary metabolite 1 in Simulator output files.}} \strong{Note:} If your
compound is a therapeutic protein or ADC, we haven't tested this very
thoroughly, so please be extra careful to check that you're getting the
correct data.}

\item{tissue}{From which tissue should the desired concentrations be
extracted? Default is plasma for typical plasma concentration-time data.
Other options are "blood" or any tissues included in "Sheet Options",
"Tissues" in the simulator. All possible options:\describe{
\item{First-order absorption models}{"plasma", "blood", "unbound blood",
"unbound plasma", "additional organ", "adipose", "bone", "brain",
"feto-placenta", "gut tissue", "heart", "kidney", "liver", "lung", "muscle",
"pancreas", "peripheral blood", "peripheral plasma", "peripheral unbound
blood", "peripheral unbound plasma", "portal vein blood", "portal vein
plasma", "portal vein unbound blood", "portal vein unbound plasma", "skin",
or "spleen".} \item{ADAM-models}{"stomach", "duodenum", "jejunum I",
"jejunum II", "ileum I", "ileum II", "ileum III", "ileum IV", "colon",
"faeces", "gut tissue", "cumulative absorption", "cumulative fraction
released", or "cumulative dissolution".} \item{ADC simulations}{NOT YET
SET UP. If you need this, please contact Laura Shireman.}} Not case sensitive.}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} or \code{\link{extractExpDetails}} to
get all the details from the "Input Sheet" (e.g., when you ran
extractExpDetails you said \code{exp_details = "Input Sheet"} or
\code{exp_details = "all"}), you can save some processing time by supplying
that object here, unquoted. If left as NA, this function will run
\code{extractExpDetails} behind the scenes to figure out some information
about your experimental set up.}

\item{returnAggregateOrIndiv}{Return aggregate and/or individual simulated
concentration-time data? Options are "aggregate", "individual", or "both"
(default). Aggregated data are not calculated here but are pulled from the
simulator output rows labeled as "Population Statistics".}
}
\value{
a data.frame of concentration-time data
}
\description{
Extract concentration-time data from a Simcyp Simulator database file --
UNDER CONSTRUCTION!!!
}
\examples{
# none yet
}
