% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractConcTime_mult.R
\name{extractConcTime_mult}
\alias{extractConcTime_mult}
\title{Pull concentration-time data from multiple Simcyp Simulator output files}
\usage{
extractConcTime_mult(
  sim_data_files,
  obs_data_files = NA,
  sim_obs_dataframe = "ConcTime",
  overwrite = FALSE,
  tissues = "plasma",
  compoundsToExtract = "substrate",
  conc_units_to_use = "ng/mL",
  time_units_to_use = "hours",
  returnAggregateOrIndiv = "aggregate",
  ...
)
}
\arguments{
\item{sim_data_files}{a character vector of the files you'd like to compare,
e.g., \code{c("MyFile1.xlsx", "MyFile2.xlsx")}. The path should be included
with the file names if they are located somewhere other than your working
directory.}

\item{obs_data_files}{UNDER CONSTRUCTION and not yet working. a character
vector of the observed data files that you'd like to compare, e.g.,
\code{c("MyObsFile1.xlsx", "MyObsFile2.xlsx")}. The path should be included
with the file names if they are located somewhere other than your working
directory. This is the file that it is ready to be converted to an XML
file, not the file that contains only the digitized time and concentration
data. The names of the observed data files are piped into
\code{\link{extractObsConcTime}}.}

\item{sim_obs_dataframe}{the data.frame that will contain the output. Because
we can see scenarios where you might want to extract some
concentration-time data, play around with those data, and then later decide
you want to pull more concentration-time data for comparisons, this
data.frame can already exist. When that is the case, this function will
\emph{add} data to that data.frame. It will \emph{not} overwrite existing
data unless \code{overwrite} is set to TRUE.}

\item{overwrite}{TRUE or FALSE on whether to re-extract the
concentration-time data from output files that are already included in
\code{sim_obs_dataframe}. Since pulling data from Excel files is slow, by
default, this will \emph{not} overwrite existing data and instead will only
add data from any Excel files that aren't already included. A situation
where you might want to set this to TRUE would be when you have changed
input parameters for simulations and re-run them.}

\item{tissues}{From which tissue(s) should the desired concentrations be
extracted? The default is plasma for typical plasma concentration-time
data. Other options are "blood" or any tissues included in "Sheet Options",
"Tissues" in the simulator. All possible options: "plasma", "blood",
"unbound blood", "unbound plasma", "additional organ", "adipose", "bone",
"brain", "feto-placenta", "GI tissue", "heart", "kidney", "liver", "lung",
"muscle", "pancreas", "peripheral blood", "peripheral plasma", "peripheral
unbound blood", "peripheral unbound plasma", "portal vein blood", "portal
vein plasma", "portal vein unbound blood", "portal vein unbound plasma",
"skin", or "spleen". Not case sensitive. List all tissues desired as a
character vector, e.g., \code{c("plasma", "blood", "liver")}.}

\item{compoundsToExtract}{For which compound(s) do you want to extract
concentration-time data? Options are "substrate" (default), "primary
metabolite 1", "primary metabolite 2", "secondary metabolite", "inhibitor
1" (this can be an inducer, inhibitor, activator, or suppresesor, but it's
labeled as "Inhibitor 1" in the simulator), "inhibitor 2" for the 2nd
inhibitor listed in the simulation, or "inhibitor 1 metabolite" for the
primary metabolite of inhibitor 1. List all desired compounds as a
character vector, e.g., \code{c("substrate", "primary metabolite 1")}.
\emph{Note:} The simulator will report up to one metabolite for the 1st
inhibitor but no other inhibitor metabolites. (Someone please correct me if
that's wrong! -LS)}

\item{conc_units_to_use}{concentration units to use so that all data will be
comparable. Options are the same as the ones in the Excel form for PE data
entry. Default is "ng/mL". NOTE: ADAM model data concentration units are
not converted because there are simply too many units to manage easily, so
please check that the units are what you expected in the end.}

\item{time_units_to_use}{time units to use so that all data will be
comparable. Options are "hours" or "minutes". Default is "hours".}

\item{returnAggregateOrIndiv}{}

\item{...}{other arguments passed to the function
\code{\link{extractConcTime}}}
}
\value{
a large data.frame with multiple sets of concentration-time data,
  formatted the same way as output from the function
  \code{\link{extractConcTime}}
}
\description{
\code{extractConcTime_mult} is meant to be used in conjunction with
\code{\link{ct_plot_overlay}} to create single graphs with overlaid
concentration-time data from multiple Simcyp Simulator output files for easy
comparisons. \strong{A couple of notes:}\enumerate{\item{If ANY of the Excel
files you wish to extract data from are open, this WILL CRASH and WILL NOT
save whatever progress it has made so far. BE SURE TO CLOSE ALL OF THE
SOURCE-DATA EXCEL FILES.} \item{If you list multiple files, multiple tissues,
and/or multiple compounds to extract (see options below), this will extract
\emph{all} possible variations of them. For example, if you ask for "File A"
and "File B" and then also ask for "substrate" and "primary metabolite 1",
you will get the substrate and primary metabolite 1 data from \emph{both}
files.}}
}
\examples{
ConcTimeData <-
      extractConcTime_mult(
            sim_data_files = c("MyFile1.xlsx", "MyFile2.xlsx"),
            sim_obs_dataframe = "ConcTimeData",
            overwrite = FALSE,
            tissue = "unbound plasma") # Note that "tissue" is passed to "extractConcTime".

}
