% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpDetails.R
\name{extractExpDetails}
\alias{extractExpDetails}
\title{Extract details about the experimental design}
\usage{
extractExpDetails(sim_data_file, exp_details = "Summary tab")
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output}

\item{exp_details}{Experiment details you want to extract from the simulator
  output file "Summary" or "Input Sheet" tabs or the tab containing
  information about that population (the tab with the same name as the
  population simulated). Options are "Summary tab" to extract details only
  from the "Summary tab" (default), "Input Sheet" to extract details only
  from the "Input Sheet" tab, "all" to extract all possible parameters, or
  any combination of the following:

  \describe{

  \item{Abs_model_sub, Abs_model_inhib, Abs_model_met1,
  Abs_model_met2}{absorption model used, e.g., "1st order", for either the
  substrate, inhibitor, metabolite 1, or metabolite 2}

  \item{Age_min, Age_max}{Minimum or maximum age in simulated population}

  \item{AGP_male or AGP_female}{AGP mean value for males or females. Values
  are pulled from the tab with information on the population simulated.
  Specifying "AGP" will return data for both sexes.}

  \item{BPratio_sub, BPratio_inhib, BPratio_met1,
  BPratio_met2}{blood-to-plasma ratio}

  \item{CLint_sub, CLint_inhib, CLint_met1, CLint_met2}{intrinsic clearance,
  Vmax, Km, fu_mic, and/or half life values used for any CYPs, UGTs, or other
  enzymes listed for the substrate, inhibitor, metabolite 1, or metabolite 2.
  Output will be labeled for each enzyme and pathway as, e.g.,
  "CLint_sub_CYP3A4_1-OH" or "Vmax_sub_UGT1A1_Pathway1". Specify, e.g.,
  "CLint_sub" and all the other values (Vmax, Km, fu,mic, half life) will
  also be returned.}

  \item{CLrenal_sub, CLrenal_inhib, CLrenal_met1, CLrenal_met2}{renal
  clearance (L/hr) for the substrate, inhibitor, metabolite 1, or metabolite
  2}

  \item{Dose_sub or Dose_inhib}{dose administered}

  \item{DoseInt_sub or DoseInt_inhib}{dosing interval}

  \item{DoseRoute_sub or DoseRoute_inhib}{dose route, e.g. oral}

  \item{fa_sub, fa_inhib, fa_met1, fa_met2}{user input value for the fraction
  absorbed for the substrate, inhibitor, metabolite 1, or metabolite 2}

  \item{fu_gut_sub, fu_gut_inhib, fu_gut_met1, fu_gut_met2}{user input value
  for the fraction escaping gut metabolism for the substrate, inhibitor,
  metabolite 1, or metabolite 2}

  \item{fu_sub, fu_inhib, fu_met1, fu_met2}{fraction unbound}

  \item{GIAbsModel_sub, GIAbsModel_inhib, GIAbsModel_met1,
  GIAbsModel_met2}{GI absorption model used}

  \item{Hematocrit or Haematocrit}{the hematocrit listed on the "Summary"
  tab}

  \item{Hematocrit_male, Hematocrit_female, Haematocrit_male, or
  Haematocrit_female}{the hematocrit listed on the tab with population info.
  Note from LS: Can someone tell me how this differs from the "Haematocrit"
  value listed on the Summary tab?}

  \item{HSA_male or HSA_female}{the HSA value for the sex specified --
  including C0, C1, and C2, where appropriate -- listed on the tab with
  population info. Specifying "HSA" will return all possible HSA details.}

  \item{Inhibitor}{inhibitor used, if applicable}

  \item{Interaction_sub, Interaction_inhib, Interaction_met1,
  Interaction_met2}{interaction parameters for any CYPs, UGTs, or other
  enzymes listed for the substrate, inhibitor, metabolite 1, or metabolite 2.
  Output will be labeled for each enzyme and interaction type.}

  \item{ka_sub, ka_inhib, ka_met1, ka_met2}{user input value for the
  absorption constant ka for the substrate, inhibitor, metabolite 1, or
  metabolite 2}

  \item{kin_sac_sub, kin_sac_inhib, kout_sac_sub, or kout_sac_inhib}{k in and
  k out for SAC (1/hr) for the substrate, inhibitor, metabolite 1, or
  metabolite 2}

  \item{kp_scalar_sub, kp_scalar_inhib, kp_scalar_met1, kp_scalar_met2}{kp
  scalar for the substrate, inhibitor, metabolite 1, or metabolite 2}

  \item{logP_sub, logP_inhib, logP_met1, logP_met2}{logP for the substrate,
  inhibitor, metabolite 1, or metabolite 2}

  \item{ModelType_sub, ModelType_inhib, ModelType_met1, ModelType_met2}{the
  type of model, e.g., full PBPK model}

  \item{MW_sub, MW_inhib, MW_met1, MW_met2}{molecular weight}

  \item{NumDoses_sub or NumDoses_inhib}{number of doses}

  \item{NumSubjTrial}{number of subjects per trial}

  \item{NumTrials}{the number of trials performed}

  \item{Ontogeny}{ontogeny profile used}

  \item{Papp_MDCK_sub, Papp_MDCK_inhib, Papp_MDCK_met1, Papp_MDCK_met2,
  Papp_Caco_sub, Papp_Caco_inhib, Papp_Caco_met1, Papp_Caco_met2}{Papp as
  determined in MDCKII or Caco-2 cells (10^-6 cm/s) for the substrate,
  inhibitor, metabolite 1, or metabolite 2}

  \item{Papp_calibrator_sub, Papp_calibrator_inhib, Papp_calibrator_met1,
  Papp_calibrator_met2}{Papp of the calibrator compound (10^-6 cm/s)}

  \item{PercFemale}{Percent of females in simulated population}

  \item{pKa1_sub, pKa2_sub, pKa1_inhib, pKa2_inhib, pKa1_met1, pKa1_met2,
  pKa2_met1, pKa2_met2}{the pertinent pKa}

  \item{Pop}{the population modeled}

  \item{PopSize}{the size of population modeled}

  \item{PrandialSt_sub or PrandialSt_inhib}{prandial state upon dosing}

  \item{Qgut_sub or Qgut_inhib}{Qgut_sub (L/hr) for the substrate or
  inhibitor}

  \item{Regimen_sub or Regimen_inhib}{dosing regimen}

  \item{SimEndDayTime}{ending day and time of the simulation}

  \item{SimStartDayTime}{starting day and time of the simulation}

  \item{SimulatorVersion}{Simulator version number}

  \item{StartDayTime_sub or StartDayTime_inhib}{Starting day and time for
  administering the substrate or inhbitor}

  \item{StartHr_sub or StartHr_inhib}{Starting time in hours since the start
  of the simulation for administering the substrate or inhibitor.
  \strong{NOTE:} Unlike the other parameters listed here, this is
  \emph{calculated} from StartDayTime_sub or StartDayTime_inhib rather than
  being simply pulled from the simulator output file.}

  \item{StudyDuration}{study duration}

  \item{Substrate}{the substrate used}

  \item{tlag_sub or tlag_inhib}{user input value for the lag time for the
  substrate or inhibitor}

  \item{Type_sub or Type_inhib}{type of compound, e.g., monoprotic base}

  \item{Units_dose_sub, Units_dose_inhib, Units_AUC, Units_Cmax, Units_tmax,
  Units_CL}{Units for substrate dose, inhibitor dose, AUC, Cmax, tmax, or CL}

  \item{UserAddnOrgan_sub or UserAddnOrgan_inhib}{yes or no: Was a
  user-defined additional organ included for the substrate, inhibitor,
  metabolite 1, or metabolite 2?}

  \item{Vsac_sub, Vsac_inhib, Vsac_met1, Vsac_met2}{V sac (L/kg) for the
  substrate, inhibitor, metabolite 1, or metabolite 2}

  \item{Vss_input_sub, Vss_input_inhib, Vss_input_met1, Vss_input_met2}{Vss
  used}

  \item{VssPredMeth_sub, VssPredMeth_inhib, VssPredMeth_met1,
  VssPredMeth_met2}{method used for predicting Vss}

  } \emph{NOTE:} The default pulls only parameters that are listed on the
  "Summary" tab. (There are ~50 of them, so I'm not listing them here for
  brevity. -LS)}
}
\value{
Returns a named list of the experimental details
}
\description{
\code{extractExpDetails} looks up experimental design details from the
"Summary" or "Input Sheet" tabs of a Simcyp simulator output file. This is
currently set up to extract details on the substrate, the first inhibitor,
metabolite 1, and/or metabolite 2.
}
\examples{

extractExpDetails(sim_data_file = "../Example simulator output.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output MD + inhibitor.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output.xlsx",
                  exp_details = "all")



}
