% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpDetails.R
\name{extractExpDetails}
\alias{extractExpDetails}
\title{Extract details about the experimental design}
\usage{
extractExpDetails(sim_data_file, exp_details = "Summary tab")
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output}

\item{exp_details}{Experiment details you want to extract from the simulator
  output file. Options are \describe{

  \item{"Summary tab"}{Extract details only from the "Summary tab" (default)}

  \item{"Input Sheet"}{Extract details only from the "Input Sheet" tab}

  \item{"population tab"}{Extract details about the population used (data
  come from the tab with the same name as the population simulated)}

  \item{"all"}{Extract all possible parameters}

  \item{a string of the specific parameters you want}{For a complete list,
  type \code{data(AllExpDetails)} into the console. Parameters are reported
  with a suffix depending on which compound they pertain to: "_sub" for the
  substrate, "_met1" for the primary metabolite, "_met2" for the second
  primary metabolite, "_secmet" for the secondary metabolite, "_inhib" for
  the 1st inhibitor or inducer listed, "_inhib2" for the 2nd inhibitor or
  inducer listed, or "_inh1met" for the inhibitor 1 metabolite. An example of
  acceptable input: \code{c("pKa1_sub", "fa_inhib2", "Regimen_sub")}}}

  \emph{NOTE:} The default pulls only parameters that are listed on the
  "Summary" tab. If you want experimental details on a second inhibitor or
  more information on metabolites, try pulling them from the "Input sheet"
  instead of the "Summary" tab, which doesn't have as much information on
  those compounds.}
}
\value{
Returns a named list of the experimental details
}
\description{
\code{extractExpDetails} looks up experimental design details from the
"Summary" or "Input Sheet" tabs or the population tab of a Simcyp simulator
output file.
}
\examples{

extractExpDetails(sim_data_file = "../Example simulator output.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output MD + inhibitor.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output.xlsx",
                  exp_details = "all")



}
