% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpDetails.R
\name{extractExpDetails}
\alias{extractExpDetails}
\title{Extract details about the experimental design}
\usage{
extractExpDetails(
  sim_data_file,
  exp_details = c("BPratio_sub", "BPratio_inhib", "Dose_sub", "Dose_inhib",
    "DoseInt_sub", "DoseInt_inhib", "DoseRoute_sub", "DoseRoute_inhib", "DoseUnits_sub",
    "DoseUnits_inhib", "fu_sub", "fu_inhib", "GIAbsModel_sub", "GIAbsModel_inhib",
    "Hematocrit", "Haematocrit", "Inhibitor", "logP_sub", "logP_inhib", "ModelType_sub",
    "ModelType_inhib", "MW_sub", "MW_inhib", "NumDoses_sub", "NumDoses_inhib",
    "NumTrials", "NumSubjTrial", "pKa1_sub", "pKa1_inhib", "pKa2_sub", "pKa2_inhib",
    "Pop", "PopSize", "PrandialSt_sub",      "PrandialSt_inhib", "Regimen_sub",
    "Regimen_inhib", "SimEndDayTime", "SimStartDayTime", "StartDayTime_sub",
    "StartDayTime_inhib", "StudyDuration", "Substrate", "Type_sub", "Type_inhib",
    "VssInput_sub", "VssInput_inhib", "VssPredMeth_sub", "VssPredMeth_inhib")
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output}

\item{exp_details}{Experiment details you want to extract from the simulator
  output file "Summary" or "Input Sheet" tabs. Options are "all" to extract
  all possible parameters or any combination of the following:

  \describe{

  \item{AbsorptionModel}{absorption model used, e.g., "1st order"}

  \item{BPratio_sub or BPratio_inhib}{blood-to-plasma ratio}

  \item{CLint}{intrinsic clearance, Vmax, Km, or half life values used for
  any CYPs, UGTs, or other enzymes listed. Output will be labeled for each
  enzyme and pathway as, e.g., "CLint_CYP3A4_1-OH" or
  "Vmax_UGT1A1_Pathway1".}

  \item{Dose_sub or Dose_inhib}{dose administered}

  \item{DoseInt_sub or DoseInt_inhib}{dosing interval}

  \item{DoseRoute_sub or DoseRoute_inhib}{dose route, e.g. oral}

  \item{DoseUnits_sub or DoseUnits_inhib}{dose units}

  \item{fa_input}{user input value for the fraction absorbed}

  \item{fu_gut_input}{user input value for the fraction escaping gut
  metabolism}

  \item{fu_sub or fu_inhib}{fraction unbound}

  \item{Hematocrit or Heamatocrit}{the hematocrit}

  \item{GIAbsModel_sub or GIAbsModel_inhib}{GI absorption model used}

  \item{Inhibitor}{inhibitor used, if applicable}

  \item{ka_input}{user input value for the absorption constant ka}

  \item{lag_input}{user input value for the lag time}

  \item{logP_sub or logP_inhib}{logP of substrate or inhibitor}

  \item{ModelType_sub or ModelType_inhib}{the type of model, e.g., full PBPK
  model}

  \item{MW_sub or MW_inhib}{molecular weight of substrate or inhibitor}

  \item{NumDoses_sub or NumDoses_inhib}{number of doses}

  \item{NumSubjTrial}{number of subjects per trial}

  \item{NumTrials}{the number of trials performed}

  \item{Peff}{Peff,man Cap(10^-4 cm/s)}

  \item{pKa1_sub, pKa2_sub, pKa1_inhib, or pKa2_inhib}{the pertinent pKa}

  \item{Pop}{the population modeled}

  \item{PopSize}{the size of population modeled}

  \item{PrandialSt_sub or PrandialSt_inhib}{prandial state upon dosing}

  \item{Regimen_sub or Regimen_inhib}{dosing regimen}

  \item{SimEndDayTime}{ending day and time of the simulation}

  \item{SimStartDayTime}{starting day and time of the simulation}

  \item{SimulatorVersion}{Simulator version number}

  \item{StartDayTime_sub or StartDayTime_inhib}{Starting day and time for
  administering the substrate or inhbitor}

  \item{StudyDuration}{study duration}

  \item{Substrate}{the substrate used}

  \item{Type_sub or Type_inhib}{type of compound, e.g., monoprotic base}

  \item{UserAddnOrgan}{yes or no: Was a user-defined additional organ
  included?}

  \item{VssInput_sub or VssInput_inhib}{Vss used}

  \item{VssPredMeth_sub or VssPredMeth_inhib}{method used for predicting Vss}

  } \emph{NOTE:} The default only pulls parameters that are listed on the
  "Summary" tab. (There are 50 of them, so I'm not listing them here for
  brevity. -LS)}
}
\value{
Returns a named list of the experimental details
}
\description{
\code{extractExpDetails} looks up experimental design details from the
"Summary" or "Input Sheet" tabs of a Simcyp simulator output file.
}
\examples{

extractExpDetails(sim_data_file = "../Example simulator output.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output MD + inhibitor.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output.xlsx",
                  exp_details = "all")



}
