% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpDetails.R
\name{extractExpDetails}
\alias{extractExpDetails}
\title{Extract details about the experimental design}
\usage{
extractExpDetails(sim_data_file, exp_details = "Summary tab")
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output}

\item{exp_details}{Experiment details you want to extract from the simulator
  output file "Summary" or "Input Sheet" tabs. Options are "Summary tab" to
  extract details only from the "Summary tab" (default), "Input Sheet" to
  extract details only from the "Input Sheet" tab, "all" to extract all
  possible parameters, or any combination of the following:

  \describe{

  \item{AbsorptionModel}{absorption model used, e.g., "1st order"}

  \item{Age_min, Age_max}{Minimum or maximum age in simulated population}

  \item{BPratio_sub or BPratio_inhib}{blood-to-plasma ratio}

  \item{CLint}{intrinsic clearance, Vmax, Km, fu_mic, and/or half life values
  used for any CYPs, UGTs, or other enzymes listed. Output will be labeled
  for each enzyme and pathway as, e.g., "CLint_CYP3A4_1-OH" or
  "Vmax_UGT1A1_Pathway1". Specify "CLint" and all the other values (Vmax, Km,
  fu,mic, half life) will also be returned.}

  \item{Dose_sub or Dose_inhib}{dose administered}

  \item{DoseInt_sub or DoseInt_inhib}{dosing interval}

  \item{DoseRoute_sub or DoseRoute_inhib}{dose route, e.g. oral}

  \item{fa_input}{user input value for the fraction absorbed}

  \item{fu_gut_input}{user input value for the fraction escaping gut
  metabolism}

  \item{fu_sub or fu_inhib}{fraction unbound}

  \item{Hematocrit or Heamatocrit}{the hematocrit}

  \item{GIAbsModel_sub or GIAbsModel_inhib}{GI absorption model used}

  \item{Inhibitor}{inhibitor used, if applicable}

  \item{ka_input}{user input value for the absorption constant ka}

  \item{kin_sac and kout_sac}{k in and k out for SAC (1/hr)}

  \item{kp_scalar}{kp scalar}

  \item{tlag_input}{user input value for the lag time}

  \item{logP_sub or logP_inhib}{logP of substrate or inhibitor}

  \item{ModelType_sub or ModelType_inhib}{the type of model, e.g., full PBPK
  model}

  \item{MW_sub or MW_inhib}{molecular weight of substrate or inhibitor}

  \item{NumDoses_sub or NumDoses_inhib}{number of doses}

  \item{NumSubjTrial}{number of subjects per trial}

  \item{NumTrials}{the number of trials performed}

  \item{Ontogeny}{ontogeny profile used}

  \item{Peff}{Peff,human Cap(10^-4 cm/s)}

  \item{PercFemale}{Percent of females in simulated population}

  \item{pKa1_sub, pKa2_sub, pKa1_inhib, or pKa2_inhib}{the pertinent pKa}

  \item{Pop}{the population modeled}

  \item{PopSize}{the size of population modeled}

  \item{PrandialSt_sub or PrandialSt_inhib}{prandial state upon dosing}

  \item{Qgut}{Qgut (L/hr)}

  \item{Regimen_sub or Regimen_inhib}{dosing regimen}

  \item{SimEndDayTime}{ending day and time of the simulation}

  \item{SimStartDayTime}{starting day and time of the simulation}

  \item{SimulatorVersion}{Simulator version number}

  \item{StartDayTime_sub or StartDayTime_inhib}{Starting day and time for
  administering the substrate or inhbitor}

  \item{StudyDuration}{study duration}

  \item{Substrate}{the substrate used}

  \item{Type_sub or Type_inhib}{type of compound, e.g., monoprotic base}

  \item{Units_dose_sub, Units_dose_inhib, Units_AUC, Units_Cmax, Units_tmax,
  Units_CL}{Units for substrate dose, inhibitor dose, AUC, Cmax, tmax, or CL}

  \item{UserAddnOrgan}{yes or no: Was a user-defined additional organ
  included?}

  \item{Vsac}{V sac (L/kg)}

  \item{Vss_input_sub or Vss_input_inhib}{Vss used}

  \item{VssPredMeth_sub or VssPredMeth_inhib}{method used for predicting Vss}

  } \emph{NOTE:} The default only pulls parameters that are listed on the
  "Summary" tab. (There are ~50 of them, so I'm not listing them here for
  brevity. -LS)}
}
\value{
Returns a named list of the experimental details
}
\description{
\code{extractExpDetails} looks up experimental design details from the
"Summary" or "Input Sheet" tabs of a Simcyp simulator output file.
}
\examples{

extractExpDetails(sim_data_file = "../Example simulator output.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output MD + inhibitor.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output.xlsx",
                  exp_details = "all")



}
