% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpDetails.R
\name{extractExpDetails}
\alias{extractExpDetails}
\title{Extract details about the experimental design}
\usage{
extractExpDetails(
  sim_data_file,
  exp_details = "Summary and Input",
  save_output = NA
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes. \strong{A note:} There are just a few items that we will attempt
to extract from the matching workspace file; for that information, we will
look for a workspace file that is named \emph{identically} to the Excel
file except for the file extension. It will ignore the date/time stamp that
the autorunner adds as long as that stamp is in a format like this: "myfile
- 2023-10-31 07-23-15.xlsx".}

\item{exp_details}{Experiment details you want to extract from the simulator
  output file. Options are \describe{

  \item{"Summary and Input"}{Extract details available from the "Summary tab"
  and the "Input Sheet" (default)}

  \item{"population tab"}{Extract details about the population used (data
  come from the tab with the same name as the population simulated)}

  \item{"Simcyp inputs"}{Extract all the details that you normally fill out
  on the "Simcyp inputs (and QC)" tab of a compound data sheet plus trial
  design information}

  \item{"workspace"}{Extract a limited set of details directly
  from the Simcyp Simulator workspace files. The set of possible details may
  be viewed by entering \code{view(AllWorkspaceDetails)} in the console. This
  \emph{only} works if the workspace file name perfectly matches the Excel
  results file name and is located in the same folder. Otherwise, this step
  in the data extraction will be skipped.}

  \item{"all"}{Extract all possible parameters}}

  \strong{NOTES:} \enumerate{\item{The default pulls parameters from the
  "Summary" tab and the "Input Sheet" tab. Note that the
  "Summary" tab does not include information on any compounds beyond the
  substrate and inhibitor 1.} \item{There are a few places where
  requesting one item as input will get you multiple items as output:
  intrinsic clearance, interaction parameters, and transport parameters. For
  example, if you request intrinsic clearance values (ex: "CLint_sub"),
  you'll get all the intrinsic clearance values for that compound, and
  they'll be named according to which parameter it is, which enzyme it's for,
  etc. Same thing with requesting interaction parameters (ex:
  "Interaction_inhib" to get all the interaction parameters for inhibitor 1)
  and transporter parameters (ex: "Transport_sub").}}}

\item{save_output}{optionally save the output by supplying a csv or Excel
file name in quotes here, e.g., "Simulation details.csv" or "Simulation
details.xlsx".  Do not include any slashes, dollar signs, or periods in the
file name. If you leave off the file extension, it will be saved as a csv
file.}
}
\value{
Returns a named list of the experimental details
}
\description{
\code{extractExpDetails} looks up experimental design details from a Simcyp
Simulator output file. For detailed instructions and examples, please see the
SharePoint file "Simcyp PBPKConsult R Files - Simcyp PBPKConsult R
Files/SimcypConsultancy function examples and instructions/Checking
simulation experimental
details/Checking-simulation-experimental-details.docx". (Sorry, we are unable
to include a link to it here.)
}
\examples{

extractExpDetails(sim_data_file = "../Example simulator output.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output MD + inhibitor.xlsx")
extractExpDetails(sim_data_file = "../Example simulator output.xlsx",
                  exp_details = "all")



}
