% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpDetails_DB.R
\name{extractExpDetails_DB}
\alias{extractExpDetails_DB}
\title{Extract simulation experimental details from a database file - UNDER
CONSTRUCTION!!!!!}
\usage{
extractExpDetails_DB(sim_data_file)
}
\arguments{
\item{sim_data_file}{the database file to get information from}
}
\value{
a list of information, same as other extractExpDetails_x functions
}
\description{
\code{extractExpDetails_DB} reads a Simcyp Simulator database file for
information about how the simulation was set up. This uses the Simcyp package
under the hood.
}
\examples{
# none yet

}
