% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFmFe.R
\name{extractFmFe}
\alias{extractFmFe}
\title{Extract fm and fe valuesthat change with time from a Simulator output Excel
file}
\usage{
extractFmFe(
  sim_data_file,
  returnOnlyMaxMin = TRUE,
  returnAggregateOrIndiv = "both",
  existing_exp_details = NA
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulated dynamic
fm and fe data, in quotes}

\item{returnOnlyMaxMin}{TRUE (default) or FALSE for whether to return only
maximum and minimum fm values -- basically, return the table in the upper
left corner of the "Time variant \%fm and fe" tab.}

\item{returnAggregateOrIndiv}{Return aggregate and/or individual simulated fm
and fe data? This currently only applies to the dynamic fms and has not yet
been developed for the static fms, which will only return aggregate data at
present. Options are "individual", "aggregate", or "both" (default).
Aggregated data are not calculated here but are pulled from the simulator
output rows labeled as "mean".}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} or \code{\link{extractExpDetails}} to
get all the details from the "Input Sheet" (e.g., when you ran
extractExpDetails you said \code{exp_details = "Input Sheet"} or
\code{exp_details = "all"}), you can save some processing time by supplying
that object here, unquoted. If left as NA, this function will run
\code{extractExpDetails} behind the scenes to figure out some information
about your experimental set up.}
}
\value{
a data.frame
}
\description{
\code{extractFmFe} extracts fm and fe data from a simulator output Excel
file. A tab named something like "Time variant \%fm and fe" must be present
to get dynamic fm and fe values, but you can get the static values if the tab
"% fm and fe SS" is present. Dynamic fm and fe data are required if you want
to use these data with the function \code{\link{fm_plot}} to make a graph of
how the fm values change over time.
}
\examples{
# None yet

}
