% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractForestData.R
\name{extractForestData}
\alias{extractForestData}
\title{Extract pertinent data from Simulator output files for creating forest plots}
\usage{
extractForestData(
  sim_data_files = NA,
  PKparameters = c("AUCinf_ratio_dose1", "AUCt_ratio_dose1", "Cmax_ratio_dose1",
    "AUCtau_ratio_last", "Cmax_ratio_last"),
  sheet = NA,
  checkDataSource = TRUE,
  save_output = NA
)
}
\arguments{
\item{sim_data_files}{a character vector of simulator output files; leaving
this as the default NA will extract forest data from all simulator files in
the current working directory}

\item{PKparameters}{PK parameters to extract from simulator output files;
default is all possible AUC and Cmax geometric mean ratios for both dose 1
and the last dose simulated. Input must be among "AUCinf_ratio_dose1",
"AUCt_ratio_dose1", "Cmax_ratio_dose1", "AUCtau_ratio_last", or
"Cmax_ratio_last". List them in the order you'd like the columns to appear
in the output.}

\item{sheet}{optionally specify the name of the sheet where you'd like to
pull the PK data, in quotes; for example, specify the tab where you have a
user-defined AUC integration. \emph{Note:} Unless you want a very specific
Excel sheet that's not what the usual sheet name would be for a first or
last dose, this function will work best if this is left as NA. Also, since
we don't know which dose these data were for, you'll see that the output
parameter names do not include the suffixes "_last" or "_dose1".}

\item{checkDataSource}{TRUE (default) or FALSE: Include in the output a
data.frame that lists exactly where the data were pulled from the simulator
output file. Useful for QCing.}

\item{save_output}{optionally save the output by supplying a file name in
quotes here, e.g., "My forest graph data.csv". If you leave off ".csv", it
will still be saved as a csv file.}
}
\description{
\code{extractForestData} automatically extract the data needed for generating
forest plots from Simulator output files and format it for using the forest
plot shiny app. This will take some time to run since it needs to open
multiple Excel files. NB: This has only been set up to work with geometric
means so far.
}
\examples{

extractForestData(sim_data_files = NA,
                  save_output = "Forest data.csv")

}
