% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractForestData.R
\name{extractForestData}
\alias{extractForestData}
\title{Extract pertinent data from Simulator output files for creating forest plots}
\usage{
extractForestData(
  sim_data_files,
  PKparameters = c("AUCinf_ratio_dose1", "AUCt_ratio_dose1", "Cmax_ratio_dose1",
    "AUCtau_ratio_ss", "Cmax_ratio_ss"),
  checkDataSource = TRUE,
  save_output = NA
)
}
\arguments{
\item{sim_data_files}{a character vector of simulator output files}

\item{PKparameters}{PK parameters to extract from simulator output files.
Defaults to all possible AUC and Cmax ratios. List them in the order you'd
like the columns to appear in the output.}

\item{checkDataSource}{TRUE or FALSE: Include in the output a data.frame that
lists exactly where the data were pulled from the simulator output file.
Useful for QCing.}

\item{save_output}{optionally save the output by supplying a file name in
quotes here, e.g., "My experimental details.csv". If you leave off ".csv",
it will still be saved as a csv file.}
}
\value{

}
\description{
Extract pertinent data from Simulator output files for creating forest plots
}
