% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractObsConcTime.R
\name{extractObsConcTime}
\alias{extractObsConcTime}
\title{Extract observed concentration-time data from an Excel file}
\usage{
<<<<<<< HEAD
extractObsConcTime(obs_data_file, returnDosingInfo = FALSE)
=======
extractObsConcTime(obs_data_file, add_t0 = FALSE, returnDosingInfo = FALSE)
>>>>>>> beta
}
\arguments{
\item{obs_data_file}{name of the Excel file containing the observed
concentration-time data, in quotes. This is the file that is \emph{ready}
to be converted to an XML file, not a file that contains only digitized
time and concentration data and not the XML file itself that you would
include in a Simulator workspace for observed data.}

<<<<<<< HEAD
=======
\item{add_t0}{TRUE or FALSE (default) for whether to add t0 points if they're
missing. Sometimes, observed data do not include a measurement at t0
because, presumably, the concentration should always be 0 at that time. If
you're using these data to calculate PK, though, you'll miss that initial
part of the AUC if t0 is missing. If \code{add_t0} is set to TRUE, this
will add a concentration of 0 and time 0 for all of the individual
concentration-time profiles.}

>>>>>>> beta
\item{returnDosingInfo}{TRUE or FALSE (default) for whether to return a
second data.frame with dosing and demographic information from the Excel
file.}
}
\value{
a data.frame or a list of two data.frames. The observed
  concentration-time data.frame, which is named "ObsCT" if the output is a
  list, has the following columns:
  \describe{\item{Individual}{the individual ID}

  \item{CompoundID}{the compound ID listed in the observed file, e.g., "Sub
  Plasma", "Sub PM1 Plasma", "Sub (Inb) Plasma" will become "substrate" and so on}

  \item{Tissue}{the tissue}

  \item{Time}{time since dosing}

  \item{Conc}{concentration}

  \item{Time_units}{the units of measurement for the time column}

  \item{Conc_units}{the units of measurement for the concentration column}

  \item{Period, Age, Weight_kg, Height_cm, Sex, SerumCreatinine_umolL,
  HSA_gL, Haematocrit, PhenotypeCYP2D6, SmokingStatus}{the columns in the
  template for "Period" and "Covariates" but with R-friendly names.}}
  If the user requested dosing information, the second item in the list will
  be a data.frame titled "ObsDosing" and will include both dosing and
  demographic information from the Excel file.
}
\description{
Extract observed data from an Excel file that follows the Simcyp Simulator
template for converting concentration-time data into an XML file.
}
\examples{
extractObsConcTime(obs_data_file = "My observed data.xlsx")

}
