% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPK.R
\name{extractPK}
\alias{extractPK}
\title{extractPK}
\usage{
extractPK(
  sim_data_file,
  PKparameters = c("AUCtau_dose1", "AUCtau_lastdose", "AUCinf_dose1", "HalfLife_dose1",
    "CL_dose1")
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output}

\item{PKparameters}{PK parameters you want to extract from the simulator
output file, all of which will be pulled from the "AUC" tab. Options are
any combination of the following: \describe{ \item{"AUCtau_dose1"}{AUC from
0 to tau for dose 1. This is the column "AUCt(0) (mg/L.h)".}
\item{"AUCtau_lastdose"}{AUC from 0 to tau for the last dose in the
simulation. This is the column "AUC (mg/L.h)" under the subheading "AUC
integrated from X to Y".} \item{"AUCinf_dose1}{AUC from 0 to infinity for
dose 1 (column: "AUC_INF (mg/L.h)")} \item{"HalfLife_dose1"}{half life,
estimated for dose 1 (column: "Half-life (h)")} \item{"CL_dose1"}{Clearance
as calculated by dose / AUCinf for dose 1 (column: "CL (Dose/AUC_INF)
(L/h)")} }}
}
\value{
Depending on the options selected, returns a list of numerical
  vectors of whichever PK parameters were chosen. If only one PK parameter
  was requested, output is a numerical vector.
}
\description{
Extract AUCinf, AUCtau, half life, and/or clearance data from the "AUC" tab
of a simulator output Excel file.
}
\examples{

sim_data_file <- "../Example simulator output.xlsx"
extractPK(sim_data_file)
extractPK(sim_data_file, PKparameters = "AUCinf_dose1")


}
