% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPK.R
\name{extractPK}
\alias{extractPK}
\title{extractPK}
\usage{
extractPK(
  sim_data_file,
  PKparameters = c("AUCinf_dose1", "AUCtau_dose1", "AUCtau_lastdose",
    "AUCtau_lastdoseToEnd", "AUCinf_dose1_withEffector", "AUCtau_lastdose_withEffector",
    "CL_dose1", "CL_dose1_withEffector", "CL_lastdose", "CL_lastdose_withEffector",
    "CL_lastdoseToEnd", "Cmax_dose1", "Cmax_dose1_withEffector", "Cmax_lastdose",
    "Cmax_lastdose_withEffector", "HalfLife_dose1", "tmax_dose1"),
  returnAggregateOrIndiv = "individual"
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output}

\item{PKparameters}{PK parameters you want to extract from the simulator
  output file. Options are any combination of the following:

  \describe{

  \item{"AUCinf_dose1"}{AUC from 0 to infinity for dose 1. Data are pulled
  from the tab "AUC", column titled, e.g., "AUC_INF (mg/L.h)"}

  \item{AUCinf_dose1_withEffector}{AUC from 0 to infinity for dose 1 in the
  presence of an effector (Inhibitor 1 in the simulator). Data are pulled
  from the tab "AUC", column titled, e.g., "AUC_INF_Inh (mg/L.h)" after the
  subheading "Extrapolated AUC_INF for the first dose in the presence of
  inhibitor"}

  \item{"AUCtau_dose1"}{AUC from 0 to tau for dose 1. Data are pulled from
  tab "AUC0(Sub)(CPlasma)", column titled, e.g., "AUC (mg/L.h)". IMPORTANT:
  This will be AUCtau for dose 1 if you have a done a multiple-dose
  simulation, but, if you have done a single-dose simulation, this will be
  the AUC from 0 to whatever time you stopped simulating.}

  \item{"AUCtau_lastdose"}{AUC from 0 to tau for the last dose in the
  simulation. Data are pulled from tab "AUC" from the column titled, e.g.,
  "AUC (mg/L.h)" under the subheading "Truncated AUCt for the last dose".
  Nota bene: These data were integrated by the simulator using whatever
  sampling points you selected in the simulation. If you did not adequately
  sample the dosing interval, the estimate could be off. To get this
  parameter, you must have checked the box for "AUCt for the last dose" to be
  calculated under Outputs-->Data Analysis-->AUC to be calculated}

  \item{AUCtau_lastdose_withEffector}{AUC from 0 to tau for the last dose in
  the presence of an effector (Inhibitor 1 in the simulator). Data are pulled
  from the tab "AUC", column titled, e.g., "AUCt(n)_Inh (mg/L.h)" after the
  subheading "Truncated AUCt for the last dose in the presence of inhibitor"}

  \item{"AUCtau_lastdoseToEnd"}{AUC from the last dose to the end of the
  simulation. Data are pulled from tab "AUCX(Sub)(CPlasma)", where "X" is the
  largest dose for which there is a tab, from the column titled, e.g., "AUC
  (mg/L.h)". \emph{Nota bene:} These data were calculated by the simulator
  from the beginning of the last dose to whenever the simulation ended, so
  that interval may not be tau if you did not set up the simulation that way.
  See the options for "AUCtau_lastdose" as an alternative.}

  \item{"CL_dose1"}{Clearance as calculated by dose / AUCinf for dose 1. Data
  are pulled from the tab "AUC" and the column titled, e.g., "CL
  (Dose/AUC_INF) (L/h)", subheading "Extrapolated AUC_INF for the first
  dose".}

  \item{"CL_dose1_withEffector"}{Clearance as calculated by dose / AUCinf for
  dose 1 in the presence of an effector (Inhibitor 1 in the simulator). Data
  are pulled from the tab "AUC" and the column titled, e.g., "CL
  (Dose/AUC_INF_Inh) (L/h)", subheading "Extrapolated AUC_INF for the first
  dose in the presence of inhibitor".}

  \item{"CL_lastdose"}{Clearance as calculated by dose / AUCtau for the last
  dose simulated. Data are pulled from the tab "AUC", column titled, e.g.,
  "CL (Dose/AUC) (L/h)")}

  \item{"CL_lastdose_withEffector"}{Clearance as calculated by dose / AUCtau
  for the last dose in the presence of an effector (Inhibitor 1 in the
  simulator). Data are pulled from the tab "AUC" and the column titled, e.g.,
  "CL (Dose/AUC_INF_Inh) (L/h)", subheading "Truncated AUCt for the last dose
  in the presence of inhibitor".}

  \item{"CL_lastdoseToEnd"}{CL for the last dose calculated as Dose / AUCtau.
  Data are pulled from tab "AUCX(Sub)(CPlasma)", where "X" is the largest
  dose for which there is a tab, from the column titled, e.g., "AUC
  (mg/L.h)". \emph{Nota bene:} These data were calculated by the simulator
  from the beginning of the last dose to whenever the simulation ended, so
  that interval may not be tau if you did not set up the simulation that way.
  See the options for "CL_lastdose_int" as an alternative.}

  \item{"Cmax_dose1"}{Cmax for dose 1. Data are pulled from tab
  "AUC0(Sub)(CPlasma)", column titled, e.g., "CMax (mg/L)".}

  \item{"Cmax_dose1_withEffector"}{Cmax for the last dose in the presence of
  an inhibitor. Data are pulled from tab "AUC0(Sub)(CPlasma)", column titled,
  e.g., "CMaxinh (mg/L)".}

  \item{"Cmax_lastdose"}{Cmax for the last dose. Data are pulled from tab
  "AUC", column titled, e.g., "CMax (mg/L)", under the subheading "Truncated
  AUCt for the last dose.}

  \item{"HalfLife_dose1"}{half life for dose 1. Data are pulled from the tab
  "AUC", column titled, e.g., "Half-life (h)")}

  \item{"tmax_dose1"}{tmax for dose 1. Data are pulled from tab
  "AUC0(Sub)(CPlasma)", column titled, e.g., "TMax (h)".} }}

\item{returnAggregateOrIndiv}{Return aggregate (geometric mean) and/or
individual PK parameters? Options are "aggregate" or "individual".}
}
\value{
Depending on the options selected, returns a list of numerical
  vectors of whichever PK parameters were chosen. If only one PK parameter
  was requested, output is a numerical vector.
}
\description{
Extract PK data for specific parameters from a simulator output Excel file.
}
\examples{

sim_data_file <- "../Example simulator output MD + inhibitor.xlsx"
extractPK(sim_data_file)
extractPK(sim_data_file, PKparameters = "AUCinf_dose1")


}
