% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPK.R
\name{extractPK}
\alias{extractPK}
\title{Extract PK data for specific parameters from a simulator output Excel file}
\usage{
extractPK(
  sim_data_file,
  PKparameters = "AUC tab",
  sheet = NA,
  tissue = "plasma",
  compoundID = "substrate",
  returnAggregateOrIndiv = "aggregate",
  includeTrialInfo = TRUE,
  returnExpDetails = FALSE,
  checkDataSource = TRUE
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes}

\item{PKparameters}{PK parameters you want to extract from the simulator
  output file. Options are: \describe{

  \item{"all"}{all possible parameters}

  \item{"AUC tab"}{only those parameters on the "AUC" tab (default). The
  "AUC_CI" tab or "AUC_SD" tab will be used if "AUC" tab is not present.}

  \item{"Absorption tab"}{only those parameters on the "Absorption" or
  "Overall Fa Fg" tab}

  \item{"Regional ADAM"}{regional fraction absorbed and fraction metabolized
  from intestinal segments; only applies to ADAM models where the tab
  "Regional ADAM Fractions (Sub)" is included in the Excel file and currently
  only applies to substrate}

  \item{a vector of any combination of specific, individual parameters, each
  surrounded by quotes and encapsulated with \code{c(...)}}{An example:
  \code{c("Cmax_dose1", "AUCtau_last")}. To see the full set of possible
  parameters to extract, enter \code{view(PKParameterDefinitions)} into the
  console. Not case sensitive. If you use "_first" instead of "_dose1", that
  will also work.}}

  Currently, the PK data are only for the substrate unless noted, although
  you can sometimes hack around this by supplying a specific sheet to extract
  for a compound other than the substrate, e.g. sheet = "AUC(Sub Pri Met1)".
  This has NOT been as well tested, though, so be sure to check that you're
  getting what you expected!}

\item{sheet}{optionally specify the name of the sheet where you'd like to
pull the PK data, in quotes; for example, specify the tab where you have a
user-defined AUC integration. \emph{Note:} Unless you want a very specific
Excel sheet that's not what the usual sheet name would be for a first or
last dose, this function will work best if this is left as NA. Also, since
we don't know which dose these data were for, you'll see that the output
parameter names do not include the suffixes "_last" or "_dose1".}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are "plasma" (default) or "blood" (possible but not as thoroughly
tested).}

\item{compoundID}{For which compund ID would you like the PK parameters to be
pulled? Options are "substrate", "primary metabolite 1", "primary
metabolite 2", "inhibitor 1", or "inhibitor 2". Note from the package
authors: We haven't found an instance where the Simulator exports PK
parameters for the secondary metabolite or the inhibitor 1 metabolite, so
those are not set up.}

\item{returnAggregateOrIndiv}{return aggregate (default) and/or individual PK
parameters? Options are "aggregate", "individual", or "both". For aggregate
data, values are pulled from simulator output -- not calculated -- and the
output will be a data.frame with the PK parameters in columns and the
statistics reported exactly as in the simulator output file.}

\item{includeTrialInfo}{TRUE or FALSE (default) for whether to include which
individual and trial the data describe. This only applies when
\code{returnAggregateOrIndiv} is "individual" or "both".}

\item{returnExpDetails}{TRUE or FALSE: Since, behind the scenes, this
function extracts experimental details from the "Summary" tab of the
simulator output file, would you like those details to be returned here? If
set to TRUE, output will be a list of 1. the requested PK data, 2. the
experimental details provided on the Summary tab, and, if you have set
checkDataSource to TRUE, 3. info for checking the data. (See
"checkDataSource".)}

\item{checkDataSource}{TRUE (default) or FALSE for whether to include in the
output a data.frame that lists exactly where the data were pulled from the
simulator output file. Useful for QCing.}
}
\value{
Depending on the options selected, returns a list of numerical
  vectors or a list of data.frames: "individual" and "aggregate". If
  \code{checkDataSource} is TRUE, this will also return a data.frame named
  "QC" that indicates where in the simulator output file the data came from.
}
\description{
Pull calculated PK parameters from a Simcyp simulation output Excel file.
\strong{Note:} Nearly all parameters are for the SUBSTRATE We're still
validating this for extracting PK for an effector. \strong{A request for
assistance:} If you extract PK data for an effector by specifying an Excel
sheet for that compound, please check the values and tell Laura Shireman how
well it works! Also, absorption parameters are for first-order absorption
models only; we haven't developed this yet to pull ADAM-model absorption
parameters, but it's on our to-do list.
}
\examples{

sim_data_file <- "../Example simulator output MD + inhibitor.xlsx"
extractPK(sim_data_file)
extractPK(sim_data_file, PKparameters = "Absorption tab")
extractPK(sim_data_file, PKparameters = c("AUCinf_dose1", "Cmax_dose1"))


}
