% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPK.R
\name{extractPK}
\alias{extractPK}
\title{Extract PK data for specific parameters from a simulator output Excel file}
\usage{
extractPK(
  sim_data_file,
  PKparameters = "AUC tab",
  sheet = NA,
  tissue = "plasma",
  returnAggregateOrIndiv = "aggregate",
  includeTrialInfo = TRUE,
  checkDataSource = TRUE
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output}

\item{PKparameters}{PK parameters you want to extract from the simulator
output file. Options are "all" for all possible parameters, "AUC tab" for
only those parameters on the "AUC" tab (default, "AUC_CI" tab will be used
if "AUC" is not present), "Absorption tab" for only those parameters on the
"Absorption" tab, or any combination of specific, individual parameters.
Currently, the PK data are only for the substrate unless noted. To see the
full set of possible parameters to extract, enter
\code{data(AllPKParameters)} into the console.}

\item{sheet}{optionally specify the name of the sheet where you'd like to
pull the PK data; for example, specify the tab where you have a
user-defined AUC integration. \emph{Note:} Unless you want a very specific
Excel sheet that's not what the usual sheet name would be for a first or
last dose, this function will work best if this is left as NA. Also, since
we don't know which dose these data were for, you'll see that the output
parameter names do not include the suffixes "_ss" or "_dose1".}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are "plasma" or "blood".}

\item{returnAggregateOrIndiv}{Return aggregate and/or individual PK
parameters? Options are "aggregate", "individual", or "both". For aggregate
data, values are pulled from simulator output -- not calculated -- and the
output will be a data.frame with the PK parameters in columns and the
statistics reported exactly as in the simulator output file.}

\item{includeTrialInfo}{TRUE or FALSE: Include which individual and trial the
data describe? This only applies when \code{returnAggregateOrIndiv}
includes "individual".}

\item{checkDataSource}{TRUE or FALSE: Include in the output a data.frame that
lists exactly where the data were pulled from the simulator output file.
Useful for QCing.}
}
\value{
Depending on the options selected, returns a list of numerical
  vectors or a list of data.frames: "individual" and "aggregate". If
  \code{checkDataSource} is TRUE, this will also return a data.frame named
  "QC" that indicates where in the simulator output file the data came from.
}
\description{
Pull calculated PK parameters from a Simcyp simulation output Excel file.
Note: Nearly all parameters are for the substrate.
}
\examples{

sim_data_file <- "../Example simulator output MD + inhibitor.xlsx"
extractPK(sim_data_file)
extractPK(sim_data_file, PKparameters = "Absorption tab")
extractPK(sim_data_file, PKparameters = "AUCinf_dose1")


}
