% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPK.R
\name{extractPK}
\alias{extractPK}
\title{Extract substrate PK data for specific parameters from a simulator output
Excel file}
\usage{
extractPK(
  sim_data_file,
  PKparameters = "AUC tab",
  sheet = NA,
  tissue = "plasma",
  returnAggregateOrIndiv = "aggregate",
  includeTrialInfo = TRUE
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output}

\item{PKparameters}{PK parameters you want to extract from the simulator
  output file. Options are "all" for all possible parameters, "AUC tab" for
  only those parameters on the "AUC" tab (default), "Absorption tab" for only
  those parameters on the "Absorption" tab, or any combination of the
  following:

  \describe{

  \item{AccumulationIndex, AccumulationIndex_withEffector}{Accumulation index
  (AUC infinity for dose 1 / AUC tau at steady state). By default, data are
  pulled from the sheet "AUC", column titled "Accumulation Index" or, for the
  accumulation index in the presence of an effector, "Accumulation
  Index_Inh". UNDER CONSTRUCTION.}

  \item{AccumulationRatio, AccumulationRatio_withEffector}{Accumulation ratio
  (AUC tau for dose 1 / AUC tau at steady state). By default, data are pulled
  from the sheet "AUC", column titled "Accumluation Ratio" or, for the
  accumulation ratio in the presence of an effector, "Accumulation
  Ratio_Inh".}

  \item{AUCinf_dose1, AUCinf_dose1_withEffector}{AUC from 0 to infinity for
  dose 1, in the absence or presence of an effector, respectively. By
  default, data are pulled from the sheet "AUC", column titled, e.g.,
  "AUC_INF (mg/L.h)"}

  \item{AUCinf_ratio_dose1, AUCtau_ratio_dose1, AUCtau_ratio_ss,
  Cmax_ratio_dose1, Cmax_ratio_ss}{The ratio of either the AUC or Cmax for
  dose 1 or at steady state with effector present / without effector
  present.}

  \item{AUCtau_dose1, AUCtau_dose1_withEffector, AUCtau_ss,
  AUCtau_ss_withEffector}{AUC from 0 to tau for dose 1 or the last dose in
  the absence or presence of an effector. By default, dose 1 data are pulled
  from sheet "AUC0(Sub)(CPlasma)", column titled, e.g., "AUC (mg/L.h)", and
  steady-state data are pulled from the "AUC" tab}

  \item{CL_dose1, CL_dose1_withEffector}{Clearance as calculated by dose /
  AUCinf for dose 1. Data are pulled from the sheet "AUC" and the column
  titled, e.g., "CL (Dose/AUC_INF) (L/h)", subheading "Extrapolated AUC_INF
  for the first dose".}

  \item{CL_ss, CL_ss_withEffector}{Clearance as calculated by dose / AUCtau
  for the last dose simulated. By default, data are pulled from the sheet
  "AUC", column titled, e.g., "CL (Dose/AUC) (L/h)")}

  \item{CL_hepatic}{I'm not actually positive that this is the total hepatic
  clearance, but I think that's what it is... By default, data are pulled
  from the "Clearance Trials SS" tab. Sorry for the uncertain explanation;
  I'll update this when I know better what this is! -LS}

  \item{Cmax_dose1, Cmax_dose1_withEffector}{Cmax for dose 1 with or without
  an effector. By default, data are pulled from sheet "AUC0(Sub)(CPlasma)",
  column titled, e.g., "CMax (mg/L)".}

  \item{Cmax_ss}{Cmax for the last dose. By default, data are pulled from
  sheet "AUC", column titled, e.g., "CMax (mg/L)", under the subheading
  "Truncated AUCt for the last dose.}

  \item{HalfLife_dose1}{half life for dose 1. By default, data are pulled
  from the sheet "AUC", column titled, e.g., "Half-life (h)")}

  \item{F_sub}{bioavailability (F) of substrate. By default, data are pulled
  from the sheet "Clearance Trials SS".}

  \item{fa_sub or fa_inhib}{fraction absorbed for the substrate or inhibitor.
  By default, data are pulled from the sheet "Absorption".}

  \item{fg_sub}{fraction of substrate escaping gut metabolism. By default,
  data are pulled from the sheet "Clearance Trials SS".}

  \item{fh_sub}{fraction of substrate escaping hepatic metabolism. By
  default, data are pulled from the sheet "Clearance Trials SS".}

  \item{ka_sub or ka_inhib}{absorption rate constant ka for the substrate or
  the 1st inhibitor. By default, data are pulled from the sheet
  "Absorption".}

  \item{tlag_sub or tlag_inhib}{lag time for the substrate or inhibitor 1. By
  default, data are pulled from the sheet "Absorption".}

  \item{tmax_dose1}{tmax for dose 1. By default, data are pulled from sheet
  "AUC0(Sub)(CPlasma)", column titled, e.g., "TMax (h)".}

  \item{tmax_ss}{tmax for the last dose. By default, data are pulled from
  sheet "AUCX(Sub)(CPlasma)", where "X" is the largest dose for which there
  is a sheet, from the column titled, e.g., "TMax (h)".}

  } The default is only those parameters present on the "AUC" sheet in the
  simulator output.}

\item{sheet}{Which sheet should be used for pulling the PK parameters?
\strong{Note:} Unless you want a very specific Excel sheet that's not what
the usual sheet name would be for a first or last dose, this function will
work best if this is left as NA.}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are "plasma" or "blood".}

\item{returnAggregateOrIndiv}{Return aggregate and/or individual PK
parameters? Options are "aggregate" and/or "individual". For aggregate
data, values are pulled from simulator output -- not calculated -- and the
output will be a data.frame with the PK parameters in columns and the
statistics reported exactly as in the simulator output file.}

\item{includeTrialInfo}{TRUE or FALSE: Include which individual and trial the
data describe? This only applies when \code{returnAggregateOrIndiv}
includes "individual".}
}
\value{
Depending on the options selected, returns a list of numerical
  vectors or a list of data.frames.
}
\description{
Pull calculated substrate PK parameters from a Simcyp simulation output Excel
file.
}
\examples{

sim_data_file <- "../Example simulator output MD + inhibitor.xlsx"
extractPK(sim_data_file)
extractPK(sim_data_file, PKparameters = "Absorption tab")
extractPK(sim_data_file, PKparameters = "AUCinf_dose1")


}
