% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChanges.R
\name{findXLChanges}
\alias{findXLChanges}
\title{findXLChanges}
\usage{
findXLChanges(File1, File2, outputAllSheets = TRUE)
}
\arguments{
\item{File1}{An Excel file you want to compare}

\item{File2}{The other Excel file you want to compare}

\item{outputAllSheets}{TRUE or FALSE: Do you want to have the output include
a list of all the loaded Excel sheets?}
}
\value{
If \code{ouputAllSheets} is FALSE, output is a list of which rows and
  columns differ between File1 and File2. If \code{ouputAllSheets} is TRUE,
  output is both the list and also a list of all the sheet contents from both
  files.
}
\description{
This function reads in all sheets in 2 Excel files and finds any differences
between them. It is NOT smart, so if the only difference on a given sheet is
that everything has moved down a row, it will find LOTS of changes. It is
also SLOW b/c it has to read everything from Excel.
}
