% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_treemap.R
\name{fm_treemap}
\alias{fm_treemap}
\title{Make a treemap of fm values as opposed to the odious and easily
misinterpreted pie chart}
\usage{
fm_treemap(
  fm_dataframe,
  color_set = "default",
  label_fm_cutoff = 0.05,
  save_graph = NA,
  fig_height = 3,
  fig_width = 4
)
}
\arguments{
\item{fm_dataframe}{a data.frame containing columns for
\describe{\item{DME}{the drug-metabolizing enzyme} \item{fm}{the fraction
metaboilized}}}

\item{color_set}{the set of colors to use. Options are "default", which
matches Simcyp colors in PowerPoint presentations and is rather nice, "set
1" for Brewer set 1, "set 2" for Brewer set 2 (colorblind friendly),
"rainbow", which works well when there are a LOT of fm's to visualize, or
"blue-green", which makes a pretty gradient of blues and greens.}

\item{label_fm_cutoff}{cutoff to use for having a label show up underneath
the graph; default is 0.05. You may want to play around with this a bit in
the final graphic file because how this graph looks is extremely dependent
on its dimensions.}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My conc time graph.png". If you do not designate a
file extension, it will be saved as a png file, but if you specify a
different file extension, it will be saved as that file format. Acceptable
extensions are "eps", "ps", "jpeg", "jpg", "tiff", "png", "bmp", or "svg".
Leaving this as NA means the file will not be automatically saved to disk.}

\item{fig_height}{figure height in inches; default is 3}

\item{fig_width}{figure width in inches; default is 4}
}
\value{

}
\description{
Create a treemap -- filled rectangles -- to indicate the fraction of
metabolism due to various pathways. The size of the rectangles are
proportional to the given fm. This is a (in Laura's opinion) superior option
compared to a pie chart since rectangles are easier for the human eye to
interpret and compare than pie slices. Rectangles will include labels for the
drug-metabolizing enzyme involved and the percentage that enzyme is
responsible for. Any rectangles that would be smaller than some
user-specified cutoff (see arugment "label_fm_cutoff") will be pulled out
underneath the graph for readability.
}
\examples{

Lenv_fm <- data.frame(DME = c("CYP3A4", "other CYPs", "aldehyde oxidase"),
                      fm = c(0.79, 0.20, 0.01))
fm_treemap(fm_dataframe = Lenv_fm)
fm_treemap(fm_dataframe = Lenv_fm, color_set = "blue-green")
fm_treemap(fm_dataframe = Lenv_fm, color_set = "set 1", 
    label_fm_cutoff = 0.01)

}
