% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Create a forest plot}
\usage{
forest_plot(
  forest_dataframe,
  perp_or_victim,
  PKparameters = NA,
  y_axis_order = NA,
  y_axis_order_secondary,
  x_axis_limits = NA,
  x_axis_label = NA,
  facet_column_x,
  prettify_compound_names = TRUE,
  legend_position = "none",
  save_graph = NA,
  fig_height = 6,
  fig_width = 5
)
}
\arguments{
\item{forest_dataframe}{a data.frame with extracted forest-plot data,
generated from running \code{\link{extractForestData}} on Simulator output
files or a csv file with the same data. If this is a forest plot of
perpetrator DDIs, the column "Substrate" will be what is used to label the
y axis. If it is a forest plot of victim DDIs, it will be the column
"Inhibitor1". If, instead of the compound or inhibitor 1 names, you would
like some other label to appear on the y axis, please see the argument
\code{y_axis_labels}.}

\item{perp_or_victim}{specify whether the drug of interest is a "victim" or
"perpetrator". This will determine whether the graphs will be labeled on
the y axis by the substrate name (for perpetrator forest plots) or by the
effector name (for victim forest plots).}

\item{PKparameters}{optionally specify which PK parameters included in
\code{forest_dataframe} to use as input. If left as NA, all the PK
parameters you extracted with \code{\link{extractForestData}} will be
included. If you try to include a parameter that's not already present in
forest_dataframe, it will be ignored. Enclose the parameters with
\code{c(...)}.}

\item{y_axis_order}{optionally supply a character vector to specify the order
of the files or compounds on the y axis. If your character vector contains
"xlsx" in the any of the text, we'll assume you want to sort by the file
names; otherwise, we'll assume you want to sort by the compound names -- by
substrate if it's a perpetrator or by inhibitor 1 if it's a victim. If
\code{y_axis_order} is unspecified, the y axis will be sorted according to
the geometric mean AUC ratio with inhibitors on top and inducers on the
bottom. If there's more than one row for a substrate or inhibitor (say
you're including scenarios where you've reduced the Ki or the IndC50, for
example), this may not give you the order you want and you might be better
off specifying manually. If you supply a named character vector of file
names, in addition to using that for setting the y axis order, we'll use
those values for the y axis labels instead of the compound names. This can
be useful when you want to add notes about changes in parameters such as
when you're including sensitivity analyses. An example of the syntax to
use: \code{y_axis_labels = c("abc1a-mdz.xlsx" = "Midazolam with in vitro
induction", "abc-1a-mdz-10x.xlsx" = "Midazolam with 10-fold higher
Indmax")} Please see the example section at the bottom of this help file
for more examples.}

\item{y_axis_order_secondary}{optionally break up the graphs along the y axis
by an additional column other than substrate (for perpetrator graphs) or
inhibitor 1 (for victim graphs). For example, say your drug of interest is
a perpetrator and you've administered each of the substrates on different
days. This will break up the graphs by substrate and, if you specify, e.g.,
\code{y_axis_order_secondary = DoseDay} where "DoseDay" is a column in
forest_dataframe listing what day the substrate was dosed, it will also
break up the graphs by what day the substrate was administered.}

\item{x_axis_limits}{the x axis limits to use; default is 0.06 to 12.}

\item{x_axis_label}{optionally supply a character vector or an expression to
use for the x axis label}

\item{facet_column_x}{optionally break up the graph horizontally into small
multiples. The designated column name should be unquoted, e.g.,
\code{facet_column_x = Dose_sub}}

\item{prettify_compound_names}{TRUE (default) or FALSE on whether to make
compound names prettier. This was designed for simulations where the
substrates or effectors are among the standard options for the simulator,
and leaving \code{prettify_compound_names = TRUE} will make the name of
those compounds something more human readable. For example,
"SV-Rifampicin-MD" will become "rifampicin", and "Sim-Midazolam" will
become "midazolam".}

\item{legend_position}{specify where you want the legend to be. Options are
"left", "right", "bottom", "top", or "none" (default) if you don't want one
at all. \emph{Note:} We're still working on the legend position when
there's something specified for \code{y_axis_order_secondary}; choosing
that option requires us to lay out the graphs differently, and when we do
that, the legend position doesn't work well for anything other than "none"
or "right".}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My conc time graph.png" or "My conc time
graph.docx". If you leave off ".png" or ".docx" from the file name, it will
be saved as a png file, but if you specify a different graphical file
extension, it will be saved as that file format. Acceptable graphical file
extensions are "eps", "ps", "jpeg", "jpg", "tiff", "png", "bmp", or "svg".
Leaving this as NA means the file will not be saved to disk.
\strong{WARNING:} SAVING TO WORD DOES NOT WORK ON SHAREPOINT. This is a
Microsoft permissions issue, not an R issue. If you try to save on
SharePoint, you will get a warning that R will save your file instead to
your Documents folder.}

\item{fig_height}{figure height in inches; default is 6}

\item{fig_width}{figure width in inches; default is 5}
}
\value{
Output is a graph.
}
\description{
\code{forest_plot} creates forest plots of AUC and Cmax ratios. Please use
the function \code{\link{extractForestData}} to generate the input data for
\code{forest_dataframe}. The data will be broken up by the file name on the y
axis and, optionally, whatever you specify for \code{facet_column_x} in the
horizontal direction. WARNING: This is a relatively new function and should
be rigorously QC'ed.
}
\examples{

# We'll use some example forest-plot data for the substrate bufuralol
# with various effectors.
forest_plot(forest_dataframe = ForestData,
            perp_or_victim = "victim",
            x_axis_limits = c(0.9, 5))

# If there were multiple dosing levels of bufuralol, it might be
# nice to break up the graph by the substrate dose like this:
forest_plot(forest_dataframe = ForestData,
            perp_or_victim = "victim",
            facet_column_x = Dose_sub,
            x_axis_limits = c(0.9, 5))


# If you want to maybe add some units in the graph to the substrate dose
# or maybe also say "Dose = X mg MyCompound", you can modify the column
# Dose_sub into a new column, Dose_sub_pretty, and use it like this:
ForestData <- ForestData \%>\%
                  mutate(Dose_sub_pretty = paste("Dose =", Dose_sub,
                                                 "mg bufuralol"))

forest_plot(forest_dataframe = ForestData,
            perp_or_victim = "victim",
            facet_column_x = Dose_sub_pretty,
            x_axis_limits = c(0.9, 5))

}
