% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_scripts.R
\name{format_scripts}
\alias{format_scripts}
\title{Format common PBPK model parameters to have appropriate subscripts and
superscripts. UNDER CONSTRUCTION.}
\usage{
format_scripts(
  DF,
  parameter_column,
  value_column,
  remove_compound_suffix = TRUE
)
}
\arguments{
\item{DF}{a data.frame with a column containing, e.g., "ka (h^-1)", "fu,p",
or other PBPK model parameters with syntax that matches what's in the
column "Notes" in the object AllExpDetails, which is included with this
package and is used extensively for figuring out how simulations were set
up.}

\item{parameter_column}{the name of the column in DF that contains your PK
parameters, unquoted. For example, supply \code{parameter_column = Notes}}

\item{value_column}{optionally supply the name of the column in DF that
contains values for PK parameters, unquoted.}

\item{remove_compound_suffix}{TRUE (default) or FALSE to remove the suffix
that shows up in many PBPK elimination and interaction parameters to
indicate which compound ID the parameter refers to. For example, "Vsac_sub"
will be just "Vsac" and "Peff_human_inhib" will become "Peff,human" with
the "eff,human" as a subscript. You should ONLY do this if your whole table
refers to the same compound; otherwise, it will be completely unclear
what's what.}
}
\value{
a flextable
}
\description{
Format common PBPK model parameters to have appropriate subscripts and
superscripts. UNDER CONSTRUCTION.
}
\examples{
DF <- data.frame(Parameter = c("MW", "ka (h^-1)",
                               "fu,p", "fu,gut", "fu,mic CYP3A4"),
                  Value = c(350, 0.5, 0.01, 0.8, 1))

format_scripts(DF = DF)

}
