% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table_simple.R
\name{format_table_simple}
\alias{format_table_simple}
\title{Format a table to look nice in a Word file or pdf - UNDER CONSTRUCTION}
\usage{
format_table_simple(DF, shading_column, merge_shaded_cells = TRUE)
}
\arguments{
\item{DF}{a data.frame or a flextable}

\item{shading_column}{If you would like to alternate the shading of the rows
in the output table, supply here the unquoted name of the column to check
for when to change the shading; every time that column's value changes, the
shading will alternate between white and light gray. For example, if you
have a table with PK values for multiple files and you have more than one
row per file (an example of this would be the output from the function
\code{\link{pksummary_mult}}), setting \code{shading_column = File} will
cause the shading of the rows to alternate between white and light gray
whenever the file changes. Please see the examples at the bottom of this
help file.}

\item{merge_shaded_cells}{TRUE (default) or FALSE for whether to merge the
cells that have the same shade. This only applies when one of the columns
in the input data.frame is used for deciding when to alternate shading,
that is, \code{shading_column} has a value.}
}
\value{
a formatted table
}
\description{
\code{format_table_simple} is meant to for use with writing compound summary
pdfs. It includes minimal error catching at this point.
}
\examples{
# None yet 

}
