% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table_simple.R
\name{format_table_simple}
\alias{format_table_simple}
\title{Format a table to look nice in a Word file or pdf - UNDER CONSTRUCTION}
\usage{
format_table_simple(
  DF,
  shading_column,
  merge_shaded_cells = TRUE,
  font = "Palatino Linotype",
  fontsize = 11,
  save_table = NA,
  page_orientation = "portrait",
  title_document = NA,
  table_caption = NA
)
}
\arguments{
\item{DF}{a data.frame or a flextable}

\item{shading_column}{If you would like to alternate the shading of the rows
in the output table, supply here the unquoted name of the column to check
for when to change the shading; every time that column's value changes, the
shading will alternate between white and light gray. For example, if you
have a table with PK values for multiple files and you have more than one
row per file (an example of this would be the output from the function
\code{\link{pksummary_mult}}), setting \code{shading_column = File} will
cause the shading of the rows to alternate between white and light gray
whenever the file changes. Please see the examples at the bottom of this
help file.}

\item{merge_shaded_cells}{TRUE (default) or FALSE for whether to merge the
cells that have the same shade. This only applies when one of the columns
in the input data.frame is used for deciding when to alternate shading,
that is, \code{shading_column} has a value.}

\item{font}{font to use. Default is "Palatino Linotype" and any fonts
available on your machine in either Word or PowerPoint should be
acceptable. If you get Times New Roman in your table when you asked for
something else, it means that that font isn't available or maybe wasn't
spelled the way R is expecting it. For example, "Calibri" works but
"Calibri (Body)" doesn't even though the latter is listed in PowerPoint and
Word.}

\item{fontsize}{the numeric font size for the output table. Default is 11
point.}

\item{save_table}{optionally save the output table by supplying a file name
in quotes here, e.g., "My nicely formatted table.docx".  Do not include any
slashes, dollar signs, or periods in the file name. If you leave off the
file extension, we'll assume you want it to be ".docx". If there is a
column titled "File" in your table, we'll add a caption listing which files
were included.}

\item{page_orientation}{set the page orientation for the Word file output to
"portrait" (default) or "landscape"}

\item{title_document}{optionally specify a title for the Word document
output. If you don't save the table, this will be ignored.}

\item{table_caption}{optionally add some text for a table caption. If the
table you supply contains a column titled "File", there will already be a
caption listing the source files; this would add some additional text
before that.}
}
\value{
a formatted table
}
\description{
\code{format_table_simple} is meant to for use with writing compound summary
pdfs. It includes minimal error catching at this point.
}
\examples{
# None yet

}
