% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTable_Simcyp.R
\name{formatTable_Simcyp}
\alias{formatTable_Simcyp}
\title{Format tables according to Simcyp Consultancy Team specfications}
\usage{
formatTable_Simcyp(DF, fontsize = 11)
}
\arguments{
\item{DF}{a data.frame, usually output from \code{\link{pksummary_table}} or
\code{\link{pksummary_mult}}}

\item{fontsize}{the numeric font size for the output. Default is 11 point.}
}
\value{

}
\description{
\code{formatTable_Simcyp} was designed to work with output from
\code{\link{pksummary_table}} and \code{\link{pksummary_mult}}, and it
formats tables so that the column headings and the first column are bold, and
the second through the last columns are centered. Column headings with, e.g.,
"AUCinf" or "Cmax" will have the "inf" or the "max" subscripted, and the
table will automatically expand to fit the contents.
}
\examples{

pksummary_table(sim_data_file = "My simulation.xlsx")$Table \%>\% 
    formatTable_Simcyp()

}
