% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTable_Simcyp.R
\name{formatTable_Simcyp}
\alias{formatTable_Simcyp}
\title{Format tables according to Simcyp Consultancy Team specifications}
\usage{
formatTable_Simcyp(
  DF,
  fontsize = 11,
  shading_column,
  merge_shaded_cells = TRUE,
  bold_cells = list(c(0, NA), c(NA, 1)),
  center_1st_column = FALSE,
  highlight_so_cutoffs = NA,
  highlight_so_colors = "yellow to red",
  highlight_cells = NA,
  highlight_color = "yellow",
  save_table = NA,
  title_document = NA
)
}
\arguments{
\item{DF}{a data.frame, usually output from \code{\link{pksummary_table}} or
\code{\link{pksummary_mult}}}

\item{fontsize}{the numeric font size for the output table. Default is 11
point.}

\item{shading_column}{If you would like to alternate the shading of the rows
in the output table, supply here the unquoted name of the column to check
for when to change the shading; every time that column's value changes, the
shading will alternate between white and light gray. For example, if you
have a table with PK values for multiple files and you have more than one
row per file (an example of this would be the output from the function
\code{\link{pksummary_mult}}), setting \code{shading_column = File} will
cause the shading of the rows to alternate between white and light gray
whenever the file changes. Please see the examples at the bottom of this
help file.}

\item{merge_shaded_cells}{TRUE (default) or FALSE for whether to merge the
cells that have the same shade. This only applies when one of the columns
in the input data.frame is used for deciding when to alternate shading,
that is, \code{shading_column} has a value.}

\item{bold_cells}{optionally specify cells in the table to be in bold-face
text with a numeric vector where the 1st number is the row number and the
2nd number is the column number (just like regular row and column
specifications in R). For example, \code{bold_cells = c(1, 2)} will make
the cell in row 1 and column 2 bold face. Use "0" for the row number if you
want to use bold face for something in the header row, and use NA in place
of a row or column number to make everything in that row or column bold
face. If you want to specify multiple places to use bold face, use a list
of numeric vectors. By default, the header row and the 1st column will be
bold. Set \code{bold_cells = NA} to make \emph{nothing} bold. Please see
the examples at the bottom of the help file.}

\item{center_1st_column}{TRUE or FALSE (default) for whether to make the
alignment of the first column centered}

\item{highlight_so_cutoffs}{optionally specify cutoffs for highlighting any
simulated-to-observed ratios. Anything that is above those values or below
the inverse of those values will be highlighted. To figure out what cells
to highlight, this looks for a column titled "Statistic" or "Stat", then
looks for what row contains "S/O" or "simulated (something something)
observed" (as in, we'll use some wildcards to try to match your specific
text). Next, it looks for any values in that same row that are above those
cutoffs. This overrides anything else you specified for highlighting. The
default is NA, for \emph{not} highlighting based on S/O value. Acceptable
input for, say, highlighting values that are > 125\% or < 80\% of the
observed and also, with a second color, values that are > 150\% or < 66\%
would be: \code{highlight_so_cutoffs = c(1.25, 1.5)}. If you would like the
middle range of values to be highlighted, include 1 in your cutoffs. For
example, say you would like everything that's < 80\% or > 125\% to be
highlighted red but you'd like the "good" values from 80\% to 125\% to be
green, you can get that by specifying
\code{highlight_so_cutoffs = c(1, 1.25)} and \code{highlight_so_colors =
c("green", "red")}}

\item{highlight_so_colors}{optionally specify a set of colors to use for
highlighting S/O values outside the limits you specified with
\code{highlight_so_cutoffs}. The default is "yellow to red", which will
color your values in that range based on how they compare to your the
cutoffs you specified. Alternatively, specify specific colors using any
R-acceptable color, e.g., \code{highlight_so_colors = c("yellow", "orange",
"red")}. If you do specify your own bespoke colors, you'll need to make
sure that you supply one color for every value in
\code{highlight_so_cutoffs}. If you have included 1 in your cutoffs and you
leave \code{highlight_so_colors} with the default setting, values in the
middle, "good" range of S/O values will be highlighted a light green.}

\item{highlight_cells}{optionally specify cells in the table to be
highlighted with a numeric vector where the 1st number is the row number
and the 2nd number is the column number (just like regular row and column
specifications in R). For example, \code{highlight_cells = c(1, 2)} will
make the cell in row 1 and column 2 highlighted. Use "0" for the row number
if you want to highlight something in the header row, and use NA in place
of a row or column number to highlight everything in that row or column. If
you want to specify multiple places to highlight, use a list of numeric
vectors. Please see the examples at the bottom of the help file.}

\item{highlight_color}{color to use for highlighting; default is yellow.
Color can be specified using any R-friendly color name or hex code, e.g.,
"red" or "#D8212D".}

\item{save_table}{optionally save the output table by supplying a file name
in quotes here, e.g., "My nicely formatted table.docx". If you leave off
the file extension, we'll assume you want it to be ".docx". If there is a
column titled "File" in your table, we'll add a caption listing which files
were included.}

\item{title_document}{optionally specify a title for the Word document
output. If you don't save the table, this will be ignored.}
}
\value{
a formatted table
}
\description{
\code{formatTable_Simcyp} makes a nicely formatted table from a data.frame or
tibble. It was primarily designed to work with output from
\code{\link{pksummary_table}} and \code{\link{pksummary_mult}}, so, by
default, it formats tables so that the column headings and the first column
are bold, and the second through the last columns are centered. Column
headings with, e.g., "AUCinf" or "Cmax" will have the "inf" or the "max"
subscripted, and the table will automatically expand to fit the contents. You
can save the output to a Word file with the argument \code{save_table}.
}
\examples{
MyPKTable <- tibble(Statistic = c("Simulated", "CV\%", "Observed", "S/O"),
                        AUCinf = c(2756, 32.5, 1801, 1.53), 
                        Cmax = c(852, 45.8, 775, 1.1), 
                        `Half life` = c(7.75, 5.7, 6.05, 1.28))
formatTable_Simcyp(MyPKTable)
formatTable_Simcyp(MyPKTable, center_1st_column = TRUE)
formatTable_Simcyp(MyPKTable, fontsize = 18)
formatTable_Simcyp(MyPKTable, shading_column = Statistic)

# Highlighting examples
## Highlighting S/O values outside bioequivalence of 125\%.
formatTable_Simcyp(MyPKTable,
                   highlight_so_cutoffs = 1.25)

## Highlighting S/O values with a few more colors based on the S/O.
formatTable_Simcyp(MyPKTable,
                   highlight_so_cutoffs = c(1.25, 1.5))

## Highlighting S/O values and shading the "good" values green.
formatTable_Simcyp(MyPKTable,
                   highlight_so_cutoffs = c(1, 1.25, 1.5))

## Highlight exactly the cells you want, e.g., row 1, column 2
formatTable_Simcyp(MyPKTable, highlight_cells = c(1, 2))

## Highlight all of column 2
formatTable_Simcyp(MyPKTable, highlight_cells = c(NA, 2))

## Highlight all of row 1
formatTable_Simcyp(MyPKTable, highlight_cells = c(1, NA))

## Highlight the 2nd column in the header
formatTable_Simcyp(MyPKTable, highlight_cells = c(0, 2))

## Set the highlight color to light blue instead of yellow
formatTable_Simcyp(MyPKTable, highlight_cells = c(1, NA),
                   highlight_color = "lightblue")

## Highlighting multiple cells
                   highlight_color = "lightblue")

# Bold-face examples
## Make only the cell in row 4 and column 2 be bold face. This will
## override the default of having the header row and the 1st column in bold.
formatTable_Simcyp(MyPKTable, bold_cells = c(4, 2))

## Make the cell in row 4 and column 2 be bold face AND include the original
## defaults of having the header row and the 1st column be in bold.
formatTable_Simcyp(MyPKTable, bold_cells = list(c(0, NA), c(NA, 1), c(4, 2)))

# Saving
## Adding a column called "File" so that there will be a caption in the Word
## document listing which files were included in the table. Also setting 
## the document title.
MyPKTable$File <- "abc-1a.xlsx"
formatTable_Simcyp(MyPKTable, 
                   highlight_so_cutoffs = c(1, 1.25, 1.5),
                   save_table = "My data.docx", 
                   title_document = "PK data")


}
