% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReportInputForm.R
\name{generateReportInputForm}
\alias{generateReportInputForm}
\title{Generate a form for entering data for a report}
\usage{
generateReportInputForm(filename)
}
\arguments{
\item{filename}{the Excel file name that you'd like for the form you're
creating, ending in ".xlsx". Only list the file name here rather than the
full path.}
}
\value{
This does not return an R object; it saves an Excel file that serves
  as a form for report information.
}
\description{

}
\details{
The tabs in the Excel file this function creates are: \describe{

\item{how to use this file}{Instructions for how to use the Excel template
form.}

\item{overall report info}{a form for entering overall data about the project
-- information that will be consistent throughout the whole report such as
the drug name and the complaint it is meant to treat. There should be only
ONE copy of this tab when you're finished. You do \emph{not} need to fill
this out for using the \code{\link{so_table}} function.}

\item{observed data - no DDI}{a form for entering observed PK data about a
clinical study that was \emph{not} a drug-drug interaction study. Make as
many copies of this tab as there are clinical studies you want to compare and
name the tabs according to which study they describe.}

\item{observed data - DDI}{a form for entering observed PK data about a
clinical study that \emph{was} a drug-drug interaction study. Make as many
copies of this tab as there are clinical studies you want to compare and name
the tabs according to which study they describe.}

\item{simulated data}{a form for entering information which simulation file
you'd like to use and which observed data you want to compare. As with the
observed data tabs, make as many copies of this tab as you require.}}
}
\examples{
setwd(paste0(SimcypDir$SharePtDir,
             "Research/R working group/SimcypConsultancy function examples and instructions/so_table examples"))
generateReportInputForm("Ultraconazole report input.xlsx")


}
