% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReportInputForm.R
\name{generateReportInputForm}
\alias{generateReportInputForm}
\title{Generate a form for entering data for a report}
\usage{
generateReportInputForm(filename)
}
\arguments{
\item{filename}{the Excel file name that you'd like for the form you're
creating, ending in ".xlsx"}
}
\value{
This does not return an R object; it saves an Excel file that serves
  as a form for recording observed data and certain general information for a
  report.
}
\description{
Creates an Excel file with forms to fill out about observed data. Those forms
can be fed into \code{\link{so_table}} to automate some of the calculations
for writing a report.
}
\details{
\strong{A few notes:} \enumerate{ \item{This generates a warning that we're
just not able to get rid of that says "Workbook has no sheets!" Please
disregard that.} \item{This does not work on the SharePoint drive because R
doesn't have permission to write there. Please set your working directory to
a different location. It \emph{will} work on the Large File Store drive.}}

The tabs in the Excel file this function creates are: \describe{

\item{how to use this file}{Instructions for how to use the Excel template
form.}

\item{overall report info}{a form for entering overall data about the project
-- information that will be consistent throughout the whole report such as
the drug name and the complaint it is meant to treat. There should be only
ONE copy of this tab when you're finished. You do \emph{not} need to fill
this out for using the \code{\link{so_table}} function.}

\item{study info - no DDI}{a form for entering observed PK data about a
clinical study that was \emph{not} a drug-drug interaction study. Make as
many copies of this tab as there are clinical data sets that you want to
compare and name the tabs according to which data sets they describe.}

\item{study info - DDI}{a form for entering observed PK data about a clinical
study that \emph{was} a drug-drug interaction study. Make as many copies of
this tab as there are clinical data sets you want to compare and name the
tabs according to which data sets they describe.} }
}
\examples{
generateReportInputForm(paste0(SimcypDir$SharePtDir,
             "Research/R working group/SimcypConsultancy function examples and instructions/so_table examples/",
             "Ultraconazole report input.xlsx"))


}
