% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReportInputForm.R
\name{generateReportInputForm}
\alias{generateReportInputForm}
\title{Generate a form for entering data for a report}
\usage{
generateReportInputForm(filename)
}
\arguments{
\item{filename}{the Excel file name that you'd like for the form you're
creating, ending in ".xlsx".}
}
\value{
This does not return an R object; it saves an Excel file that serves
  as a form for report information.
}
\description{
Creates an Excel file with three forms to fill out. Those forms can be fed
into \code{\link{getSectionInfo}} and \code{\link{so_table}} to automate some
of the calculations and graph creation for writing a report.
}
\details{
The three tabs in the Excel file this function creates are: \describe{

\item{Overall report form}{a form for entering overall data about the project
-- information that will be consistent throughout the whole report such as
the drug name and the complaint it is meant to treat. There should be only
ONE copy of this tab when you're finished.}

\item{Observed data form}{a form for entering data about a clinical study.
Make as many copies of this tab as there are clinical studies you want to
compare and name the tabs according to which study they describe.}

\item{Section input form}{a form for entering information about a given
section such as which simulation file you'd like to graph and compare to
clinical data. As with the tab "Observed data form", make as many copies of
this tab as you'd like -- one for each standard results subsection in the
report.}}
}
\examples{
generateReportInputForm("//certara.com/data/sites/SHF/Consult/abc-1a/Ultraconazole report input.xlsx")
generateReportInputForm(paste0(SimcypDir$LgFileDir, "abc-1a/Ultraconazole report input.xlsx"))
generateReportInputForm(paste0(SimcypDir$SharePtDir, "def-2b/Superstatin report input.xlsx"))


}
