% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReportInputForm.R
\name{generateReportInputForm}
\alias{generateReportInputForm}
\title{Generate a form for entering data for a report}
\usage{
generateReportInputForm(filename)
}
\arguments{
\item{filename}{the Excel file name that you'd like for the form you're
creating, ending in ".xlsx"}
}
\value{
This does not return an R object; it saves an Excel file that serves
  as a form for recording observed data and certain general information for a
  report.
}
\description{
Creates an Excel file with forms to fill out about observed data. Those forms
can be fed into \code{\link{pksummary_table}} to automate some of the
calculations for writing a report.
}
\details{
The tabs in the Excel file this function creates are: \describe{

\item{how to use this file}{Instructions for how to use the Excel template
form.}

\item{study info - no DDI}{a form for entering observed PK data about a
clinical study that was \emph{not} a drug-drug interaction study. Make as
many copies of this tab as there are clinical data sets that you want to
compare and name the tabs according to which data sets they describe.}

\item{study info - DDI}{a form for entering observed PK data about a clinical
study that \emph{was} a drug-drug interaction study. Make as many copies of
this tab as there are clinical data sets you want to compare and name the
tabs according to which data sets they describe.} }
}
\examples{
generateReportInputForm("Ultraconazole report input.xlsx")


}
