% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSectionInfo.R
\name{getSectionInfo}
\alias{getSectionInfo}
\title{Get information about a section for a report}
\usage{
getSectionInfo(report_input_file = NA, sheet = NA, section_input_DF = NA)
}
\arguments{
\item{report_input_file}{the name of the Excel file formatted exactly like
"Report input template.xlsx", including the path if it's in any other
directory than the current one}

\item{sheet}{the sheet in the Excel file that contains information about this
section of the report}

\item{section_input_DF}{a data.frame object that contains information about
this section of the report and is the filled-out version of
ReportInputForm[["Section input form"]]. If this is NOT set to NA, then
this will be used for section details instead of reading the Excel sheet
listed. You still must fill out \code{report_input_file}, though, because
that will be used to read in data about the clinical study.}
}
\value{
a list object
}
\description{
\code{getSectionInfo} pulls information from an Excel file generated by
\code{\link{generateReportInputForm}} to feed into the function
\code{\link{so_table}}. Since this requires comparing observed data to a
simulator output file, setting up the input for this function requires a few
steps: \enumerate{ \item{Use the function
\code{\link{generateReportInputForm}} to create an Excel file where you can
enter information about your project.} \item{Go to the tab "observed data"
and enter details about your observed data. It's ok if you don't have all the
information; anything that's missing won't be included in the final S/O
table. It's also ok to rename this tab and/or make copies of it within the
same Excel file for making other S/O tables.} \item{Go to the tab "section
input" and fill out information here for the specific section you're making
this S/O table for. Make sure that whatever you list for the tab that
contains information about the observed data is exactly the same as the
actual tab name that you filled out in step 2. Also make sure that the file
names include the full file path.} \item{Save your Excel file.} \item{Here,
within RStudio (or within the shiny app that we plan to make!), run this
function using the name of that Excel file as input for
\code{report_input_file} and the name of the "section input" tab as the input
for \code{sheet}.} }
}
\examples{
# No examples yet.

}
