% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSectionInfo.R
\name{getSectionInfo}
\alias{getSectionInfo}
\title{Get information about a section for a report}
\usage{
getSectionInfo(report_input_file = NA, sheet = NA, report_input_DF = NA)
}
\arguments{
\item{report_input_file}{the name of the Excel file formatted exactly like
"Report input template.xlsx", including the path if it's in any other
directory than the current one. This should be filled out with your
specific project details.}

\item{sheet}{the sheet to read in that Excel file}

\item{report_input_DF}{the data.frame object that is laid out exactly like a
tab from the "Report input template.xlsx" file}
}
\value{
a list object
}
\description{
\code{getSectionInfo} pulls information from an Excel file that is formatted
\emph{exactly} like the file "Report input template.xlsx" and organizes those
data plus some further data it pulls from the simulator output Excel file.
The output of this function serves as part of the input for the function
\code{so_table}.
}
\examples{
# No examples yet.

}
