% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_boxplot.R
\name{graph_boxplot}
\alias{graph_boxplot}
\title{Make boxplots or boxplots overlaid with individual points}
\usage{
graph_boxplot(
  DF,
  category_column,
  value_column,
  facet_column1 = NA,
  facet_column2 = NA,
  graph_type = "boxplot",
  xlabel = NA,
  ylabel = NA,
  color = "default"
)
}
\arguments{
\item{DF}{the data.frame you want to graph}

\item{category_column}{the name of the column with categorical data in quotes}

\item{value_column}{the name of the column with value data in quotes}

\item{facet_column1}{(optional) the name of a column by which you might break
up your graph into small multiples. Please see the example if you're
uncertain what this does.}

\item{facet_column2}{(optional) the name of a second column by which you
might break up the graph into small multiples.}

\item{graph_type}{the type of graph to plot. Options: \describe{
  \item{"boxplot"}{standard boxplots or box-and-whisker plots}

  \item{"boxplot with error bars"}{standard boxplots or box-and-whisker plots
  plus error bars on the whiskers}

  \item{"jittered points"}{boxplots overlaid with points depicting each
  individual observation.}}}

\item{xlabel}{the label to  use for the x axis}

\item{ylabel}{the label to use for the y axis}

\item{color}{the set of colors to use. Options: \describe{
\item{"default"}{colors selected from the color brewer palette "set 1"}
\item{"black"}{black and white} \item{"rainbow"}{colors selected from a
rainbow palette. The default palette is limited to something like 6 groups,
so if you have more than that, that's when this palette is most useful.}
\item{"blue-green"}{a set of blues and greens} \item{"Tableau"}{uses the
standard Tableau palette; requires the "ggthemes" package}}}
}
\value{

}
\description{
\code{graph_boxplot} takes as input a data.frame of numeric values and the
categories those values are in and creates boxplots of those data. Options
are included for type of graph to make and some aesthetics.
}
\examples{
AUCs <- data.frame(AgeGroup = LETTERS[1:5],
                   AUC = rnorm(n = 500, mean = 20000, sd = 2000))

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC")

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
               xlabel = "Age group", graph_type = "boxplot with error bars")

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
               color = "rainbow")

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
               color = "blue-green", graph_type =  "jittered points")

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
              color = "Tableau", graph_type =  "jittered points")

graph_boxplot(AUCs \%>\% filter(AgeGroup \%in\% c("A", "B", "C")),
               category_column = "AgeGroup", value_column = "AUC",
               graph_type =  "jittered points")

# Adding a couple of example columns to use the "facet" options.
AUCs$Sex <- c("M", "F")
AUCs$Metabolizer <- sample(c("poor", "extensive"), 100, replace = TRUE)

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
               facet_column1 = "Sex",
               color = "blue-green", graph_type =  "jittered points")

graph_boxplot(AUCs, category_column = "AgeGroup", value_column = "AUC",
               facet_column1 = "Sex", facet_column2 = "Metabolizer",
               color = "blue-green", graph_type =  "jittered points")


}
