% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_details.R
\name{harmonize_details}
\alias{harmonize_details}
\title{FOR INTERNAL PACKAGE USE. Take as input an existing data.frame or list of
experimental details and make it be a list with standard items}
\usage{
harmonize_details(existing_exp_details)
}
\arguments{
\item{existing_exp_details}{an object created from running extractExpDetails,
extractExpDetails_mult, annotateDetails, or deannotateDetails}
}
\value{
Returns a list
}
\description{
\code{harmonize_details} takes as input an object created from running
extractExpDetails, extractExpDetails_mult, annotateDetails, or
deannotateDetails and makes it conform to the data structure the rest of the
package needs: a list with the following items, which can be added to in the
future: 1) MainDetails, 2) CustomDosing_sub, 3) CustomDosing_inhib, and 4)
CustomDosing_inhib2.
}
\examples{
# none
}
