% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inductFit.R
\name{inductFit}
\alias{inductFit}
\title{Fit induction data to calculate IndC50, Indmax, and/or slope}
\usage{
inductFit(
  DF,
  conc_column = Concentration_uM,
  fold_change_column = FoldInduction,
  model = "IndmaxSlope",
  measurement = "mRNA",
  donor_column = DonorID,
  fitByDonor = TRUE,
  weights = "1/y^2",
  color_set = "default",
  y_axis_limits = NA,
  hline_foldinduct1 = FALSE,
  num_sigfig = NA,
  save_graph = NA,
  fig_height = 5,
  fig_width = 5.5,
  save_output = NA
)
}
\arguments{
\item{DF}{the data.frame containing induction data with a column for the drug
concentration, a column for the fold induction observed, and, if you want
to fit the data by individual, a column for the donor ID. For an example,
please see \code{data(IndData)}. This should not be in quotes.}

\item{conc_column}{the name of the column within DF that contains
concentration data. This should be unquoted.}

\item{fold_change_column}{the name of the column within DF that contains
fold-change data, e.g., mRNA measurements or activity. This should be
unquoted.}

\item{model}{which model(s) you would like to use. The four model options
  are: \describe{

  \item{"Indmax"}{the Indmax model. This assumes a hyperbolic shape to the
  interaction when inducer concentration is plotted on the x axis and fold
  induction is plotted on the y. \deqn{Ind = (Indmax * I)/(IndC50 + I)} where
  Ind = the y axis observed fold induction, I is the inducer concentration,
  and Indmax (the maximum induction), and IndC50 (the inducer concentration
  at 1/2 Indmax) are the fitted parameters.}

  \item{"IndmaxSlope"}{the Indmax model with the addition of a slope n
  \deqn{fold induction = 1 + Indmax * I^n / (IndC50^n + I^n)}}

  \item{"Slope"}{slope only: \deqn{fold induction = I * n}}

  \item{"Sig3Param"}{Sigmoidal 3-parameter model (often used by Xenotech):
  \deqn{fold induction = Indmax / (1 + exp( -(I - IndC50)/n ))}}

  \item{"all"}{All 4 models will be fitted to the data.} }}

\item{measurement}{the type of measurement used. Options are "mRNA" or
"activity". This only affects the y axis labels on the output graph(s).}

\item{donor_column}{the name of the column within DF that contains the donor
IDs, unquoted}

\item{fitByDonor}{TRUE (default) or FALSE for whether to fit the data by
individual donor}

\item{weights}{weighting scheme to use for the regression. User may supply a
numeric vector of weights to use or choose from "none", "1/x", "1/x^2",
"1/y" or "1/y^2" (default). Be careful that you don't have any infinite
values or this will fail!}

\item{color_set}{the set of colors to use. Options: \describe{

  \item{"default"}{a set of colors from Cynthia Brewer et al. from Penn State
  that are friendly to those with red-green colorblindness. The first three
  colors are green, orange, and purple. This can also be referred to as
  "Brewer set 2".}

  \item{"Brewer set 1"}{colors selected from the Brewer palette "set 1". The
  first three colors are red, blue, and green.}

  \item{"ggplot2 default"}{the default set of colors used in ggplot2 graphs
  (ggplot2 is an R package for graphing.)}

  \item{"rainbow"}{colors selected from a rainbow palette. The default
  palette is limited to something like 6 colors, so if you have more than
  that, that's when this palette is most useful. It's \emph{not} very useful
  when you only need a couple of colors.}

  \item{"blue-green"}{a set of blues and greens}

  \item{"Tableau"}{uses the standard Tableau palette; requires the "ggthemes"
  package}

  \item{"viridis"}{from the eponymous package by Simon Garnier and ranges
  colors from purple to blue to green to yellow in a manner that is
  "printer-friendly, perceptually uniform and easy to read by those with
  colorblindness", according to the package author}}}

\item{y_axis_limits}{optionally set the Y axis limits, e.g., \code{c(1, 5)}.
If left as NA, the Y axis limits will be automatically selected. (Reminder:
Numeric data should not be in quotes.)}

\item{hline_foldinduct1}{TRUE or FALSE (default) on whether to include a
dotted red line where the fold induction = 1.}

\item{num_sigfig}{optionally specify the number of significant figures you
would like any output rounded to. If left as NA, no rounding will be
performed.}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My induction graph.png". If you leave off ".png", it
will be saved as a png file, but if you specify a different file extension,
it will be saved as that file format. Acceptable extensions are "eps",
"ps", "jpeg", "jpg", "tiff", "png", "bmp", or "svg". Leaving this as NA
means the file will not be automatically saved to disk.}

\item{fig_height}{figure height in inches; default is 5}

\item{fig_width}{figure width in inches; default is 5.5}

\item{save_output}{optionally save the output by supplying a file name in
quotes here, e.g., "My fitted induction parameters.csv". If you leave off
".csv", it will still be saved as a csv file.}
}
\value{
Returns a list of \describe{ \item{Fit}{the fitted parameters}
  \item{Fit_means}{the mean fitted parameters for all donors} \item{Graph}{a
  graph or set of graphs of the data with the fitted parameters depicted as
  lines} \item{Curve}{a data.frame of numeric data used to graph the fitted
  parameters in case you'd like to plot the fitted parameters in some other
  way}}
}
\description{
\code{inductFit} fits induction data -- either activity or mRNA expression --
to one or all of four models for calculating Ind_max, Ind50, and, when
appropriate, a slope. Like Howie's R script for fitting induction data
("Induction_fit_script_ver2.r"), default weighting is by 1/y^2, although you
can change that with the "weights" argument. IMPORTANT: One way in which this
function differs from Howie's script is that no upper or lower bounds for
parameter estimates are in use. This means that, if sufficient data to
describe the curve do not exist, the function will fail to deliver any fitted
parameters, which is meant to be a benefit but could be an annoyance
depending on your perspective. Because this does not include those
boundaries, you may get different results between this function and Howie's
script.
}
\examples{
data(IndData)

inductFit(IndData, model = "Indmax")$Graph
inductFit(IndData \%>\% rename(Conc_uM = Concentration_uM),
          conc_column = Conc_uM, model = "IndmaxSlope")$Graph
inductFit(IndData, model = "Slope")$Graph
inductFit(IndData, model = "Sig3Param")

inductFit(IndData, model = "IndmaxSlope", measurement = "activity")

MyFit <- inductFit(IndData, model = "all")
MyFit$Graph; MyFit$Fit

MyFit <- inductFit(IndData, model = "Sig3Param", fitByDonor = TRUE, donor_column = DonorID)
MyFit$Fit; MyFit$Fit_means; MyFit$Graph

MyFit <- inductFit(IndData, model = "all", fitByDonor = TRUE, donor_column = DonorID)
MyFit$Fit; MyFit$Fit_means; MyFit$Graph



}
