% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inductFit.R
\name{inductFit}
\alias{inductFit}
\title{Fit induction data to calculate EC50, Emax, and/or slope}
\usage{
inductFit(
  DF,
  conc_column = Concentration_uM,
  fold_change_column = FoldInduction,
  donor_column = DonorID,
  fitByDonor = TRUE,
  model = "all",
  graph_mean_of_fits = FALSE,
  measurement = "mRNA",
  enzyme = "CYP3A4",
  drug = NA,
  weights = "1/y^2",
  bounds_Emax = NA,
  bounds_EC50 = NA,
  bounds_slope = NA,
  omit = NA,
  color_set = "default",
  y_axis_limits = NA,
  hline_foldinduct1 = FALSE,
  vert_line = NA,
  Imaxu_line = NA,
  rounding = NA,
  graph_title = NA,
  graph_title_size = 14,
  save_graph = NA,
  fig_height = 5,
  fig_width = 5.5,
  save_output = NA
)
}
\arguments{
\item{DF}{the data.frame containing induction data with a column for the drug
concentration, a column for the fold induction observed, and, if you want
to fit the data by individual, a column for the donor ID. For an example,
please see \code{data(IndData)}. This should not be in quotes.}

\item{conc_column}{the name of the column within DF that contains
concentration data. This should be unquoted.}

\item{fold_change_column}{the name of the column within DF that contains
fold-change data, e.g., mRNA measurements or activity. This should be
unquoted.}

\item{donor_column}{the name of the column within DF that contains the donor
IDs, unquoted}

\item{fitByDonor}{TRUE (default) or FALSE for whether to fit the data by
individual donor}

\item{model}{which model(s) you would like to use. The four model options
  are: \describe{

  \item{"Emax"}{the Emax model. This assumes a hyperbolic shape to the
  interaction when inducer concentration is plotted on the x axis and fold
  induction is plotted on the y. \deqn{fold induction = 1 + (Emax * I)/(EC50
  + I)} where I is the inducer concentration, and Emax (the maximum
  induction), and EC50 (the inducer concentration at 1/2 Emax) are the fitted
  parameters.}

  \item{"EmaxSlope"}{the Emax model with the addition of a slope n \deqn{fold
  induction = 1 + Emax * I^n / (EC50^n + I^n)}}

  \item{"slope"}{slope only: \deqn{fold induction = 1 + I * n}}

  \item{"Sig3Param"}{Sigmoidal 3-parameter model (often used by Xenotech):
  \deqn{fold induction = Indmax / (1 + exp( -(I - IndC50)/n ))} SPECIAL NOTE:
  The sigmoidal 3-parameter model is the \emph{only} model that determines
  Indmax instead of Emax!}

  \item{"all" (default)}{All 4 models will be fitted to the data.} }}

\item{graph_mean_of_fits}{TRUE or FALSE (default) for whether to graph a line
representing the mean of all donors' fitted parameters. Please note that
this is NOT line showing the model fitted to the mean data; instead, this
is the mean of all the donors individually fitted parameters. This only
applies when \code{fitByDonor = TRUE}. If \code{graph_mean_of_fits =
FALSE}, graph will show a line for each donor's fitted parameters.}

\item{measurement}{the type of measurement used. Options are "mRNA" or
"activity". This only affects the y axis labels on the output graph(s).}

\item{enzyme}{the enzyme involved. Default is "CYP3A4" but can be set to NA
to leave as unspecified. This only affects the figure and table headings
and captions in the Word document output.}

\item{drug}{the drug used in the incubations. Default is NA to leave this
unspecified, but supplying a value will include the drug name in the x-axis
label on the graphs as well as the figure and table headings and captions
in the Word document output.}

\item{weights}{weighting scheme to use for the regression. User may supply a
numeric vector of weights to use or choose from "none", "1/x", "1/x^2",
"1/y" or "1/y^2" (default). Be careful that you don't have any infinite
values or this will fail!}

\item{bounds_Emax}{upper and lower boundaries for fitting Indmax or Emax,
specified as a vector of numbers, e.g., \code{c(1, 20)}; default is no
boundaries. If you only want to specify an upper limit, set the first
number to \code{-Inf} and the second number to that value. If you only want
to specify a lower limit, set the first number to that value and the second
number to \code{Inf}.}

\item{bounds_EC50}{upper and lower boundaries for fitting IndC50 or EC50,
specified as a vector of numbers, e.g., \code{c(1, 20)}; default is no
boundaries. If you only want to specify an upper limit, set the first
number to \code{-Inf} and the second number to that value. If you only want
to specify a lower limit, set the first number to that value and the second
number to \code{Inf}.}

\item{bounds_slope}{upper and lower boundaries for fitting the slope,
specified as a vector of numbers, e.g., \code{c(1, 20)}; default is no
boundaries. If you only want to specify an upper limit, set the first
number to \code{-Inf} and the second number to that value. If you only want
to specify a lower limit, set the first number to that value and the second
number to \code{Inf}.}

\item{omit}{An index of which, if any, samples to omit from regression(s).
These samples will be depicted as open circles in the graph but will not be
included in the regression. An example of acceptable input where, say, you
want to omit samples where the concentration was greater than 100 uM but
still want to see those points on the graph(s): \code{omit =
which(DF$Concentration > 100)}}

\item{color_set}{the set of colors to use. Options: \describe{

  \item{"default"}{a set of colors from Cynthia Brewer et al. from Penn State
  that are friendly to those with red-green colorblindness. The first three
  colors are green, orange, and purple. This can also be referred to as
  "Brewer set 2".}

  \item{"Brewer set 1"}{colors selected from the Brewer palette "set 1". The
  first three colors are red, blue, and green.}

  \item{"ggplot2 default"}{the default set of colors used in ggplot2 graphs
  (ggplot2 is an R package for graphing.)}

  \item{"rainbow"}{colors selected from a rainbow palette. The default
  palette is limited to something like 6 colors, so if you have more than
  that, that's when this palette is most useful. It's \emph{not} very useful
  when you only need a couple of colors.}

  \item{"blue-green"}{a set of blues fading into greens. This palette can be
  especially useful if you are comparing a systematic change in some
  continuous variable -- for example, increasing dose or predicting how a
  change in intrinsic solubility will affect concentration-time profiles --
  because the direction of the trend will be clear.}

  \item{"blues"}{a set of blues fading light blue to dark blue. Like
  "blue-green", this palette can be especially useful if you are comparing a
  systematic change in some continuous variable.}

  \item{"Tableau"}{uses the standard Tableau palette; requires the "ggthemes"
  package}

  \item{"viridis"}{from the eponymous package by Simon Garnier and ranges
  colors from purple to blue to green to yellow in a manner that is
  "printer-friendly, perceptually uniform and easy to read by those with
  colorblindness", according to the package author}}}

\item{y_axis_limits}{optionally set the Y axis limits, e.g., \code{c(1, 5)}.
If left as NA, the Y axis limits will be automatically selected. (Reminder:
Numeric data should not be in quotes.)}

\item{hline_foldinduct1}{TRUE or FALSE (default) on whether to include a
dotted red line where the fold induction = 1.}

\item{vert_line}{optionally include a vertical dotted red line at some
concentration. Use a numeric value for the concentration you want or leave
as NA (default) for no vertical line.}

\item{Imaxu_line}{optionally specify a value for 30*Imax,u (see p. 5 USFDA
Guidance (2020) "In Vitro Drug Interaction Studies - Cytochrome P450 Enzyme
and Transporter Mediated Drug Interactions") and this will create red
dotted line that goes down from the top of the graph at that concentration
and then across to the y axis at the 2-fold change level. Points within
that upper left corner are a potential concern for induction. Use a numeric
value for the concentration you want or leave as NA (default) for no line.}

\item{rounding}{option for what rounding to perform, if any. Options are:
\describe{\item{NA or "Consultancy"}{All output will be rounded according
to Simcyp Consultancy Team standards: to three significant figures when the
value is < 100 or to the ones place if the value is >= 100. Please see the
function \code{\link{round_consultancy}}, which does the rounding here.}
\item{"none"}{No rounding will be performed.} \item{"significant X" where
"X" is a number}{Output will be rounded to X significant figures. "signif
X" also works fine.} \item{"round X" where "X" is a number}{Output will be
rounded to X digits} \item{"Word only"}{Output saved to Word or a csv file
will be rounded using the function \code{\link{round_consultancy}}, but
nothing will be rounded in the output R object. This can be useful when you
want to have nicely rounded and formatted output in a Word file but you
\emph{also} want to use the results from \code{calc_PK_ratios} to make
forest plots, which requires numbers that are \emph{not} rounded.}}}

\item{graph_title}{optionally specify a title that will be centered across
your graph or set of graphs}

\item{graph_title_size}{the font size for the graph title if it's included;
default is 14. This also determines the font size of the graph labels.}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My induction graph.png". If you leave off ".png", it
will be saved as a png file, but if you specify a different file extension,
it will be saved as that file format. Acceptable extensions are "eps",
"ps", "jpeg", "jpg", "tiff", "png", "bmp", or "svg". Leaving this as NA
means the file will not be automatically saved to disk.}

\item{fig_height}{figure height in inches; default is 5}

\item{fig_width}{figure width in inches; default is 5.5}

\item{save_output}{optionally save the output by supplying a file name in
quotes here, e.g., "My fitted induction parameters.csv". Saving as a csv
file will save only the fitted parameters. However, if you would like a
Word file with a) the equations used, b) the fitted parameters, and c) the
graphs, end this file name with ".docx" instead.}
}
\value{
Returns a list of \describe{ \item{Fit}{the fitted parameters}
  \item{Fit_means}{the mean fitted parameters for all donors} \item{Graph}{a
  graph or set of graphs of the data with the fitted parameters depicted as
  lines} \item{Curve}{a data.frame of numeric data used to graph the fitted
  parameters in case you'd like to plot the fitted parameters in some other
  way}}
}
\description{
\code{inductFit} fits induction data -- either activity or mRNA expression --
to one or all of four models for calculating Indmax or Emax, EC50, and, when
appropriate, a slope. \strong{Two important notes:} \itemize{\item{With the
exception of the sigmoidal 3-parameter model, the fitted parameter describing
maximal induction is Emax and \emph{not} Indmax, the parameter used in the
Simcyp Simulator. Note that \strong{Indmax = Emax + 1}.} \item{If you're
familiar with Howie's R script for fitting induction data
("Induction_fit_script_ver2.r"), one way in which this function differs is
that no upper or lower bounds for parameter estimates are in use. This means
that, if sufficient data to describe the curve do not exist, the function
will fail to deliver any fitted parameters, which is meant to be a benefit
but could be an annoyance depending on your perspective. Because this does
not include those boundaries, you may get different results between this
function and Howie's script.}}
}
\examples{
data(IndData)

inductFit(IndData, model = "Emax")$Graph
inductFit(IndData \%>\% rename(Conc_uM = Concentration_uM),
          conc_column = Conc_uM, model = "EmaxSlope")$Graph
inductFit(IndData, model = "slope")$Graph
inductFit(IndData, model = "Sig3Param")

inductFit(IndData, model = "EmaxSlope", measurement = "activity")

MyFit <- inductFit(IndData, model = "all")
MyFit$Graph; MyFit$Fit

MyFit <- inductFit(IndData, model = "Sig3Param", fitByDonor = TRUE, donor_column = DonorID)
MyFit$Fit; MyFit$Fit_means; MyFit$Graph

MyFit <- inductFit(IndData, model = "all", fitByDonor = TRUE, donor_column = DonorID)
MyFit$Fit; MyFit$Fit_means; MyFit$Graph



}
