% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_copyright.R
\name{insert_copyright}
\alias{insert_copyright}
\title{Insert the standard Certara UK Limited copyright box that appears in compound
summary files}
\usage{
insert_copyright(font = "Fourier")
}
\arguments{
\item{font}{font to use. Default is "Fourier" and any fonts available on your
machine in either Word or PowerPoint should be acceptable. If you get Times
New Roman in your table when you asked for something else, it means that
that font isn't available or maybe wasn't spelled the way R is expecting
it. For example, "Calibri" works but "Calibri (Body)" doesn't even though
the latter is listed in PowerPoint and Word.}
}
\value{
a flextable with the current copyright information in it
}
\description{
Insert the standard Certara UK Limited copyright box that appears in compound
summary files
}
\examples{
insert_copyright()

}
