% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ct_caption.R
\name{make_ct_caption}
\alias{make_ct_caption}
\title{INTERNAL - make caption for conc-time plots}
\usage{
make_ct_caption(
  ct_dataframe,
  single_or_multiple_profiles = "single",
  plot_type = "concentration-time",
  existing_exp_details = NA,
  mean_type = "arithmetic",
  linear_or_log = "linear",
  tissue = "plasma",
  compoundID = "substrate",
  figure_type = "percentiles",
  prettify_compound_names = TRUE,
  name_clinical_study = NA,
  hline_position = NA,
  hline_style = "red dotted",
  vline_position = NA,
  vline_style = "red dotted"
)
}
\arguments{
\item{ct_dataframe}{ct_dataframe}

\item{single_or_multiple_profiles}{whether this is for a ct_plot graph (only
1 compound, tissue, file, etc.) or a graph with multiple profiles (multiple
compounds, tissues, files, etc.). Options are "single" (default) or
"multiple".}

\item{plot_type}{options are "concentration-time", "enzyme-abundance", "dissolution-profile", "release-profile". May
change these to be prettier.}

\item{existing_exp_details}{OR supply: a data.frame with the following
columns: File, Population, Substrate, Inhibitor1, Inhibitor1Metabolite,
Inhibitor2, ... If left as NA, we'll note some generic parameters and make
them bold to note that they should be changed.}

\item{mean_type}{"arithmetic" (default) or "geometric"}

\item{linear_or_log}{"linear" (default), "both", "both horizontal", "both
vertical", "log", "semi-log", "horizontal and vertical"}

\item{tissue}{default is plasma}

\item{compoundID}{default is substrate}

\item{figure_type}{default is percentiles}

\item{prettify_compound_names}{default is TRUE}

\item{name_clinical_study}{optionally specify the name of the clinical study
for any observed data. This only affects the caption of the graph. For
example, specifying \code{name_clinical_study = "101, fed cohort"} will
result in a figure caption that reads in part "Clinical Study 101, fed
cohort".}

\item{hline_position}{hline_position}

\item{hline_style}{hline_style}

\item{vline_position}{vline_position}

\item{vline_style}{vline_style}

\item{study_design_matches_obs}{optionally specify whether the study design
for the simulated data matched that of any observed data. This only affects
the caption of the graph. If set to TRUE, this assumes that the number of
subjects in the clinical study matches the number of subjects per trial in
the simulated data.}
}
\value{
a string of text for making a caption
}
\description{
INTERNAL - make caption for conc-time plots
}
\examples{
# none 
}
