% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_summary_table.R
\name{make_model_summary_table}
\alias{make_model_summary_table}
\title{Summarize a model pretty simply - UNDER CONSTRUCTION}
\usage{
make_model_summary_table(
  existing_exp_details,
  sims_to_include = "all",
  compoundID = "substrate"
)
}
\arguments{
\item{existing_exp_details}{the output from running
\code{\link{extractExpDetails_mult}}}

\item{sims_to_include}{optionally specify which simulation files you'd like
to include in the annotated output. Acceptable input: "all" to get
everything or else a character vector of the sim results files you want. If
these are not present in existing_exp_details, they will be ignored.}

\item{compoundID}{For which compound do you want to summarize the model
parameters? Options are "substrate", "primary metabolite 1", "primary
metabolite 2", "secondary metabolite", "inhibitor 1", "inhibitor 2", or
"inhibitor 1 metabolite".}
}
\value{
a formatted table
}
\description{
\code{make_model_summary_table} is meant for use with writing compound
summary pdfs. It includes minimal error catching at this point. NB: This
rounds numeric values to the 5th digit.
}
\examples{
# none yet
}
