% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simulation_directory.R
\name{make_simulation_directory}
\alias{make_simulation_directory}
\title{Make a directory of simulations indicating which simulation file is located
in which folder}
\usage{
make_simulation_directory(
  project_folder = NA,
  recursive = TRUE,
  which_files = NA,
  existing_exp_details = NA,
  relative_paths = TRUE,
  save_table = NA,
  wrap_text = FALSE
)
}
\arguments{
\item{project_folder}{location of the project folder to search. Please note
that R requires forward slashes ("/") in file paths rather than the back
slashes Microsoft uses. If left as NA (default), we'll assume your current
working directory is the top level of your project folders. This function
\emph{only} searches for simulation files in the project folder and any
subfolders.}

\item{recursive}{TRUE (default) or FALSE for whether to recursively check for
simulation files in subfolders}

\item{which_files}{which files to include in the directory of simulations.
  Options are:
  \describe{\item{NA}{all xlsx or wksz files in the current folder and any
  subfolders OR, if you supply something for \code{existing_exp_details},
  the files included in that object}

  \item{a single text string such as "abc"}{only include files with that
  specific text}

  \item{a character vector of specific file names, e.g., \code{which_files =
  c("abc.xlsx", "def.xlsx")}}{only include files listed}}}

\item{existing_exp_details}{the output from running
\code{\link{extractExpDetails_mult}}}

\item{relative_paths}{TRUE (default) or FALSE for whether to put relative
(rather than full) paths in the table. This means "relative to your current
working directory", so it matters what folder you'r in when you call this
function.}

\item{save_table}{optionally specify either an Excel or csv file name for
saving your simulation directory}

\item{wrap_text}{TRUE or FALSE (default) for whether to wrap text in the
Excel output}
}
\value{
a data.frame of simulation files and their respective file paths
}
\description{
\code{make_simulation_directory} will create a data.frame of
  simulations in a given project folder, which XML files match which
  simulations when applicable (you'll have to provide that information with
  the argument \code{existing_exp_details}), and, optionally, save that
  data.frame to an Excel file. It will also perform a few checks such as
  checking whether the file names comply with USFDA and Consultancy Team
  standards.

  This is wicked fast.
}
\examples{
# none yet

}
