% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simulation_directory.R
\name{make_simulation_directory}
\alias{make_simulation_directory}
\title{Make a directory of simulations indicating which simulation file is located
in which subfolder}
\usage{
make_simulation_directory(
  project_folder = NA,
  sim_data_files = "recursive",
  existing_exp_details = NA,
  search_workspaces_for_obsfile = FALSE,
  save_table = NA,
  overwrite = "ask"
)
}
\arguments{
\item{project_folder}{location of the project folder to search. Please note
that R requires forward slashes ("/") in file paths rather than the back
slashes Microsoft uses. If left as NA (default), we'll assume your current
working directory is the top level of your project folders.}

\item{sim_data_files}{which files to include in the directory of simulations.
  This only applies when you don't supply anything for the argument
  \code{existing_exp_details}. Options are:
  \describe{

  \item{"recursive" (default)}{all .xlsx, .db, or .wksz files in the current folder and any
  subfolders}

  \item{NA}{all .xlsx, .db, or .wksz files in the current folder}

  \item{a single text string such as "dev"}{include any files in the current
  folder or any folders below it that have that specific text}

  \item{a character vector of specific file names, e.g., \code{sim_data_files =
  c("abc.xlsx", "def.xlsx")}}{only include the files listed. We'll figure
  out which folder they're in.}}}

\item{existing_exp_details}{optionally supply the output from running
\code{\link{extractExpDetails_mult}} to get only the simulation files
included there in your simulation directory. If you supply something here,
whatever you supply for \code{sim_data_files} will be ignored. This will
also be used to figure out which XML files go with which simulations.}

\item{search_workspaces_for_obsfile}{TRUE or FALSE (default) for whether to
search through any workspace files and check for possible XML overlay
files. This runs considerably slower when set to TRUE and will take about
an additional 2 seconds per simulation. If you have supplied something for
'existing_exp_details', you don't need this and we'll ignore it if you set
this to TRUE.}

\item{save_table}{optionally specify an Excel file name for saving your
simulation directory. If you don't include the file extension ".xlsx",
we'll add it.}

\item{overwrite}{Should we overwrite if your Excel file already exists and
already has a tab named "Simulation directory"? Options are "yes" to always
overwrite, "no" to never overwrite, or "ask" (default), which means that we
will ask you whether to overwrite and give you a chance to supply a
different file name if the one you supplied already exists.}
}
\value{
a data.frame of simulation files and their respective file paths
}
\description{
The function \code{make_simulation_directory} will create a
  data.frame of simulations in a given project folder, the associated XML
  files when applicable (you'll have to provide that information with the
  argument \code{existing_exp_details}), and, optionally, save that
  data.frame to an Excel file. It will also check whether the file names
  comply with USFDA and Consultancy Team standards and check whether the same
  simulation is saved in more than one place.

  This is wicked fast.*

  * Caveat: Unless you search all the workspaces for XML observed overlay data files. Then it
  takes about 2 seconds per simulation.
}
\examples{
# none yet
}
