% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simulation_directory.R
\name{make_simulation_directory}
\alias{make_simulation_directory}
\title{Make a directory of simulations indicating which simulation file is located
in which folder}
\usage{
make_simulation_directory(
  project_folder = NA,
  recursive = TRUE,
  which_files = NA,
  existing_exp_details = NA,
  relative_paths = TRUE,
  save_table = NA
)
}
\arguments{
\item{project_folder}{location of the project folder to search. Please note
that R requires forward slashes ("/") in file paths rather than the back
slashes Microsoft uses. If left as NA (default), we'll assume your current
working directory is the top level of your project folders. This function
\emph{only} searches for simulation files in the project folder and any
subfolders.}

\item{recursive}{TRUE (default) or FALSE for whether to recursively check for
simulation files in subfolders}

\item{which_files}{which files to include in the directory of simulations.
  Options are:
  \describe{\item{NA}{all xlsx or wksz files in the current folder and any
  subfolders}

  \item{"use existing_exp_details"}{only include simulations in the object
  supplied for \code{existing_exp_details}, which you can generate by running
  \code{\link{extractExpDetails_mult}}.}

  \item{a single text string such as "abc"}{only include files with that
  specific text}

  \item{a character vector of specific file names, e.g., \code{which_files =
  c("abc.xlsx", "def.xlsx")}}{only include files listed}}}

\item{existing_exp_details}{the output from running
\code{\link{extractExpDetails_mult}}; only applicable when
\code{which_files = "use existing_exp_details"}}

\item{relative_paths}{TRUE (default) or FALSE for whether to put relative
(rather than full) paths in the table}

\item{save_table}{optionally specify either an Excel or csv file name for
saving your simulation directory}
}
\value{
a data.frame of simulation files and their respective file paths
}
\description{
Make a directory of simulations indicating which simulation file is located
in which folder
}
\examples{
# none yet

}
