% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_annotations.R
\name{make_table_annotations}
\alias{make_table_annotations}
\title{Make table headings and captions for use internally when saving PK tables to
Word}
\usage{
make_table_annotations(
  MyPKResults,
  MyFile = NA,
  MyCompoundID = NA,
  prettify_compound_names,
  Deets = NA,
  MeanType,
  tissue = NA,
  DosesIncluded = NA,
  name_clinical_study = NA,
  return_all_objects = FALSE
)
}
\arguments{
\item{MyPKResults}{only the PK table}

\item{MyFile}{the specific sim_data_file in question}

\item{MyCompoundID}{substrate, inhibitor 1, etc. for this specific table. If
length > 1, then generic info will be used.}

\item{prettify_compound_names}{pass through from parent function}

\item{Deets}{experimental details for only the single sim_data_file in
question rather than for all possible files that may have been contained in
existing_exp_details}

\item{MeanType}{mean type used}

\item{tissue}{tissue in which the PK parameters were measured}

\item{DosesIncluded}{which doses were included? options: "Dose1 Last" when
both 1st and last-dose or user-defined interval info included, "Dose1" when
it's only dose 1, "Last" when it's either last or user-defined info
included. Leave as NA to make a guess, which works well when
include_dose_num = TRUE in original PK table call but less so when
include_dose_num = FALSE. Set to "no dose num included" to use the
most-generic text for the annotations.}

\item{name_clinical_study}{optionally specify the name(s) of the clinical
study or studies for any observed data. This only affects the caption of
the graph. For example, specifying \code{name_clinical_study = "101, fed
cohort"} will result in a figure caption that reads in part "clinical study
101, fed cohort". If you have more than one study, that's fine; we'll take
care of stringing them together appropriately. Just list them as a
character vector, e.g., \code{name_clinical_study = c("101",
"102", "103")} will become "clinical studies 101, 102, and 103."}

\item{return_all_objects}{T or F (default) for whether to return a ton of
objects for use downstream}
}
\value{
a list with a) table_heading text and b) table_caption text
}
\description{
\code{make_table_annotations} is meant to make a table heading and caption
for a table with a single set of results but will return more generic info
for tables with multiple sets of results. UNDER CONSTRUCTION.
}
\examples{
# None yet


}
