% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_annotations.R
\name{make_table_annotations}
\alias{make_table_annotations}
\title{Make table headings and captions for use internally when saving PK tables to
Word}
\usage{
make_table_annotations(
  MyPKResults,
  MyFile,
  PKpulled,
  MyCompoundID,
  prettify_compound_names,
  Deets,
  MeanType,
  tissue
)
}
\arguments{
\item{MyPKResults}{only the PK table}

\item{MyFile}{the specific sim_data_file in question}

\item{PKpulled}{PK that was pulled -- the coded version, i.e., "AUCinf_dose1"
or "Cmax_last"}

\item{MyCompoundID}{substrate, inhibitor 1, etc. for this specific table}

\item{prettify_compound_names}{pass through from parent function}

\item{Deets}{experimental details for only the single sim_data_file in
question rather than for all possible files that may have been contained in
existing_exp_details}

\item{MeanType}{mean type used}

\item{tissue}{tissue in which the PK parameters were measured}
}
\value{
a list with a) TableHeading text and b) TableCaption text
}
\description{
\code{make_table_annotations} is meant to make a table heading and caption
for a table with a single set of results. This is meant for use with
pksummary_table or for pksummary_mult when the results are split into
multiple tables rather than all together in one. Note that this is designed
for scenarios where, if there's an effector, there's only one. The text won't
be quite right if there are 2 effectors. UNDER CONSTRUCTION.
}
\examples{
# None yet


}
