% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_units.R
\name{match_units}
\alias{match_units}
\title{Match concentration and time units as needed}
\usage{
match_units(DF_to_adjust, goodunits)
}
\arguments{
\item{goodunits}{either a data.frame that has the desired concentration and
time units OR a named character vector with the desired units such as
\code{c("Conc_units" = "ng/mL", "Time_units" = "hours")}}

\item{DF}{the data.frame of concentration-time data with units that may need
to be adjusted. This must include the columns "Conc", "Conc_units", "Time",
and "Time_units".}
}
\value{
a data.frame with the corrected units
}
\description{
\code{match_units} adjusts the units for a data.frame containing
concentration-time data as necessary to match some desired set of
concentration and time units.
}
\examples{
DF_to_adjust <- match_units(DF_to_adjust = SimulatedData,
                            goodunits = ObsData)

}
