% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MDZ_Keto}
\alias{MDZ_Keto}
\title{Plasma concentration-time data for midazolam and ketoconazole}
\format{
A data.frame
}
\usage{
MDZ_Keto
}
\description{
Plasma concentration-time data of multiple-dose midazolam plus multiple-dose
ketoconazole, including data for both substrate and inhibitor and observed
data. Format of the data are the same as the output from
\code{\link{extractConcTime_mult}}.
}
\keyword{datasets}
