% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontogeny_plot.R
\name{ontogeny_plot}
\alias{ontogeny_plot}
\title{Graph the ontogenies of drug-metabolizing enzymes and transporters}
\usage{
ontogeny_plot(
  DME = NA,
  DME_type = NA,
  simulator_version = 22,
  compare_to_no_ontogeny = FALSE,
  age_range = c(0, 18),
  x_axis_interval = 3,
  facet1_column,
  facet1_title = NA,
  facet2_column,
  facet2_title = NA,
  colorBy_column
)
}
\arguments{
\item{DME}{drug-metabolizing enzymes (DME) and transporters to plot. Options
are listed in the column "DME" of the data.frame OntogenyEquations, which
is saved with this package. To see it, type into the console
\code{view(OntogenyEquations)}}

\item{DME_type}{type of DME or transporter}

\item{simulator_version}{Simcyp Simulator version to display. Options are 21,
22 (default), or 23.}

\item{compare_to_no_ontogeny}{TRUE or FALSE (default) for whether to show a
line on the graph for no ontogeny, which would be a horizontal line at 1.}

\item{age_range}{age range in years as a numeric vector. Default is
\code{age_range = c(0, 18)}, which will include ages 0 to 18 years in the
graph.}

\item{x_axis_interval}{optionally specify the x-axis interval to use. Default
is an interval of 3 years.}

\item{facet1_column}{optionally break up the graph into small multiples; this
specifies the first of up to two columns to break up the data by, and the
designated column name should be unquoted, e.g., \code{facet1_column =
Tissue}. If \code{floating_facet_scale} is FALSE and you haven't specified
\code{facet_ncol} or  \code{facet_nrow}, then \code{facet1_column} will
designate the rows of the output graphs.}

\item{facet1_title}{optionally specify a title to describe facet 1. This is
ignored if \code{floating_facet_scale} is TRUE or if you have specified
\code{facet_ncol} or \code{facet_nrow}.}

\item{facet2_column}{optionally break up the graph into small multiples; this
specifies the second of up to two columns to break up the data by, and the
designated column name should be unquoted, e.g., \code{facet2_column =
CompoundID}. If \code{floating_facet_scale} is FALSE and you haven't
specified \code{facet_ncol} or  \code{facet_nrow}, then
\code{facet2_column} will designate the columns of the output graphs.}

\item{facet2_title}{optionally specify a title to describe facet 2. This is
ignored if \code{floating_facet_scale} is TRUE or if you have specified
\code{facet_ncol} or \code{facet_nrow}.}

\item{colorBy_column}{(optional) the column in \code{OntogenyEquations} that
should be used for determining which color lines should be.
This should be unquoted, e.g., \code{colorBy_column = Tissue}.}
}
\value{
a ggplot2 graph
}
\description{
\code{ontogeny_plot} uses the betas from the Simcyp Simulator to make graphs
of the ontogenies of all the DME and transporters included in the Simulator.
\strong{Note:} Only the beta version of the package includes the object
OntogenyEquations, which is required for this function to work.
}
\examples{
# none yet

}
