% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontogeny_plot.R
\name{ontogeny_plot}
\alias{ontogeny_plot}
\title{Graph the ontogenies of drug-metabolizing enzymes and transporters}
\usage{
ontogeny_plot(
  enzyme = NA,
  enzyme_type = NA,
  simulator_version = 22,
  compare_to_no_ontogeny = FALSE,
  age_range = c(0, 18),
  x_axis_interval = 3,
  facet1_column,
  facet1_title = NA,
  facet2_column,
  facet2_title = NA,
  colorBy_column,
  color_set = "default",
  legend_position = "right",
  save_graph = NA,
  fig_height = NA,
  fig_width = NA
)
}
\arguments{
\item{enzyme}{drug-metabolizing enzymes and transporters to plot, which
should be supplied as either a single string, in quotes, that we'll try to
match (e.g., "CYP3A4" will get you both the Upreti and Salem CYP3A4
ontogeny profiles as well as "CYP3A4 in vitro" and "Intestinal CYP3A4") or
as a character vector of the enzymes or transporters you want. Options are
listed in the data.frame OntogenyEquations. Whatever you supply, we'll look
for it in either the column "Enzyme", which is just the actual isoform
name, or in the column EnzymeDescription, which specifies exactly which
ontogeny profile it is. To see the object OntogenyEquations and get a
better idea of what we mean, type this into the console: 
\code{view(OntogenyEquations)}}

\item{enzyme_type}{type of enzyme or transporter. Instead of graphing only a
specific enzyme or enzymes, you can request any enzymes or transporters of
a specific type. Please see the options in "OntogenyEquations" in the
column "EnzymeType". This should be supplied as a character vector, e.g.,
\code{enzyme_type = c("transporters", "UGTs")}}

\item{simulator_version}{Simcyp Simulator version to display. Options are 21,
22 (default), or 23.}

\item{compare_to_no_ontogeny}{TRUE or FALSE (default) for whether to show a
line on the graph for no ontogeny, which would be a horizontal line at 1.}

\item{age_range}{age range in years as a numeric vector. Default is
\code{age_range = c(0, 18)}, which will include ages 0 to 18 years in the
graph.}

\item{x_axis_interval}{optionally specify the x-axis interval to use. Default
is an interval of 3 years.}

\item{facet1_column}{optionally break up the graph into small multiples; this
specifies the first of up to two columns to break up the data by, and the
designated column name should be unquoted, e.g., \code{facet1_column =
Tissue}. If \code{floating_facet_scale} is FALSE and you haven't specified
\code{facet_ncol} or  \code{facet_nrow}, then \code{facet1_column} will
designate the rows of the output graphs.}

\item{facet1_title}{optionally specify a title to describe facet 1. This is
ignored if \code{floating_facet_scale} is TRUE or if you have specified
\code{facet_ncol} or \code{facet_nrow}.}

\item{facet2_column}{optionally break up the graph into small multiples; this
specifies the second of up to two columns to break up the data by, and the
designated column name should be unquoted, e.g., \code{facet2_column =
CompoundID}. If \code{floating_facet_scale} is FALSE and you haven't
specified \code{facet_ncol} or  \code{facet_nrow}, then
\code{facet2_column} will designate the columns of the output graphs.}

\item{facet2_title}{optionally specify a title to describe facet 2. This is
ignored if \code{floating_facet_scale} is TRUE or if you have specified
\code{facet_ncol} or \code{facet_nrow}.}

\item{colorBy_column}{(optional) the column in \code{OntogenyEquations} that
should be used for determining which color lines should be. This should be
unquoted, e.g., \code{colorBy_column = Tissue}.}

\item{color_set}{the set of colors to use. Options: \describe{

  \item{"default"}{a set of colors from Cynthia Brewer et al. from Penn State
  that are friendly to those with red-green colorblindness. The first three
  colors are green, orange, and purple. This can also be referred to as
  "Brewer set 2". If there are only two unique values in the colorBy_column,
  then Brewer set 1 will be used since red and blue are still easily
  distinguishable but also more aesthetically pleasing than green and
  orange.}

  \item{"Brewer set 1"}{colors selected from the Brewer palette "set 1". The
  first three colors are red, blue, and green.}

  \item{"ggplot2 default"}{the default set of colors used in ggplot2 graphs
  (ggplot2 is an R package for graphing.)}

  \item{"rainbow"}{colors selected from a rainbow palette. The default
  palette is limited to something like 6 colors, so if you have more than
  that, that's when this palette is most useful. It's \emph{not} very useful
  when you only need a couple of colors.}

  \item{"blue-green"}{a set of blues fading into greens. This palette can be
  especially useful if you are comparing a systematic change in some
  continuous variable -- for example, increasing dose or predicting how a
  change in intrinsic solubility will affect concentration-time profiles --
  because the direction of the trend will be clear.}

  \item{"blues"}{a set of blues fading light blue to dark blue. Like
  "blue-green", this palette can be especially useful if you are comparing a
  systematic change in some continuous variable.}

  \item{"Tableau"}{uses the standard Tableau palette; requires the "ggthemes"
  package}

  \item{"viridis"}{from the eponymous package by Simon Garnier and ranges
  colors from purple to blue to green to yellow in a manner that is
  "printer-friendly, perceptually uniform and easy to read by those with
  colorblindness", according to the package author}

  \item{a character vector of colors}{If you'd prefer to set all the colors
  yourself to \emph{exactly} the colors you want, you can specify those
  colors here. An example of how the syntax should look: \code{color_set =
  c("dodgerblue3", "purple", "#D8212D")} or, if you want to specify exactly
  which item in \code{colorBy_column} gets which color, you can supply a
  named vector. For example, if you're coloring the lines by the compound ID,
  you could do this: \code{color_set = c("substrate" = "dodgerblue3",
  "inhibitor 1" = "purple", "primary metabolite 1" = "#D8212D")}. If you'd
  like help creating a specific gradation of colors, please talk to a member
  of the R Working Group about how to do that using
  \link{colorRampPalette}.}}}

\item{legend_position}{Specify where you want the legend to be. Options are
"left", "right" (default in most scenarios), "bottom", "top", or "none" if
you don't want one at all.}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My ontogeny graph.png"If you leave off ".png" or
".docx", the graph will be saved as a png file, but if you specify a
different graphical file extension, it will be saved as that file format.
Acceptable graphical file extensions are "eps", "ps", "jpeg", "jpg",
"tiff", "png", "bmp", or "svg". Do not include any slashes, dollar signs,
or periods in the file name. Leaving this as NA means the file will not be
saved to disk.}

\item{fig_height}{figure height in inches}

\item{fig_width}{figure width in inches}
}
\value{
a ggplot2 graph
}
\description{
\code{ontogeny_plot} uses the betas from the Simcyp Simulator to make graphs
of the ontogenies of all the enzymes and transporters included in the
Simulator. \strong{Note:} Only the beta version of the package includes the
object OntogenyEquations, which is required for this function to work.
}
\examples{
# none yet

}
