% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_csv.R
\name{pdf_to_csv}
\alias{pdf_to_csv}
\title{Convert a page from a pdf file to a csv file}
\usage{
pdf_to_csv(pdf_file, page, save_csv = "csv")
}
\arguments{
\item{pdf_file}{the pdf file you want to read, in quotes, e.g.,
\code{pdf_file = "Table data that I want to graph but is in pdf form.pdf"}}

\item{page}{page number(s) of the pdf you want to read as a numeric vector.
For example, to read page 10, use \code{page = 10}. To read pages 1, 3, and
5 and put them into a single csv file: \code{page = c(1, 3, 5)}. To read
pages 9 through 12: \code{page = 9:12}. If you use multiple pages, keep in
mind that, if the columns don't line up well in the pdf, they also won't
line up well in the output csv file.}

\item{save_csv}{optionally specify a file name for saving the output. If left
as "csv", the file name will be the pdf file name plus the 1st page number
requested. If set to NA, no file will be saved.}
}
\value{
a data.frame of pdf table content
}
\description{
\code{pdf_to_csv} takes as input a pdf file and a page number
and converts that to a data.frame and, optionally, saves that to a csv file.
This can be useful for getting client-supplied data into a useful format.

NOTE: This requires the package pdftools, which you may have to download and
install with \code{install.packages("pdftools")}. You only have to do that
once.
}
