% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pksummary_mult.R
\name{pksummary_mult}
\alias{pksummary_mult}
\title{Make PK summary tables from multiple simulator output files at once}
\usage{
pksummary_mult(sim_data_files, observed_PK, ...)
}
\arguments{
\item{observed_PK}{(optional) If you have a data.frame or an Excel or csv
file with observed PK parameters, supply the data.frame or the full file
name in quotes here, and the simulated-to-observed mean ratios will be
calculated. The supplied data.frame or file \emph{must} include columns for
the simulator output Excel file (title this "File") and each of the PK
parameters you would like to compare, and those column names \emph{must}
match the PK parameter options listed in \code{data(AllPKParameters)}. If
you would like the output table to include the observed data CV for any of
the parameters, add "_CV" to the end of the parameter name, e.g.,
"AUCinf_dose1_CV". Note: Whatever you list for "File" will override
anything specified for the argument \code{sim_data_file}.}

\item{...}{other arguments passed to the function \code{\link{pksummary_table}}}
}
\value{
Returns a data.frame with summary PK parameters from multiple
  simulator output files
}
\description{
Make PK summary tables from multiple simulator output files at once
}
\examples{

# No examples yet

}
