% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify_compound_name.R
\name{prettify_compound_name}
\alias{prettify_compound_name}
\title{Make a compound name prettier and more human readable}
\usage{
prettify_compound_name(CompoundName, case = "lower")
}
\arguments{
\item{CompoundName}{a compound name or vector of names to be prettified such
as "Sim-Midazolam" or "Sim-Ketoconazole-400 mg QD", which would become
"midazolam" and "ketoconazole". This works fine with tidyverse syntax such
as "mutate". .}

\item{case}{set the case to be "lower" (default), "title" (only the first
letter in each word is capitalized), or "original" (nothing about the case
changes). If you set the case to "lower" or "title", compounds with "OH-",
"O-", "S-", or "R-" as a prefix will still have that capitalized, and if
anyone has any other common exceptions they'd like, email Laura Shireman.}
}
\value{
a vector of prettier compound names
}
\description{
Make a compound name prettier and more human readable
}
\examples{

prettify_compound_name(c("Sim-Midazolam",
                         "Sim-Ketoconazole-400 mg QD",
                         "none"))

}
