% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectors.R
\name{purples}
\alias{purples}
\title{Create a vector of purples}
\usage{
purples(ncolors)
}
\arguments{
\item{ncolors}{number of colors desired}
}
\value{
a character vector of colors
}
\description{
\code{purples} is useful for getting a set of yellows to greens for
graphs. These are from the RColorBrewer package, color set "YlGn", just to
give credit where it's due, but they can be any number in length rather than
only up to 9 colors.
}
\examples{
# Create a set of 10 purples.
purples(10)

# Try using scales::show_col() to visualize the colors, ex:
scales::show_col(purples(10))

}
