% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorVectors.R
\name{rainbow}
\alias{rainbow}
\title{Create a vector of a rainbow of colors}
\usage{
rainbow(ncolors)
}
\arguments{
\item{ncolors}{number of colors desired}
}
\value{
a character vector of colors
}
\description{
\code{rainbow} is useful for getting a set of colors for graphs when you need
a LOT of colors; there are prettier and more easily distinguishable options
if you only have a couple of comparisons.
}
\examples{
# Create a rainbow of 10 colors.
rainbow(10)

# Try using scales::show_col() to visualize the colors, ex:
scales::show_col(rainbow(10))

}
