% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ReportInputForm}
\alias{ReportInputForm}
\title{Report input forms to be used in conjunction with
\code{\link{getSectionInfo}} and \code{\link{pk_table}}}
\format{
A named list of the data that fill the tabs for
  "generateReportInputForm"
}
\usage{
ReportInputForm
}
\description{
Report input forms to be used in conjunction with
\code{\link{getSectionInfo}} and \code{\link{pk_table}}
}
\keyword{datasets}
