% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_opt.R
\name{round_opt}
\alias{round_opt}
\title{INTERNAL: Rounding function}
\usage{
round_opt(x, round_fun)
}
\arguments{
\item{x}{vector to be rounded}

\item{round_fun}{which function to use to round}
}
\value{
rounded numbers that might, after running this, be character data
}
\description{
INTERNAL: Rounding function
}
\examples{
# None.

}
