% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_plot.R
\name{sensitivity_plot}
\alias{sensitivity_plot}
\title{Make graphs of sensitivity analysis results}
\usage{
sensitivity_plot(
  SA_file,
  dependent_variable,
  ind_var_label = NA,
  target_DV = NA,
  linear_or_log = "linear",
  y_axis_limits_lin = NA,
  y_axis_limits_log = NA,
  title = NA,
  save_graph = NA,
  fig_height = 4,
  fig_width = 5
)
}
\arguments{
\item{SA_file}{sensitivity analysis Excel file}

\item{dependent_variable}{dependent variable to plot. Options are: "CL",
"Dose over AUC", "AUC over Dose", "Vss", "Fg", "Fh", "fa", "CLpo", "Cmax",
"AUC", "tmax", or "plasma concentration" (just "plasma" will also work).
Other than "plasma concentration", that parameter \emph{must} be one of the
ones you requested when you ran the sensitivity analysis. Not case
sensitive.}

\item{ind_var_label}{optionally specify what text use for labeling the
independent variable (the x axis). If left as NA, R will find the value
listed next to "Run Number" on the "ASA Summary" tab, which may be
something not terribly pretty like "Fugut Value (Sub) (No Units)", and
attempt to prettify it. If we haven't set things up to prettify that value,
which is likely as we haven't used this function with very many
sensitivity-analysis scenarios yet, the value will be unchanged. (If you
have an independent variable you'd like to add to our list, email Laura
Shireman!)}

\item{target_DV}{optionally specify a target value for the dependent
variable, which will add a horizontal red dotted line to the graph where
the target lies.}

\item{linear_or_log}{make the y axis "linear" (default) or "log" for plasma
concentration-time plots (this is ignored for all other plots)}

\item{y_axis_limits_lin}{optionally set the Y axis limits for the linear
plot, e.g., \code{c(10, 1000)}. If left as the default NA, the Y axis
limits for the linear plot will be automatically selected.}

\item{y_axis_limits_log}{optionally set the Y axis limits for the semi-log
plot, e.g., \code{c(10, 1000)}. Values will be rounded down and up,
respectively, to a round number. If left as the default NA, the Y axis
limits for the semi-log plot will be automatically selected.}

\item{title}{(optional) a title to include on your graph in quotes}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My conc time graph.png" or "My conc time
graph.docx". If you leave off ".png" or ".docx" from the file name, it will
be saved as a png file, but if you specify a different graphical file
extension, it will be saved as that file format. Acceptable graphical file
extensions are "eps", "ps", "jpeg", "jpg", "tiff", "png", "bmp", or "svg".
Leaving this as NA means the file will not be saved to disk.
\strong{WARNING:} SAVING TO WORD DOES NOT WORK ON SHAREPOINT. This is a
Microsoft permissions issue, not an R issue. If you try to save on
SharePoint, you will get a warning that R will save your file instead to
your local (not OneDrive) Documents folder.}

\item{fig_height}{figure height in inches; default is 4}

\item{fig_width}{figure width in inches; default is 5}
}
a ggplot2 graph
}
\description{
Doesn't detect units yet. Would need to check that sheet names are consistent
b/c I bet they're not. Would like to add options for axis breaks and limits
as well as color schemes. Will need to add more options for labeling the
sensitivity parameter in the graphs. Does not do 3D graphs and I have no
plans to add option for 3D graphs b/c A) they're just not easy to interpret
and B) there aren't great R packages out there for 3D graphs.
}
\examples{

sensitivity_plot(SA_file = "SA example.xlsx",
                 dependent_variable = "Cmax",
                 title = "My pretty sensitivity-analysis graph that's not pink",
                 save_graph = "SA graph")

}
