% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aesthet.R
\name{set_aesthet}
\alias{set_aesthet}
\title{Set up certain aesthetics a conc-time plot}
\usage{
set_aesthet(
  line_type,
  figure_type,
  MyEffector,
  compoundToExtract,
  obs_shape,
  line_color,
  obs_color,
  obs_line_trans,
  obs_fill_trans
)
}
\arguments{
\item{line_type}{user-specified line type}

\item{figure_type}{user-specified figure type}

\item{MyEffector}{concatenated string of all effectors involved}

\item{compoundToExtract}{compound user wants}

\item{obs_shape}{user-specified observed data shape}

\item{line_color}{user-specified line color}

\item{obs_color}{user-specified observed-data color}
}
\description{
This function is specifically for setting certain aesthetic options for a
concentration-time graph and is NOT meant to be called on its own.
}
