% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aesthet.R
\name{set_aesthet}
\alias{set_aesthet}
\title{INTERNAL PACKAGE USE: Set up certain aesthetics for a conc-time plot}
\usage{
set_aesthet(
  figure_type,
  from_ct_plot,
  DDI,
  AESCols,
  line_type,
  n_line_type,
  linetype_levels = NA,
  line_color,
  n_line_color,
  color_levels = NA,
  obs_shape,
  n_obs_shape,
  obs_color,
  n_obs_color,
  obs_line_trans,
  obs_fill_trans
)
}
\arguments{
\item{figure_type}{user-specified figure type}

\item{from_ct_plot}{T or F on whether this is from ct_plot (alternative is
ct_plot_overlay). Using this to determine whether line color is black or
colors.}

\item{DDI}{T or F - only applies to ct_plot; might not even need this}

\item{AESCols}{AESCols}

\item{line_type}{user-specified line type}

\item{n_line_type}{number of line types needed}

\item{linetype_levels}{levels in linetype_column}

\item{line_color}{user-specified line color}

\item{n_line_color}{number line colors needed}

\item{color_levels}{levels in colorBy_column}

\item{obs_shape}{user-specified observed data shape}

\item{n_obs_shape}{number obs shapes needed}

\item{obs_color}{user-specified observed-data color}

\item{n_obs_color}{number obs colors needed}

\item{obs_line_trans}{obs line alpha}

\item{obs_fill_trans}{obs fill alpha}
}
\value{
several objects to use with setting graph aesthetics
}
\description{
Determine line colors, line types, point shapes, and point colors.
}
