% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_boundary_colors.R
\name{set_boundary_colors}
\alias{set_boundary_colors}
\title{INTERNAL: Create a set of boundary colors based on the color set and the
boundaries the user needs ---- CONSIDERING WHETHER IT'S WORTH IT TO MAKE
THIS. It doesn't actually save me much copying and pasting since I only use
it once in formatTable_Simcyp, where it has one set of color_set options and
boundaries and number of breaks, and only twice in so_graph, where it has a
different set of color_set options and boundaries and number of breaks.}
\usage{
set_boundary_colors(color_set, boundaries, break_type = "GMR")
}
\arguments{
\item{color_set}{either a single value, e.g., "red green" or a character
  vector of colors. Acceptable single-value color set names:\itemize{
  \item{"yellow to red" - for highlighting in PK tables}

  \item{"traffic" - for highlighting in PK tables}

  \item{"red black lines" or "red black fill" - for S/O boundary lines or fill}

  \item{"red green lines" or "red green fill" - for S/O boundary lines or fill}

  \item{"muted red green lines" or "muted red green fill" - for S/O boundary lines or fill}}}

\item{boundaries}{numeric vector of boundaries}

\item{max_num_breaks}{maximum number of breaks. For so_graph, this should be
3 and for formatTable_Simcyp, this should be 4.}
}
\value{
a character vector of colors, one for each numeric boundary
}
\description{
INTERNAL: Create a set of boundary colors based on the color set and the
boundaries the user needs ---- CONSIDERING WHETHER IT'S WORTH IT TO MAKE
THIS. It doesn't actually save me much copying and pasting since I only use
it once in formatTable_Simcyp, where it has one set of color_set options and
boundaries and number of breaks, and only twice in so_graph, where it has a
different set of color_set options and boundaries and number of breaks.
}
\examples{
# None. Internal use only.

}
