% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimcypDir}
\alias{SimcypDir}
\title{Simcyp paths}
\format{
A two-item list: \describe{ \item{LgFileDir}{the large-file store}
  \item{SharePtDir}{the SharePoint directory} \item{RDir}{the folder on the
  large-file store for sharing SimcypConsultancy R files}}
}
\usage{
SimcypDir
}
\description{
Paths for Simcyp's large-file store, SharePoint directory, and the specific
folder on the large-file store for sharing R-related files
}
\keyword{datasets}
