% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_table.R
\name{so_table}
\alias{so_table}
\title{Make simulated vs. observed tables for reports}
\usage{
so_table(
  report_input_file = NA,
  sheet_report = NA,
  PKparameters = NA,
  sheet_PKparameters = NA,
  mean_type = NA,
  variability_option = "90\% CI",
  concatVariability = FALSE,
  includeHalfLife = FALSE,
  includeTrialMeans = FALSE,
  includeCV = TRUE,
  prettify_columns = TRUE,
  checkDataSource = TRUE,
  sim_data_file = NA,
  save_table = NA,
  sectionInfo = NA
)
}
\arguments{
\item{report_input_file}{the name of the Excel file created by running
\code{\link{generateReportInputForm}}, which you have now filled out,
including the path if it's in any other directory than the current one}

\item{sheet_report}{the sheet in the Excel report template file that contains
information about this section of the report. In the original template,
this was the tab titled "table and graph input".}

\item{PKparameters}{the PK parameters to include as a character vector. To
see the full set of possible parameters to extract, enter
\code{data(AllPKParameters)} into the console. By default, if you supply a
file for \code{report_input_file}, the PK parameters included are only
those included for the observed data in that file. Otherwise, the PK
parameters will be automatically selected. An example of acceptable input
here: \code{c("AUCtau_ss", "AUCtau_ss_withInhib", "Cmax_ss",
"Cmax_ss_withInhib", "AUCtau_ratio_ss", "Cmax_ratio_ss")}. Parameters that
don't make sense for your scenario -- like asking for
\code{AUCinf_ss_withInhib} when your simulation did not include an
inhibitor or effector -- will not be included.}

\item{sheet_PKparameters}{(optional) If you want the PK parameters to be
pulled from a specific tab in the simulator output file, list that tab
here. Most of the time, this should be left as NA.}

\item{mean_type}{return "arithmetic" or "geometric" (default) means and CVs.
Only specify this if you'd like to override the value listed in
\code{sectionInfo}. If no value is specified here or in \code{sectionInfo},
the default is "geometric".}

\item{variability_option}{What type of variability would you like the table
to include? Options are: "90\% CI", "95\% CI", "95th percentiles", or any
combination of those, e.g. \code{variability_option = c("90\% CI", "95th
percentiles")} as long as they were included in your simulator output. Note
that the confidence intervals are geometric since that's what the simulator
outputs (see an AUC tab and the summary statistics; these values are the
ones for, e.g., "90% confidence interval around the geometric mean(lower
limit)"). Setting \code{variability_option = NA} will omit any of the
variability other than the CV. The CV will automatically be included unless
you omit it with \code{includeCV = FALSE}.}

\item{concatVariability}{Would you like to have the variability concatenated?
TRUE or FALSE. If "TRUE", the output will be formatted into a single row
and listed as the lower confidence interval or percentile to the upper CI
or percentile. Ex: "2400 to 2700"}

\item{includeHalfLife}{TRUE or FALSE for whether to include half life as a
parameter in the output table}

\item{includeTrialMeans}{TRUE or FALSE for whether to include the range of
trial means for a given parameter. Note: This is calculated from individual
values rather than being pulled directly from the output.}

\item{includeCV}{TRUE or FALSE for whether to include rows for CV in the
table}

\item{prettify_columns}{TRUE or FALSE for whether to make easily
human-readable column names. TRUE makes pretty column names such as "AUC0
to inf (h*ng/mL)" whereas FALSE leaves the column with the R-friendly name
from \code{\link{extractPK}}, e.g., "AUCinf_dose1".}

\item{checkDataSource}{TRUE or FALSE: Include in the output a data.frame that
lists exactly where the data were pulled from the simulator output file.
Useful for QCing.}

\item{sim_data_file}{HISTORICAL, BACK-COMPATIBILITY PURPOSES ONLY: the
simulator output file. This is for when you DON'T fill out a report input
form and instead plan to add information about the observed data later
manually.}

\item{save_table}{optionally save the output table and, if requested, the QC
info, by supplying a file name in quotes here, e.g., "My nicely formatted
table.csv". If you leave off ".csv", it will still be saved as a csv file.
If you requested both the table and the QC info, the QC file will have "-
QC" added to the end of the file name.}

\item{sectionInfo}{information about the simulated and observed data. This is
output from the function \code{\link{getSectionInfo}} and can be used
instead of listing the Excel file and report template sheet name as input.}
}
\value{
a data.frame of S/O values or a list of that data.frame (named
  "Table") plus information on where the values came from for QCing (named
  "QC")
}
\description{
\code{so_table} creates simulated vs. observed tables for reports and
presentations, including reporting means, CVs, confidence intervals or
percentiles, and ratios of simulated vs. observed mean values. Because we
need to have a standardized way to input observed data, setting up the input
for this function requires filling out an Excel template form. In this Excel
file, there will be one tab for entering your observed PK parameters and a
second tab for entering information about your simulation file. Here are the
steps to take: \enumerate{\item{Use the function
\code{\link{generateReportInputForm}} to create an Excel file where you can
enter information about your project. Example:
\code{generateReportInputForm("My report input form.xlsx")}} \item{Go to the
tab "observed data - DDI" or "observed data - no DDI", whichever is
appropriate for your situation, and enter details about your observed data.
It's ok if you don't have all the information; anything that's missing won't
be included in the final S/O table. It's also ok to rename this tab and/or
make copies of it within the same Excel file for making other S/O tables.}
\item{Go to the tab "simulated data" and fill out information here for the
specific simulation you want to compare. Make sure that whatever you list as
the tab that contains information about the observed data is \emph{exactly}
the same as the actual tab name that you filled out in step 2. Also make sure
that the file names include the full file path.} \item{Save your Excel file.}
\item{Here, within RStudio (or within the shiny app that we plan to make!),
run this function using the name of that Excel file as input for
\code{report_input_file} and the name of the "table and graph input" tab as
the input for \code{sheet}. Note: If the Excel file lives on SharePoint,
you'll need to close it or this function will just keep running and not
generate any output while it waits for access to the file.} }
}
\examples{
# so_table(report_input_file = "//certara.com/data/sites/SHF/Consult/abc-1a/Report input.xlsx",
#          sheet_report = "table and graph input", includeTrialMeans = TRUE)
}
