% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_table.R
\name{so_table}
\alias{so_table}
\title{Make simulated vs. observed tables for reports}
\usage{
so_table(
  report_input_file = NA,
  sheet_report = NA,
  sim_data_file = NA,
  PKparameters = NA,
  sheet_PKparameters = NA,
  observed_PK = NA,
  mean_type = NA,
  includeCV = TRUE,
  includeConfInt = TRUE,
  includePerc = FALSE,
  includeTrialMeans = FALSE,
  concatVariability = FALSE,
  prettify_columns = TRUE,
  checkDataSource = TRUE,
  save_table = NA
)
}
\arguments{
\item{report_input_file}{the name of the Excel file created by running
\code{\link{generateReportInputForm}}, which you have now filled out,
including the path if it's in any other directory than the current one}

\item{sheet_report}{the sheet in the Excel report template file that contains
information about the study, e.g., "study info - DDI" or "study info - no
DDI" if you haven't renamed the tab.}

\item{sim_data_file}{the simulator output file. This is only for when you
don't fill out a report input form because you either have no observed data
or you want to compare those data later manually.}

\item{PKparameters}{the PK parameters to include as a character vector.
Notes: \itemize{ \item{To see the full set of possible parameters to
extract, enter \code{data(AllPKParameters)} into the console.} \item{By
default, if you supply a file for \code{report_input_file}, the PK
parameters included are only those included for the observed data in that
file. Otherwise, the PK parameters will be automatically selected.}
\item{Parameters that don't make sense for your scenario -- like asking for
\code{AUCinf_ss_withInhib} when your simulation did not include an
inhibitor or effector -- will not be included.} \item{tmax will be listed
as median, min, and max rather than mean, lower and higher X\% confidence
interval or X percentiles. Similarly, if you request trial means, the
values for tmax will be the range of medians for the trials rather than the
range of means.}} An example of acceptable input here: \code{c("AUCtau_ss",
"AUCtau_ss_withInhib", "Cmax_ss", "Cmax_ss_withInhib", "AUCtau_ratio_ss",
"Cmax_ratio_ss")}.}

\item{sheet_PKparameters}{(optional) If you want the PK parameters to be
pulled from a specific tab in the simulator output file, list that tab
here. Most of the time, this should be left as NA.}

\item{observed_PK}{(optional) If you have a data.frame or an Excel or csv
file with observed PK parameters, supply the data.frame or the full file
name in quotes here, and the simulated-to-observed mean ratios will be
calculated. The supplied data.frame or file \emph{must} include columns for
the simulator output Excel file (title this "File") and each of the PK
parameters you would like to compare, and those column names \emph{must}
match the PK parameter options listed in \code{data(AllPKParameters)}. If
you would like the output table to include the observed data CV for any of
the parameters, add "_CV" to the end of the parameter name, e.g.,
"AUCinf_dose1_CV". Note: Whatever you list for "File" will override
anything specified for the argument \code{sim_data_file}.}

\item{mean_type}{return "arithmetic" or "geometric" (default) means and CVs.
If you supplied a report input form, only specify this if you'd like to
override the value listed there. If no value is specified here or in
\code{sectionInfo}, the default is "geometric".}

\item{includeCV}{TRUE or FALSE for whether to include rows for CV in the
table}

\item{includeConfInt}{TRUE or FALSE for whether to include whatever
confidence intervals were included in the simulator output file. Note that
the confidence intervals are geometric since that's what the simulator
outputs (see an AUC tab and the summary statistics; these values are the
ones for, e.g., "90\% confidence interval around the geometric mean(lower
limit)").}

\item{includePerc}{TRUE or FALSE for whether to include 5th to 95th
percentiles}

\item{includeTrialMeans}{TRUE or FALSE for whether to include the range of
trial means for a given parameter. Note: This is calculated from individual
values rather than being pulled directly from the output.}

\item{concatVariability}{Would you like to have the variability concatenated?
TRUE or FALSE. If "TRUE", the output will be formatted into a single row
and listed as the lower confidence interval or percentile to the upper CI
or percentile. Ex: "2400 to 2700"}

\item{prettify_columns}{TRUE or FALSE for whether to make easily
human-readable column names. TRUE makes pretty column names such as "AUC0
to inf (h*ng/mL)" whereas FALSE leaves the column with the R-friendly name
from \code{\link{extractPK}}, e.g., "AUCinf_dose1".}

\item{checkDataSource}{TRUE or FALSE: Include in the output a data.frame that
lists exactly where the data were pulled from the simulator output file.
Useful for QCing.}

\item{save_table}{optionally save the output table and, if requested, the QC
info, by supplying a file name in quotes here, e.g., "My nicely formatted
table.csv". If you leave off ".csv", it will still be saved as a csv file.
If you requested both the table and the QC info, the QC file will have "-
QC" added to the end of the file name.}
}
\value{
a data.frame of S/O values or a list of that data.frame (named
  "Table") plus information on where the values came from for QCing (named
  "QC")
}
\description{
\code{so_table} creates simulated vs. observed tables for reports and
presentations, including reporting means, CVs, confidence intervals or
percentiles, and ratios of simulated vs. observed mean values. Please see the
notes at the bottom of this help file for options for how to supply observed
data in a standardized fashion that this function can read.
}
\details{
Because we need to have a standardized way to input observed data, setting up
the input for this function requires filling out an Excel template form. Here
are the steps to take: \enumerate{\item{Use the function
\code{\link{generateReportInputForm}} to create an Excel file where you can
enter information about your project. Example:
\code{generateReportInputForm("My report input form.xlsx")}} \item{Go to the
tab "study info - DDI" or "study info - no DDI", whichever is appropriate for
your situation. Under the heading "Observed data", enter details about your
observed data. It's ok if you don't have all the information; anything that's
missing won't be included in the final S/O table. It's also ok to rename this
tab and/or make copies of it within the same Excel file for making other S/O
tables.} \item{Under the heading "Simulated data" on that same tab, fill out
the name of the specific simulator output Excel file you want to compare.}
\item{Save the report form.} \item{Back in RStudio (or within the shiny app
that we plan to make!), run this function using the file name of that Excel
report form as input for \code{report_input_file} and the name of the "study
info - DDI/no DDI" tab as the input for \code{sheet_report}. Note: If the
Excel file lives on SharePoint, you'll need to close it or this function will
just keep running and not generate any output while it waits for access to
the file.} }
}
\examples{
# so_table(report_input_file = "//certara.com/data/sites/SHF/Consult/abc-1a/Report input.xlsx",
#          sheet_report = "table and graph input", includeTrialMeans = TRUE)

# An example of how to format observed data:
observed_PK <- data.frame(File = "mdz-5mg-sd.xlsx",
                    AUCinf_dose1 = 60,
                    AUCinf_dose1_CV = 0.38,
                    Cmax_dose1 = 22,
                    Cmax_dose1_CV = 0.24)


}
