% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_table.R
\name{so_table}
\alias{so_table}
\title{Make simulated vs. observed tables for reports}
\usage{
so_table(
  Info,
  sheet = NA,
  mean_type = NA,
  variability_option = "90\% CI",
  concatVariability = FALSE,
  includeT12 = FALSE
)
}
\arguments{
\item{Info}{information about the simulated and observed data. This is output
from the function \code{\link{getSectionInfo}} or the name of the Excel
file to be piped to \code{\link{getSectionInfo}}.}

\item{sheet}{the sheet name to read if "Info" is not the output from
\code{getSectionInfo} but the name of an Excel file}

\item{mean_type}{Arithmetic or geometric means? Only specify this if you'd
like to override the value listed in "Info". If no value is specified here
or in "Info", the default is "geometric".}

\item{variability_option}{What type of variability would you like the table
to include? Options are: "90% CI", "95% CI", "95th percentiles", or any
combination of those, e.g. \code{variability_option = c("90% CI", "95th
percentiles")}}

\item{concatVariability}{Would you like to have the variability concatenated?
TRUE or FALSE. If "TRUE", the output will be formatted into a single row
and listed as the lower confidence interval or percentile to the upper. Ex:
"2400 to  2700"}

\item{includeT12}{TRUE or FALSE on whether to include t1/2 as a parameter in
the output table}
}
\value{
a data.frame
}
\description{
\code{so_table} creates simulated vs. observed tables for reports and
presentations, including reporting means, CVs, confidence intervals or
percentiles, and ratios of simulated vs. observed mean values. The data it
uses for these calculations are standard information about the observed and
simulated data in the format generated by running
\code{\link{getSectionInfo}}.
}
\examples{
# No examples yet.
}
