% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_table.R
\name{so_table}
\alias{so_table}
\title{Make simulated vs. observed tables for reports}
\usage{
so_table(
  report_input_file = NA,
  sheet = NA,
  sectionInfo = NA,
  PKparameters = NA,
  mean_type = NA,
  variability_option = "90\% CI",
  concatVariability = FALSE,
  includeHalfLife = FALSE,
  includeTrialMeans = FALSE,
  includeCV = TRUE,
  checkDataSource = TRUE,
  sim_data_file = NA
)
}
\arguments{
\item{report_input_file}{the name of the Excel file created by running
\code{\link{generateReportInputForm}}, which you have now filled out,
including the path if it's in any other directory than the current one}

\item{sheet}{the sheet in the Excel file that contains information about this
section of the report. In the original template, this was the tab titled
"table and graph input".}

\item{sectionInfo}{information about the simulated and observed data. This is
output from the function \code{\link{getSectionInfo}} and can be used
instead of listing the Excel file and sheet name as input.}

\item{PKparameters}{the PK parameters to include as a character vector. To
see the full set of possible parameters to extract, enter
\code{data(AllPKParameters)} into the console. By default, if you supply a
file for \code{report_input_file}, the PK parameters included are only
those included for the observed data in that file. Otherwise, the PK
parameters will be automatically selected. An example of acceptable input
here: \code{c("AUCtau_ss", "AUCtau_ss_withInhib", "Cmax_ss",
"Cmax_ss_withInhib", "AUCtau_ratio_ss", "Cmax_ratio_ss")}. Parameters that
don't make sense for your scenario -- like asking for
\code{AUCinf_ss_withInhib} when your simulation did not include an
inhibitor or effector -- will not be included.}

\item{mean_type}{return "arithmetic" or "geometric" (default) means and CVs.
Only specify this if you'd like to override the value listed in
\code{sectionInfo}. If no value is specified here or in \code{sectionInfo},
the default is "geometric".}

\item{variability_option}{What type of variability would you like the table
to include? Options are: "90\% CI", "95\% CI", "95th percentiles", or any
combination of those, e.g. \code{variability_option = c("90\% CI", "95th
percentiles")} as long as they were included in your simulator output. Note
that the confidence intervals are geometric since that's what the simulator
outputs (see an AUC tab and the summary statistics; these values are the
ones for, e.g., "90% confidence interval around the geometric mean(lower
limit)"). Setting \code{variability_option = NA} will omit any of the
variability other than the CV. The CV will automatically be included unless
you omit it with \code{includeCV = FALSE}.}

\item{concatVariability}{Would you like to have the variability concatenated?
TRUE or FALSE. If "TRUE", the output will be formatted into a single row
and listed as the lower confidence interval or percentile to the upper CI
or percentile. Ex: "2400 to 2700"}

\item{includeHalfLife}{TRUE or FALSE for whether to include half life as a
parameter in the output table}

\item{includeTrialMeans}{TRUE or FALSE for whether to include the range of
trial means for a given parameter. Note: This is calculated from individual
values rather than being pulled directly from the output.}

\item{includeCV}{TRUE or FALSE for whether to include rows for CV in the
table}

\item{checkDataSource}{TRUE or FALSE: Include in the output a data.frame that
lists exactly where the data were pulled from the simulator output file.
Useful for QCing.}

\item{sim_data_file}{HISTORICAL, BACK-COMPATIBILITY PURPOSES ONLY: the
simulator output file. This is for when you DON'T fill out a report input
form and instead plan to add information about the observed data later
manually.}
}
\value{
a data.frame of S/O values or a list of that data.frame plus
  information on where the values came from for QCing
}
\description{
\code{so_table} creates simulated vs. observed tables for reports and
presentations, including reporting means, CVs, confidence intervals or
percentiles, and ratios of simulated vs. observed mean values. There are two
main ways to approach using this function 1. Select which PK parameters you
want to compare (\code{PKparameters} argument) and, later, manually fill out
rows for observed data that you've calculated elsewhere, or 2. Fill out an
Excel form with information about your observed data, in which case R will
calculate the comparisons for you here. Setting up the input for this
approach requires a few steps:\enumerate{\item{Use the function
\code{\link{generateReportInputForm}} to create an Excel file where you can
enter information about your project.} \item{Go to the tab "observed data"
and enter details about your observed data. It's ok if you don't have all the
information; anything that's missing won't be included in the final S/O
table. It's also ok to rename this tab and/or make copies of it within the
same Excel file for making other S/O tables.} \item{Go to the tab "table and
graph input" and fill out information here for the specific report section
you're making this S/O table for. Make sure that whatever you list as the tab
that contains information about the observed data is \emph{exactly} the same
as the actual tab name that you filled out in step 2. Also make sure that the
file names include the full file path.}\item{Save your Excel file.}
\item{Here, within RStudio (or within the shiny app that we plan to make!),
run this function using the name of that Excel file as input for
\code{report_input_file} and the name of the "table and graph input" tab as
the input for \code{sheet}. Note: If the Excel file lives on SharePoint,
you'll need to close it or this function will just keep running and not
generate any output while it waits for access to the file.} }
}
\examples{
# so_table(report_input_file = "//certara.com/data/sites/SHF/Consult/abc-1a/Report input.xlsx",
#          sheet = "table and graph input", includeTrialMeans = TRUE)
}
