% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_table.R
\name{so_table}
\alias{so_table}
\title{Make simulated vs. observed tables for reports}
\usage{
so_table(
  report_input_file = NA,
  sheet_report = NA,
  sim_data_file = NA,
  PKparameters = NA,
  sheet_PKparameters = NA,
  mean_type = NA,
  includeCV = TRUE,
  includeConfInt = TRUE,
  includePerc = FALSE,
  includeTrialMeans = FALSE,
  concatVariability = FALSE,
  prettify_columns = TRUE,
  prettify_effector_name = TRUE,
  checkDataSource = TRUE,
  save_table = NA
)
}
\arguments{
\item{report_input_file}{the name of the Excel file created by running
\code{\link{generateReportInputForm}}, which you have now filled out,
including the path if it's in any other directory than the current one}

\item{sheet_report}{the sheet in the Excel report template file that contains
information about the study, e.g., "study info - DDI" or "study info - no
DDI" if you haven't renamed the tab.}

\item{sim_data_file}{the simulator output file. This is only for when you
don't fill out a report input form because you either have no observed data
or you want to compare those data later manually.}

\item{PKparameters}{the PK parameters to include as a character vector.
  Notes: \itemize{

  \item{By default, if you have a single-dose simulation, the parameters will
  include AUC and Cmax for dose 1, and, if you have a multiple-dose
  simulation, AUC and Cmax for the last dose. Also by default, if you have an
  effector present, the parameters will include the AUC and Cmax values with
  and without the effector as well as those ratios.}

  \item{Alternatively, you can specify a vector of any combination of
  specific, individual parameters, e.g., \code{c("Cmax_dose1",
  "AUCtau_last").} Be sure to encapsulate the parameters you want with
  \code{c(...)}! To see the full set of possible parameters to extract, enter
  \code{data(PKParameterDefinitions); view(PKParameterDefinitions)} into the
  console.}

  \item{By default, if you supply a file for \code{report_input_file}, the PK
  parameters included are only those included for the observed data in that
  file. Otherwise, the PK parameters will be automatically selected.}

  \item{Parameters that don't make sense for your scenario -- like asking for
  \code{AUCinf_last_withInhib} when your simulation did not include an
  inhibitor or effector -- will not be included.}

  \item{tmax will be listed as median, min, and max rather than mean, lower
  and higher X\% confidence interval or X percentiles. Similarly, if you
  request trial means, the values for tmax will be the range of medians for
  the trials rather than the range of means.}} An example of acceptable input
  here: \code{c("AUCtau_last", "AUCtau_last_withInhib", "Cmax_last",
  "Cmax_last_withInhib", "AUCtau_ratio_last", "Cmax_ratio_last")}.}

\item{sheet_PKparameters}{(optional) If you want the PK parameters to be
pulled from a specific tab in the simulator output file, list that tab
here. Most of the time, this should be left as NA.}

\item{mean_type}{return "arithmetic" or "geometric" (default) means and CVs.
If you supplied a report input form, only specify this if you'd like to
override the value listed there. If no value is specified here or in
\code{sectionInfo}, the default is "geometric".}

\item{includeCV}{TRUE or FALSE for whether to include rows for CV in the
table}

\item{includeConfInt}{TRUE or FALSE for whether to include whatever
confidence intervals were included in the simulator output file. Note that
the confidence intervals are geometric since that's what the simulator
outputs (see an AUC tab and the summary statistics; these values are the
ones for, e.g., "90\% confidence interval around the geometric mean(lower
limit)").}

\item{includePerc}{TRUE or FALSE for whether to include 5th to 95th
percentiles}

\item{includeTrialMeans}{TRUE or FALSE for whether to include the range of
trial means for a given parameter. Note: This is calculated from individual
values rather than being pulled directly from the output.}

\item{concatVariability}{TRUE or FALSE (default) for whether to concatenate
the variability. If "TRUE", the output will be formatted into a single row
and listed as the lower confidence interval or percentile to the upper CI
or percentile, e.g., "2400 to 2700". Please note that the current
SimcypConsultancy template lists one row for each of the upper and
lower values, so this should be set to FALSE for official reports.}

\item{prettify_columns}{TRUE or FALSE for whether to make easily
human-readable column names. TRUE makes pretty column names such as "AUCinf
(h*ng/mL)" whereas FALSE leaves the column with the R-friendly name from
\code{\link{extractPK}}, e.g., "AUCinf_dose1".}

\item{prettify_effector_name}{TRUE (default) or FALSE on whether to make
effector name prettier in the prettified column titles. This was designed
for simulations where the effector is one of the standard options for the
simulator, and leaving \code{prettify_effector_name = TRUE} will make the
name of that effector (or effectors if there are any effector metabolites
or other effectors present) be something more human readable. For example,
"SV-Rifampicin-MD" will become "rifampicin", and "Sim-Ketoconazole-200 mg
BID" will become "ketoconazole". Set it to the name you'd prefer to see in
your column titles if you would like something different. For example,
\code{prettify_effector_name = "Drug ABC"}}

\item{checkDataSource}{TRUE or FALSE: Include in the output a data.frame that
lists exactly where the data were pulled from the simulator output file.
Default is TRUE to include it. Useful for QCing.}

\item{save_table}{optionally save the output table and, if requested, the QC
info, by supplying a file name in quotes here, e.g., "My nicely formatted
table.csv". If you leave off ".csv", it will still be saved as a csv file.
If you requested both the table and the QC info, the QC file will have "-
QC" added to the end of the file name.}
}
\value{
a data.frame of S/O values or a list of that data.frame (named
  "Table") plus information on where the values came from for QCing (named
  "QC")
}
\description{
\code{so_table} creates simulated vs. observed tables for reports and
presentations, including reporting means, CVs, confidence intervals or
percentiles, and ratios of simulated vs. observed mean values. This function
automatically finds the correct tab and the correct cells to pull those data.
\strong{Notes:} \itemize{\item{Please see the notes at the bottom of this
help file for options for supplying observed data in a standardized fashion
that this function can read.} \item{Nearly all parameters are for the
\emph{substrate}. We're still validating this for extracting PK for an
effector. \strong{A request for assistance:} If you extract PK data for an
effector by specifying an Excel sheet for that compound, please check the
values and tell Laura Shireman how well it works!} \item{Currently, the
output column titles list units of ng, mL, and h for AUC and Cmax, and the
function doesn't actually check what units are present in the data. If your
units are something else, our apologies, but please change the units in the
column titles when you use the output table. (The values in the table are
fine.) We're working on making this detect what the units were and print
those.} \item{ If the simulator output Excel file lives on SharePoint, you'll
need to close it or this function will just keep running and not generate any
output while it waits for access to the file.}}
}
\details{
Please see the notes at the bottom of this help file for options for how to
supply observed data in a standardized fashion that this function can read.
\strong{Note:} If the simulator output Excel file lives on SharePoint, you'll
need to close it or this function will just keep running and not generate any
output while it waits for access to the file.

Because we need to have a standardized way to input observed data, setting up
the input for this function requires filling out an Excel template form. Here
are the steps to take: \enumerate{\item{Use the function
\code{\link{generateReportInputForm}} to create an Excel file where you can
enter information about your project. Example:
\code{generateReportInputForm("My report input form.xlsx")}} \item{Go to the
tab "study info - DDI" or "study info - no DDI", whichever is appropriate for
your situation. Under the heading "Observed data", enter details about your
observed data. It's ok if you don't have all the information; anything that's
missing won't be included in the final S/O table. It's also ok to rename this
tab and/or make copies of it within the same Excel file for making other S/O
tables.} \item{Under the heading "Simulated data" on that same tab, fill out
the name of the specific simulator output Excel file you want to compare.}
\item{Save the report form.} \item{Back in RStudio (or within the shiny app
that we plan to make!), run this function using the file name of that Excel
report form as input for \code{report_input_file} and the name of the "study
info - DDI/no DDI" tab as the input for \code{sheet_report}. Note: If the
Excel file lives on SharePoint, you'll need to close it or this function will
just keep running and not generate any output while it waits for access to
the file.} }
}
\examples{
# so_table(report_input_file = "//certara.com/data/sites/SHF/Consult/abc-1a/Report input.xlsx",
#          sheet_report = "table and graph input", includeTrialMeans = TRUE)
}
