% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyPop.R
\name{tidyPop}
\alias{tidyPop}
\title{Tidying Simcyp simulator population names for reports, etc.}
\usage{
tidyPop(input_pop, dialect = "British")
}
\arguments{
\item{input_pop}{input population to tidy up; this has been designed to work
with \code{\link{extractExpDetails}} with the output item named
"Population"}

\item{dialect}{Use "British" (default) or "American" spelling when there's a
difference. Currently only affects "paediatric" vs. "pediatric"
populations.}
}
\value{
a tidied set of names for a simulated population
}
\description{
\code{tidyPop}
}
\examples{
# No examples yet
}
