% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pop.R
\name{tidyPop}
\alias{tidyPop}
\title{Tidying Simcyp simulator population names for reports, etc.}
\usage{
tidyPop(input_pop)
}
\arguments{
\item{input_pop}{input population to tidy up; this has been designed to work
with \code{extractExpDetails} with the output item named "Pop"}
}
\value{

}
\description{
\code{tidyPop}
}
\examples{
# No examples yet
}
