% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_means_plot.R
\name{trial_means_plot}
\alias{trial_means_plot}
\title{Make graphs comparing the center statistic and variability in PK across
trials and, optionally, against observed PK}
\usage{
trial_means_plot(
  sim_data_file,
  PKparameter = "Cmax_dose1",
  compoundToExtract = "substrate",
  tissue = "plasma",
  sheet = NA,
  mean_type = "geometric",
  variability_type = "percentiles",
  observed_PK = NA,
  lines_for_population_stats = "none",
  color_set = "default",
  color_option = "by study",
  point_size = NA,
  bar_width = NA,
  y_axis_limits_lin = NA,
  legend_position = "right",
  include_dose_num = FALSE,
  existing_exp_details = NA,
  return_caption = FALSE,
  prettify_compound_names = TRUE,
  name_clinical_study = NA,
  save_graph = NA,
  fig_height = NA,
  fig_width = NA
)
}
\arguments{
\item{sim_data_file}{name of the Excel file containing the simulator output,
in quotes}

\item{PKparameter}{PK parameter you want to extract from the simulator output
file. To see the full set of possible parameters to extract, enter
\code{view(PKParameterDefinitions)} into the console. Not case sensitive.}

\item{compoundToExtract}{For which compound do you want to extract
PK data? Options are: \itemize{\item{"substrate"
(default),} \item{"primary metabolite 1",} \item{"primary metabolite 2",}
\item{"secondary metabolite",} \item{"inhibitor 1" -- this can be an
inducer, inhibitor, activator, or suppresesor, but it's labeled as
"Inhibitor 1" in the simulator,} \item{"inhibitor 2" for the 2nd inhibitor
listed in the simulation,} \item{"inhibitor 1 metabolite" for the primary
metabolite of inhibitor 1}}}

\item{tissue}{For which tissue would you like the PK parameters to be pulled?
Options are "plasma" (default), "unbound plasma", "blood", "unbound blood",
"peripheral plasma", or "peripheral blood". \strong{NOTE: PK for peripheral
sampling is not as well tested as for other tissues and is only set up for
V21+. Please check your results carefully.}}

\item{sheet}{optionally specify the name of the sheet where you'd like to
pull the PK data, in quotes. For example, if you have a user-defined AUC
interval, specify the tab where those data are. \strong{NOTE:} If you want
typical first- or last-dose PK, this should be left as NA. This \emph{only}
applies for user-defined AUC intervals.}

\item{mean_type}{What kind of means do you want to use for the center point
on the graph? Options are "geometric" (default), "arithmetic", or "median".}

\item{variability_type}{What variability statistic to you want to use for the
error bars? Options are "percentiles" for the 5th to 95th percentiles
(default), "90\% CI", "SD", "CV", "GCV" (for geometric CV), or "range".}

\item{observed_PK}{a data.frame of observed PK that you would like to
  compare. This must include the columns "ObsValue" or "Value" for the center
  value for the observed data and "ObsVariability" or "Variability" for the
  error bars for the observed data. Some filtering this will do
  automatically:

  \itemize{\item{If you include a column titled "File", we will only use PK
  data that have the same value in that column as what you supply for
  \code{sim_data_file}.}

  \item{If you include a column titled "PKparameter", we will
  only use PK data that match what you supply for \code{PKparameter}.}

  \item{If you want to include comparisons to more than one clinical study, add a column
  titled "Study" to your data. The value listed in that column will be used
  for labeling the study on your graph. If there is not a column titled
  "Study" in your observed PK, we will label the data in the graph as "observed".}

  \item{If you include a column titled "CompoundID", we will only include PK
  data that match what you supplied for \code{compoundToExtract}.}

  \item{If you include a column titled "Tissue", we will only include PK
  data that match what you supplied for \code{tissue}.}

  }}

\item{lines_for_population_stats}{optionally include horizontal lines for the
overall simulated population statistics by specifying the desired line
color and type. If left as "none" (default), no lines will be included. If
set to, e.g., "red solid dotted" or "gray50 dashed dashed" or even "#87A896
solid E2", then lines will be added to the graph at whatever central
statistic you specified for "mean_type" and whatever variability statistic
you specified for "variability_type". The first word must be a legitimate
color in R (hex codes are fine), the second word, separated by a space,
must be the line type you want for the central statistic, and the third
word must be the line type you want for the variability statistic. If you
add another space and then a fourth value, that will set the line width,
which will be 0.5 by default.}

\item{color_set}{the set of colors to use. Options: \describe{

  \item{"default"}{a set of colors from Cynthia Brewer et al. from Penn State
  that are friendly to those with red-green colorblindness. The first three
  colors are green, orange, and purple. This can also be referred to as
  "Brewer set 2". If there are only two unique values in the colorBy_column,
  then Brewer set 1 will be used since red and blue are still easily
  distinguishable but also more aesthetically pleasing than green and
  orange.}

  \item{"black and white" ("BW" also works)}{black and white only. "white" 
  means a black circle with a white center. If you set 
  \code{color_option = "S or O"}, the simulated trials will be white and the 
  observed will be black. This looks like the trial-means 
  plots in the Excel outputs from the Simulator.}

  \item{"Brewer set 1"}{colors selected from the Brewer palette "set 1". The
  first three colors are red, blue, and green.}

  \item{"ggplot2 default"}{the default set of colors used in ggplot2 graphs
  (ggplot2 is an R package for graphing.)}

  \item{"rainbow"}{colors selected from a rainbow palette. The default
  palette is limited to something like 6 colors, so if you have more than
  that, that's when this palette is most useful. It's \emph{not} very useful
  when you only need a couple of colors.}

  \item{"blue-green"}{a set of blues fading into greens. This palette can be
  especially useful if you are comparing a systematic change in some
  continuous variable -- for example, increasing dose or predicting how a
  change in intrinsic solubility will affect concentration-time profiles --
  because the direction of the trend will be clear.}

  \item{"blues"}{a set of blues fading from sky to navy. Like
  "blue-green", this palette can be especially useful if you are comparing a
  systematic change in some continuous variable.}

  \item{"greens"}{a set of greens fading from chartreuse to forest. Like
  "blue-green", this palette can be especially useful if you are comparing a
  systematic change in some continuous variable.}

  \item{"purples"}{a set of purples fading from lavender to aubergine. Like
  "blue-green", this palette can be especially useful if you are comparing a
  systematic change in some continuous variable.}

  \item{"Tableau"}{uses the standard Tableau palette; requires the "ggthemes"
  package}

  \item{"viridis"}{from the eponymous package by Simon Garnier and ranges
  colors from purple to blue to green to yellow in a manner that is
  "printer-friendly, perceptually uniform and easy to read by those with
  colorblindness", according to the package author}

  \item{a character vector of colors}{If you'd prefer to set all the colors
  yourself to \emph{exactly} the colors you want, you can specify those
  colors here. An example of how the syntax should look: \code{color_set =
  c("dodgerblue3", "purple", "#D8212D")} or, if you want to specify exactly
  which item in \code{colorBy_column} gets which color, you can supply a
  named vector. For example, if you're coloring the lines by the compound ID,
  you could do this: \code{color_set = c("substrate" = "dodgerblue3",
  "inhibitor 1" = "purple", "primary metabolite 1" = "#D8212D")}. If you'd
  like help creating a specific gradation of colors, please talk to a member
  of the R Working Group about how to do that using
  \link{colorRampPalette}.}}}

\item{color_option}{How do you want to color information in the graph?
  Options are: \describe{\item{"by study" (default)}{all simulated trials
  will be the first color from \code{color_set} and then each set of
  observed data will be a different color from that set}

  \item{"by trial"}{uses a different color from \code{color_set} for every
  trial, whether simulated or observed}}}

\item{y_axis_limits_lin}{optionally set the Y axis limits for the linear
plot, e.g., \code{c(10, 1000)}. If left as the default NA, the Y axis
limits for the linear plot will be automatically selected. (Setting up
semi-log plot y axis intervals manually is a bit tricky and is not
currently supported.)}

\item{legend_position}{specify where you want the legend to be. Options are
"left", "right" (default), "bottom", "top", or "none" if you don't want one
at all.}

\item{existing_exp_details}{If you have already run
\code{\link{extractExpDetails_mult}} to get all the details from the "Input
Sheet" (e.g., when you ran extractExpDetails you said \code{exp_details =
"Summary and Input"} or \code{exp_details = "all"}), you can save some processing
time by supplying that object here, unquoted. If left as NA, this function
will run \code{extractExpDetails} behind the scenes anyway to figure out
some information about your experimental set up.}

\item{return_caption}{TRUE or FALSE (default) for whether to return any
caption text to use with the graph. This works best if you supply something
for the argument \code{existing_exp_details}. If set to TRUE, you'll get as
output a list of the graph, the figure heading, and the figure caption.}

\item{prettify_compound_names}{TRUE (default), FALSE or a character vector:
This is asking whether to make compound names prettier in the figure
heading and caption. This was designed for simulations where the substrate
and any metabolites, perpetrators, or perpetrator metabolites are among the
standard options for the simulator, and leaving
\code{prettify_compound_names = TRUE} will make the name of those compounds
something more human readable. For example, "SV-Rifampicin-MD" will become
"rifampicin", and "Sim-Midazolam" will become "midazolam". Setting this to
FALSE will leave the compound names as is. For an approach with more
control over what the compound names will look like in legends and Word
output, set each compound to the exact name you  want with a named
character vector where the names are "substrate" and, as applicable,
"inhibitor 1", "primary metabolite 1", etc. and the values are the names
you want, e.g.,
\code{prettify_compound_names = c("inhibitor 1" = "teeswiftavir",
"substrate" = "superstatin")}.}

\item{name_clinical_study}{optionally specify the name(s) of the clinical
study or studies for any observed data. This only affects the caption of
the graph. For example, specifying \code{name_clinical_study = "101, fed
cohort"} will result in a figure caption that reads in part "clinical study
101, fed cohort". If you have more than one study, that's fine; we'll take
care of stringing them together appropriately. Just list them as a
character vector, e.g., \code{name_clinical_study = c("101",
"102", "103")} will become "clinical studies 101, 102, and 103."}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "My conc time graph.png" or "My conc time
graph.docx". The nice thing about saving to Word is that the figure title
and caption text will be filled in automatically. If you leave off ".png"
or ".docx", the graph will be saved as a png file, but if you specify a
different graphical file extension, it will be saved as that file format.
Acceptable graphical file extensions are "eps", "ps", "jpeg", "jpg",
"tiff", "png", "bmp", or "svg". Do not include any slashes, dollar signs,
or periods in the file name. Leaving this as NA means the file will not be
saved to disk.}

\item{fig_height}{figure height in inches}

\item{fig_width}{figure width in inches}
}
\value{
a ggplot2 graph
}
\description{
\code{trial_means_plot} creates a graph comparing variability
  across simulated trials and, when supplied, across observed trials as well
  for a given PK parameter. The central statistic is shown as a dot and the
  variability statistic is shown as error bars for each of the trials, and
  you can choose which statistics you want for each, e.g., geometric means
  for the central statistic and the geometric 90\% confidence interval for
  the variability statistic. UNDER CONSTRUCTION.

  \strong{IMPORTANT:} If the summary statistics used for your observed data
  don't match the summary statistics you choose for \code{mean_type} and
  \code{variability_type}, you will get a misleading graph, so be careful!
}
\examples{
# None yet

}
