% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBE_safe_space_plot.R
\name{VBE_safe_space_plot}
\alias{VBE_safe_space_plot}
\title{Make dissolution safe-space plots}
\usage{
VBE_safe_space_plot(
  VBE_dataframe,
  color_set = NA,
  safe_space_color = NA,
  safe_space_trans = NA,
  linetypes = NA,
  linewidths = NA,
  point_shapes = NA,
  point_sizes = NA,
  show_points_on_boundaries = TRUE,
  save_graph = NA,
  fig_height = NA,
  fig_width = NA
)
}
\arguments{
\item{VBE_dataframe}{a data.frame of virtual bioequivalence dissolution data
with the following columns:
\describe{
\item{Type}{the type of formulation being tested. In the graph, you can
change the color of the lines based on the values in the column "Type".}
\item{SorO}{Are the data simulated or observed? The values in this column
can be anything you prefer as long as there are only two of them, e.g.,
"simulated" and "observed" or "hypothetical" and "actual". In the graph,
you can change the point shape and the line type based on the values in
the column "SorO".}
\item{Time}{the time in hours}
\item{Dissolution}{the percent dissolution. This should be specified as,
e.g., 0.5 for 50\%.}
\item{Limit}{This column will be used for specifying which data sets are
the upper and lower limits of the safe space. Data sets that do \emph{not}
describe the upper or lower limits should have NA in this column or be blank in the .csv file, and the
upper- and lower-limit datasets should be specified as "upper" and "lower".}
} For an example, please view the object "VBE_disso_example" by running
this in your console: \code{view(VBE_disso_example)} and set up your data
like that.}

\item{color_set}{the set of colors to use. Options: \describe{

  \item{"default"}{a set of colors from Cynthia Brewer et al. from Penn State
  that are friendly to those with red-green colorblindness. The first three
  colors are green, orange, and purple. This can also be referred to as
  "Brewer set 2". If there are only two unique values in the "Type" column,
  then Brewer set 1 will be used since red and blue are still easily
  distinguishable but also more aesthetically pleasing than green and
  orange.}

  \item{"Brewer set 1"}{colors selected from the Brewer palette "set 1". The
  first three colors are red, blue, and green.}

  \item{"ggplot2 default"}{the default set of colors used in ggplot2 graphs
  (ggplot2 is an R package for graphing.)}

  \item{"rainbow"}{colors selected from a rainbow palette. The default
  palette is limited to something like 6 colors, so if you have more than
  that, that's when this palette is most useful. It's \emph{not} very useful
  when you only need a couple of colors.}

  \item{"blue-green"}{a set of blues fading into greens. This palette can be
  especially useful if you are comparing a systematic change in some
  continuous variable -- for example, increasing dose or predicting how a
  change in intrinsic solubility will affect concentration-time profiles --
  because the direction of the trend will be clear.}

  \item{"blues"}{a set of blues fading from sky to navy. Like
  "blue-green", this palette can be especially useful if you are comparing a
  systematic change in some continuous variable.}

  \item{"greens"}{a set of greens fading from chartreuse to forest. Great for showing
  systematic changes in a continuous variable.}

  \item{"purples"}{a set of purples fading from lavender to aubergine. Great for showing
  systematic changes in a continuous variable.}

  \item{"reds"}{a set of reds from pink to brick. Great for showing
  systematic changes in a continuous variable.}

  \item{"Tableau"}{uses the standard Tableau palette; requires the "ggthemes"
  package}

  \item{"viridis"}{from the eponymous package by Simon Garnier and ranges
  colors from purple to blue to green to yellow in a manner that is
  "printer-friendly, perceptually uniform and easy to read by those with
  colorblindness", according to the package author}

  \item{a character vector of colors}{If you'd prefer to set all the colors
  yourself to \emph{exactly} the colors you want, you can specify those
  colors here. An example of how the syntax should look: \code{color_set =
  c("dodgerblue3", "purple", "#D8212D")} or, if you want to specify exactly
  which item in \code{Type} gets which color, you can supply a
  named vector. For example,
  you could do this: \code{color_set = c("Formulation A" = "dodgerblue3",
  "Formulation B" = "purple", "Test 1" = "#D8212D")}. If you'd
  like help creating a specific gradation of colors, please talk to a member
  of the R Working Group about how to do that using
  \link{colorRampPalette}.}}}

\item{safe_space_color}{color to use for indicating the safe space; any color
that's acceptable in R will be acceptable here.}

\item{safe_space_trans}{transparency to use for the safe space from 0 (fully
transparent, i.e., invisible) to 1 (fully opaque)}

\item{linetypes}{the line type(s) to use for the graph. Possible options can
be seen by typing \code{ggpubr::show_line_types()} into the console.}

\item{linewidths}{the line widths to use (numeric)}

\item{point_shapes}{the point shape(s) to use for the graph to indicate
whether a given value was predicted or observed. Possible options can be
seen by typing \code{ggpubr::show_point_shapes()} into the console. If you
do not want points to show for one dataset, e.g., maybe you only want
points for the observed data and just lines for the predicted, then set the
point shape to NA.}

\item{point_sizes}{the point sizes (numeric)}

\item{show_points_on_boundaries}{TRUE (default) or FALSE to include points
on the boundaries of the dissolution safe space.}

\item{save_graph}{optionally save the output graph by supplying a file name
in quotes here, e.g., "Demographics comparisons.png". Acceptable graphical
file extensions are "eps", "ps", "jpeg", "jpg", "tiff", "png", "bmp", or
"svg". Do not include any slashes, dollar signs, or periods in the file
name. Leaving this as NA means the file will not be saved to disk.}

\item{fig_height}{figure height in inches; default is 6}

\item{fig_width}{figure width in inches; default is 8}
}
\value{
a ggplot2 graph
}
\description{
\code{VBE_safe_space_plot} makes dissolution safe-space plots
  comparing predicted dissolution values to observed and includes shading to
  indicate the safe space for bioequivalence.
}
\examples{
# Using example data included in the package
VBE_safe_space_plot(VBE_dataframe = VBE_disso_example)

# Setting some colors for actual and hypothetical datasets
MyColors_actual <- reds(4)
names(MyColors_actual) <- c("Formulation A",
                            "Formulation B",
                            "Formulation C",
                            "Formulation D")

MyColors_hyp <- blues(6)
names(MyColors_hyp) <- paste("Test", 1:6)

MyColors <- c(MyColors_actual, MyColors_hyp)

VBE_safe_space_plot(VBE_dataframe = VBE_disso_example,
                    color_set = MyColors,
                    safe_space_color = "gray80",
                    linetypes = c("solid", "longdash"),
                    save_graph = "VBE safe space.png",
                    fig_height = 4, fig_width = 6)


}
